Algorithms.Geometry.LineSegmentIntersection

Description

Documentation

hasIntersections :: (Ord r, Num r) => [LineSegment 2 p r] -> Bool

Tests if there are any intersections.

\(O(n\log n)\)

intersections :: (Ord r, Fractional r) => [LineSegment 2 p r] -> Intersections p r

Compute all intersections

\(O((n+k)\log n)\), where \(k\) is the number of intersections.

interiorIntersections :: (Ord r, Fractional r) => [LineSegment 2 p r] -> Intersections p r

Computes all intersection points p s.t. p lies in the interior of at least one of the segments.

\(O((n+k)\log n)\), where \(k\) is the number of intersections.

type Intersections p r = Map (Point 2 r) (Associated p r)

For each intersection point the segments intersecting there.

data Associated p r

The line segments that contain a given point p may either have p as the endpoint or have p in their interior.

if somehow the segment is degenerate, and p is both the start and end it is reported only as the start point.

Constructors

Associated

Fields

_startPointOf :: Set (AroundEnd (LineSegment 2 p r))

segments for which the intersection point is the start point (i.e. s^.start.core == p)

_endPointOf :: Set (AroundStart (LineSegment 2 p r))

segments for which the intersection point is the end point (i.e. s^.end.core == p)

_interiorTo :: Set (AroundIntersection (LineSegment 2 p r))

data IntersectionPoint p r

An intersection point together with all segments intersecting at this point.

Constructors

IntersectionPoint

Fields

_intersectionPoint :: !(Point 2 r)

_associatedSegs :: !(Associated p r)

mkIntersectionPoint

Arguments

:: (Ord r, Fractional r)

=> Point 2 r

-> [LineSegment 2 p r]
uncategorized

-> [LineSegment 2 p r]
segments we know contain p,

-> IntersectionPoint p r

Given a point p, and a bunch of segments that suposedly intersect at p, correctly categorize them.

hasSelfIntersections :: (Ord r, Fractional r) => Polygon t p r -> Bool

Test if the polygon has self intersections.

\(O(n \log n)\)