Algorithms.Geometry.PolygonTriangulation.MakeMonotone

Description

Documentation

makeMonotone :: (Fractional r, Ord r) => proxy s -> Polygon t p r -> PlanarSubdivision s p PolygonEdgeType PolygonFaceData r

Computes a set of diagionals that decompose the polygon into y-monotone pieces.

pre: the polygon boundary is given in counterClockwise order.

running time: \(O(n\log n)\)

computeDiagonals :: forall t r p. (Fractional r, Ord r) => Polygon t p r -> [LineSegment 2 p r]

Given a polygon, find a set of non-intersecting diagonals that partition the polygon into y-monotone pieces.

running time: \(O(n\log n)\)

data VertexType

Constructors

Start

Merge

Split

End

Regular

classifyVertices :: (Num r, Ord r) => Polygon t p r -> Polygon t (p :+ VertexType) r

assigns a vertex type to each vertex

pre: Both the outer boundary and the inner boundary of the polygon are given in CCW order.

running time: \(O(n)\).