Algorithms.Geometry.PolygonTriangulation.Types

Description

Documentation

data PolygonEdgeType

After triangulation, edges are either from the original polygon or a new diagonal.

Constructors

Original

Diagonal

constructSubdivision

Arguments

:: forall proxy r s p. (Fractional r, Ord r)

=> proxy s

-> LineSegment 2 p r
A counter-clockwise edge along the outer boundary

-> [LineSegment 2 p r]
remaining original edges

-> [LineSegment 2 p r]
diagonals

-> PlanarSubdivision s p PolygonEdgeType PolygonFaceData r

Given a list of original edges and a list of diagonals, creates a planar-subdivision

running time: \(O(n\log n)\)

constructGraph

Arguments

:: forall proxy r s p. (Fractional r, Ord r)

=> proxy s

-> LineSegment 2 p r
A counter-clockwise edge along the outer boundary

-> [LineSegment 2 p r]
remaining original edges

-> [LineSegment 2 p r]
diagonals

-> PlaneGraph s p PolygonEdgeType PolygonFaceData r

Given a list of original edges and a list of diagonals, creates a planar-subdivision

running time: \(O(n\log n)\)