Algorithms.Geometry.SmallestEnclosingBall

Description

Types to represent the smallest enclosing disk of a set of points in \(\mathbb{R}^2\)

Documentation

data DiskResult p r

The result of a smallest enclosing disk computation: The smallest ball and the points defining it

Constructors

DiskResult

Fields

_enclosingDisk :: Disk () r

_definingPoints :: TwoOrThree (Point 2 r :+ p)

enclosingDisk :: forall p r. Lens' (DiskResult p r) (Disk () r)

definingPoints :: forall p r p. Lens (DiskResult p r) (DiskResult p r) (TwoOrThree ((:+) (Point 2 r) p)) (TwoOrThree ((:+) (Point 2 r) p))

data TwoOrThree a

List of two or three elements

Constructors

Two !a !a

Three !a !a !a

twoOrThreeFromList :: [a] -> Either String (TwoOrThree a)

Construct datatype from list with exactly two or three elements.