Algorithms.Geometry.SmallestEnclosingBall.Naive

Description

Naive implementation to compute the smallest enclosing disk of a set of points in \(\mathbb{R}^2\)

Documentation

smallestEnclosingDisk :: (Ord r, Fractional r) => [Point 2 r :+ p] -> DiskResult p r

Horrible \( O(n^4) \) implementation that simply tries all disks, checks if they enclose all points, and takes the largest one. Basically, this is only useful to check correctness of the other algorithm(s)

enclosesAll :: (Num r, Ord r) => DiskResult p r -> [Point 2 r :+ q] -> Bool

check if a disk encloses all points