Algorithms.Geometry.SoS.Symbolic

Documentation

data EpsFold i

eps :: i -> EpsFold i

Creates the term \(\varepsilon(i)\)

mkEpsFold :: Ord i => [i] -> EpsFold i

hasNoPertubation :: EpsFold i -> Bool

Test if the epsfold has no pertubation at all (i.e. if it is \(\Pi_{\emptyset}\)

factors :: EpsFold i -> Bag i

Gets the factors

suitableBase :: EpsFold i -> Int

computes a base d that can be used as:

\( \varepsilon(i) = \varepsilon^{d^i} \)

data Term i r

A term 'Term c es' represents a term:

\[ c \Pi_{i \in es} \varepsilon(i) \]

for a constant c and an arbitrarily small value \(\varepsilon\), parameterized by i.

Constructors

Term r (EpsFold i)

term :: r -> i -> Term i r

Creates a singleton term

constantFactor :: Lens' (Term i r) r

Lens to access the constant c in the term.

data Symbolic i r

Represents a Sum of terms, i.e. a value that has the form:

\[ \sum c \Pi_i \varepsilon(i) \]

The terms are represented in order of decreasing significance.

The main idea in this type is that, if symbolic values contains \(\varepsilon(i)\) terms we can always order them. That is, two Symbolic terms will be equal only if:

they contain *only* a constant term (that is equal)

they contain the exact same \(\varepsilon\)-fold.

constant :: Ord i => r -> Symbolic i r

Creates a constant symbolic value

symbolic :: Ord i => r -> i -> Symbolic i r

Creates a symbolic vlaue with a single indexed term. If you just need a constant (i.e. non-indexed), use constant

perturb :: (Num r, Ord i) => r -> i -> Symbolic i r

given the value c and the index i, creates the perturbed value \(c + \varepsilon(i)\)

toTerms :: Symbolic i r -> [Term i r]

Produces a list of terms, in decreasing order of significance

signOf :: (Num r, Eq r) => Symbolic i r -> Maybe Sign

Computing the Sign of an expression. (Nothing represents zero)