Algorithms.Geometry.WSPD

Description

Algorithm to construct a well separated pair decomposition (wspd).

Documentation

fairSplitTree :: (Fractional r, Ord r, Arity d, 1 <= d, Show r, Show p) => NonEmpty (Point d r :+ p) -> SplitTree d p r ()

Construct a split tree

running time: \(O(n \log n)\)

wellSeparatedPairs :: (Floating r, Ord r, Arity d, Arity (d + 1)) => r -> SplitTree d p r a -> [WSP d p r a]

Given a split tree, generate the Well separated pairs

running time: \(O(s^d n)\)

data NodeData d r a

Data that we store in the split tree

Constructors

NodeData !Int !(Box d () r) !a

type WSP d p r a = (PointSet d p r a, PointSet d p r a)

type SplitTree d p r a = BinLeafTree (NodeData d r a) (Point d r :+ p)

nodeData :: forall d r a a. Lens (NodeData d r a) (NodeData d r a) a a

data Level

Constructors

Level

Fields

_unLevel :: Int

_widestDim :: Maybe Int

reIndexPoints :: (Arity d, 1 <= d) => Vector d (PointSeq d (Idx :+ p) r) -> Vector d (PointSeq d (Idx :+ p) r)

Given a sequence of points, whose index is increasing in the first dimension, i.e. if idx p < idx q, then p[0] < q[0]. Reindex the points so that they again have an index in the range [0,..,n'], where n' is the new number of points.

running time: O(n' * d) (more or less; we are actually using an intmap for the lookups)

alternatively: I can unsafe freeze and thaw an existing vector to pass it along to use as mapping. Except then I would have to force the evaluation order, i.e. we cannot be in reIndexPoints for two of the nodes at the same time.

so, basically, run reIndex points in ST as well.

distributePoints :: (Arity d, Show r, Show p) => Int -> Vector (Maybe Level) -> Vector d (PointSeq d (Idx :+ p) r) -> Vector (Vector d (PointSeq d (Idx :+ p) r))

Assign the points to their the correct class. The Nothing class is considered the last class

distributePoints'

Arguments

:: Int
number of classes

-> Vector (Maybe Level)
level assignment

-> PointSeq d (Idx :+ p) r
input points

-> Vector (PointSeq d (Idx :+ p) r)

Assign the points to their the correct class. The Nothing class is considered the last class