Data.Geometry.Ball

Description

\(d\)-dimensional Balls and Spheres

A d-dimensional ball

data Ball d p r

A d-dimensional ball.

Constructors

Ball

Fields

_center :: !(Point d r :+ p)

_squaredRadius :: !r

squaredRadius :: forall d p r. Lens' (Ball d p r) r

center :: forall d p r d p. Lens (Ball d p r) (Ball d p r) ((:+) (Point d r) p) ((:+) (Point d r) p)

radius :: Floating r => Lens' (Ball d p r) r

A lens to get/set the radius of a Ball

Constructing Balls

fromDiameter :: (Arity d, Fractional r) => Point d r -> Point d r -> Ball d () r

Given two points on the diameter of the ball, construct a ball.

fromCenterAndPoint :: (Arity d, Num r) => (Point d r :+ p) -> (Point d r :+ p) -> Ball d p r

Construct a ball given the center point and a point p on the boundary.

unitBall :: (Arity d, Num r) => Ball d () r

A d dimensional unit ball centered at the origin.

Querying if a point lies in a ball

inBall :: (Arity d, Ord r, Num r) => Point d r -> Ball d p r -> PointLocationResult

Query location of a point relative to a d-dimensional ball.

insideBall :: (Arity d, Ord r, Num r) => Point d r -> Ball d p r -> Bool

Test if a point lies strictly inside a ball

>>> (Point2 0.5 0.0) `insideBall` unitBall
True
>>> (Point2 1 0) `insideBall` unitBall
False
>>> (Point2 2 0) `insideBall` unitBall
False

inClosedBall :: (Arity d, Ord r, Num r) => Point d r -> Ball d p r -> Bool

Test if a point lies in or on the ball

onBall :: (Arity d, Ord r, Num r) => Point d r -> Ball d p r -> Bool

Test if a point lies on the boundary of a ball.

>>> (Point2 1 0) `onBall` unitBall
True
>>> (Point3 1 1 0) `onBall` unitBall
False

type Sphere d p r = Boundary (Ball d p r)

Spheres, i.e. the boundary of a ball.

pattern Sphere :: (Point d r :+ p) -> r -> Sphere d p r

_BallSphere :: Iso (Disk p r) (Disk p s) (Circle p r) (Circle p s)

Disks and Circles, aka 2-dimensional Balls and Spheres

type Disk p r = Ball 2 p r

pattern Disk :: (Point 2 r :+ p) -> r -> Disk p r

Given the center and the squared radius, constructs a disk

type Circle p r = Sphere 2 p r

_DiskCircle :: Iso (Disk p r) (Disk p s) (Circle p r) (Circle p s)

Iso for converting between Disks and Circles, i.e. forgetting the boundary

pattern Circle :: (Point 2 r :+ p) -> r -> Circle p r

Given the center and the squared radius, constructs a circle

disk :: (Eq r, Fractional r) => Point 2 r -> Point 2 r -> Point 2 r -> Maybe (Disk () r)

Given three points, get the disk through the three points. If the three input points are colinear we return Nothing

>>> disk (Point2 0 10) (Point2 10 0) (Point2 (-10) 0)
Just (Ball {_center = Point2 [0.0,0.0] :+ (), _squaredRadius = 100.0})

from3Points :: Fractional r => (Point 2 r :+ p) -> (Point 2 r :+ q) -> (Point 2 r :+ s) -> Circle () r

Creates a circle from three points on the boundary

newtype Touching p

Constructors

Touching p