Data.Geometry.Boundary

Description

Documentation

newtype Boundary g

The boundary of a geometric object.

Constructors

Boundary g

_Boundary :: Iso g h (Boundary g) (Boundary h)

Iso for converting between things with a boundary and without its boundary

data PointLocationResult

Result of a query that asks if something is Inside a g, *on* the boundary of the g, or outside.

Constructors

Inside

OnBoundary

Outside