Data.Geometry.Box.Corners

Description

Documentation

data Corners a

A Quadrant data type

Constructors

Corners !a !a !a !a

northWest :: forall a. Lens' (Corners a) a

northEast :: forall a. Lens' (Corners a) a

southEast :: forall a. Lens' (Corners a) a

southWest :: forall a. Lens' (Corners a) a

corners :: Num r => Rectangle p r -> Corners (Point 2 r :+ p)

Get the corners of a rectangle, the order is: (TopLeft, TopRight, BottomRight, BottomLeft). The extra values in the Top points are taken from the Top point, the extra values in the Bottom points are taken from the Bottom point

cornersInDirection :: CardinalDirection -> Corners p -> Two p

Gets the corners in a particular direction