Data.Geometry.Box.Internal

Description

Orthogonal \(d\)-dimensiontal boxes (e.g. rectangles)

Documentation

newtype CWMin a

Coordinate wize minimum

Constructors

CWMin

Fields

_cwMin :: a

cwMin :: forall a a. Iso (CWMin a) (CWMin a) a a

newtype CWMax a

Coordinate wize maximum

Constructors

CWMax

Fields

_cwMax :: a

cwMax :: forall a a. Iso (CWMax a) (CWMax a) a a

d-dimensional boxes

data Box d p r

Constructors

Box

Fields

_minP :: !(CWMin (Point d r) :+ p)

_maxP :: !(CWMax (Point d r) :+ p)

minP :: forall d p r. Lens' (Box d p r) ((:+) (CWMin (Point d r)) p)

maxP :: forall d p r. Lens' (Box d p r) ((:+) (CWMax (Point d r)) p)

box :: (Point d r :+ p) -> (Point d r :+ p) -> Box d p r

Given the point with the lowest coordinates and the point with highest coordinates, create a box.

grow :: (Num r, Arity d) => r -> Box d p r -> Box d p r

grows the box by x on all sides

fromExtent :: Arity d => Vector d (Range r) -> Box d () r

Build a d dimensional Box given d ranges.

fromCenter :: (Arity d, Fractional r) => Point d r -> Vector d r -> Box d () r

Given a center point and a vector specifying the box width's, construct a box.

centerPoint :: (Arity d, Fractional r) => Box d p r -> Point d r

Center of the box

Functions on d-dimensonal boxes

minPoint :: Box d p r -> Point d r :+ p

maxPoint :: Box d p r -> Point d r :+ p

inBox :: (Arity d, Ord r) => Point d r -> Box d p r -> Bool

Check if a point lies a box

>>> origin `inBox` (boundingBoxList' [Point3 1 2 3, Point3 10 20 30] :: Box 3 () Int)
False
>>> origin `inBox` (boundingBoxList' [Point3 (-1) (-2) (-3), Point3 10 20 30] :: Box 3 () Int)
True

extent :: Arity d => Box d p r -> Vector d (Range r)

Get a vector with the extent of the box in each dimension. Note that the resulting vector is 0 indexed whereas one would normally count dimensions starting at zero.

>>> extent (boundingBoxList' [Point3 1 2 3, Point3 10 20 30] :: Box 3 () Int)
Vector3 (Range (Closed 1) (Closed 10)) (Range (Closed 2) (Closed 20)) (Range (Closed 3) (Closed 30))

size :: (Arity d, Num r) => Box d p r -> Vector d r

Get the size of the box (in all dimensions). Note that the resulting vector is 0 indexed whereas one would normally count dimensions starting at zero.

>>> size (boundingBoxList' [origin, Point3 1 2 3] :: Box 3 () Int)
Vector3 1 2 3

widthIn :: forall proxy p i d r. (Arity d, Arity (i - 1), Num r, ((i - 1) + 1) <= d) => proxy i -> Box d p r -> r

Given a dimension, get the width of the box in that dimension. Dimensions are 1 indexed.

>>> widthIn (C :: C 1) (boundingBoxList' [origin, Point3 1 2 3] :: Box 3 () Int)
1
>>> widthIn (C :: C 3) (boundingBoxList' [origin, Point3 1 2 3] :: Box 3 () Int)
3

widthIn' :: (Arity d, Num r) => Int -> Box d p r -> Maybe r

Same as widthIn but with a runtime int instead of a static dimension.

>>> widthIn' 1 (boundingBoxList' [origin, Point3 1 2 3] :: Box 3 () Int)
Just 1
>>> widthIn' 3 (boundingBoxList' [origin, Point3 1 2 3] :: Box 3 () Int)
Just 3
>>> widthIn' 10 (boundingBoxList' [origin, Point3 1 2 3] :: Box 3 () Int)
Nothing

Rectangles, aka 2-dimensional boxes

type Rectangle = Box 2

width :: Num r => Rectangle p r -> r

>>> width (boundingBoxList' [origin, Point2 1 2] :: Rectangle () Int)
1
>>> width (boundingBoxList' [origin] :: Rectangle () Int)
0

height :: Num r => Rectangle p r -> r

>>> height (boundingBoxList' [origin, Point2 1 2] :: Rectangle () Int)
2
>>> height (boundingBoxList' [origin] :: Rectangle () Int)
0

Constructing bounding boxes

class IsBoxable g where

Methods

boundingBox :: Ord (NumType g) => g -> Box (Dimension g) () (NumType g)

boundingBoxList :: (IsBoxable g, Foldable1 c, Ord (NumType g), Arity (Dimension g)) => c g -> Box (Dimension g) () (NumType g)

Create a bounding box that encapsulates a list of objects.

boundingBoxList' :: (IsBoxable g, Foldable c, Ord (NumType g), Arity (Dimension g)) => c g -> Box (Dimension g) () (NumType g)

Unsafe version of boundingBoxList, that does not check if the list is non-empty