Data.Geometry.Interval

Description

1 dimensional Intervals

data Interval a r where

An Interval is essentially a Range but with possible payload

We can think of an interval being defined as:

>>> data Interval a r = Interval (EndPoint (r :+ a)) (EndPoint (r :+ a))

Bundled Patterns

pattern Interval :: EndPoint (r :+ a) -> EndPoint (r :+ a) -> Interval a r

pattern OpenInterval :: (r :+ a) -> (r :+ a) -> Interval a r

pattern ClosedInterval :: (r :+ a) -> (r :+ a) -> Interval a r

fromRange :: Range (r :+ a) -> Interval a r

Constrct an interval from a Range

toRange :: Interval a r -> Range (r :+ a)

_Range :: Lens' (Interval a r) (Range (r :+ a))

querying the start and end of intervals

class HasStart t where

Associated Types

type StartCore t

type StartExtra t

Methods

start :: Lens' t (StartCore t :+ StartExtra t)

class HasEnd t where

Associated Types

type EndCore t

type EndExtra t

Methods

end :: Lens' t (EndCore t :+ EndExtra t)

Working with intervals

inInterval :: Ord r => r -> Interval a r -> Bool

Test if a value lies in an interval. Note that the difference between inInterval and inRange is that the extra value is *not* used in the comparison with inInterval, whereas it is in inRange.

shiftLeft' :: Num r => r -> Interval a r -> Interval a r

Shifts the interval to the left by delta

asProperInterval :: Ord r => Interval p r -> Interval p r

Makes sure the start and endpoint are oriented such that the starting value is smaller than the ending value.

flipInterval :: Interval a r -> Interval a r

Flips the start and endpoint of the interval.

module Data.Range