Data.Geometry.IntervalTree

Description

Documentation

data NodeData i r

Information stored in a node of the Interval Tree

Constructors

NodeData

Fields

_splitPoint :: !r

_intervalsLeft :: !(Map (L r) [i])

_intervalsRight :: !(Map (R r) [i])

splitPoint :: forall i r. Lens' (NodeData i r) r

intervalsLeft :: forall i r. Lens' (NodeData i r) (Map (L r) [i])

intervalsRight :: forall i r. Lens' (NodeData i r) (Map (R r) [i])

newtype IntervalTree i r

IntervalTree type, storing intervals of type i

Constructors

IntervalTree

Fields

_unIntervalTree :: BinaryTree (NodeData i r)

unIntervalTree :: forall i r i r. Iso (IntervalTree i r) (IntervalTree i r) (BinaryTree (NodeData i r)) (BinaryTree (NodeData i r))

class IntervalLike i where

Anything that looks like an interval

Methods

asRange :: i -> Range (NumType i)

createTree :: Ord r => [r] -> IntervalTree i r

Given an ordered list of points, create an interval tree

\(O(n)\)

fromIntervals :: (Ord r, IntervalLike i, NumType i ~ r) => [i] -> IntervalTree i r

Build an interval tree

\(O(n \log n)\)

insert :: (Ord r, IntervalLike i, NumType i ~ r) => i -> IntervalTree i r -> IntervalTree i r

Insert : pre: the interval intersects some midpoint in the tree

\(O(\log n)\)

delete :: (Ord r, IntervalLike i, NumType i ~ r, Eq i) => i -> IntervalTree i r -> IntervalTree i r

Delete an interval from the Tree

\(O(\log n)\) (under some general position assumption)

stab :: Ord r => r -> IntervalTree i r -> [i]

Find all intervals that stab x

\(O(\log n + k)\), where k is the output size

search :: Ord r => r -> IntervalTree i r -> [i]

Find all intervals that stab x

\(O(\log n + k)\), where k is the output size

toList :: IntervalTree i r -> [i]

Lists the intervals. We don't guarantee anything about the order

running time: \(O(n)\).