Data.Geometry.Line.Internal

Description

\(d\)-dimensional lines.

d-dimensional Lines

data Line d r

A line is given by an anchor point and a vector indicating the direction.

Constructors

Line

Fields

_anchorPoint :: !(Point d r)

_direction :: !(Vector d r)

anchorPoint :: Lens' (Line d r) (Point d r)

Line anchor point.

direction :: Lens' (Line d r) (Vector d r)

Line direction.

Functions on lines

lineThrough :: (Num r, Arity d) => Point d r -> Point d r -> Line d r

A line may be constructed from two points.

verticalLine :: Num r => r -> Line 2 r

Vertical line with a given X-coordinate.

horizontalLine :: Num r => r -> Line 2 r

Horizontal line with a given Y-coordinate.

perpendicularTo :: Num r => Line 2 r -> Line 2 r

Given a line l with anchor point p and vector v, get the line perpendicular to l that also goes through p. The resulting line m is oriented such that v points into the left halfplane of m.

>>> perpendicularTo $ Line (Point2 3 4) (Vector2 (-1) 2)
Line (Point2 3 4) (Vector2 [-2,-1])

isPerpendicularTo :: (Num r, Eq r) => Vector 2 r -> Line 2 r -> Bool

Test if a vector is perpendicular to the line.

isIdenticalTo :: (Eq r, Arity d) => Line d r -> Line d r -> Bool

Test if two lines are identical, meaning; if they have exactly the same anchor point and directional vector.

isParallelTo :: (Eq r, Fractional r, Arity d) => Line d r -> Line d r -> Bool

Test if the two lines are parallel.

>>> lineThrough origin (Point2 1 0) `isParallelTo` lineThrough (Point2 1 1) (Point2 2 1)
True
>>> lineThrough origin (Point2 1 0) `isParallelTo` lineThrough (Point2 1 1) (Point2 2 2)
False

onLine :: (Eq r, Fractional r, Arity d) => Point d r -> Line d r -> Bool

Test if point p lies on line l

>>> origin `onLine` lineThrough origin (Point2 1 0)
True
>>> Point2 10 10 `onLine` lineThrough origin (Point2 2 2)
True
>>> Point2 10 5 `onLine` lineThrough origin (Point2 2 2)
False

onLine2 :: (Ord r, Num r) => Point 2 r -> Line 2 r -> Bool

Specific 2d version of testing if apoint lies on a line.

pointAt :: (Num r, Arity d) => r -> Line d r -> Point d r

Get the point at the given position along line, where 0 corresponds to the anchorPoint of the line, and 1 to the point anchorPoint .+^ directionVector

toOffset :: (Eq r, Fractional r, Arity d) => Point d r -> Line d r -> Maybe r

Given point p and a line (Line q v), Get the scalar lambda s.t. p = q + lambda v. If p does not lie on the line this returns a Nothing.

toOffset' :: (Eq r, Fractional r, Arity d) => Point d r -> Line d r -> r

Given point p *on* a line (Line q v), Get the scalar lambda s.t. p = q + lambda v. (So this is an unsafe version of toOffset)

pre: the input point p lies on the line l.

sqDistanceTo :: (Fractional r, Arity d) => Point d r -> Line d r -> r

Squared distance from point p to line l

sqDistanceToArg :: (Fractional r, Arity d) => Point d r -> Line d r -> (r, Point d r)

The squared distance between the point p and the line l, and the point m realizing this distance.

Supporting Lines

class HasSupportingLine t where

Types for which we can compute a supporting line, i.e. a line that contains the thing of type t.

Methods

supportingLine :: t -> Line (Dimension t) (NumType t)

Convenience functions on Two dimensional lines

fromLinearFunction :: Num r => r -> r -> Line 2 r

Create a line from the linear function ax + b

toLinearFunction :: forall r. (Fractional r, Eq r) => Line 2 r -> Maybe (r, r)

get values a,b s.t. the input line is described by y = ax + b. returns Nothing if the line is vertical

data SideTestUpDown

Result of a side test

Constructors

Below

On

Above

class OnSideUpDownTest t where

Methods

onSideUpDown :: (d ~ Dimension t, r ~ NumType t, Ord r, Num r) => Point d r -> t -> SideTestUpDown

data SideTest

Result of a side test

Constructors

LeftSide

OnLine

RightSide

onSide :: (Ord r, Num r) => Point 2 r -> Line 2 r -> SideTest

Given a point q and a line l, compute to which side of l q lies. For vertical lines the left side of the line is interpeted as below.

>>> Point2 10 10 `onSide` (lineThrough origin $ Point2 10 5)
LeftSide
>>> Point2 10 10 `onSide` (lineThrough origin $ Point2 (-10) 5)
RightSide
>>> Point2 5 5 `onSide` (verticalLine 10)
LeftSide
>>> Point2 5 5 `onSide` (lineThrough origin $ Point2 (-3) (-3))
OnLine

liesAbove :: (Ord r, Num r) => Point 2 r -> Line 2 r -> Bool

Test if the query point q lies (strictly) above line l

bisector :: Fractional r => Point 2 r -> Point 2 r -> Line 2 r

Get the bisector between two points

cmpSlope :: (Num r, Ord r) => Line 2 r -> Line 2 r -> Ordering

Compares the lines on slope. Vertical lines are considered larger than anything else.

>>> (Line origin (Vector2 5 1)) `cmpSlope` (Line origin (Vector2 3 3))
LT
>>> (Line origin (Vector2 5 1)) `cmpSlope` (Line origin (Vector2 (-3) 3))
GT
>>> (Line origin (Vector2 5 1)) `cmpSlope` (Line origin (Vector2 0 1))
LT