Data.Geometry.PlanarSubdivision.Basic

Description

Documentation

type VertexId' (s :: k) = VertexId s 'Primal

Shorthand for vertices in the primal.

type FaceId' (s :: k) = FaceId s 'Primal

Shorthand for FaceId's in the primal.

data VertexData r v

Note that the functor instance is in v

Constructors

VertexData !(Point 2 r) !v

vData :: forall r v v. Lens (VertexData r v) (VertexData r v) v v

location :: forall r v r. Lens (VertexData r v) (VertexData r v) (Point 2 r) (Point 2 r)

data FaceData h f

The Face data consists of the data itself and a list of holes

Constructors

FaceData (Seq h) !f

holes :: forall h f h. Lens (FaceData h f) (FaceData h f) (Seq h) (Seq h)

fData :: forall h f f. Lens (FaceData h f) (FaceData h f) f f

data PlanarSubdivision s v e f r

A planarsubdivision is essentially a bunch of plane-graphs; one for every connected component. These graphs store the global ID's (darts, vertexId's, faceId's) in their data values. This essentially gives us a mapping between the two.

note that a face may actually occur in multiple graphs, hence when we store the edges to the the holes, we store the global edgeId's rather than the local edgeId (dart)'s.

invariant: the outerface has faceId 0

Constructors

PlanarSubdivision (Vector (Component s r)) (Vector (Raw s (VertexId' (Wrap s)) v)) (Vector (Raw s (Dart (Wrap s)) e)) (Vector (RawFace s f))

type family Wrap (s :: k) :: k where ...

Equations

Wrap s = Wrap' s

type Component s r = PlaneGraph (Wrap s) (VertexId' s) (Dart s) (FaceId' s) r

A connected component.

For every face f, and every hole in this face, the facedata points to a dart d on the hole s.t. this dart has the face f on its left. i.e. leftFace d = f

data ComponentId s

ComponentId type

data PolygonFaceData

Data type that expresses whether or not we are inside or outside the polygon.

Constructors

Inside

Outside

data PlanarGraph (s :: k) (w :: World) v e f

A *connected* Planar graph with bidirected edges. I.e. the edges (darts) are directed, however, for every directed edge, the edge in the oposite direction is also in the graph.

The types v, e, and f are the are the types of the data associated with the vertices, edges, and faces, respectively.

The orbits in the embedding are assumed to be in counterclockwise order. Therefore, every dart directly bounds the face to its right.

data PlaneGraph s v e f r

Embedded, *connected*, planar graph

fromSimplePolygon

Arguments

:: (Ord r, Fractional r)

=> proxy s

-> SimplePolygon p r

-> f
data inside

-> f
data outside the polygon

-> PlanarSubdivision s p () f r

Construct a plane graph from a simple polygon. It is assumed that the polygon is given in counterclockwise order.

the interior of the polygon will have faceId 0

pre: the input polygon is given in counterclockwise order running time: \(O(n)\).

fromConnectedSegments :: (Foldable f, Ord r, Fractional r) => proxy s -> f (LineSegment 2 p r :+ e) -> PlanarSubdivision s (NonEmpty p) e () r

Constructs a connected planar subdivision.

pre: the segments form a single connected component running time: \(O(n\log n)\)

fromPlaneGraph :: forall s v e f r. (Ord r, Fractional r) => PlaneGraph s v e f r -> PlanarSubdivision s v e f r

Constructs a planarsubdivision from a PlaneGraph

runningTime: \(O(n)\)

fromPlaneGraph' :: forall s v e f r. PlaneGraph s v e f r -> Dart s -> PlanarSubdivision s v e f r

Given a (connected) PlaneGraph and a dart that has the outerface on its left | Constructs a planarsubdivision

runningTime: \(O(n)\)

numComponents :: PlanarSubdivision s v e f r -> Int

Get the number of vertices

>>> numVertices myGraph
1

numVertices :: PlanarSubdivision s v e f r -> Int

Get the number of vertices

>>> numVertices myGraph
4

numEdges :: PlanarSubdivision s v e f r -> Int

Get the number of Edges

>>> numEdges myGraph
6

numFaces :: PlanarSubdivision s v e f r -> Int

\( O(1) \). Get the number of faces

>>> numFaces myGraph
4

numDarts :: PlanarSubdivision s v e f r -> Int

Get the number of Darts

>>> numDarts myGraph
12

dual :: forall k (s :: k) (w :: World) v e f. Getter (PlanarGraph s w v e f) (PlanarGraph s (DualOf w) f e v)

Get the dual graph of this graph.

components :: forall s v e f r r. Lens (PlanarSubdivision s v e f r) (PlanarSubdivision s v e f r) (Vector (Component s r)) (Vector (Component s r))

component :: ComponentId s -> Lens' (PlanarSubdivision s v e f r) (Component s r)

vertices' :: PlanarSubdivision s v e f r -> Vector (VertexId' s)

Enumerate all vertices

>>> vertices' myGraph
[VertexId 0,VertexId 1,VertexId 2,VertexId 3]

vertices :: PlanarSubdivision s v e f r -> Vector (VertexId' s, VertexData r v)

Enumerate all vertices, together with their vertex data

edges' :: PlanarSubdivision s v e f r -> Vector (Dart s)

Enumerate all edges. We report only the Positive darts

edges :: PlanarSubdivision s v e f r -> Vector (Dart s, e)

Enumerate all edges with their edge data. We report only the Positive darts.

>>> mapM_ print $ edges myGraph
(Dart (Arc 2) +1,"c+")
(Dart (Arc 1) +1,"b+")
(Dart (Arc 0) +1,"a+")
(Dart (Arc 5) +1,"g+")
(Dart (Arc 4) +1,"e+")
(Dart (Arc 3) +1,"d+")

faces' :: PlanarSubdivision s v e f r -> Vector (FaceId' s)

\( O(n) \). Vector of all primal faces.

faces :: PlanarSubdivision s v e f r -> Vector (FaceId' s, FaceData (Dart s) f)

\( O(n) \). Vector of all primal faces with associated data.

internalFaces :: PlanarSubdivision s v e f r -> Vector (FaceId' s, FaceData (Dart s) f)

Enumerates all faces with their face data exlcluding the outer face

darts' :: PlanarSubdivision s v e f r -> Vector (Dart s)

Enumerate all darts

headOf :: Dart s -> PlanarSubdivision s v e f r -> VertexId' s

The vertex this dart is heading in to

running time: \(O(1)\)

tailOf :: Dart s -> PlanarSubdivision s v e f r -> VertexId' s

The tail of a dart, i.e. the vertex this dart is leaving from

running time: \(O(1)\)

twin :: forall k (s :: k). Dart s -> Dart s

Get the twin of this dart (edge)

>>> twin (dart 0 "+1")
Dart (Arc 0) -1
>>> twin (dart 0 "-1")
Dart (Arc 0) +1

endPoints :: Dart s -> PlanarSubdivision s v e f r -> (VertexId' s, VertexId' s)

endPoints d g = (tailOf d g, headOf d g)

running time: \(O(1)\)

incidentEdges :: VertexId' s -> PlanarSubdivision s v e f r -> Vector (Dart s)

All edges incident to vertex v, in counterclockwise order around v.

running time: \(O(k)\), where \(k\) is the number of edges reported.

incomingEdges :: VertexId' s -> PlanarSubdivision s v e f r -> Vector (Dart s)

All edges incident to vertex v in incoming direction (i.e. pointing into v) in counterclockwise order around v.

running time: \(O(k)\), where (k) is the total number of incident edges of v

outgoingEdges :: VertexId' s -> PlanarSubdivision s v e f r -> Vector (Dart s)

All edges incident to vertex v in outgoing direction (i.e. pointing away from v) in counterclockwise order around v.

running time: \(O(k)\), where (k) is the total number of incident edges of v

nextIncidentEdge :: Dart s -> PlanarSubdivision s v e f r -> Dart s

Given a dart d that points into some vertex v, report the next dart e in the cyclic order around v.

running time: \(O(1)\)

neighboursOf :: VertexId' s -> PlanarSubdivision s v e f r -> Vector (VertexId' s)

Gets the neighbours of a particular vertex, in counterclockwise order around the vertex.

running time: \(O(k)\), where \(k\) is the output size

leftFace :: Dart s -> PlanarSubdivision s v e f r -> FaceId' s

The face to the left of the dart

running time: \(O(1)\).

rightFace :: Dart s -> PlanarSubdivision s v e f r -> FaceId' s

The face to the right of the dart

running time: \(O(1)\).

outerBoundaryDarts :: FaceId' s -> PlanarSubdivision s v e f r -> Vector (Dart s)

The darts on the outer boundary of the face, for internal faces the darts are in clockwise order. For the outer face the darts are in counterclockwise order, and the darts from various components are in no particular order.

running time: \(O(k)\), where \(k\) is the output size.

boundaryVertices :: FaceId' s -> PlanarSubdivision s v e f r -> Vector (VertexId' s)

The vertices of the outer boundary of the face, for internal faces in clockwise order, for the outer face in counter clockwise order.

running time: \(O(k)\), where \(k\) is the output size.

holesOf :: FaceId' s -> PlanarSubdivision s v e f r -> Seq (Dart s)

Lists the holes in this face, given as a list of darts to arbitrary darts on those faces. The returned darts are on the outside of the hole, i.e. they are incident to the given input face:

all (\d -> leftFace d ps == fi) $ holesOf fi ps

running time: \(O(k)\), where \(k\) is the number of darts returned.

outerFaceId :: PlanarSubdivision s v e f r -> FaceId' s

gets the id of the outer face

running time: \(O(1)\)

boundary' :: Dart s -> PlanarSubdivision s v e f r -> Vector (Dart s)

Generates the darts incident to a face, starting with the given dart.

\(O(k)\), where \(k\) is the number of darts reported

locationOf :: VertexId' s -> Lens' (PlanarSubdivision s v e f r) (Point 2 r)

class HasDataOf g i where

Associated Types

type DataOf g i

Methods

dataOf :: i -> Lens' g (DataOf g i)

get the data associated with the value i.

running time: \(O(1)\) to read the data, \(O(n)\) to write it.

endPointsOf :: Dart s -> Getter (PlanarSubdivision s v e f r) (VertexData r v, VertexData r v)

Getter for the data at the endpoints of a dart

running time: \(O(1)\)

endPointData :: Dart s -> PlanarSubdivision s v e f r -> (VertexData r v, VertexData r v)

data corresponding to the endpoints of the dart

running time: \(O(1)\)

faceDataOf :: FaceId' s -> Lens' (PlanarSubdivision s v e f r) (FaceData (Dart s) f)

edgeSegment :: Dart s -> PlanarSubdivision s v e f r -> LineSegment 2 v r :+ e

Given a dart and the subdivision constructs the line segment representing it

\(O(1)\)

edgeSegments :: PlanarSubdivision s v e f r -> Vector (Dart s, LineSegment 2 v r :+ e)

Reports all edges as line segments

rawFacePolygon :: FaceId' s -> PlanarSubdivision s v e f r -> SomePolygon v r :+ f

Constructs the boundary of the given face

\(O(k)\), where \(k\) is the complexity of the face

rawFaceBoundary :: FaceId' s -> PlanarSubdivision s v e f r -> SimplePolygon v r :+ f

Constructs the outer boundary of the face

\(O(k)\), where \(k\) is the complexity of the outer boundary of the face

rawFacePolygons :: PlanarSubdivision s v e f r -> Vector (FaceId' s, SomePolygon v r :+ f)

Lists all *internal* faces of the planar subdivision.

newtype VertexId (s :: k) (w :: World)

A vertex in a planar graph. A vertex is tied to a particular planar graph by the phantom type s, and to a particular world w.

Constructors

VertexId

Fields

_unVertexId :: Int

newtype FaceId (s :: k) (w :: World)

The type to represent FaceId's

Constructors

FaceId

Fields

_unFaceId :: VertexId s (DualOf w)

data Dart (s :: k)

A dart represents a bi-directed edge. I.e. a dart has a direction, however the dart of the oposite direction is always present in the planar graph as well.

data World

The world in which the graph lives

Constructors

Primal

Dual

rawVertexData :: forall s v e f r v. Lens (PlanarSubdivision s v e f r) (PlanarSubdivision s v e f r) (Vector (Raw s (VertexId' (Wrap s)) v)) (Vector (Raw s (VertexId' (Wrap s)) v))

rawDartData :: forall s v e f r e. Lens (PlanarSubdivision s v e f r) (PlanarSubdivision s v e f r) (Vector (Raw s (Dart (Wrap s)) e)) (Vector (Raw s (Dart (Wrap s)) e))

rawFaceData :: forall s v e f r f. Lens (PlanarSubdivision s v e f r) (PlanarSubdivision s v e f r) (Vector (RawFace s f)) (Vector (RawFace s f))

vertexData :: Lens (PlanarSubdivision s v e f r) (PlanarSubdivision s v' e f r) (Vector v) (Vector v')

Lens to the facedata of the vertexdata themselves. The indices correspond to the vertexId's

dartData :: Lens (PlanarSubdivision s v e f r) (PlanarSubdivision s v e' f r) (Vector (Dart s, e)) (Vector (Dart s, e'))

lens to access the Dart Data

faceData :: Lens (PlanarSubdivision s v e f r) (PlanarSubdivision s v e f' r) (Vector f) (Vector f')

Lens to the facedata of the faces themselves. The indices correspond to the faceIds

dataVal :: Lens (Raw s ia a) (Raw s ia b) a b

get the dataVal of a Raw

dartMapping :: PlanarSubdivision s v e f r -> Vector (Dart (Wrap s), Dart s)

data Raw s ia a

Helper type for the data that we store in a planar subdivision

Constructors

Raw

Fields

_compId :: !(ComponentId s)

_idxVal :: !ia

_dataVal :: !a