Data.Geometry.PlanarSubdivision.Raw

Description

Documentation

data Wrap' s

Helper data type and type family to Wrap a proxy type.

type family Wrap (s :: k) :: k where ...

Equations

Wrap s = Wrap' s

newtype ComponentId s

ComponentId type

Constructors

ComponentId

Fields

unCI :: Int

data Raw s ia a

Helper type for the data that we store in a planar subdivision

Constructors

Raw

Fields

_compId :: !(ComponentId s)

_idxVal :: !ia

_dataVal :: !a

dataVal :: Lens (Raw s ia a) (Raw s ia b) a b

get the dataVal of a Raw

data FaceData h f

The Face data consists of the data itself and a list of holes

Constructors

FaceData

Fields

_holes :: Seq h

_fData :: !f

holes :: forall h f h. Lens (FaceData h f) (FaceData h f) (Seq h) (Seq h)

fData :: forall h f f. Lens (FaceData h f) (FaceData h f) f f

data RawFace s f

Face data, if the face is an inner face, store the component and faceId of it. If not, this face must be the outer face (and thus we can find all the face id's it correponds to through the FaceData).

Constructors

RawFace

Fields

_faceIdx :: !(Maybe (ComponentId s, FaceId' (Wrap s)))

_faceDataVal :: !(FaceData (Dart s) f)

faceIdx :: forall s f. Lens' (RawFace s f) (Maybe (ComponentId s, FaceId' (Wrap s)))

faceDataVal :: forall s f f. Lens (RawFace s f) (RawFace s f) (FaceData (Dart s) f) (FaceData (Dart s) f)