Data.Geometry.Point

Description

\(d\)-dimensional points.

Documentation

newtype Point d r

A d-dimensional point.

There are convenience pattern synonyms for 1, 2 and 3 dimensional points.

>>> let f (Point1 x) = x in f (Point1 1)
1
>>> let f (Point2 x y) = x in f (Point2 1 2)
1
>>> let f (Point3 x y z) = z in f (Point3 1 2 3)
3
>>> let f (Point3 x y z) = z in f (Point $ Vector3 1 2 3)
3

Constructors

Point

Fields

toVec :: Vector d r

Bundled Patterns

pattern Point1 :: r -> Point 1 r
A bidirectional pattern synonym for 1 dimensional points.

pattern Point2 :: r -> r -> Point 2 r
A bidirectional pattern synonym for 2 dimensional points.

pattern Point3 :: r -> r -> r -> Point 3 r
A bidirectional pattern synonym for 3 dimensional points.

origin :: (Arity d, Num r) => Point d r

Point representing the origin in d dimensions

>>> origin :: Point 4 Int
Point4 0 0 0 0

vector :: Lens (Point d r) (Point d r') (Vector d r) (Vector d r')

Lens to access the vector corresponding to this point.

>>> (Point3 1 2 3) ^. vector
Vector3 [1,2,3]
>>> origin & vector .~ Vector3 1 2 3
Point3 1 2 3

pointFromList :: Arity d => [r] -> Maybe (Point d r)

Constructs a point from a list of coordinates. The length of the list has to match the dimension exactly.

>>> pointFromList [1,2,3] :: Maybe (Point 3 Int)
Just (Point3 1 2 3)
>>> pointFromList [1] :: Maybe (Point 3 Int)
Nothing
>>> pointFromList [1,2,3,4] :: Maybe (Point 3 Int)
Nothing

projectPoint :: (Arity i, Arity d, i <= d) => Point d r -> Point i r

Project a point down into a lower dimension.

xCoord :: (1 <= d, Arity d, AsAPoint point) => Lens' (point d r) r

Shorthand to access the first coordinate C 1

>>> Point3 1 2 3 ^. xCoord
1
>>> Point2 1 2 & xCoord .~ 10
Point2 10 2

yCoord :: (2 <= d, Arity d, AsAPoint point) => Lens' (point d r) r

Shorthand to access the second coordinate C 2

>>> Point2 1 2 ^. yCoord
2
>>> Point3 1 2 3 & yCoord %~ (+1)
Point3 1 3 3

zCoord :: (3 <= d, Arity d, AsAPoint point) => Lens' (point d r) r

Shorthand to access the third coordinate C 3

>>> Point3 1 2 3 ^. zCoord
3
>>> Point3 1 2 3 & zCoord %~ (+1)
Point3 1 2 4

class PointFunctor g where

Types that we can transform by mapping a function on each point in the structure

Methods

pmap :: (Point (Dimension (g r)) r -> Point (Dimension (g s)) s) -> g r -> g s

data CCW

Data type for expressing the orientation of three points, with the option of allowing Colinearities.

ccw :: (Ord r, Num r) => Point 2 r -> Point 2 r -> Point 2 r -> CCW

Given three points p q and r determine the orientation when going from p to r via q.

ccw' :: (Ord r, Num r) => (Point 2 r :+ a) -> (Point 2 r :+ b) -> (Point 2 r :+ c) -> CCW

Given three points p q and r determine the orientation when going from p to r via q.

pattern CCW :: CCW

CounterClockwise orientation. Also called a left-turn.

pattern CW :: CCW

Clockwise orientation. Also called a right-turn.

pattern CoLinear :: CCW

CoLinear orientation. Also called a straight line.

ccwCmpAround :: (Num r, Ord r) => (Point 2 r :+ qc) -> (Point 2 r :+ p) -> (Point 2 r :+ q) -> Ordering

Counter clockwise ordering of the points around c. Points are ordered with respect to the positive x-axis.

cwCmpAround :: (Num r, Ord r) => (Point 2 r :+ qc) -> (Point 2 r :+ p) -> (Point 2 r :+ q) -> Ordering

Clockwise ordering of the points around c. Points are ordered with respect to the positive x-axis.

ccwCmpAroundWith :: (Ord r, Num r) => Vector 2 r -> (Point 2 r :+ c) -> (Point 2 r :+ a) -> (Point 2 r :+ b) -> Ordering

Given a zero vector z, a center c, and two points p and q, compute the ccw ordering of p and q around c with this vector as zero direction.

pre: the points p,q /= c

cwCmpAroundWith :: (Ord r, Num r) => Vector 2 r -> (Point 2 r :+ a) -> (Point 2 r :+ b) -> (Point 2 r :+ c) -> Ordering

Given a zero vector z, a center c, and two points p and q, compute the cw ordering of p and q around c with this vector as zero direction.

pre: the points p,q /= c

sortAround :: (Ord r, Num r) => (Point 2 r :+ q) -> [Point 2 r :+ p] -> [Point 2 r :+ p]

Sort the points arround the given point p in counter clockwise order with respect to the rightward horizontal ray starting from p. If two points q and r are colinear with p, the closest one to p is reported first. running time: O(n log n)

insertIntoCyclicOrder :: (Ord r, Num r) => (Point 2 r :+ q) -> (Point 2 r :+ p) -> CList (Point 2 r :+ p) -> CList (Point 2 r :+ p)

Given a center c, a new point p, and a list of points ps, sorted in counter clockwise order around c. Insert p into the cyclic order. The focus of the returned cyclic list is the new point p.

running time: O(n)

data Quadrant

Quadrants of two dimensional points. in CCW order

Constructors

TopRight

TopLeft

BottomLeft

BottomRight

quadrantWith :: (Ord r, 1 <= d, 2 <= d, Arity d) => (Point d r :+ q) -> (Point d r :+ p) -> Quadrant

Quadrants around point c; quadrants are closed on their "previous" boundary (i..e the boundary with the previous quadrant in the CCW order), open on next boundary. The origin itself is assigned the topRight quadrant

quadrant :: (Ord r, Num r, 1 <= d, 2 <= d, Arity d) => (Point d r :+ p) -> Quadrant

Quadrants with respect to the origin

partitionIntoQuadrants :: (Ord r, 1 <= d, 2 <= d, Arity d) => (Point d r :+ q) -> [Point d r :+ p] -> ([Point d r :+ p], [Point d r :+ p], [Point d r :+ p], [Point d r :+ p])

Given a center point c, and a set of points, partition the points into quadrants around c (based on their x and y coordinates). The quadrants are reported in the order topLeft, topRight, bottomLeft, bottomRight. The points are in the same order as they were in the original input lists. Points with the same x-or y coordinate as p, are "rounded" to above.

cmpByDistanceTo :: (Ord r, Num r, Arity d) => (Point d r :+ c) -> (Point d r :+ p) -> (Point d r :+ q) -> Ordering

Compare by distance to the first argument

squaredEuclideanDist :: (Num r, Arity d) => Point d r -> Point d r -> r

Squared Euclidean distance between two points

euclideanDist :: (Floating r, Arity d) => Point d r -> Point d r -> r

Euclidean distance between two points

coord :: (1 <= i, i <= d, KnownNat i, Arity d, AsAPoint p) => proxy i -> Lens' (p d r) r

Get the coordinate in a given dimension

>>> Point3 1 2 3 ^. coord (C :: C 2)
2
>>> Point3 1 2 3 & coord (C :: C 1) .~ 10
Point3 10 2 3
>>> Point3 1 2 3 & coord (C :: C 3) %~ (+1)
Point3 1 2 4

unsafeCoord :: (Arity d, AsAPoint p) => Int -> Lens' (p d r) r

Get the coordinate in a given dimension. This operation is unsafe in the sense that no bounds are checked. Consider using coord instead.

>>> Point3 1 2 3 ^. unsafeCoord 2
2