Data.Geometry.Polygon

Description

A Polygon data type and some basic functions to interact with them.

Types

data PolygonType

We distinguish between simple polygons (without holes) and polygons with holes.

Constructors

Simple

Multi

data Polygon (t :: PolygonType) p r where

Polygons are sequences of points and may or may not contain holes.

Degenerate polygons (polygons with self-intersections or fewer than 3 points) are only possible if you use functions marked as unsafe.

Constructors

SimplePolygon :: Vertices (Point 2 r :+ p) -> SimplePolygon p r

MultiPolygon :: SimplePolygon p r -> [SimplePolygon p r] -> MultiPolygon p r

_SimplePolygon :: Prism' (Polygon Simple p r) (Vertices (Point 2 r :+ p))

Prism to test if we are a simple polygon

>>> is _SimplePolygon simplePoly
True

_MultiPolygon :: Prism' (Polygon Multi p r) (Polygon Simple p r, [Polygon Simple p r])

Prism to test if we are a Multi polygon

>>> is _MultiPolygon multiPoly
True

type SimplePolygon = Polygon Simple

Polygon without holes.

type MultiPolygon = Polygon Multi

Polygon with zero or more holes.

type SomePolygon p r = Either (Polygon Simple p r) (Polygon Multi p r)

Either a simple or multipolygon

Conversion

fromPoints :: forall p r. (Eq r, Num r) => [Point 2 r :+ p] -> SimplePolygon p r

\( O(n) \) Creates a polygon from the given list of vertices.

The points are placed in CCW order if they are not already. Overlapping edges and repeated vertices are allowed.

fromCircularVector :: forall p r. (Eq r, Num r) => CircularVector (Point 2 r :+ p) -> SimplePolygon p r

\( O(n) \) Creates a polygon from the given vector of vertices.

The points are placed in CCW order if they are not already. Overlapping edges and repeated vertices are allowed.

simpleFromPoints :: forall p r. (Ord r, Fractional r) => [Point 2 r :+ p] -> SimplePolygon p r

\( O(n \log n) \) Creates a simple polygon from the given list of vertices.

The points are placed in CCW order if they are not already. Overlapping edges and repeated vertices are not allowed and will trigger an exception.

simpleFromCircularVector :: forall p r. (Ord r, Fractional r) => CircularVector (Point 2 r :+ p) -> SimplePolygon p r

\( O(n \log n) \) Creates a simple polygon from the given vector of vertices.

The points are placed in CCW order if they are not already. Overlapping edges and repeated vertices are not allowed and will trigger an exception.

unsafeFromPoints :: [Point 2 r :+ p] -> SimplePolygon p r

\( O(n) \) Creates a simple polygon from the given list of vertices.

pre: the input list constains no repeated vertices.

unsafeFromCircularVector :: CircularVector (Point 2 r :+ p) -> SimplePolygon p r

\( O(1) \) Creates a simple polygon from the given vector of vertices.

pre: the input list constains no repeated vertices.

unsafeFromVector :: Vector (Point 2 r :+ p) -> SimplePolygon p r

\( O(1) \) Creates a simple polygon from the given vector of vertices.

pre: the input list constains no repeated vertices.

toVector :: Polygon t p r -> Vector (Point 2 r :+ p)

\( O(n) \) Polygon points, from left to right.

toPoints :: Polygon t p r -> [Point 2 r :+ p]

\( O(n) \) Polygon points, from left to right.

Accessors

size :: Polygon t p r -> Int

\( O(1) \) Vertex count. Includes the vertices of holes.

polygonVertices :: Polygon t p r -> NonEmpty (Point 2 r :+ p)

\( O(n) \) The vertices in the polygon. No guarantees are given on the order in which they appear!

listEdges :: Polygon t p r -> [LineSegment 2 p r]

\( O(n) \) Lists all edges. The edges on the outer boundary are given before the ones on the holes. However, no other guarantees are given on the order.

outerBoundary :: forall t p r. Lens' (Polygon t p r) (SimplePolygon p r)

\( O(1) \) Lens access to the outer boundary of a polygon.

outerBoundaryVector :: forall t p r. Getter (Polygon t p r) (CircularVector (Point 2 r :+ p))

Getter access to the outer boundary vector of a polygon.

>>> toList (simpleTriangle ^. outerBoundaryVector)
[Point2 0 0 :+ (),Point2 2 0 :+ (),Point2 1 1 :+ ()]

unsafeOuterBoundaryVector :: forall t p r. Lens' (Polygon t p r) (CircularVector (Point 2 r :+ p))

Unsafe lens access to the outer boundary vector of a polygon.

>>> toList (simpleTriangle ^. unsafeOuterBoundaryVector)
[Point2 0 0 :+ (),Point2 2 0 :+ (),Point2 1 1 :+ ()]

>>> simpleTriangle & unsafeOuterBoundaryVector .~ CV.singleton (Point2 0 0 :+ ())
SimplePolygon [Point2 0 0 :+ ()]

outerBoundaryEdges :: Polygon t p r -> CircularVector (LineSegment 2 p r)

\( O(n) \) The edges along the outer boundary of the polygon. The edges are half open.

outerVertex :: Int -> Getter (Polygon t p r) (Point 2 r :+ p)

\( O(1) \) Access the i^th vertex on the outer boundary

>>> simplePoly ^. outerVertex 0
Point2 0 0 :+ ()

outerBoundaryEdge :: Int -> Polygon t p r -> LineSegment 2 p r

\( O(1) \) Get the n^th edge along the outer boundary of the polygon. The edge is half open.

polygonHoles :: forall p r. Lens' (Polygon Multi p r) [Polygon Simple p r]

Lens access for polygon holes.

>>> multiPoly ^. polygonHoles
[SimplePolygon [Point2 0 0 :+ (),Point2 2 0 :+ (),Point2 1 1 :+ ()]]

polygonHoles' :: Traversal' (Polygon t p r) [Polygon Simple p r]

\( O(1) \). Traversal lens for polygon holes. Does nothing for simple polygons.

holeList :: Polygon t p r -> [Polygon Simple p r]

Get all holes in a polygon

Properties

area :: Fractional r => Polygon t p r -> r

Compute the area of a polygon

signedArea :: Fractional r => SimplePolygon p r -> r

Compute the signed area of a simple polygon. The the vertices are in clockwise order, the signed area will be negative, if the verices are given in counter clockwise order, the area will be positive.

centroid :: Fractional r => SimplePolygon p r -> Point 2 r

Compute the centroid of a simple polygon.

Queries

inPolygon :: forall t p r. (Fractional r, Ord r) => Point 2 r -> Polygon t p r -> PointLocationResult

Check if a point lies inside a polygon, on the boundary, or outside of the polygon. Running time: O(n).

>>> Point2 1 1 `inPolygon` simplePoly
Inside
>>> Point2 0 0 `inPolygon` simplePoly
OnBoundary
>>> Point2 10 0 `inPolygon` simplePoly
OnBoundary
>>> Point2 5 13 `inPolygon` simplePoly
Inside
>>> Point2 5 10 `inPolygon` simplePoly
Inside
>>> Point2 10 5 `inPolygon` simplePoly
OnBoundary
>>> Point2 20 5 `inPolygon` simplePoly
Outside

TODO: Add some testcases with multiPolygons TODO: Add some more onBoundary testcases

insidePolygon :: (Fractional r, Ord r) => Point 2 r -> Polygon t p r -> Bool

Test if a point lies strictly inside the polgyon.

onBoundary :: (Fractional r, Ord r) => Point 2 r -> Polygon t p r -> Bool

\( O(n) \) Test if q lies on the boundary of the polygon.

>>> Point2 1 1 `onBoundary` simplePoly
False
>>> Point2 0 0 `onBoundary` simplePoly
True
>>> Point2 10 0 `onBoundary` simplePoly
True
>>> Point2 5 13 `onBoundary` simplePoly
False
>>> Point2 5 10 `onBoundary` simplePoly
False
>>> Point2 10 5 `onBoundary` simplePoly
True
>>> Point2 20 5 `onBoundary` simplePoly
False

TODO: testcases multipolygon

isTriangle :: Polygon p t r -> Bool

\( O(1) \) Test if the polygon is a triangle

isStarShaped :: (MonadRandom m, Ord r, Fractional r) => SimplePolygon p r -> m (Maybe (Point 2 r))

Test if a Simple polygon is star-shaped. Returns a point in the kernel (i.e. from which the entire polygon is visible), if it exists.

\(O(n)\) expected time

isCounterClockwise :: (Eq r, Num r) => Polygon t p r -> Bool

\( O(n) \) Test if the outer boundary of the polygon is in clockwise or counter clockwise order.

toCounterClockWiseOrder :: (Eq r, Num r) => Polygon t p r -> Polygon t p r

\( O(n) \) Make sure that every edge has the polygon's interior on its left, by orienting the outer boundary into counter-clockwise order, and the inner borders (i.e. any holes, if they exist) into clockwise order.

toCounterClockWiseOrder' :: (Eq r, Num r) => Polygon t p r -> Polygon t p r

\( O(n) \) Orient the outer boundary into counter-clockwise order. Leaves any holes as they are.

toClockwiseOrder :: (Eq r, Num r) => Polygon t p r -> Polygon t p r

\( O(n) \) Make sure that every edge has the polygon's interior on its right, by orienting the outer boundary into clockwise order, and the inner borders (i.e. any holes, if they exist) into counter-clockwise order.

toClockwiseOrder' :: (Eq r, Num r) => Polygon t p r -> Polygon t p r

\( O(n) \) Orient the outer boundary into clockwise order. Leaves any holes as they are.

reverseOuterBoundary :: Polygon t p r -> Polygon t p r

Reorient the outer boundary from clockwise order to counter-clockwise order or from counter-clockwise order to clockwise order. Leaves any holes as they are.

Misc

pickPoint :: (Ord r, Fractional r) => Polygon p t r -> Point 2 r

\( O(n) \) Pick a point that is inside the polygon.

(note: if the polygon is degenerate; i.e. has <3 vertices, we report a vertex of the polygon instead.)

pre: the polygon is given in CCW order

findDiagonal :: (Ord r, Fractional r) => Polygon t p r -> LineSegment 2 p r

\( O(n) \) Find a diagonal of the polygon.

pre: the polygon is given in CCW order

withIncidentEdges :: Polygon t p r -> Polygon t (Two (LineSegment 2 p r)) r

Pairs every vertex with its incident edges. The first one is its predecessor edge, the second one its successor edge (in terms of the ordering along the boundary).

>>> mapM_ print . polygonVertices $ withIncidentEdges simplePoly
Point2 0 0 :+ V2 LineSegment (Closed (Point2 1 11 :+ ())) (Closed (Point2 0 0 :+ ())) LineSegment (Closed (Point2 0 0 :+ ())) (Closed (Point2 10 0 :+ ()))
Point2 10 0 :+ V2 LineSegment (Closed (Point2 0 0 :+ ())) (Closed (Point2 10 0 :+ ())) LineSegment (Closed (Point2 10 0 :+ ())) (Closed (Point2 10 10 :+ ()))
Point2 10 10 :+ V2 LineSegment (Closed (Point2 10 0 :+ ())) (Closed (Point2 10 10 :+ ())) LineSegment (Closed (Point2 10 10 :+ ())) (Closed (Point2 5 15 :+ ()))
Point2 5 15 :+ V2 LineSegment (Closed (Point2 10 10 :+ ())) (Closed (Point2 5 15 :+ ())) LineSegment (Closed (Point2 5 15 :+ ())) (Closed (Point2 1 11 :+ ()))
Point2 1 11 :+ V2 LineSegment (Closed (Point2 5 15 :+ ())) (Closed (Point2 1 11 :+ ())) LineSegment (Closed (Point2 1 11 :+ ())) (Closed (Point2 0 0 :+ ()))

numberVertices :: Polygon t p r -> Polygon t (SP Int p) r

assigns unique integer numbers to all vertices. Numbers start from 0, and are increasing along the outer boundary. The vertices of holes will be numbered last, in the same order.

>>> numberVertices simplePoly
SimplePolygon [Point2 0 0 :+ SP 0 (),Point2 10 0 :+ SP 1 (),Point2 10 10 :+ SP 2 (),Point2 5 15 :+ SP 3 (),Point2 1 11 :+ SP 4 ()]

extremesLinear :: (Ord r, Num r) => Vector 2 r -> Polygon t p r -> (Point 2 r :+ p, Point 2 r :+ p)

Finds the extreme points, minimum and maximum, in a given direction

running time: \(O(n)\)

cmpExtreme :: (Num r, Ord r) => Vector 2 r -> (Point 2 r :+ p) -> (Point 2 r :+ q) -> Ordering

Comparison that compares which point is larger in the direction given by the vector u.