Data.Geometry.Polygon.Inflate

Description

Documentation

data Arc r

Points annotated with an Arc indicate that the edge from this point to the next should not be a straight line but instead an arc with a given center and a given border edge.

Constructors

Arc

Fields

arcCenter :: Point 2 r

arcEdge :: (Point 2 r, Point 2 r)

inflate :: (Real r, Fractional r) => Double -> SimplePolygon () r -> SimplePolygon (Arc r) r

\( O(n \log n) \)