Data.Geometry.QuadTree.Split

Description

Documentation

data Split i v p

Data Type to Decide if we should continue splitting the current cell

Constructors

No !p

Yes !v (Quadrants i)

_Yes :: forall i v p i v. Prism (Split i v p) (Split i v p) (v, Quadrants i) (v, Quadrants i)

_No :: forall i v p p. Prism (Split i v p) (Split i v p) p p

type Splitter r i v p = Cell r -> i -> Split i v p

A splitter is a function that determines weather or not we should the given cell corresponding to the given input (i).

type Limiter r i v p = Splitter r i v p -> Splitter r i v (Either i p)

Transformer that limits the depth of a splitter

limitWidthTo

Arguments

:: WidthIndex
smallest allowed width of a cell (i.e. width of a leaf)

-> Limiter r i v p

Split only when the Cell-width is at least wMin