Data.Geometry.RangeTree.Generic

Description

Documentation

data NodeData v r

Constructors

NodeData

Fields

_minVal :: !(Min r)

_maxVal :: !(Max r)

_assoc :: !v

newtype RangeTree v q r

A generic (1D) range tree. The r parameter indicates the type of the coordinates of the points. The q represents any associated data values with those points (stored in the leaves), and the v types represents the data stored at internal nodes.

Constructors

RangeTree

Fields

_unRangeTree :: BinLeafTree (NodeData v r) (NodeData (v, q) r)

createTree :: (Ord r, Measured v p, Semigroup p) => NonEmpty (r :+ p) -> RangeTree v p r

Creates a range tree

createTree' :: (Ord r, Measured v p) => NonEmpty (r :+ p) -> RangeTree v p r

pre: input is sorted and grouped by x-coord

Converting to a List

toAscList :: RangeTree v p r -> NonEmpty (r :+ p)

Lists all points in increasing order

running time: \(O(n)\)

Querying x

search :: (Ord r, Monoid v) => Range r -> RangeTree v p r -> v

Range search

running time: \(O(\log n)\)

search' :: Ord r => Range r -> RangeTree v p r -> [v]

Range search, report the (associated data structures of the) \(O(\log n)\) nodes that form the disjoint union of the range we are querying with.

running time: \(O(\log n)\)

search'' :: Ord r => Range r -> BinLeafTree (NodeData v r) (NodeData (v, q) r) -> [v]

The actual search

rangeOf :: BinLeafTree (NodeData v r) (NodeData v' r) -> Range r

Helper function to get the range of a binary leaf tree

rangeOf' :: NodeData v r -> Range r

Get the range of a node

Updates

createReportingTree :: Ord r => NonEmpty (r :+ [p]) -> RangeTree (Report p) (Report p) r

report :: Ord r => Range r -> RangeTree (Report p) q r -> [p]

newtype CountOf p

Constructors

CountOf [p]

createCountingTree :: Ord r => NonEmpty (r :+ [p]) -> RangeTree (Count p) (CountOf p) r

count :: Ord r => Range r -> RangeTree (Count p) q r -> Int

Perform a counting query