Data.Geometry.SegmentTree.Generic

Description

Documentation

data NodeData v r

Internal nodes store a split point, the range, and an associated data structure

Constructors

NodeData

Fields

_splitPoint :: !(EndPoint r)

_range :: !(Range r)

_assoc :: !v

splitPoint :: forall v r. Lens' (NodeData v r) (EndPoint r)

range :: forall v r. Lens' (NodeData v r) (Range r)

assoc :: forall v r v. Lens (NodeData v r) (NodeData v r) v v

data LeafData v r

Leaf nodes store an atomic range, and an associated data structure.

Constructors

LeafData

Fields

_atomicRange :: !(AtomicRange r)

_leafAssoc :: !v

atomicRange :: forall v r r. Lens (LeafData v r) (LeafData v r) (AtomicRange r) (AtomicRange r)

leafAssoc :: forall v r v. Lens (LeafData v r) (LeafData v r) v v

newtype SegmentTree v r

Segment tree on a Fixed set of endpoints

Constructors

SegmentTree

Fields

_unSegmentTree :: BinLeafTree (NodeData v r) (LeafData v r)

unSegmentTree :: forall v r v r. Iso (SegmentTree v r) (SegmentTree v r) (BinLeafTree (NodeData v r) (LeafData v r)) (BinLeafTree (NodeData v r) (LeafData v r))

class Measured v i => Assoc v i where

Class for associcated data structures

Methods

insertAssoc :: i -> v -> v

deleteAssoc :: i -> v -> v

createTree :: NonEmpty r -> v -> SegmentTree v r

Given a sorted list of endpoints, without duplicates, construct a segment tree

\(O(n)\) time

fromIntervals :: (Ord r, Eq p, Assoc v i, IntervalLike i, Monoid v, NumType i ~ r) => (Interval p r -> i) -> NonEmpty (Interval p r) -> SegmentTree v r

Build a SegmentTree

\(O(n \log n)\)

insert :: (Assoc v i, NumType i ~ r, Ord r, IntervalLike i) => i -> SegmentTree v r -> SegmentTree v r

Pre: the interval should have one of the endpoints on which the tree is built.

delete :: (Assoc v i, NumType i ~ r, Ord r, IntervalLike i) => i -> SegmentTree v r -> SegmentTree v r

Delete an interval from the tree

pre: The segment is in the tree!

search :: (Ord r, Monoid v) => r -> SegmentTree v r -> v

Search for all intervals intersecting x

\(O(\log n + k)\) where \(k\) is the output size

stab :: Ord r => r -> SegmentTree v r -> [v]

Returns the associated values of the nodes on the search path to x

\(O(\log n)\)

newtype I a

Interval

Constructors

I

Fields

_unI :: a

fromIntervals' :: (Eq p, Ord r) => NonEmpty (Interval p r) -> SegmentTree [I (Interval p r)] r

newtype Count

Constructors

Count

Fields

getCount :: Word