Data.Geometry.SubLine

Description

SubLine; a part of a line

Documentation

data SubLine d p s r

Part of a line. The interval is ranged based on the vector of the line l, and s.t.t zero is the anchorPoint of l.

Constructors

SubLine

Fields

_line :: Line d r

_subRange :: Interval p s

subRange :: forall d p s r p s. Lens (SubLine d p s r) (SubLine d p s r) (Interval p s) (Interval p s)

line :: forall d p s r d r. Lens (SubLine d p s r) (SubLine d p s r) (Line d r) (Line d r)

fixEndPoints :: (Num r, Arity d) => SubLine d p r r -> SubLine d (Point d r :+ p) r r

Annotate the subRange with the actual ending points

dropExtra :: SubLine d p s r -> SubLine d () s r

forget the extra information stored at the endpoints of the subline.

_unBounded :: Prism' (SubLine d p (UnBounded r) r) (SubLine d p r r)

toUnbounded :: SubLine d p r r -> SubLine d p (UnBounded r) r

Transform into an subline with a potentially unbounded interval

fromUnbounded :: SubLine d p (UnBounded r) r -> Maybe (SubLine d p r r)

Try to make a potentially unbounded subline into a bounded one.

onSubLine :: (Ord r, Fractional r, Arity d) => Point d r -> SubLine d p r r -> Bool

given point p, and a Subline l r such that p lies on line l, test if it lies on the subline, i.e. in the interval r

onSubLineUB :: (Ord r, Fractional r) => Point 2 r -> SubLine 2 p (UnBounded r) r -> Bool

given point p, and a Subline l r such that p lies on line l, test if it lies on the subline, i.e. in the interval r

onSubLine2 :: (Ord r, Num r) => Point 2 r -> SubLine 2 p r r -> Bool

given point p, and a Subline l r such that p lies on line l, test if it lies on the subline, i.e. in the interval r

onSubLine2UB :: (Ord r, Fractional r) => Point 2 r -> SubLine 2 p (UnBounded r) r -> Bool

given point p, and a Subline l r such that p lies on line l, test if it lies on the subline, i.e. in the interval r

getEndPointsUnBounded :: (Num r, Arity d) => SubLine d p (UnBounded r) r -> Interval p (UnBounded (Point d r))

Get the endpoints of an unbounded interval

fromLine :: Arity d => Line d r -> SubLine d () (UnBounded r) r

testz :: SubLine 2 () Rational Rational