Data.Geometry.Vector.VectorFamily

Description

Implementation of \(d\)-dimensional vectors. The implementation automatically selects an optimized representation for small (up to size 4) vectors.

d dimensional Vectors

newtype Vector (d :: Nat) (r :: *)

Datatype representing d dimensional vectors. The default implementation is based n VectorFixed. However, for small vectors we automatically select a more efficient representation.

Constructors

MKVector

Fields

_unV :: VectorFamily (Peano d) r

unV :: Iso (Vector d r) (Vector d s) (VectorFamily (Peano d) r) (VectorFamily (Peano d) s)

Vectors are isomorphic to a definition determined by VectorFamily.

class (ImplicitArity (Peano d), KnownNat d) => Arity d

Convenience "constructors"

pattern Vector :: VectorFamilyF (Peano d) r -> Vector d r

Constant sized vector with d elements.

pattern Vector1 :: r -> Vector 1 r

Constant sized vector with 1 element.

pattern Vector2 :: r -> r -> Vector 2 r

Constant sized vector with 2 elements.

pattern Vector3 :: r -> r -> r -> Vector 3 r

Constant sized vector with 3 elements.

pattern Vector4 :: r -> r -> r -> r -> Vector 4 r

Constant sized vector with 4 elements.

vectorFromList :: Arity d => [r] -> Maybe (Vector d r)

\( O(n) \) Convert from a list to a non-empty vector.

vectorFromListUnsafe :: Arity d => [r] -> Vector d r

\( O(n) \) Convert from a list to a non-empty vector.

destruct :: (Arity d, Arity (d + 1)) => Vector (d + 1) r -> (r, Vector d r)

\( O(n) \) Pop the first element off a vector.

head :: (Arity d, 1 <= d) => Vector d r -> r

\( O(1) \) First element. Since arity is at least 1, this function is total.

Indexing vectors

element :: forall proxy i d r. (Arity d, KnownNat i, (i + 1) <= d) => proxy i -> Lens' (Vector d r) r

Lens into the i th element

element' :: forall d r. Arity d => Int -> Traversal' (Vector d r) r

Similar to element above. Except that we don't have a static guarantee that the index is in bounds. Hence, we can only return a Traversal

Snoccing and consindg

cons :: (Arity d, Arity (d + 1)) => r -> Vector d r -> Vector (d + 1) r

\( O(n) \) Prepend an element.

snoc :: (Arity (d + 1), Arity d) => Vector d r -> r -> Vector (d + 1) r

Add an element at the back of the vector

init :: (Arity d, Arity (d + 1)) => Vector (d + 1) r -> Vector d r

Get a vector of the first d - 1 elements.

last :: forall d r. (KnownNat d, Arity (d + 1)) => Vector (d + 1) r -> r

\( O(1) \) Last element. Since the vector is non-empty, runtime bounds checks are bypassed.

prefix :: forall i d r. (Arity d, Arity i, i <= d) => Vector d r -> Vector i r

Get a prefix of i elements of a vector

Specific on 3-dimensional vectors

cross :: Num r => Vector 3 r -> Vector 3 r -> Vector 3 r

Cross product of two three-dimensional vectors