Data.Geometry.Vector.VectorFamilyPeano

Description

Documentation

class (ImplicitPeano d, Arity (FromPeano d)) => ImplicitArity d

newtype VectorFamily (d :: PeanoNum) (r :: *)

Datatype representing d dimensional vectors. The default implementation is based n VectorFixed. However, for small vectors we automatically select a more efficient representation.

Constructors

VectorFamily (VectorFamilyF d r)

type family VectorFamilyF (d :: PeanoNum) :: * -> * where ...

Mapping between the implementation type, and the actual implementation.

Equations

VectorFamilyF Z = Const ()

VectorFamilyF One = Identity

VectorFamilyF Two = V2

VectorFamilyF Three = V3

VectorFamilyF Four = V4

VectorFamilyF (Many d) = Vector (FromPeano (Many d))

type family FromPeano (d :: PeanoNum) :: Nat where ...

Equations

FromPeano Z = 0

FromPeano (S d) = 1 + FromPeano d

type Two = S One