Data.Geometry.Vector.VectorFixed

Documentation

data C (n :: Nat)

A proxy which can be used for the coordinates.

Constructors

C

d dimensional Vectors

newtype Vector (d :: Nat) (r :: *)

Datatype representing d dimensional vectors. Our implementation wraps the implementation provided by fixed-vector.

Constructors

Vector

Fields

_unV :: Vec d r

unV :: Lens' (Vector d r) (Vec d r)

element :: forall proxy i d r. (Arity d, Arity i, (i + 1) <= d) => proxy i -> Lens' (Vector d r) r

Lens into the i th element

element' :: forall d r. Arity d => Int -> Traversal' (Vector d r) r

Similar to element above. Except that we don't have a static guarantee that the index is in bounds. Hence, we can only return a Traversal

vectorFromList :: Arity d => [a] -> Maybe (Vector d a)

vectorFromListUnsafe :: Arity d => [a] -> Vector d a

destruct :: (Arity d, Arity (d + 1), 1 <= (d + 1)) => Vector (d + 1) r -> (r, Vector d r)

Get the head and tail of a vector

cross :: Num r => Vector 3 r -> Vector 3 r -> Vector 3 r

Cross product of two three-dimensional vectors

toV2 :: Vector 2 a -> V2 a

Vonversion to a Linear.V2

toV3 :: Vector 3 a -> V3 a

Conversion to a Linear.V3

fromV3 :: V3 a -> Vector 3 a

Conversion from a Linear.V3

snoc :: (Arity (d + 1), Arity d) => Vector d r -> r -> Vector (d + 1) r

Add an element at the back of the vector

init :: (Arity d, Arity (d + 1)) => Vector (d + 1) r -> Vector d r

Get a vector of the first d - 1 elements.

last :: forall d r. (Arity d, Arity (d + 1)) => Vector (d + 1) r -> r

prefix :: forall i d r. (Arity d, Arity i, i <= d) => Vector d r -> Vector i r

Get a prefix of i elements of a vector

Functions specific to two and three dimensional vectors.

v2 :: r -> r -> Vector 2 r

Construct a 2 dimensional vector

v3 :: r -> r -> r -> Vector 3 r

Construct a 3 dimensional vector

_unV2 :: Vector 2 r -> (r, r)

Destruct a 2 dim vector into a pair

_unV3 :: Vector 3 r -> (r, r, r)

pattern Vector2 :: r -> r -> Vector 2 r

Pattern synonym for two and three dim vectors

pattern Vector3 :: r -> r -> r -> Vector 3 r

pattern Vector4 :: r -> r -> r -> r -> Vector 4 r