Data.PlaneGraph

Description

Data type for planar graphs embedded in \(\mathbb{R}^2\). For functions that export faces and edges etc, we assume the graph has a (planar) straight line embedding.

Documentation

newtype PlaneGraph s v e f r

Embedded, *connected*, planar graph

Constructors

PlaneGraph (PlanarGraph s Primal (VertexData r v) e f)

graph :: forall s v e f r s v e f r. Iso (PlaneGraph s v e f r) (PlaneGraph s v e f r) (PlanarGraph s 'Primal (VertexData r v) e f) (PlanarGraph s 'Primal (VertexData r v) e f)

data PlanarGraph (s :: k) (w :: World) v e f

A *connected* Planar graph with bidirected edges. I.e. the edges (darts) are directed, however, for every directed edge, the edge in the oposite direction is also in the graph.

The types v, e, and f are the are the types of the data associated with the vertices, edges, and faces, respectively.

The orbits in the embedding are assumed to be in counterclockwise order. Therefore, every dart directly bounds the face to its right.

data VertexData r v

Note that the functor instance is in v

Constructors

VertexData !(Point 2 r) !v

vData :: forall r v v. Lens (VertexData r v) (VertexData r v) v v

location :: forall r v r. Lens (VertexData r v) (VertexData r v) (Point 2 r) (Point 2 r)

vtxDataToExt :: VertexData r v -> Point 2 r :+ v

fromSimplePolygon

Arguments

:: proxy s

-> SimplePolygon p r

-> f
data inside

-> f
data outside the polygon

-> PlaneGraph s p () f r

Construct a plane graph from a simple polygon. It is assumed that the polygon is given in counterclockwise order.

the interior of the polygon will have faceId 0

pre: the input polygon is given in counterclockwise order running time: \(O(n)\).

fromConnectedSegments :: (Foldable f, Ord r, Num r) => proxy s -> f (LineSegment 2 p r :+ e) -> PlaneGraph s (NonEmpty p) e () r

Constructs a connected plane graph

pre: The segments form a single connected component

running time: \(O(n\log n)\)

toAdjRep :: PlaneGraph s v e f r -> Gr (Vtx v e r) (Face f)

Transforms the plane graph into adjacency lists. For every vertex, the adjacent vertices are given in counter clockwise order.

See toAdjacencyLists for notes on how we handle self-loops.

running time: \(O(n)\)

fromAdjRep :: proxy s -> Gr (Vtx v e r) (Face f) -> PlaneGraph s v e f r

Given the AdjacencyList representation of a plane graph, construct the plane graph representing it. All the adjacencylists should be in counter clockwise order.

running time: \(O(n)\)

numVertices :: PlaneGraph s v e f r -> Int

Get the number of vertices

>>> numVertices smallG
4

numEdges :: PlaneGraph s v e f r -> Int

Get the number of Edges

>>> numEdges smallG
5

numFaces :: PlaneGraph s v e f r -> Int

Get the number of faces

>>> numFaces smallG
3

numDarts :: PlaneGraph s v e f r -> Int

Get the number of Darts

>>> numDarts smallG
10

dual :: forall k (s :: k) (w :: World) v e f. Getter (PlanarGraph s w v e f) (PlanarGraph s (DualOf w) f e v)

Get the dual graph of this graph.

vertices' :: PlaneGraph s v e f r -> Vector (VertexId' s)

Enumerate all vertices

>>> vertices' smallG
[VertexId 0,VertexId 1,VertexId 2,VertexId 3]

vertices :: PlaneGraph s v e f r -> Vector (VertexId' s, VertexData r v)

Enumerate all vertices, together with their vertex data

>>> mapM_ print $ vertices smallG
(VertexId 0,VertexData {_location = Point2 0 0, _vData = 0})
(VertexId 1,VertexData {_location = Point2 2 2, _vData = 1})
(VertexId 2,VertexData {_location = Point2 2 0, _vData = 2})
(VertexId 3,VertexData {_location = Point2 (-1) 4, _vData = 3})

edges' :: PlaneGraph s v e f r -> Vector (Dart s)

Enumerate all edges. We report only the Positive darts

edges :: PlaneGraph s v e f r -> Vector (Dart s, e)

Enumerate all edges with their edge data. We report only the Positive darts.

>>> mapM_ print $ edges smallG
(Dart (Arc 0) +1,"0->2")
(Dart (Arc 1) +1,"0->1")
(Dart (Arc 2) +1,"0->3")
(Dart (Arc 4) +1,"1->2")
(Dart (Arc 3) +1,"1->3")

faces' :: PlaneGraph s v e f r -> Vector (FaceId' s)

Enumerate all faces in the plane graph

faces :: PlaneGraph s v e f r -> Vector (FaceId' s, f)

All faces with their face data.

>>> mapM_ print $ faces smallG
(FaceId 0,"OuterFace")
(FaceId 1,"A")
(FaceId 2,"B")

internalFaces :: (Ord r, Fractional r) => PlaneGraph s v e f r -> Vector (FaceId' s, f)

Reports all internal faces. running time: \(O(n)\)

faces'' :: (Ord r, Fractional r) => PlaneGraph s v e f r -> ((FaceId' s, f), Vector (FaceId' s, f))

Reports the outerface and all internal faces separately. running time: \(O(n)\)

darts' :: PlaneGraph s v e f r -> Vector (Dart s)

Enumerate all darts

darts :: PlaneGraph s v e f r -> Vector (Dart s, e)

Get all darts together with their data

traverseVertices :: Applicative m => (VertexId' s -> v -> m v') -> PlaneGraph s v e f r -> m (PlaneGraph s v' e f r)

Traverse the vertices

(^.vertexData) $ traverseVertices (i x -> Just (i,x)) smallG Just [(VertexId 0,0),(VertexId 1,1),(VertexId 2,2),(VertexId 3,3)] >>> traverseVertices (i x -> print (i,x)) smallG >> pure () (VertexId 0,0) (VertexId 1,1) (VertexId 2,2) (VertexId 3,3)

traverseDarts :: Applicative m => (Dart s -> e -> m e') -> PlaneGraph s v e f r -> m (PlaneGraph s v e' f r)

Traverses the darts

>>> traverseDarts (\d x -> print (d,x)) smallG >> pure ()
(Dart (Arc 0) +1,"0->2")
(Dart (Arc 0) -1,"2->0")
(Dart (Arc 1) +1,"0->1")
(Dart (Arc 1) -1,"1->0")
(Dart (Arc 2) +1,"0->3")
(Dart (Arc 2) -1,"3->0")
(Dart (Arc 3) +1,"1->3")
(Dart (Arc 3) -1,"3->1")
(Dart (Arc 4) +1,"1->2")
(Dart (Arc 4) -1,"2->1")

traverseFaces :: Applicative m => (FaceId' s -> f -> m f') -> PlaneGraph s v e f r -> m (PlaneGraph s v e f' r)

Traverses the faces

>>> traverseFaces (\i x -> print (i,x)) smallG >> pure ()
(FaceId 0,"OuterFace")
(FaceId 1,"A")
(FaceId 2,"B")

headOf :: Dart s -> PlaneGraph s v e f r -> VertexId' s

The vertex this dart is heading in to

running time: \(O(1)\)

>>> headOf (dart 0 "+1") smallG
VertexId 2

tailOf :: Dart s -> PlaneGraph s v e f r -> VertexId' s

The tail of a dart, i.e. the vertex this dart is leaving from

running time: \(O(1)\)

>>> tailOf (dart 0 "+1") smallG
VertexId 0

twin :: forall k (s :: k). Dart s -> Dart s

Get the twin of this dart (edge)

>>> twin (dart 0 "+1")
Dart (Arc 0) -1
>>> twin (dart 0 "-1")
Dart (Arc 0) +1

endPoints :: Dart s -> PlaneGraph s v e f r -> (VertexId' s, VertexId' s)

endPoints d g = (tailOf d g, headOf d g)

running time: \(O(1)\)

>>> endPoints (dart 0 "+1") smallG
(VertexId 0,VertexId 2)

incidentEdges :: VertexId' s -> PlaneGraph s v e f r -> Vector (Dart s)

All edges incident to vertex v, in counterclockwise order around v.

running time: \(O(k)\), where \(k\) is the output size

>>> incidentEdges (VertexId 1) smallG
[Dart (Arc 1) -1,Dart (Arc 4) +1,Dart (Arc 3) +1]

incomingEdges :: VertexId' s -> PlaneGraph s v e f r -> Vector (Dart s)

All edges incident to vertex v in incoming direction (i.e. pointing into v) in counterclockwise order around v.

running time: \(O(k)\), where (k) is the total number of incident edges of v

>>> incomingEdges (VertexId 1) smallG
[Dart (Arc 1) +1,Dart (Arc 4) -1,Dart (Arc 3) -1]

outgoingEdges :: VertexId' s -> PlaneGraph s v e f r -> Vector (Dart s)

All edges incident to vertex v in incoming direction (i.e. pointing into v) in counterclockwise order around v.

running time: \(O(k)\), where (k) is the total number of incident edges of v

>>> outgoingEdges (VertexId 1) smallG
[Dart (Arc 1) -1,Dart (Arc 4) +1,Dart (Arc 3) +1]

neighboursOf :: VertexId' s -> PlaneGraph s v e f r -> Vector (VertexId' s)

Gets the neighbours of a particular vertex, in counterclockwise order around the vertex.

running time: \(O(k)\), where \(k\) is the output size

>>> neighboursOf (VertexId 1) smallG
[VertexId 0,VertexId 2,VertexId 3]

nextIncidentEdge :: Dart s -> PlaneGraph s v e f r -> Dart s

Given a dart d that points into some vertex v, report the next dart in the cyclic order around v in clockwise direction.

running time: \(O(1)\)

>>> nextIncidentEdge (dart 1 "+1") smallG
Dart (Arc 2) +1

prevIncidentEdge :: Dart s -> PlaneGraph s v e f r -> Dart s

Given a dart d that points into some vertex v, report the next dart in the cyclic order around v (in clockwise order)

running time: \(O(1)\)

>>> prevIncidentEdge (dart 1 "+1") smallG
Dart (Arc 0) +1

leftFace :: Dart s -> PlaneGraph s v e f r -> FaceId' s

The face to the left of the dart

running time: \(O(1)\).

>>> leftFace (dart 1 "+1") smallG
FaceId 2
>>> leftFace (dart 1 "-1") smallG
FaceId 1
>>> leftFace (dart 2 "+1") smallG
FaceId 0
>>> leftFace (dart 2 "-1") smallG
FaceId 2

rightFace :: Dart s -> PlaneGraph s v e f r -> FaceId' s

The face to the right of the dart

running time: \(O(1)\).

>>> rightFace (dart 1 "+1") smallG
FaceId 1
>>> rightFace (dart 1 "-1") smallG
FaceId 2
>>> rightFace (dart 2 "+1") smallG
FaceId 2
>>> rightFace (dart 2 "-1") smallG
FaceId 0

nextEdge :: Dart s -> PlaneGraph s v e f r -> Dart s

Get the next edge along the face

running time: \(O(1)\).

>>> nextEdge (dart 1 "+1") smallG
Dart (Arc 4) +1

prevEdge :: Dart s -> PlaneGraph s v e f r -> Dart s

Get the previous edge along the face

running time: \(O(1)\).

>>> prevEdge (dart 1 "+1") smallG
Dart (Arc 0) -1

boundary :: FaceId' s -> PlaneGraph s v e f r -> Vector (Dart s)

The darts bounding this face, for internal faces in clockwise order, for the outer face in counter clockwise order.

running time: \(O(k)\), where \(k\) is the output size.

boundary' :: Dart s -> PlaneGraph s v e f r -> Vector (Dart s)

Generates the darts incident to a face, starting with the given dart.

\(O(k)\), where \(k\) is the number of darts reported

boundaryDart :: FaceId' s -> PlaneGraph s v e f r -> Dart s

Gets a dart bounding this face. I.e. a dart d such that the face lies to the right of the dart.

boundaryVertices :: FaceId' s -> PlaneGraph s v e f r -> Vector (VertexId' s)

The vertices bounding this face, for internal faces in clockwise order, for the outer face in counter clockwise order.

running time: \(O(k)\), where \(k\) is the output size.

outerFaceId :: (Ord r, Fractional r) => PlaneGraph s v e f r -> FaceId' s

gets the id of the outer face

running time: \(O(n)\)

outerFaceDart :: (Ord r, Fractional r) => PlaneGraph s v e f r -> Dart s

gets a dart incident to the outer face (in particular, that has the outerface on its left)

running time: \(O(n)\)

vertexDataOf :: VertexId' s -> Lens' (PlaneGraph s v e f r) (VertexData r v)

locationOf :: VertexId' s -> Lens' (PlaneGraph s v e f r) (Point 2 r)

class HasDataOf g i where

General interface to accessing vertex data, dart data, and face data.

Associated Types

type DataOf g i

Methods

dataOf :: i -> Lens' g (DataOf g i)

get the data associated with the value i.

running time: \(O(1)\) to read the data, \(O(n)\) to write it.

endPointsOf :: Dart s -> Getter (PlaneGraph s v e f r) (VertexData r v, VertexData r v)

Getter for the data at the endpoints of a dart

running time: \(O(1)\)

endPointData :: Dart s -> PlaneGraph s v e f r -> (VertexData r v, VertexData r v)

Data corresponding to the endpoints of the dart

running time: \(O(1)\)

vertexData :: Lens (PlaneGraph s v e f r) (PlaneGraph s v' e f r) (Vector v) (Vector v')

faceData :: Lens (PlaneGraph s v e f r) (PlaneGraph s v e f' r) (Vector f) (Vector f')

Lens to access face data

dartData :: Lens (PlaneGraph s v e f r) (PlaneGraph s v e' f r) (Vector (Dart s, e)) (Vector (Dart s, e'))

lens to access the Dart Data

rawDartData :: Lens (PlaneGraph s v e f r) (PlaneGraph s v e' f r) (Vector e) (Vector e')

Lens to access the raw dart data, use at your own risk

edgeSegment :: Dart s -> PlaneGraph s v e f r -> LineSegment 2 v r :+ e

Given a dart and the graph constructs the line segment representing the dart. The segment \(\overline{uv})\) is has \(u\) as its tail and \(v\) as its head.

\(O(1)\)

edgeSegments :: PlaneGraph s v e f r -> Vector (Dart s, LineSegment 2 v r :+ e)

Reports all edges as line segments

>>> mapM_ print $ edgeSegments smallG
(Dart (Arc 0) +1,LineSegment (Closed (Point2 0 0 :+ 0)) (Closed (Point2 2 0 :+ 2)) :+ "0->2")
(Dart (Arc 1) +1,LineSegment (Closed (Point2 0 0 :+ 0)) (Closed (Point2 2 2 :+ 1)) :+ "0->1")
(Dart (Arc 2) +1,LineSegment (Closed (Point2 0 0 :+ 0)) (Closed (Point2 (-1) 4 :+ 3)) :+ "0->3")
(Dart (Arc 4) +1,LineSegment (Closed (Point2 2 2 :+ 1)) (Closed (Point2 2 0 :+ 2)) :+ "1->2")
(Dart (Arc 3) +1,LineSegment (Closed (Point2 2 2 :+ 1)) (Closed (Point2 (-1) 4 :+ 3)) :+ "1->3")

rawFacePolygon :: FaceId' s -> PlaneGraph s v e f r -> SimplePolygon v r :+ f

Alias for rawFace Boundary

runningtime: \(O(k)\), where \(k\) is the size of the face.

rawFaceBoundary :: FaceId' s -> PlaneGraph s v e f r -> SimplePolygon v r :+ f

The polygon describing the face

runningtime: \(O(k)\), where \(k\) is the size of the face.

rawFacePolygons :: PlaneGraph s v e f r -> Vector (FaceId' s, SimplePolygon v r :+ f)

Lists all faces of the plane graph.

newtype VertexId (s :: k) (w :: World)

A vertex in a planar graph. A vertex is tied to a particular planar graph by the phantom type s, and to a particular world w.

Constructors

VertexId

Fields

_unVertexId :: Int

newtype FaceId (s :: k) (w :: World)

The type to represent FaceId's

Constructors

FaceId

Fields

_unFaceId :: VertexId s (DualOf w)

data Dart (s :: k)

A dart represents a bi-directed edge. I.e. a dart has a direction, however the dart of the oposite direction is always present in the planar graph as well.

data World

The world in which the graph lives

Constructors

Primal

Dual

type VertexId' (s :: k) = VertexId s 'Primal

Shorthand for vertices in the primal.

type FaceId' (s :: k) = FaceId s 'Primal

Shorthand for FaceId's in the primal.

withEdgeDistances :: (Point 2 r -> Point 2 r -> a) -> PlaneGraph s p e f r -> PlaneGraph s p (a :+ e) f r

Labels the edges of a plane graph with their distances, as specified by the distance function.

writePlaneGraph :: (ToJSON v, ToJSON e, ToJSON f, ToJSON r) => PlaneGraph s v e f r -> ByteString

Writes a plane graph to a bytestring

readPlaneGraph :: (FromJSON v, FromJSON e, FromJSON f, FromJSON r) => proxy s -> ByteString -> Either ParseException (PlaneGraph s v e f r)

Reads a plane graph from a bytestring