Geometry.Ellipse

Description

Data type for representing an Ellipse.

Documentation

newtype Ellipse r

A type representing planar ellipses

Constructors

Ellipse (Transformation 2 r)

affineTransformation :: forall r r. Iso (Ellipse r) (Ellipse r) (Transformation 2 r) (Transformation 2 r)

ellipseMatrix :: Iso (Ellipse r) (Ellipse s) (Matrix 3 3 r) (Matrix 3 3 s)

Get the matrix describing the ellipse.

unitEllipse :: Num r => Ellipse r

Ellipse representing the unit circle

circleToEllipse :: Floating r => Circle p r -> Ellipse r

Convert a circle to an ellipse

ellipseToCircle :: (Num r, Eq r) => Ellipse r -> Maybe (Circle () r)

Try to Convert a circle into an ellipse. Returns a Nothing if the ellipse is not an actual circle.

_EllipseCircle :: (Floating r, Eq r) => Prism' (Ellipse r) (Circle () r)

Converting between ellipses and circles

Prims to convert between an Ellipse and a Circle.