Geometry.HalfSpace

Description

\(d\)-dimensional HalfSpaces

Documentation

>>> :{
let myVector :: Vector 3 Int
    myVector = Vector3 1 2 3
    myPoint = Point myVector
:}

newtype HalfSpace d r

A Halfspace in \(d\) dimensions. Note that the intended side of the halfspace is already indicated by the normal vector of the bounding plane.

Constructors

HalfSpace

Fields

_boundingPlane :: HyperPlane d r

boundingPlane :: forall d r d r. Iso (HalfSpace d r) (HalfSpace d r) (HyperPlane d r) (HyperPlane d r)

type HalfPlane = HalfSpace 2

leftOf :: Num r => Line 2 r -> HalfPlane r

Get the halfplane left of a line (i.e. "above") a line

>>> leftOf $ horizontalLine 4
HalfSpace {_boundingPlane = HyperPlane {_inPlane = Point2 0 4, _normalVec = Vector2 0 1}}

rightOf :: Num r => Line 2 r -> HalfPlane r

Get the halfplane right of a line (i.e. "below") a line

>>> rightOf $ horizontalLine 4
HalfSpace {_boundingPlane = HyperPlane {_inPlane = Point2 0 4, _normalVec = Vector2 0 (-1)}}

above :: Num r => Line 2 r -> HalfPlane r

below :: Num r => Line 2 r -> HalfPlane r

inHalfSpace :: (Num r, Ord r, Arity d) => Point d r -> HalfSpace d r -> PointLocationResult

Test if a point lies in a halfspace