Geometry.SoS.Orientation

Description

Orientation/side tests tests using Simulation of Simplicity.

Documentation

type SoS d = (Arity d, HasDeterminant (d + 1))

A dimension d has support for SoS when we can: compute a dterminant of a d+1 by d+1 dimensional matrix.

sideTest :: (SoS d, Num r, Ord r, ToAPoint point d r, HasIndex point) => point -> Vector d point -> Sign

Given a query point q, and a vector of d points defining a hyperplane test if q lies above or below the hyperplane. Each point is assumed to have an unique index of type i that can be used to disambiguate it in case of degeneracies.

some 1D examples:

>>> sideTest (Point1 0 `with` 0) (Vector1 $ Point1 2 `with` 1)
Negative
>>> sideTest (Point1 10 `with` 0) (Vector1 $ Point1 2 `with` 1)
Positive
>>> sideTest (Point1 2 `with` 0) (Vector1 $ Point1 2 `with` 1)
Positive
>>> sideTest (Point1 2 `with` 3) (Vector1 $ Point1 2 `with` 1)
Negative

some 2D examples:

>>> sideTest (Point2 1 2 `with` 0) $ Vector2 (Point2 0 0 `with` 1) (Point2 2 2 `with` 3)
Positive
>>> sideTest (Point2 1 (-2) `with` 0) $ Vector2 (Point2 0 0 `with` 1) (Point2 2 2 `with` 3)
Negative
>>> sideTest (Point2 1 1 `with` 0) $ Vector2 (Point2 0 0 `with` 1) (Point2 2 2 `with` 3)
Negative
>>> sideTest (Point2 1 1 `with` 10) $ Vector2 (Point2 0 0 `with` 1) (Point2 2 2 `with` 3)
Positive
>>> sideTest (Point2 1 1 `with` 10) $ Vector2 (Point2 0 0 `with` 3) (Point2 2 2 `with` 1)
Positive

sideTest' :: (Num r, Ord r, Ord i, SoS d, Arity (d + 1)) => Point d (Symbolic i r) -> Vector d (Point d (Symbolic i r)) -> Sign

Given a point q and a vector of d points defining a hyperplane, test on which side of the hyperplane q lies.

TODO: Specify what the sign means