HGeometry

Description

Main starting point for HGeometry

Primitive Geometric Data Types

diskFromPoints :: (Point_ point 2 r, Ord r, Num r) => point -> point -> point -> Maybe (BallByPoints point)

_BallSphere :: forall point point' p f. (Profunctor p, Functor f) => p (Sphere point) (f (Sphere point')) -> p (Ball point) (f (Ball point'))

_DiskCircle :: forall point point' p f. (Profunctor p, Functor f) => p (Circle point) (f (Circle point')) -> p (Disk point) (f (Disk point'))

fromBoundaryPoints :: (Point_ point 2 r, Fractional r, Ord r) => Vector 3 point -> Maybe (Disk (Point 2 r))

fromCenterAndPoint :: forall r point (d :: Nat) center. (Num r, Point_ point d r, Point_ center d r, Has_ Metric_ d r) => center -> point -> Ball center

fromDiametralPair :: forall r point (d :: Nat). (Fractional r, Point_ point d r, Has_ Metric_ d r) => point -> point -> Ball (Point d r)

squaredRadius' :: forall point f. Functor f => (NumType point -> f (NumType point)) -> Ball point -> f (Ball point)

pattern Ball_ :: ConstructableBall_ ball point => point -> NumType ball -> ball

pattern Disk_ :: (Disk_ ball point, ConstructableBall_ ball point) => point -> NumType ball -> ball

inDisk :: (Point_ point 2 r, HasInBall disk, Ord r, Num r, Dimension disk ~ 2, NumType disk ~ r) => point -> disk -> PointLocationResult

radius :: (Radical r, Ball_ ball point, NumType ball ~ r) => ball -> r

boundingHyperPlaneLens :: forall boundingHyperPlane boundingHyperPlane' f. Functor f => (boundingHyperPlane -> f boundingHyperPlane') -> HalfSpaceF boundingHyperPlane -> f (HalfSpaceF boundingHyperPlane')

pattern Plane_ :: Plane_ plane r => r -> r -> r -> plane

isParallelTo :: forall hyperPlane (d :: Natural) r hyperPlane'. (HyperPlane_ hyperPlane d r, HyperPlane_ hyperPlane' d r, Has_ Metric_ d r, Num r, Eq r, 1 <= d) => hyperPlane -> hyperPlane' -> Bool

pointOn :: forall hyperPlane (d :: Natural) r. (HyperPlane_ hyperPlane d r, Eq r, Fractional r, Has_ Additive_ d r, FoldableWithIndex Int (Vector d)) => hyperPlane -> Point d r

showPlaneEquation :: (Plane_ plane r, Show r) => plane -> String

cmpInDirection :: forall point (d :: Natural) r. (Ord r, Num r, Has_ Metric_ (d + 1) r, Has_ Metric_ d r, Point_ point d r, d < (d + 1)) => Vector d r -> point -> point -> Ordering

asNonVerticalHyperPlane :: forall hyperPlane (d :: Natural) r. (HyperPlane_ hyperPlane d r, Fractional r, Eq r, 1 <= d) => hyperPlane -> Maybe (NonVerticalHyperPlane d r)

fromLineEQ :: Num r => LineEQ r -> LinePV 2 r

horizontalLine :: forall r line. (Line_ line 2 r, Num r) => r -> line

lineThrough :: forall line point (d :: Nat) r. (Line_ line d r, Point_ point d r, Num r) => point -> point -> line

verticalLine :: forall r line. (Line_ line 2 r, Num r) => r -> line

evalAt' :: Num r => r -> LineEQ r -> r

pattern NVLine2_ :: NonVerticalHyperPlane_ line 2 r => r -> r -> line

intercept :: NonVerticalHyperPlane_ line 2 r => Lens' line r

slope :: NonVerticalHyperPlane_ line 2 r => Lens' line r

anchorPoint :: forall (d :: Nat) r f. Functor f => (Point d r -> f (Point d r)) -> LinePV d r -> f (LinePV d r)

bisector :: (Fractional r, Point_ point 2 r) => point -> point -> LinePV 2 r

cmpSlope :: (Num r, Ord r) => LinePV 2 r -> LinePV 2 r -> Ordering

fromLinearFunction :: Num r => r -> r -> LinePV 2 r

isIdenticalTo :: forall (d :: Nat) r. Eq (Vector d r) => LinePV d r -> LinePV d r -> Bool

isPerpendicularTo :: (Num r, Eq r) => Vector 2 r -> LinePV 2 r -> Bool

leftHalfPlane :: (Num r, Ord r) => LinePV 2 r -> HalfSpaceF (LinePV 2 r)

liesAbove :: (Ord r, Num r, Point_ point 2 r) => point -> LinePV 2 r -> Bool

liesBelow :: (Ord r, Num r, Point_ point 2 r) => point -> LinePV 2 r -> Bool

onSide :: (Ord r, Num r, Point_ point 2 r) => point -> LinePV 2 r -> SideTest

perpendicularTo :: Num r => LinePV 2 r -> LinePV 2 r

rightHalfPlane :: (Num r, Ord r) => LinePV 2 r -> HalfSpaceF (LinePV 2 r)

toLinearFunction :: (Fractional r, Ord r) => LinePV 2 r -> Maybe (LineEQ r)

pattern LineSegment_ :: LineSegment_ lineSegment point => point -> point -> lineSegment

interpolate :: forall lineSegment (d :: Nat) point r. (Fractional r, LineSegment_ lineSegment point, ConstructablePoint_ point d r) => r -> lineSegment -> point

ordAtX :: (Num r, Ord r, LineSegment_ lineSegment point, Point_ point 2 r) => r -> lineSegment -> lineSegment -> Ordering

ordAtY :: (Num r, Ord r, LineSegment_ lineSegment point, Point_ point 2 r) => r -> lineSegment -> lineSegment -> Ordering

orientBT :: forall lineSegment point (d :: Nat) r. (LineSegment_ lineSegment point, Point_ point d r, 2 <= d, Ord r) => lineSegment -> lineSegment

orientLR :: forall lineSegment point (d :: Nat) r. (LineSegment_ lineSegment point, Point_ point d r, 1 <= d, Ord r) => lineSegment -> lineSegment

xCoordAt :: (Fractional r, Ord r, LineSegment_ lineSegment point, Point_ point 2 r) => r -> lineSegment -> r

yCoordAt :: (Fractional r, Ord r, LineSegment_ lineSegment point, Point_ point 2 r) => r -> lineSegment -> r

asALineSegment :: LineSegment_ segment point => segment -> LineSegment AnEndPoint point

spanIn :: forall point (d :: Nat) r (endPoint :: Type -> Type). (Point_ point d r, Ord r, IxValue (endPoint point) ~ point, EndPoint_ (endPoint point)) => Getter point r -> LineSegment endPoint point -> Interval AnEndPoint r

asPoint :: forall point (d :: Nat) r. Point_ point d r => Lens' point (Point d r)

projectPoint :: forall (i :: Nat) point (d :: Nat) r. (Point_ point d r, i <= d, Has_ Vector_ i r) => point -> Point i r

pattern Point1_ :: Point_ point 1 r => r -> point

pattern Point2_ :: Point_ point 2 r => r -> r -> point

pattern Point3_ :: Point_ point 3 r => r -> r -> r -> point

pattern Point4_ :: Point_ point 4 r => r -> r -> r -> r -> point

coord :: forall (i :: Natural) point (d :: Natural) r. (1 <= i, i <= d, KnownNat i, Point_ point d r) => IndexedLens' Int point r

dCoord :: forall point (d :: Natural) r. (1 <= d, Point_ point d r) => IndexedLens' Int point r

origin :: forall point (d :: Nat) r. (Num r, ConstructablePoint_ point d r) => point

pointFromList :: forall point (d :: Nat) r. (ConstructablePoint_ point d r, Vector_ (Vector d r) d r) => [r] -> Maybe point

wCoord :: forall (d :: Natural) point r. (4 <= d, Point_ point d r) => IndexedLens' Int point r

xCoord :: forall (d :: Natural) point r. (1 <= d, Point_ point d r) => IndexedLens' Int point r

yCoord :: forall (d :: Natural) point r. (2 <= d, Point_ point d r) => IndexedLens' Int point r

zCoord :: forall (d :: Natural) point r. (3 <= d, Point_ point d r) => IndexedLens' Int point r

cmpByDistanceTo :: forall r point (d :: Nat) center. (Ord r, Num r, Point_ point d r, Point_ center d r, Metric_ (Vector d r) d r) => center -> point -> point -> Ordering

euclideanDist :: forall r point (d :: Nat). (Radical r, Point_ point d r, Metric_ (Vector d r) d r) => point -> point -> r

squaredEuclideanDist :: forall r point (d :: Nat) point'. (Num r, Point_ point d r, Point_ point' d r, Metric_ (Vector d r) d r) => point -> point' -> r

cmpInDirection2 :: (Num r, Ord r, Point_ point 2 r) => Vector 2 r -> point -> point -> Ordering

ccw :: (Point_ point 2 r, Point_ point' 2 r, Point_ point'' 2 r, Num r, Ord r) => point -> point' -> point'' -> CCW

ccwCmpAround :: (Point_ center 2 r, Point_ point 2 r, Ord r, Num r) => center -> point -> point -> Ordering

ccwCmpAroundWith :: (Point_ center 2 r, Point_ point 2 r, Ord r, Num r) => Vector 2 r -> center -> point -> point -> Ordering

cwCmpAround :: (Point_ center 2 r, Point_ point 2 r, Ord r, Num r) => center -> point -> point -> Ordering

cwCmpAroundWith :: (Point_ center 2 r, Point_ point 2 r, Ord r, Num r) => Vector 2 r -> center -> point -> point -> Ordering

insertIntoCyclicOrder :: (Point_ center 2 r, Point_ point 2 r, Ord r, Num r) => center -> point -> CList point -> CList point

isCoLinear :: (Point_ point 2 r, Num r, Eq r) => point -> point -> point -> Bool

sortAround :: (Point_ center 2 r, Point_ point 2 r, Num r, Ord r) => center -> [point] -> [point]

partitionIntoQuadrants :: (Ord r, Point_ point 2 r) => point -> [point] -> ([point], [point], [point], [point])

quadrant :: (Ord r, Num r, ConstructablePoint_ point 2 r) => point -> Quadrant

quadrantWith :: (Ord r, Point_ point 2 r) => point -> point -> Quadrant

isScalarMultipleOf :: forall r vector (d :: Nat). (Eq r, Num r, Metric_ vector d r) => vector -> vector -> Bool

sameDirection :: forall vector (d :: Nat) r. (Additive_ vector d r, Num r, Eq r) => vector -> vector -> Bool

scalarMultiple :: forall r vector (d :: Nat). (Eq r, Fractional r, Additive_ vector d r) => vector -> vector -> Maybe r

(*^) :: forall r vector (d :: Nat). (Num r, Vector_ vector d r) => r -> vector -> vector

(^*) :: forall r vector (d :: Nat). (Num r, Vector_ vector d r) => vector -> r -> vector

(^+^) :: forall r vector (d :: Nat). (Num r, Additive_ vector d r) => vector -> vector -> vector

(^-^) :: forall r vector (d :: Nat). (Num r, Additive_ vector d r) => vector -> vector -> vector

(^/) :: forall vector (d :: Nat) r. (Vector_ vector d r, Fractional r) => vector -> r -> vector

basis :: forall vector (d :: Nat) r. (Additive_ vector d r, Num r) => [vector]

component :: forall (i :: Natural) vector r (d :: Natural). (i <= (d - 1), KnownNat i, Vector_ vector d r) => IndexedLens' Int vector r

cons :: forall vector' vector (d :: Nat) r. (Vector_ vector d r, Vector_ vector' (d + 1) r) => r -> vector -> vector'

foldMapZip :: forall vector (d :: Nat) r m. (Additive_ vector d r, Semigroup m) => (r -> r -> m) -> vector -> vector -> m

generate :: forall vector (d :: Nat) r. Vector_ vector d r => (Int -> r) -> vector

lerp :: forall r vector (d :: Nat). (Num r, Additive_ vector d r) => r -> vector -> vector -> vector

liftI2 :: forall vector (d :: Nat) r. Additive_ vector d r => (r -> r -> r) -> vector -> vector -> vector

negated :: forall r vector (d :: Nat). (Num r, Vector_ vector d r) => vector -> vector

prefix :: forall (i :: Nat) (d :: Nat) vector vector' r. (i <= d, Vector_ vector d r, Vector_ vector' i r) => vector -> vector'

snoc :: forall vector' vector (d :: Nat) r. (Vector_ vector d r, Vector_ vector' (d + 1) r) => vector -> r -> vector'

suffix :: forall (i :: Nat) (d :: Nat) vector vector' r. (i <= d, Vector_ vector d r, Vector_ vector' i r) => vector -> vector'

sumV :: forall f vector (d :: Nat) r. (Foldable f, Additive_ vector d r, Num r) => f vector -> vector

uncons :: forall vector' vector (d :: Natural) r. (Vector_ vector (d + 1) r, Vector_ vector' d r, 0 <= ((d + 1) - 1), d <= Dimension vector) => vector -> (r, vector')

unit :: forall vector (d :: Nat) r. (Additive_ vector d r, Num r) => vector

unsnoc :: forall vector' vector (d :: Natural) r. (Vector_ vector (d + 1) r, Vector_ vector' d r, d <= ((d + 1) - 1), d <= Dimension vector) => vector -> (vector', r)

vectorFromList :: forall vector (d :: Nat) r. Vector_ vector d r => [r] -> Maybe vector

wComponent :: forall vector (d :: Nat) r. (Vector_ vector d r, 4 <= d) => IndexedLens' Int vector r

xComponent :: forall vector (d :: Nat) r. (Vector_ vector d r, 1 <= d) => IndexedLens' Int vector r

yComponent :: forall vector (d :: Nat) r. (Vector_ vector d r, 2 <= d) => IndexedLens' Int vector r

zComponent :: forall vector (d :: Nat) r. (Vector_ vector d r, 3 <= d) => IndexedLens' Int vector r

zero :: forall r vector (d :: Nat). (Num r, Additive_ vector d r) => vector

cross :: Num r => Vector 3 r -> Vector 3 r -> Vector 3 r

type BallByPoints point = BallByPoints' (1 + Dimension point) point

newtype BallByPoints' (k :: Nat) point

Constructors

BoundaryPoints (Vector k point)

data Ball point

Constructors

Ball !point !(NumType point)

Bundled Patterns

pattern Disk :: point -> NumType point -> Disk point

type Circle = Sphere

type Disk = Ball

newtype Sphere point

Constructors

MkSphere (Ball point)

Bundled Patterns

pattern Circle :: point -> NumType point -> Circle point

pattern Sphere :: point -> NumType point -> Sphere point

class (HasCenter ball point, NumType ball ~ NumType point, Point_ point (Dimension ball) (NumType ball)) => Ball_ ball point | ball -> point where

Methods

squaredRadius :: Getter ball (NumType ball)

class Ball_ ball point => ConstructableBall_ ball point where

Minimal complete definition

fromCenterAndSquaredRadius

Methods

fromCenterAndSquaredRadius :: point -> NumType ball -> ball

unitBall :: forall r (d :: Nat). (r ~ NumType ball, d ~ Dimension ball, Num r, ConstructablePoint_ point d r) => ball

class (Ball_ disk point, Dimension disk ~ 2) => Disk_ disk point

class HasCenter geom point | geom -> point where

Methods

center :: Lens' geom point

class HasInBall ball where

Methods

inBall :: forall point (d :: Nat) r. (Point_ point d r, Ord r, Num r, NumType ball ~ r, Dimension ball ~ d) => point -> ball -> PointLocationResult

newtype DiametralBall point

Constructors

MkDiametralBall (Vector 2 point)

Bundled Patterns

pattern DiametralPoints :: point -> point -> DiametralBall point

class IsBoxable g where

Minimal complete definition

Nothing

Methods

boundingBox :: forall (d :: Nat) r. (d ~ Dimension g, r ~ NumType g, Ord r) => g -> Box (Point d r)

class (HasMinPoint box point, HasMaxPoint box point, Point_ point (Dimension box) (NumType box)) => Box_ box point | box -> point where

Methods

extent :: forall r (d :: Nat). (r ~ NumType box, d ~ Dimension box, Num r) => box -> Vector d (ClosedInterval r)

data Box point where

Bundled Patterns

pattern Box :: point -> point -> Box point

pattern Rect :: Num r => r -> r -> r -> r -> Rectangle (Point 2 r)

pattern Rectangle :: Dimension point ~ 2 => point -> point -> Box point

type family BoundingHyperPlane halfSpace (d :: Nat) r

type HalfPlane_ halfPlane r = HalfSpace_ halfPlane 2 r

class (Dimension halfSpace ~ d, Dimension (BoundingHyperPlane halfSpace d r) ~ d, NumType halfSpace ~ r, NumType (BoundingHyperPlane halfSpace d r) ~ r) => HalfSpace_ halfSpace (d :: Nat) r | halfSpace -> d, halfSpace -> r where

Associated Types

type BoundingHyperPlane halfSpace (d :: Nat) r

Methods

boundingHyperPlane :: Lens' halfSpace (BoundingHyperPlane halfSpace d r)

halfSpaceSign :: Lens' halfSpace Sign

type HalfPlaneF line = HalfSpaceF line

type HalfSpace (d :: Natural) r = HalfSpaceF (HyperPlane d r)

data HalfSpaceF boundingHyperPlane

Constructors

HalfSpace !Sign boundingHyperPlane

data Point_x_HalfSpace_Intersection boundaryPoint interiorPoint

Constructors

Point_x_HalfSpace_OnBoundary boundaryPoint

Point_x_HalfSpace_Interior interiorPoint

class HyperPlane_ hyperPlane d r => ConstructableHyperPlane_ hyperPlane (d :: Natural) r where

Minimal complete definition

hyperPlaneFromEquation

Associated Types

type HyperPlaneFromEquationConstraint hyperPlane (d :: Natural) r

type HyperPlaneFromEquationConstraint hyperPlane (d :: Natural) r = ()

Methods

hyperPlaneFromEquation :: Vector (d + 1) r -> hyperPlane

fromPointAndNormal :: (Point_ point d r, Num r) => point -> Vector d r -> hyperPlane

type family HyperPlaneFromEquationConstraint hyperPlane (d :: Natural) r

class HyperPlaneFromPoints hyperPlane where

Methods

hyperPlaneThrough :: forall point (d :: Nat) r. (Point_ point d r, HyperPlane_ hyperPlane d r, Num r) => Vector d point -> hyperPlane

class (NumType hyperPlane ~ r, Dimension hyperPlane ~ d, Has_ Vector_ d r, Has_ Vector_ (1 + d) r) => HyperPlane_ hyperPlane (d :: Natural) r | hyperPlane -> d, hyperPlane -> r where

Minimal complete definition

Nothing

Methods

evalHyperPlaneEquation :: (Num r, Point_ point d r) => hyperPlane -> point -> r

hyperPlaneEquation :: hyperPlane -> Vector (d + 1) r

normalVector :: hyperPlane -> Vector d r

onHyperPlane :: (Point_ point d r, Eq r, Num r) => point -> hyperPlane -> Bool

onSideTest :: (Point_ point d r, Ord r, Num r) => point -> hyperPlane -> Ordering

class HyperPlane_ hyperPlane d r => NonVerticalHyperPlane_ hyperPlane (d :: Natural) r where

Minimal complete definition

hyperPlaneCoefficients

Methods

evalAt :: (Num r, 1 <= d, Point_ point (d - 1) r) => point -> hyperPlane -> r

hyperPlaneCoefficients :: Lens' hyperPlane (Vector d r)

verticalSideTest :: (1 <= d, Point_ point d r, Ord r, Num r) => point -> hyperPlane -> Ordering

type Plane_ plane = NonVerticalHyperPlane_ plane 3

newtype HyperPlane (d :: Natural) r

Constructors

HyperPlane (Vector (d + 1) r)

Bundled Patterns

pattern HyperPlane2 :: r -> r -> r -> HyperPlane 2 r

pattern HyperPlane3 :: r -> r -> r -> r -> HyperPlane 3 r

type MkHyperPlaneConstraints (d :: Natural) r = (d < (d + 1), KnownNat d, Has_ Metric_ d r, Has_ Metric_ (d + 1) r, Has_ Vector_ d r, Has_ Vector_ (d + 1) r)

data PlanePlaneIntersection plane line

Constructors

Plane_x_Plane_Line line

Plane_x_Plane_Plane plane

newtype NonVerticalHyperPlane (d :: Nat) r

Constructors

NonVerticalHyperPlane (Vector d r)

Bundled Patterns

pattern Plane :: r -> r -> r -> Plane r

type Plane = NonVerticalHyperPlane 3

data family IntersectionOf g h

type family EndPointOf interval

class HasEnd seg p | seg -> p where

Methods

end :: Lens' seg p

class HasEndPoint seg p | seg -> p where

Methods

endPoint :: Lens' seg p

class HasStart seg p | seg -> p where

Methods

start :: Lens' seg p

class HasStartPoint seg p | seg -> p where

Methods

startPoint :: Lens' seg p

type family StartPointOf interval

data AnEndPoint r

Constructors

AnEndPoint !EndPointType !r

Bundled Patterns

pattern AnClosedE :: r -> AnEndPoint r

pattern AnOpenE :: r -> AnEndPoint r

newtype EndPoint (et :: EndPointType) r

Constructors

EndPoint r

Bundled Patterns

pattern ClosedE :: r -> EndPoint 'Closed r

pattern OpenE :: r -> EndPoint 'Open r

class IsEndPoint endPoint endPoint => EndPoint_ endPoint where

Methods

endPointType :: endPoint -> EndPointType

mkEndPoint :: IxValue endPoint -> endPoint

class HasOnLine line (d :: Nat) | line -> d where

Methods

onLine :: (Point_ point d r, Num r, Eq r, r ~ NumType line, d ~ Dimension line) => point -> line -> Bool

type Line2_ line r = (Line_ line 2 r, HyperPlane_ line 2 r)

class Line_ line (d :: Nat) r | line -> d, line -> r where

Methods

fromPointAndVec :: (Point_ point d r, Line_ line d r, Num r) => point -> Vector d r -> line

type LineLineIntersection line = LineLineIntersectionG (NumType line) line

data LineLineIntersectionG r line

Constructors

Line_x_Line_Point (Point 2 r)

Line_x_Line_Line line

newtype LineEQ r

Constructors

MkLineEQ (NonVerticalHyperPlane 2 r)

Bundled Patterns

pattern LineEQ :: r -> r -> LineEQ r

type NVLine2_ line r = (Line_ line 2 r, NonVerticalHyperPlane_ line 2 r)

class HasDirection t where

Methods

direction :: forall (d :: Nat) r. (Dimension t ~ d, NumType t ~ r) => Lens' t (Vector d r)

class HasSupportingLine t where

Methods

supportingLine :: t -> LinePV (Dimension t) (NumType t)

data LinePV (d :: Nat) r

Constructors

LinePV

Fields

_anchorPoint :: !(Point d r)

_direction :: !(Vector d r)

class OnSideUpDownTest t where

Methods

onSideUpDown :: forall (d :: Nat) r point. (d ~ Dimension t, r ~ NumType t, Ord r, Num r, Point_ point d r) => point -> t -> SideTestUpDown

data SideTest

Constructors

LeftSide

OnLine

RightSide

data SideTestUpDown

Constructors

Below

On

Above

class (LineSegment_ lineSegment point, StartPointOf lineSegment ~ EndPoint 'Closed point, EndPointOf lineSegment ~ EndPoint 'Closed point) => ClosedLineSegment_ lineSegment point

class LineSegment_ lineSegment point => ConstructableLineSegment_ lineSegment point where

Minimal complete definition

uncheckedLineSegment

Methods

uncheckedLineSegment :: point -> point -> lineSegment

mkLineSegment :: point -> point -> Maybe lineSegment

class HasOnSegment lineSegment (d :: Nat) | lineSegment -> d where

Methods

onSegment :: (Ord r, Point_ point d r, r ~ NumType lineSegment, d ~ Dimension lineSegment) => point -> lineSegment -> Bool

class (IntervalLike_ lineSegment point, Point_ point (Dimension lineSegment) (NumType lineSegment)) => LineSegment_ lineSegment point | lineSegment -> point

class (LineSegment_ lineSegment point, StartPointOf lineSegment ~ EndPoint 'Open point, EndPointOf lineSegment ~ EndPoint 'Open point) => OpenLineSegment_ lineSegment point

type ClosedLineSegment point = LineSegment (EndPoint 'Closed) point

data LineSegment (endPoint :: Type -> Type) point where

Bundled Patterns

pattern ClosedLineSegment :: point -> point -> ClosedLineSegment point

pattern LineSegment :: endPoint point -> endPoint point -> LineSegment endPoint point

pattern OpenLineSegment :: point -> point -> OpenLineSegment point

type OpenLineSegment point = LineSegment (EndPoint 'Open) point

data ClosedSegmentHalfSpaceIntersection extra point

Constructors

ClosedLineSegment_x_HalfSpace_Point point

ClosedLineSegment_x_HalfSpace_SubSegment (ClosedLineSegment (OriginalOrExtra point extra))

ClosedLineSegment_x_HalfSpace_CompleteSegment (ClosedLineSegment point)

data HalfLineLineSegmentIntersection point segment

Constructors

HalfLine_x_LineSegment_Point point

HalfLine_x_LineSegment_LineSegment segment

data LineLineSegmentIntersection lineSegment

Constructors

Line_x_LineSegment_Point (Point 2 (NumType lineSegment))

Line_x_LineSegment_LineSegment lineSegment

data LineSegmentLineSegmentIntersection lineSegment

Constructors

LineSegment_x_LineSegment_Point (Point 2 (NumType lineSegment))

LineSegment_x_LineSegment_LineSegment lineSegment

data Triangle point where

Bundled Patterns

pattern Triangle :: point -> point -> point -> Triangle point

type CanonicalPoint geom = Point (Dimension geom) (NumType geom)

type Point (d :: Nat) r = PointF (Vector d r)

class (Additive_ (Vector d r) d r, HasCoordinates point point, d ~ Dimension point, r ~ NumType point) => Affine_ point (d :: Nat) r | point -> d, point -> r where

Minimal complete definition

Nothing

Methods

(.-.) :: point -> point -> Vector d r

(.+^) :: point -> Vector d r -> point

(.-^) :: point -> Vector d r -> point

class Point_ point d r => ConstructablePoint_ point (d :: Nat) r where

Methods

fromVector :: Vector d r -> point

class (Has_ Vector_ (Dimension point) (NumType point), Has_ Vector_ (Dimension point') (NumType point'), HasComponents (Vector (Dimension point') (NumType point)) (Vector (Dimension point') (NumType point')), Dimension point ~ Dimension point', HasVector point point') => HasCoordinates point point' where

Minimal complete definition

Nothing

Methods

coordinates :: IndexedTraversal1 Int point point' (NumType point) (NumType point')

class HasPoints s t point point' | s -> point, t -> point' where

Methods

allPoints :: forall (d :: Nat) r r'. (Point_ point d r, Point_ point' d r', NumType s ~ r, NumType t ~ r', Dimension s ~ d, Dimension t ~ d) => Traversal1 s t point point'

type HasPoints' s point = HasPoints s s point point

class HasVector point point' where

Methods

vector :: forall (d :: Nat) r s. (Dimension point ~ d, NumType point ~ r, Dimension point' ~ d, NumType point' ~ s) => Lens point point' (Vector d r) (Vector d s)

class (Affine_ point d r, HasVector point point) => Point_ point (d :: Nat) r where

Minimal complete definition

Nothing

Methods

coord' :: Int -> IndexedTraversal' Int point r

class Metric_ (Vector (Dimension g) (NumType g)) (Dimension g) (NumType g) => HasSquaredEuclideanDistance g where

Minimal complete definition

pointClosestToWithDistance | pointClosestTo

Methods

squaredEuclideanDistTo :: forall r (d :: Nat) point. (r ~ NumType g, d ~ Dimension g, Num r, Point_ point d r) => point -> g -> r

pointClosestTo :: forall r (d :: Nat) point. (r ~ NumType g, d ~ Dimension g, Num r, Point_ point d r) => point -> g -> Point d r

pointClosestToWithDistance :: forall r (d :: Nat) point. (r ~ NumType g, d ~ Dimension g, Num r, Point_ point d r) => point -> g -> (Point d r, r)

data CCW where

Bundled Patterns

pattern CCW :: CCW

pattern CW :: CCW

pattern CoLinear :: CCW

newtype PointF v

Constructors

Point

Fields

toVec :: v

Bundled Patterns

pattern Point1 :: r -> Point 1 r

pattern Point2 :: r -> r -> Point 2 r

pattern Point3 :: r -> r -> r -> Point 3 r

pattern Point4 :: r -> r -> r -> r -> Point 4 r

data Quadrant

Constructors

TopRight

TopLeft

BottomLeft

BottomRight

class Vector_ vector d r => Additive_ vector (d :: Nat) r where

Methods

liftU2 :: (r -> r -> r) -> vector -> vector -> vector

liftI2A :: Apply f => (r -> r -> f r) -> vector -> vector -> f vector

class (r ~ IxValue vector, s ~ IxValue vector', d ~ Dimension vector, d ~ Dimension vector') => AsVector_ vector vector' (d :: Nat) r s | vector -> d, vector -> r, vector' -> s where

Methods

_Vector :: Iso vector vector' (Vector d r) (Vector d s)

type Has_ (c :: Type -> Nat -> Type -> k) (d :: Nat) r = c (Vector d r) d r

class Additive_ vector d r => Metric_ vector (d :: Nat) r where

Minimal complete definition

Nothing

Methods

dot :: vector -> vector -> r

quadrance :: vector -> r

qd :: vector -> vector -> r

norm :: vector -> r

signorm :: vector -> vector

class (HasComponents vector vector, AsVector_ vector vector d r r, KnownNat d) => Vector_ vector (d :: Nat) r where

Minimal complete definition

generateA

Methods

generateA :: Applicative f => (Int -> f r) -> f vector

component' :: Int -> IndexedTraversal' Int vector r

class HasComponents vector vector' where

Methods

components :: IndexedTraversal1 Int vector vector' (IxValue vector) (IxValue vector')

data family Vector (d :: Nat) r

data Sign

The sign of an expression

Constructors

Negative

Positive