HGeometry.Ball.BoundaryPoints

Description

A ball in \(d\) dimensional space represented by \(d+1\) points on the boundary.

Documentation

type BallByPoints point = BallByPoints' (1 + Dimension point) point

A ball specified by d+1 points which form a CCW simplex.

newtype BallByPoints' (k :: Nat) point

A ball specified by k points which form a CCW simplex.

This is just a helper type to implement the BallByPoints type.

Constructors

BoundaryPoints (Vector k point)

diskFromPoints :: (Point_ point 2 r, Ord r, Num r) => point -> point -> point -> Maybe (BallByPoints point)

smart constructor that makes sure the points are given in the right orientation may fail if the points are colinear