HGeometry.Ball.CenterAndRadius

Description

Balls in d-dimensional space, represented by their center and squared radius.

Documentation

data Ball point

d-dimensional ball, specified by center and squared radius.

Constructors

Ball !point !(NumType point)

Bundled Patterns

pattern Disk :: point -> NumType point -> Disk point
Construct a disk

type Disk = Ball

Balls in 2D are also known as Disks

newtype Sphere point

A sphere, i.e. the boudary of a Ball.

Constructors

MkSphere (Ball point)

Bundled Patterns

pattern Sphere :: point -> NumType point -> Sphere point
Construct a Sphere; the boundary of a ball

pattern Circle :: point -> NumType point -> Circle point
Construct a Circle

type Circle = Sphere

A circle, i.e. the boundary of a Disk

_BallSphere :: Iso (Ball point) (Ball point') (Sphere point) (Sphere point')

Iso to convert between a ball and a sphere.

_DiskCircle :: Iso (Disk point) (Disk point') (Circle point) (Circle point')

Iso to convert between a Disk and a Circle

data family IntersectionOf g h

The data family specifying to help implement the Intersection type family.