HGeometry.Ball.Class

Description

A class of types that can act as balls in \(d\)-dimensional space.

Documentation

class (HasCenter ball point, NumType ball ~ NumType point, Point_ point (Dimension ball) (NumType ball)) => Ball_ ball point | ball -> point where

Balls in d-dimensional space

Methods

squaredRadius :: Getter ball (NumType ball)

Getter to access the squared radius of a ball

class Ball_ ball point => ConstructableBall_ ball point where

Minimal complete definition

fromCenterAndSquaredRadius

Methods

fromCenterAndSquaredRadius :: point -> NumType ball -> ball

Construct a ball from the center and *squared* radius

unitBall :: forall r (d :: Nat). (r ~ NumType ball, d ~ Dimension ball, Num r, ConstructablePoint_ point d r) => ball

A d dimensional unit ball centered at the origin.

pattern Ball_ :: ConstructableBall_ ball point => point -> NumType ball -> ball

Constructs a Ball from its center and squared radius.

class HasCenter geom point | geom -> point where

Types that have a center field lens

Methods

center :: Lens' geom point

Lens to access the center point of a geometry

radius :: (Radical r, Ball_ ball point, NumType ball ~ r) => ball -> r

A lens to get/set the radius of a Ball

class (Ball_ disk point, Dimension disk ~ 2) => Disk_ disk point

Disks are two dimensional balls

pattern Disk_ :: (Disk_ ball point, ConstructableBall_ ball point) => point -> NumType ball -> ball

Constructs a Disk by its center and squared radius.

class HasInBall ball where

Class for types that have an "in ball" test.

Methods

inBall :: forall point (d :: Nat) r. (Point_ point d r, Ord r, Num r, NumType ball ~ r, Dimension ball ~ d) => point -> ball -> PointLocationResult

Test if a query point lies inside or outside of a ball.

inDisk :: (Point_ point 2 r, HasInBall disk, Ord r, Num r, Dimension disk ~ 2, NumType disk ~ r) => point -> disk -> PointLocationResult

In Disk test