HGeometry.Box

Description

d-dimensional Boxes

Documentation

module HGeometry.Box.Class

module HGeometry.Box.Internal

module HGeometry.Box.Corners

module HGeometry.Box.Sides

class IsBoxable g where

Types for which we can compute an axis parallel boundingbox

Minimal complete definition

Nothing

Methods

boundingBox :: (d ~ Dimension g, r ~ NumType g, Ord r) => g -> Box (Point d r)

Compute the axis-parallel boundingbox of the given geometry.

default boundingBox :: (d ~ Dimension g, r ~ NumType g, Ord r, HasPoints' g point, Point_ point d r, Ord (Vector d r)) => g -> Box (Point d r)

data LineBoxIntersection d r

Data type representing the intersection of a Box and a line

Constructors

Line_x_Box_Point (Point d r)

Line_x_Box_LineSegment (ClosedLineSegment (Point d r))

data HalfLineBoxIntersection point

Constructors

HalfLine_x_Box_Point point

HalfLine_x_Box_LineSegment (ClosedLineSegment point)

Orphan instances

(Point_ point d r, IsTransformable point) => IsTransformable (Box point)

Instance details

Methods

transformBy :: Transformation (Dimension (Box point)) (NumType (Box point)) -> Box point -> Box point

(Point_ point 2 r, Ord r, Num r) => HasIntersectionWith (HalfLine point) (Rectangle point)

Instance details

Methods

intersects :: HalfLine point -> Rectangle point -> Bool

(Num r, Ord r, Point_ point 2 r) => HasIntersectionWith (LineEQ r) (Rectangle point)

Instance details

Methods

intersects :: LineEQ r -> Rectangle point -> Bool

(Point_ point 2 r, Ord r, Fractional r) => IsIntersectableWith (HalfLine point) (Rectangle point)

Instance details

Methods

intersect :: HalfLine point -> Rectangle point -> Intersection (HalfLine point) (Rectangle point)

(Fractional r, Ord r, Point_ point 2 r) => IsIntersectableWith (LineEQ r) (Rectangle point)

Instance details

Methods

intersect :: LineEQ r -> Rectangle point -> Intersection (LineEQ r) (Rectangle point)

(Num r, Ord r, Point_ point 2 r) => HasIntersectionWith (LinePV 2 r) (Rectangle point)

Instance details

Methods

intersects :: LinePV 2 r -> Rectangle point -> Bool

(Ord r, Num r, Point_ point 2 r, Point_ point' 2 r, HasIntersectionWith (LineSegment endPoint point) (ClosedLineSegment point')) => HasIntersectionWith (LineSegment endPoint point) (Boundary (Rectangle point'))

Instance details

Methods

intersects :: LineSegment endPoint point -> Boundary (Rectangle point') -> Bool

(Ord r, Num r, Point_ point 2 r, Point_ point' 2 r, IxValue (endPoint point) ~ point, EndPoint_ (endPoint point), HasIntersectionWith (LineSegment endPoint point) (ClosedLineSegment point')) => HasIntersectionWith (LineSegment endPoint point) (Rectangle point')

Instance details

Methods

intersects :: LineSegment endPoint point -> Rectangle point' -> Bool

(Point_ point d r, Num r, Ord r, Foldable1 (Vector d), Zip (Vector d)) => HasIntersectionWith (Point d r) (Box point)

Instance details

Methods

intersects :: Point d r -> Box point -> Bool

(Fractional r, Ord r, Point_ point 2 r) => IsIntersectableWith (LinePV 2 r) (Rectangle point)

Instance details

Methods

intersect :: LinePV 2 r -> Rectangle point -> Intersection (LinePV 2 r) (Rectangle point)

(Point_ point d r, Num r, Ord r, Foldable1 (Vector d), Zip (Vector d)) => IsIntersectableWith (Point d r) (Box point)

Instance details

Methods

intersect :: Point d r -> Box point -> Intersection (Point d r) (Box point)