HGeometry.Box.Boxable

Description

Typeclass that expresses that we can compute an axis parallel bounding box of an object.

Documentation

class IsBoxable g where

Types for which we can compute an axis parallel boundingbox

Minimal complete definition

Nothing

Methods

boundingBox :: (d ~ Dimension g, r ~ NumType g, Ord r) => g -> Box (Point d r)

Compute the axis-parallel boundingbox of the given geometry.

default boundingBox :: (d ~ Dimension g, r ~ NumType g, Ord r, HasPoints' g point, Point_ point d r, Ord (Vector d r)) => g -> Box (Point d r)