HGeometry.Box.Class

Description

d dimensional boxes

Documentation

class (HasMinPoint box point, HasMaxPoint box point, Point_ point (Dimension box) (NumType box)) => Box_ box point | box -> point where

d-dimensional Boxes

Methods

extent :: forall r (d :: Nat). (r ~ NumType box, d ~ Dimension box, Num r) => box -> Vector d (ClosedInterval r)

Get a vector with the extent of the box in each dimension. Note that the resulting vector is 0 indexed whereas one would normally count dimensions starting at zero.

class HasMinPoint box point | box -> point where

Types that have a minPoint field lens

Methods

minPoint :: Lens' box point

Lens to access the lexicographically smallest point

class HasMaxPoint box point | box -> point where

Types that have a maxPoint field lens

Methods

maxPoint :: Lens' box point

Lens to access the lexicographically largest point

centerPoint :: forall box point (d :: Nat) r. (Box_ box point, Point_ point d r, Fractional r) => box -> point

Get the center point of a box

>>> centerPoint myRect
Point2 5.5 10.5

type Rectangle_ rectangle point = (Box_ rectangle point, Dimension rectangle ~ 2)

Rectangles are two dimensional boxes.

width :: forall box point (d :: Nat) r. (Box_ box point, Point_ point d r, 1 <= d, Functor (Vector d), Num r) => box -> r

Get the width of a rectangle.

height :: forall box point (d :: Nat) r. (Box_ box point, Point_ point d r, 2 <= d, Functor (Vector d), Num r) => box -> r

get the height of a rectangle

size :: forall box (d :: Nat) point r. (Box_ box point, Point_ point d r, Num r, Functor (Vector d)) => box -> Vector d r

Get the size of the box (in all dimensions). Note that the resulting vector is 0 indexed whereas one would normally count dimensions starting at zero.

>>> size myRect
Vector2 9.0 19.0