HGeometry.Box.Corners

Description

The corners of a box.

Documentation

data Corners a

A data type rperesenting the corners of a box. the order of the Corners is 'northWest, northEast, southEast, southWest', i.e. in clockwise order starting from the topleft.

Constructors

Corners !a !a !a !a

northWest :: Lens' (Corners a) a

Access the northwest corner

northEast :: Lens' (Corners a) a

Access the northeast corner

southEast :: Lens' (Corners a) a

Access the southeast corner

southWest :: Lens' (Corners a) a

Access the southwest corner

corners :: (Num r, Rectangle_ rectangle point, Point_ point 2 r) => rectangle -> Corners point

Get the corners of a rectangle, the order is:

Any additional information from in the top points is taken from the Top point, Any additional information from in the bottom points is taken from the Bottom point,

cornersInDirection :: CardinalDirection -> Corners p -> Vector 2 p

Gets the corners in a particular direction