HGeometry.Box.Sides

Description

Documentation

data Sides a

The four sides of a rectangle

Constructors

Sides !a !a !a !a

north :: Lens' (Sides a) a

Access the north side

east :: Lens' (Sides a) a

Access the east side

south :: Lens' (Sides a) a

Access the south side

west :: Lens' (Sides a) a

Access the west side

topSide :: (Num r, Rectangle_ rectangle point, Point_ point 2 r) => rectangle -> ClosedLineSegment point

The top side of the box, from left to right.

bottomSide :: (Num r, Rectangle_ rectangle point, Point_ point 2 r) => rectangle -> ClosedLineSegment point

Oriented from *left to right*

leftSide :: (Num r, Rectangle_ rectangle point, Point_ point 2 r) => rectangle -> ClosedLineSegment point

Left side of the box, from bottom to top

rightSide :: (Num r, Rectangle_ rectangle point, Point_ point 2 r) => rectangle -> ClosedLineSegment point

The right side, oriented from *bottom* to top

sides :: (Num r, Rectangle_ rectangle point, Point_ point 2 r) => rectangle -> Sides (ClosedLineSegment point)

The sides of the rectangle, in order (Top, Right, Bottom, Left). The sides themselves are also oriented in clockwise order. If, you want them in the same order as the functions topSide, bottomSide, leftSide, and rightSide, use sides' instead.

sides' :: (Num r, Rectangle_ rectangle point, Point_ point 2 r) => rectangle -> Sides (ClosedLineSegment point)

The sides of the rectangle. The order of the segments is (Top, Right, Bottom, Left). Note that the segments themselves, are oriented as described by the functions topSide, bottomSide, leftSide, rightSide (basically: from left to right, and from bottom to top). If you want the segments oriented along the boundary of the rectangle, use the sides function instead.

sideDirections :: Sides CardinalDirection

Constructs a Sides value that indicates the appropriate direction.