HGeometry.ByIndex

Description

Types that have an Index, which is used to distinguish the items. Hence, Eq and Ord are *only* based on the Index.

Documentation

data ByIndex ix a

Helper data type for which the Ord and Eq instance are by index.

Note that one may use the index as some sort of optimization. i.e. if the indices are the same, then the values are considerd the same

Constructors

ByIndex !ix !a

theIndex :: forall ix1 a ix2 f. Functor f => (ix1 -> f ix2) -> ByIndex ix1 a -> f (ByIndex ix2 a)

theValue :: forall ix a1 a2 f. Functor f => (a1 -> f a2) -> ByIndex ix a1 -> f (ByIndex ix a2)