HGeometry.ClosestPair.DivideAndConquer

Description

Classical \(O(n\log n)\) time divide and conquer algorithm to compute the closest pair among a set of \(n\) points in \(\mathbb{R}^2\).

Documentation

closestPair :: (Ord r, Num r, Point_ point 2 r, Foldable1 f) => f point -> Vector 2 point

Classical divide and conquer algorithm to compute the closest pair among \(n\) points.

pre: input contains at least two points

running time: \(O(n \log n)\)

type CP point = Top (SP (Two point) (NumType point))

the closest pair and its (squared) distance

data CCP point

Type used in the closest pair computation. The fields represent the points ordered on increasing y-order and the closest pair (if we know it)

Constructors

CCP (NonEmpty point) !(CP point)

mergePairs

Arguments

:: forall point r. (Ord r, Num r, Point_ point 2 r)

=> CP point
current closest pair and its dist

-> NonEmpty point
pts on the left

-> NonEmpty point
pts on the right

-> CP point

Function that does the actual merging work