HGeometry.ClosestPair.Naive

Description

Naive O(n^2)) time algorithm to compute the closest pair of points among \(n\) points in \(\mathbb{R}^d\).

Documentation

closestPair :: (Ord r, Num r, Point_ point d r, Foldable1 f, Has_ Metric_ d r) => f point -> Vector 2 point

Naive algorithm to compute the closest pair according to the (squared) Euclidean distance in \(d\) dimensions. Note that we need at least two elements for there to be a closest pair.

pre: there are at least two! points in the input

running time: \(O(dn^2)\) time.

closestPairWith :: (Ord r, Point_ point d r, Foldable1 f) => DistanceFunction point -> f point -> SP (Vector 2 point) r

Naive algorithm to compute the closest pair of points (and the distance realized by those points) given a distance function. Note that we need at least two elements for there to be a closest pair.

running time: \(O(T(d)n^2)\), where \(T(d)\) is the time required to evaluate the distance between two points in \(\mathbb{R}^d\).

type DistanceFunction g = g -> g -> NumType g

A distance function between geometries of type g