HGeometry.Cone

Description

A Data type to represent Cones

Documentation

data Cone r point edge

A Cone

Constructors

Cone point (Vector 2 r :+ edge) (Vector 2 r :+ edge)

apex :: forall r point1 edge point2 f. Functor f => (point1 -> f point2) -> Cone r point1 edge -> f (Cone r point2 edge)

leftBoundaryVector :: forall r point edge f. Functor f => ((Vector 2 r :+ edge) -> f (Vector 2 r :+ edge)) -> Cone r point edge -> f (Cone r point edge)

rightBoundaryVector :: forall r point edge f. Functor f => ((Vector 2 r :+ edge) -> f (Vector 2 r :+ edge)) -> Cone r point edge -> f (Cone r point edge)

leftBoundary :: (Dimension point ~ 2, NumType point ~ r) => Cone r point edge -> HalfLine point :+ edge

Get the left boundary as a HalfLine starting at the apex.

rightBoundary :: (Dimension point ~ 2, NumType point ~ r) => Cone r point edge -> HalfLine point :+ edge

Get the left boundary as a HalfLine starting at the apex.

coneBisector :: (Point_ point 2 r, Num r) => Cone r point edge -> HalfLine point

Get the bisector of the cone

intersectingHalfplanes :: (Point_ point 2 r, Num r, Ord r) => Cone r point edge -> Vector 2 (HalfSpaceF (LinePV 2 r))

Get the two halfplanes so that the cone is the intersection of the two halfplanes. the first halfplane is the plane right of the left boundary, whereas the second halfplane is the plane left of the right boundary.