HGeometry.Cone.Intersection

Documentation

toConvexPolygonIn :: (Rectangle_ rectangle point, Point_ point 2 r, Ord r, Fractional r) => rectangle -> Cone r point edge -> ConvexPolygonF (Cyclic NonEmpty) (OriginalOrExtra point (Point 2 r))

turn the cone into a convex polygon by clipping it in the given rectangle.

pre: the apex lies inside the given rectangle

extraPoints :: (Rectangle_ rectangle corner, Point_ corner 2 r, Point_ point 2 r, Fractional r, Ord r, IsIntersectableWith (HalfLine point) (ClosedLineSegment corner), Intersection (HalfLine point) (ClosedLineSegment corner) ~ Maybe (HalfLineLineSegmentIntersection (Point 2 r) (ClosedLineSegment corner))) => HalfLine point -> HalfLine point -> rectangle -> NonEmpty (Point 2 r)

computes the extra vertices (i.e. corners of the box) in between the two given halflines (in CCW order) that we have to insert to make an unbounded region bounded.

pre: the starting points of the halflines lie inside the rectangle