HGeometry.ConvexHull.R3.Naive

Description

Documentation

type ConvexHull point = [Triangle point]

lowerHull' :: forall point r. (Ord r, Fractional r, Show r, Point_ point 3 r) => NonEmpty point -> ConvexHull point

Computes the lower hull without its vertical triangles.

pre: The points are in general position. In particular, no four points should be coplanar.

running time: \(O(n^4)\)

lowerHullAll :: (Ord r, Fractional r, Show r, Point_ point 3 r) => NonEmpty point -> ConvexHull point

Generates a set of triangles to be used to construct a complete convex hull. In particular, it may contain vertical triangles.

pre: The points are in general position. In particular, no four points should be coplanar.

running time: \(O(n^4)\)

isValidTriangle :: (Num r, Ord r, Point_ point 3 r) => Triangle point -> [point] -> Maybe point

Tests if this is a valid triangle for the lower envelope. That is, if all point lie above the plane through these points. Returns a Maybe; if the result is a Nothing the triangle is valid, if not it returns a counter example.

>>> let t = (Triangle origin (Point3 1 0 0) (Point3 0 1 0))
>>> isValidTriangle t [Point3 5 5 0]
Nothing
>>> let t = (Triangle origin (Point3 1 0 0) (Point3 0 1 0))
>>> isValidTriangle t [Point3 5 5 (-10)]
Just (Point3 5 5 (-10))

upperHalfSpaceOf :: (Ord r, Num r, Point_ point 3 r) => Triangle point -> HalfSpace 3 r

Computes the halfspace above the triangle.

>>> upperHalfSpaceOf (Triangle origin (Point3 10 0 0) (Point3 0 10 0))
HalfSpace Positive (HyperPlane [0,0,0,-100])