HGeometry.ConvexHull.R3.Naive.Dual

Documentation

type UpperHull point = LowerEnvelope (NonVerticalHyperPlane 3 (NumType point) :+ point) ()

upperHull :: (Point_ point 3 r, Ord r, Fractional r, Foldable1 f, Functor f, Show point, Show r, Ord point) => f point -> UpperHull point

Computes the upper hull of a set of points in R^3

\(O(n^4)\)

type Facet point = NonEmpty point

facets :: (Ord (NumType point), Ord point) => UpperHull point -> [Facet point]

Outputs the facets of the upper hull.