HGeometry.Cyclic

Description

Representing Cyclic Sequences

Documentation

newtype Cyclic (v :: k -> Type) (a :: k)

A cyclic sequence type

Constructors

Cyclic (v a)

class HasDirectedTraversals (v :: Type -> Type) where

Class that models that some type has a cyclic traversal starting from a particular index.

Methods

traverseRightFrom :: Index (v a) -> IndexedTraversal1' (Index (v a)) (v a) a

A rightward-traversal over all elements starting from the given one. Indices are taken modulo the length.

running time : \(O(n)\)

traverseLeftFrom :: Index (v a) -> IndexedTraversal1' (Index (v a)) (v a) a

A rightward-traversal over all elements starting from the given one. Indices are taken modulo the length.

running time : \(O(n)\)

class ShiftedEq t where

Class for types that have an Equality test up to cyclic shifts

Associated Types

type ElemCyclic t

The type of the elements stored in this cyclic container.

Methods

isShiftOf :: t -> t -> Bool

Given a and b, test if a is a shifted version of the other.

groupWith :: (Foldable1 cyclic, Eq b) => (a -> b) -> cyclic a -> Cyclic NonEmpty (b, NonEmpty a)

Groups the elements of a cyclic. Note in particular that this may join the first and last group, thereby changing the indices of the individual elements.

the items are reported in the same order as before.

withCyclicSuccessor :: forall (cyclic :: Type -> Type) a b. Traversable1 cyclic => Traversal1 (cyclic a) (cyclic b) (a, a) b

A traversal that associates every elemnt with its successor

withCyclicPredecessor :: forall (cyclic :: Type -> Type) a b. Traversable1 cyclic => Traversal1 (cyclic a) (cyclic b) (a, a) b

A traversal that associates every elemnt with its predecessor

withCyclicNeighbours :: forall (cyclic :: Type -> Type) a b. Traversable1 cyclic => Traversal1 (cyclic a) (cyclic b) (a, V2 a) b

Traverse a cyclic structure together with both its neighbors

iWithCyclicSuccessor :: forall (cyclic :: Type -> Type) i a b. (Traversable1 cyclic, TraversableWithIndex i cyclic) => IndexedTraversal1 (i, i) (cyclic a) (cyclic b) (a, a) b

An indexed version of withCyclicSuccessor

iWithCyclicPredecessor :: forall (cyclic :: Type -> Type) i a b. (Traversable1 cyclic, TraversableWithIndex i cyclic) => IndexedTraversal1 (i, i) (cyclic a) (cyclic b) (a, a) b

An indexed traversal that associates every elemnt with its predecessor

iWithCyclicNeighbours :: forall (cyclic :: Type -> Type) i a b. (Traversable1 cyclic, TraversableWithIndex i cyclic) => IndexedTraversal1 (i, V2 i) (cyclic a) (cyclic b) (a, V2 a) b

An indexed traversal that associates every elemnt with its neighbours

data V2 a

A 2-dimensional vector

>>> pure 1 :: V2 Int
V2 1 1

>>> V2 1 2 + V2 3 4
V2 4 6

>>> V2 1 2 * V2 3 4
V2 3 8

>>> sum (V2 1 2)
3

Constructors

V2 !a !a