HGeometry.DelaunayTriangulation

Description

Documentation

delaunayTriangulation :: (Foldable1 set, Point_ point 2 r, Ord point, Ord r, Num r) => set point -> Triangulation point

Naive \( O(n^4) \) time implementation of the delaunay triangulation. Simply tries each triple (p,q,r) and tests if it is delaunay, i.e. if there are no other points in the circle defined by p, q, and r.

pre: the input is a *SET*, i.e. contains no duplicate points.

edgesAsPoints :: Triangulation point -> [(point, point)]

List add edges as point pairs.

edgesAsVertices :: Triangulation point -> [(VertexID, VertexID)]

List add edges as VertexID pairs.

neighbours :: forall point f. Functor f => (Vector (CList VertexID) -> f (Vector (CList VertexID))) -> Triangulation point -> f (Triangulation point)

Point neighbours indexed by VertexID.

positions :: forall point f. Functor f => (Vector point -> f (Vector point)) -> Triangulation point -> f (Triangulation point)

Point positions indexed by VertexID.

toPlaneGraph :: forall {k} (s :: k) point r. (Point_ point 2 r, Ord r, Num r) => Triangulation point -> CPlaneGraph s point () ()

ST' is a strict triple (m,a,x) containing:

m: a Map, mapping edges, represented by a pair of vertexId's (u,v) with u < v, to arcId's.

a: the next available unused arcID

x: the data value we are interested in computing type ST' a = ST (SM.Map (VertexID,VertexID) ArcID) ArcID a

convert the triangulation into a plane graph

running time: \(O(n)\).

vertexIds :: forall point f. Functor f => (Map point VertexID -> f (Map point VertexID)) -> Triangulation point -> f (Triangulation point)

Mapping between triangulated points and their internal VertexID.

type Adj = IntMap (CList VertexID)

Neighbours indexed by VertexID.

type DTVertex = CList VertexID

Rotating Right - rotate clockwise

type Mapping point = (Map point VertexID, Vector point)

Bidirectional mapping between points and VertexIDs.

data Triangulation point

Neighbours are stored in clockwise order: i.e. rotating right moves to the next clockwise neighbour.

Constructors

Triangulation

Fields

_vertexIds :: Map point VertexID

_positions :: Vector point

_neighbours :: Vector (CList VertexID)

neighbours are in CW order

type VertexID = Int

Vertex identifier.