HGeometry.DelaunayTriangulation.Naive

Description

Documentation

delaunayTriangulation :: (Foldable1 set, Point_ point 2 r, Ord point, Ord r, Num r) => set point -> Triangulation point

Naive \( O(n^4) \) time implementation of the delaunay triangulation. Simply tries each triple (p,q,r) and tests if it is delaunay, i.e. if there are no other points in the circle defined by p, q, and r.

pre: the input is a *SET*, i.e. contains no duplicate points.

toAdjLists :: (Point_ point 2 r, Num r, Ord r) => Mapping point -> Set (VertexID, VertexID) -> Vector (CList VertexID)

Given a list of edges, as vertexId pairs, construct a vector with the adjacency lists, each in CW sorted order.

sortAroundMapping :: (Point_ point 2 r, Num r, Ord r) => Mapping point -> VertexID -> [VertexID] -> [VertexID]

Given a particular point u and a list of points vs, sort the points vs in CW order around u. running time: \( O(m log m) \), where m=|vs| is the number of vertices to sort.

isDelaunay :: (Point_ point 2 r, Num r, Ord r) => Mapping point -> VertexID -> VertexID -> VertexID -> Bool

\( O(n) \) Test if the given three points form a triangle in the delaunay triangulation.