HGeometry.Direction

Description

Compuations that have to do with directions

Documentation

class r ~ NumType geom => CanComputeNormalVector geom r | geom -> r where

Minimal complete definition

normalVectorAt

Methods

normalUnitVectorAt :: forall point (d :: Nat). (Point_ point d r, Has_ Metric_ d r, d ~ Dimension geom, Radical r, Fractional r) => point -> geom -> Vector d r

Given a point on the object, and the object. Compute the outward normal vector at the given point. The normal is returned as a unit vector.

pre: the query point lies on (the surface of) the object. (This is not checked)

normalVectorAt :: forall point (d :: Nat). (Point_ point d r, d ~ Dimension geom, Num r) => point -> geom -> Vector d r

Given a point on the object, and the object. Compute the outward normal vector at the given point. No Guarantees are given about the length of the resulting vector.

pre: the query point lies on (the surface of) the object. (This is not checked)

uniformDirection :: forall gen m r (d :: Nat). (StatefulGen gen m, Fractional r, Ord r, Radical r, UniformRange r, Has_ Metric_ d r) => gen -> m (Vector d r)

Generates a unit vector corresponding to a direction.

The general idea is to generate a random vector in the [-1,1]^d, if it lies inside the unit ball, scale it to be length 1, otherwise, retry.

uniformUpwardDirectionWrt :: forall gen m r (d :: Nat). (StatefulGen gen m, Fractional r, Ord r, Radical r, UniformRange r, Has_ Metric_ d r) => Vector d r -> gen -> m (Vector d r)

Generates an upward direction (with respect to the given normal/up vector) uniformly at random.

module HGeometry.Direction.Cardinal