HGeometry.Disk.Smallest.Naive

Description

Naive implementation to compute the smallest enclosing disk of a set of points in \(\mathbb{R}^2\)

Documentation

smallestEnclosingDisk :: (Point_ point 2 r, Ord r, Fractional r, Foldable set) => set point -> DiskByPoints point

Horrible \( O(n^4) \) implementation that simply tries all disks, checks if they enclose all points, and takes the largest one. Basically, this is only useful to check correctness of the other algorithm(s)

pre: the set contains at least 2 (distinct) points

smallestEnclosingDiskWith :: (Point_ point 2 r, Fractional r, Ord r, Foldable1 nonEmpty) => point -> nonEmpty point -> DiskByPoints point

Given a point p and a non-empty set of n points, computes the smallest disk that has p on the boundary and encloses all points.

pre: such a disk exists

running time: \(O(n^2)\)

enclosesAll :: (Foldable set, Point_ point 2 r, HasInBall disk, HasIntersectionWith (Point 2 r) disk) => set point -> disk -> Maybe disk

Test if the given disk encloses all points.