HGeometry.Ellipse

Description

Data type for representing an Ellipse.

Documentation

newtype Ellipse r

A type representing planar ellipses

Constructors

Ellipse (Transformation 2 r)

affineTransformation :: forall r s p f. (Profunctor p, Functor f) => p (Transformation 2 r) (f (Transformation 2 s)) -> p (Ellipse r) (f (Ellipse s))

Iso to access the transformation (applied on the unit circle) that we use to actually represent the ellipse

ellipseMatrix :: forall r s p f. (Profunctor p, Functor f) => p (Matrix 3 3 r) (f (Matrix 3 3 s)) -> p (Ellipse r) (f (Ellipse s))

Get the matrix describing the ellipse.

unitEllipse :: Num r => Ellipse r

Ellipse representing the unit circle

circleToEllipse :: (Radical r, Point_ point 2 r) => Circle point -> Ellipse r

Convert a circle to an ellipse

ellipseToCircle :: (Num r, Eq r) => Ellipse r -> Maybe (Circle (Point 2 r))

Try to Convert a circle into an ellipse. Returns a Nothing if the ellipse is not an actual circle.

_EllipseCircle :: (Radical r, Eq r) => Prism' (Ellipse r) (Circle (Point 2 r))

Converting between ellipses and circles

Prims to convert between an Ellipse and a Circle.