HGeometry.Foldable.Util

Description

Convenience classes for types that can be built from foldable collections.

Documentation

class HasFromFoldable (f :: Type -> Type) where

Types that can be built from foldable collections

Minimal complete definition

fromList

Methods

fromFoldable :: Foldable g => g a -> f a

Build the data structure from something that is foldable.

fromList :: [a] -> f a

Build the data structure from a list of elements

class HasFromFoldable1 (f :: Type -> Type) where

Types that can be built from non-empty foldable collections

Minimal complete definition

fromNonEmpty

Methods

fromFoldable1 :: Foldable1 g => g a -> f a

Build the data structure from something that is foldable and has at least one element.

fromNonEmpty :: NonEmpty a -> f a

Build the data structure from a non-empty list of elements