HGeometry.Graphics.Camera

Description

Documentation

data Camera r

A basic camera data type. The fields stored are:

the camera position,

the raw camera normal, i.e. a unit vector into the center of the screen,

the raw view up vector indicating which side points "upwards" in the scene,

the focal depth (i.e. the distance from the camera position to the plane on which we project),

the near distance (everything closer than this is clipped),

the far distance (everything further away than this is clipped), and

the viewport dimensions; i.e. the viewport (on the view plane) on which we draw

Constructors

Camera !(Point 3 r) !(Vector 3 r) !(Vector 3 r) !r !r !r !(Vector 2 r)

cameraPosition :: forall r f. Functor f => (Point 3 r -> f (Point 3 r)) -> Camera r -> f (Camera r)

Camera position.

rawCameraNormal :: forall r f. Functor f => (Vector 3 r -> f (Vector 3 r)) -> Camera r -> f (Camera r)

Raw camera normal, i.e. a vector into the center of the screen.

rawViewUp :: forall r f. Functor f => (Vector 3 r -> f (Vector 3 r)) -> Camera r -> f (Camera r)

Raw view up vector indicating which side points "upwards" in the scene.

focalDepth :: forall r f. Functor f => (r -> f r) -> Camera r -> f (Camera r)

The focal length (i.e. distance from the camera position to the plane on which we project.) Also known as viewplane depth.

nearDist :: forall r f. Functor f => (r -> f r) -> Camera r -> f (Camera r)

Near distance (everything closer than this is clipped).

farDist :: forall r f. Functor f => (r -> f r) -> Camera r -> f (Camera r)

Far distance (everything further away than this is clipped).

viewportDimensions :: forall r f. Functor f => (Vector 2 r -> f (Vector 2 r)) -> Camera r -> f (Camera r)

The viewport dimensions.

cameraNormal :: (Radical r, Fractional r) => Lens' (Camera r) (Vector 3 r)

Lens to get and set the Camera normal, makes sure that the vector remains normalized.

viewUp :: (Radical r, Fractional r) => Lens' (Camera r) (Vector 3 r)

Lens to get and set the viewUp vector. Makes sure the vector remains normalized.

cameraTransform :: Fractional r => Camera r -> Transformation 3 r

Full transformation that renders the figure

worldToView :: Fractional r => Camera r -> Transformation 3 r

Translates world coordinates into view coordinates

toViewPortTransform :: Fractional r => Camera r -> Transformation 3 r

Transformation into viewport coordinates

perspectiveProjection :: Fractional r => Camera r -> Transformation 3 r

constructs a perspective projection

rotateCoordSystem :: Num r => Camera r -> Transformation 3 r

Rotates coordinate system around the camera, such that we look in the negative z direction

flipAxes :: Num r => Transformation 3 r

Flips the y and z axis.