HGeometry.Graphics.Render

Description

Some functions that help rendering 3D objects

Documentation

newtype ProjectedTriangle r

Represent the projection of a 3D triangle in 2D space. i.e. this triangle acts as a triangle in R^2, but also has the information from where it came from.

Constructors

ProjectedTriangle (Triangle (Point 3 r))

triangle3 :: forall r p f. (Profunctor p, Functor f) => p (Triangle (Point 3 r)) (f (Triangle (Point 3 r))) -> p (ProjectedTriangle r) (f (ProjectedTriangle r))

Access the 3D triagnle

toTriangle2 :: ProjectedTriangle r -> Triangle (Point 2 r)

Renders a Projected Triangle as a 2D Triangle

renderTriangles :: forall triangle point r set. (Functor set, Triangle_ triangle point, Point_ point 3 r, Real r, Fractional r) => Camera Double -> set triangle -> set (ProjectedTriangle r :+ triangle)

Render a scene; i..e a set of triangles

this intermediately uses doubles to apply the camera transform