HGeometry.HalfLine

Description

Half-lines in \(d\)-dimensional space.

Documentation

data HalfLine point

A Halfline in R^d

Constructors

HalfLine !point !(Vector (Dimension point) (NumType point))

halfLineThrough :: (Point_ point d r, Num r) => point -> point -> HalfLine point

Given two points p and q, create a halfline from p through q.

>>> halfLineThrough (Point2 5 10) (Point2 10 30 :: Point 2 Int)
HalfLine (Point2 5 10) (Vector2 5 20)

data LineHalfLineIntersection point halfLine

Data type representing the intersection of a Line and a HalfLine

Constructors

Line_x_HalfLine_Point point

Line_x_HalfLine_HalfLine halfLine