HGeometry.HalfPlane.CommonIntersection

Documentation

data CommonIntersection halfPlane r

A type representing the Common intersection of a bunch of halfplanes; assuming this intersection is non-empty.

Constructors

SingletonPoint (Point 2 r) (Vector 3 halfPlane)
Common intersection is a singleton point, defined by the three given halfplanes

InSubLine (VerticalOrLineEQ r) (Vector 2 halfPlane) (SubLine halfPlane r)
The two halfPlanes that define the line, and the other halfplanes furthe restricitng the line.

InSlab (Slab r halfPlane)
two parallel halfPlanes l and u that form a slab;

BoundedRegion (ConvexPolygon (Point 2 r :+ halfPlane))
each vertex stores the interior halfplane of the CCW-edge it is incident to.

UnboundedRegion (Chain Seq r halfPlane)
each vertex stores the interior halfplane of the CCW-edge it is incident to.

data SubLine halfPlane r

Part of the line

Constructors

EntireLine

InHalfLine (HalfLine (Point 2 r)) halfPlane
the third halfPlane

InSegment (ClosedLineSegment (Point 2 r)) halfPlane halfPlane
the remaining two halfplanes

newtype Chain (f :: Type -> Type) r halfPlane

A polygonal chain bounding an unbounded convex region, in CCW order.

Constructors

Chain (Alternating f (Point 2 r) halfPlane)

commonIntersection :: (Foldable1 f, Functor f, HalfPlane_ halfPlane r, Fractional r, Ord r, HyperPlane_ (BoundingHyperPlane halfPlane 2 r) 2 r, Show halfPlane, Show r) => f halfPlane -> Maybe (CommonIntersection halfPlane r)

Computes the common intersection of a \(n\) halfplanes.

running time: \(O(n\log n)\)

boundaries :: (HalfPlane_ halfPlane r, Ord r, Fractional r, Show r, Show halfPlane) => NonVerticals halfPlane r -> These2 (Chain Seq r (LineEQ r :+ halfPlane))

Computes the upper boundary of the halfplanes that have negative sign, and the upper boundary of the halfplanes that have negative sign

clipUpperByLower :: (Ord r, Num r, HasIntersectionWith (Point 2 r) halfPlane) => Chain Seq r (line :+ halfPlane) -> Chain Seq r (line :+ halfPlane) -> Chain Seq r (line :+ halfPlane)

given the upper and lower chain, clip the upper thain to retain only the vertices that are above the lower chain.