HGeometry.HalfPlane.CommonIntersection.Bounded

Documentation

boundedCommonIntersection :: (Foldable1 set, HalfPlane_ halfPlane r, Ord r, Fractional r, HyperPlane_ (BoundingHyperPlane halfPlane 2 r) 2 r, Intersection (BoundingHyperPlane halfPlane 2 r) (BoundingHyperPlane halfPlane 2 r) ~ Maybe (LineLineIntersectionG r (BoundingHyperPlane halfPlane 2 r)), HasIntersectionWith (Point 2 r) halfPlane, IsIntersectableWith (BoundingHyperPlane halfPlane 2 r) (BoundingHyperPlane halfPlane 2 r)) => set halfPlane -> Maybe (ConvexPolygon (Point 2 r :+ halfPlane))

Given a (non)-empty set of n halfplanes with the guarantee that if they have a non-empty common intersection X than this common intersection is bounded. Compute the common intersection X.

pre: X is either empty or a non-empty two-dimensional set of points

this is for example useful if you want to compute the common intersection of halfplanes restricted to soem bounded domain (triangle or box).

\(O(n\log n)\)