HGeometry.HalfPlane.CommonIntersection.Chain

Documentation

newtype Chain (f :: Type -> Type) r halfPlane

A polygonal chain bounding an unbounded convex region, in CCW order.

Constructors

Chain (Alternating f (Point 2 r) halfPlane)

_ChainAlternating :: forall (f1 :: Type -> Type) r halfPlane (f' :: Type -> Type) r' halfPlane' p f2. (Profunctor p, Functor f2) => p (Alternating f1 (Point 2 r) halfPlane) (f2 (Alternating f' (Point 2 r') halfPlane')) -> p (Chain f1 r halfPlane) (f2 (Chain f' r' halfPlane'))

Iso to convert to just an Alternating

unboundedEdges :: forall halfPlane r (f :: Type -> Type). (HalfPlane_ halfPlane r, Ord r, Num r, HasIntersectionWith (Point 2 r) halfPlane, Cons (f (Point 2 r, halfPlane)) (f (Point 2 r, halfPlane)) (Point 2 r, halfPlane) (Point 2 r, halfPlane), Snoc (f (Point 2 r, halfPlane)) (f (Point 2 r, halfPlane)) (Point 2 r, halfPlane) (Point 2 r, halfPlane), GetDirection (BoundingHyperPlane halfPlane 2 r)) => Chain f r halfPlane -> Either (BoundingHyperPlane halfPlane 2 r :+ halfPlane) (Vector 2 (HalfLine (Point 2 r) :+ halfPlane))

Get the unbounded edges of the chain.

the unbounded edges are edges pointing twoards infinity.

singleton :: forall (f :: Type -> Type) r halfPlane. Default (f (Point 2 r, halfPlane)) => halfPlane -> Chain f r halfPlane

Creates a singleton chain

bimap' :: forall (f :: Type -> Type) r s halfPlane halfPlane'. Functor f => (Point 2 r -> Point 2 s) -> (halfPlane -> halfPlane') -> Chain f r halfPlane -> Chain f s halfPlane'

slightly more general version of bimap so we can easily flip the plane.

evalChainAt :: (Num r, Ord r) => r -> Chain Seq r (LineEQ r :+ halfPlane) -> r :+ halfPlane

Evaluates the chain at the given x-coordinate. Returns the value (y-coordinate) y at the given x-coordinate x, and the halfplane containing the point (x,y)

leftMost :: forall r halfPlane f. Functor f => (halfPlane -> f halfPlane) -> Chain Seq r halfPlane -> f (Chain Seq r halfPlane)

Get the leftmost line in the chain (i.e. the line bounding the first halfplane)

rightMost :: forall r halfPlane f. Functor f => (halfPlane -> f halfPlane) -> Chain Seq r halfPlane -> f (Chain Seq r halfPlane)

Get the rightmost line in the chain (i.e. the line bounding the last halfplane)

clipLeft :: Ord r => r -> Chain Seq r halfPlane -> Chain Seq r halfPlane

Given a value x, Clip the chain to the interval \([x,\infty)\)

clipRight :: Ord r => r -> Chain Seq r halfPlane -> Chain Seq r halfPlane

Given a value x, Clip the chain to the interval \((-\infty,x]\)

glueChains :: (Ord r, Fractional r) => Chain Seq r (LineEQ r :+ halfPlane) -> Chain Seq r (LineEQ r :+ halfPlane) -> Chain Seq r (LineEQ r :+ halfPlane)

Given two chains; glue them together by introducing the new appropriate intersection point.

pre: such an intersection point exists (i.e. the rightmost bounding line of the prefix) and the leftmost bounding line of the suffix intersect in a point.

clipLeftLine :: (Ord r, Num r, HasIntersectionWith (Point 2 r) halfPlane') => halfPlane' -> Chain Seq r halfPlane -> Chain Seq r halfPlane

Clip the chain from the left; removing when a vertex is not in the halfplane

clipRightLine :: (Ord r, Num r, HasIntersectionWith (Point 2 r) halfPlane') => halfPlane' -> Chain Seq r halfPlane -> Chain Seq r halfPlane

Clip on the right by a line

consChain :: forall (f :: Type -> Type) r halfPlane. Cons (f (Point 2 r, halfPlane)) (f (Point 2 r, halfPlane)) (Point 2 r, halfPlane) (Point 2 r, halfPlane) => (halfPlane, Point 2 r) -> Chain f r halfPlane -> Chain f r halfPlane

Add a new halfplane and first vertex to the chain.