HGeometry.HalfSpace

Description

Representing halfspaces

Documentation

module HGeometry.HalfSpace.Class

boundingHyperPlaneLens :: forall boundingHyperPlane boundingHyperPlane' f. Functor f => (boundingHyperPlane -> f boundingHyperPlane') -> HalfSpaceF boundingHyperPlane -> f (HalfSpaceF boundingHyperPlane')

Lens to access the hyperplane bounding the halfspace

Generally, one can just use boundingHyperPlane from the HalfSpace_ class. But this one is useful if you need a lens to change the type of the bounding hyperplane.

data Sign

The sign of an expression

Constructors

Negative

Positive

type family BoundingHyperPlane halfSpace (d :: Nat) r

type HalfPlane_ halfPlane r = HalfSpace_ halfPlane 2 r

Type synonym for halfplanes in R^2

class (Dimension halfSpace ~ d, Dimension (BoundingHyperPlane halfSpace d r) ~ d, NumType halfSpace ~ r, NumType (BoundingHyperPlane halfSpace d r) ~ r) => HalfSpace_ halfSpace (d :: Nat) r | halfSpace -> d, halfSpace -> r where

Types modelling halfspaces.

Associated Types

type BoundingHyperPlane halfSpace (d :: Nat) r

Methods

boundingHyperPlane :: Lens' halfSpace (BoundingHyperPlane halfSpace d r)

Access the bounding hyperplane

halfSpaceSign :: Lens' halfSpace Sign

Lens to access the sign of the halfspace.

type HalfPlaneF line = HalfSpaceF line

A Halfplane (which is just a Halfspace in R^2)

type HalfSpace (d :: Natural) r = HalfSpaceF (HyperPlane d r)

Arbitrary halfspaces in r^d.

data HalfSpaceF boundingHyperPlane

A HalfSpace bounded by a hyperplane. The sign indicates which side of the bounding hyperplane is indicated. Iff the sign is positive, we mean the points for which onSideTest returns a positive value.

Half spaces include their bounding hyperplane.

Constructors

HalfSpace !Sign boundingHyperPlane

data Point_x_HalfSpace_Intersection boundaryPoint interiorPoint

Point x HalfSpace intersection

Constructors

Point_x_HalfSpace_OnBoundary boundaryPoint

Point_x_HalfSpace_Interior interiorPoint

class GetDirection line where

Methods

inLineVector :: forall r (d :: Nat). (r ~ NumType line, d ~ Dimension line) => line -> Vector d r

Get a vector v that lies in the line; i.e. given a point p that lies on the line; the point p .+^ v lies also in the line.

data HalfPlaneIntersection r halfPlane

The non-empty intersection of two halfPlanes

Constructors

HalfPlane_x_HalfPlane_Line (BoundingHyperPlane halfPlane 2 r)

HalfPlane_x_HalfPlane_Slab (Slab r halfPlane)

HalfPlane_x_HalfPlane_Cone (Cone r (Point 2 r) halfPlane)

HalfPlane_x_HalfPlane_HalfPlane halfPlane

data LineHalfPlaneIntersection r line

Constructors

Line_x_HalfPlane_Line line

Line_x_HalfPlane_HalfLine (HalfLine (Point 2 r))

leftBoundingVector :: (HalfPlane_ halfPlane r, Ord r, Num r, HasIntersectionWith (Point 2 r) halfPlane, GetDirection (BoundingHyperPlane halfPlane 2 r)) => Point 2 r -> halfPlane -> Vector 2 r

given a point a on the bounding hyperplane; compute a vector pointing in the direction of the bounding line so that the halfspace is to its left.

rightBoundingVector :: (HalfPlane_ halfPlane r, Ord r, Num r, HasIntersectionWith (Point 2 r) halfPlane, GetDirection (BoundingHyperPlane halfPlane 2 r)) => Point 2 r -> halfPlane -> Vector 2 r

given a point a on the bounding hyperplane; compute a vector pointing in the direction of the bounding line so that the halfspace is to its right