HGeometry.HyperPlane

Description

Hyperplanes in \(d\)-dimensional space.

Documentation

module HGeometry.HyperPlane.Class

newtype HyperPlane (d :: Natural) r

A Hyperplane h in d-dimensions, described by a vector of coefficients (a_0,..,a_d).

a point \( (p_1,..,p_d) \) lies on (h) iff: \( a_0 + \sum_i=1^d a_i*p_i = 0 \)

Constructors

HyperPlane (Vector (d + 1) r)

Bundled Patterns

pattern HyperPlane2 :: r -> r -> r -> HyperPlane 2 r
Construct a Hyperplane, i.e. a line in R^2

HyperPlane2 c a b represents the line ax + by + c = 0

pattern HyperPlane3 :: r -> r -> r -> r -> HyperPlane 3 r
Construct a plane in R^3

HyperPlane3 d a b c represnest the plane ax + by + cz + d = 0

type MkHyperPlaneConstraints (d :: Natural) r = (d < (d + 1), KnownNat d, Has_ Metric_ d r, Has_ Metric_ (d + 1) r, Has_ Vector_ d r, Has_ Vector_ (d + 1) r)

Constraints on d needed to be able to construct hyperplanes; pretty much all of these are satisfied by default, it is just that the typechecker does not realize that.

cmpInDirection :: forall point (d :: Natural) r. (Ord r, Num r, Has_ Metric_ (d + 1) r, Has_ Metric_ d r, Point_ point d r, d < (d + 1)) => Vector d r -> point -> point -> Ordering

Compare points with respect to the direction given by the vector, i.e. by taking planes whose normal is the given vector.

module HGeometry.HyperPlane.NonVertical

data PlanePlaneIntersection plane line

The intersection between two planes in R^3.

Constructors

Plane_x_Plane_Line line

Plane_x_Plane_Plane plane