HGeometry.HyperPlane.Class

Description

Classes for representing hyperplanes in d-dimensional space.

Documentation

class (NumType hyperPlane ~ r, Dimension hyperPlane ~ d, Has_ Vector_ d r, Has_ Vector_ (1 + d) r) => HyperPlane_ hyperPlane (d :: Natural) r | hyperPlane -> d, hyperPlane -> r where

A class to represent hyperplanes in d-dimensional space.

Minimal complete definition

Nothing

Methods

evalHyperPlaneEquation :: (Num r, Point_ point d r) => hyperPlane -> point -> r

Evalute the expression \(a_0 + \sum_i=1^d a_i*p_i \) at the given point.

>>> evalHyperPlaneEquation myLine (Point2 5 10)
-3.0

default evalHyperPlaneEquation :: (Num r, Point_ point d r, Has_ Metric_ (d + 1) r) => hyperPlane -> point -> r

hyperPlaneEquation :: hyperPlane -> Vector (d + 1) r

A hyperplane \(h) has coefficients \(a_i \in \mathbb{R}\) so that a point \( (p_1,..,p_d) \) lies on (h) iff: \( a_0 + \sum_i=1^d a_i*p_i = 0 \)

this fuction returns the vector of these coefficients \(\langle a_0,..,a_d \rangle\).

>>> hyperPlaneEquation myLine
Vector3 2.0 1.0 (-1.0)
>>> hyperPlaneEquation myLineAgain
Vector3 2.0 1.0 (-1.0)
>>> hyperPlaneEquation myLineAsNV
Vector3 2.0 1.0 (-1.0)

so for example myLineAsNV actually a line with slope 1 through the point (0,2)

default hyperPlaneEquation :: (NonVerticalHyperPlane_ hyperPlane d r, Num r, 1 <= d) => hyperPlane -> Vector (d + 1) r

normalVector :: hyperPlane -> Vector d r

Get the normal vector of the hyperplane. The vector points into the positive halfspace.

>>> normalVector myVerticalLine
Vector2 1.0 (-0.0)
>>> normalVector myLine
Vector2 (-1.0) 1.0

default normalVector :: (KnownNat d, Num r, Eq r, 1 <= d) => hyperPlane -> Vector d r

onHyperPlane :: (Point_ point d r, Eq r, Num r) => point -> hyperPlane -> Bool

Test if a point lies on a hyperplane.

>>> Point2 0 2 `onHyperPlane` myLineAgain
True
>>> Point2 1 3 `onHyperPlane` myLineAgain
True
>>> Point2 1 5 `onHyperPlane` myLineAgain
False

>>> Point2 0 2 `onHyperPlane` myLineAsNV
True
>>> Point2 1 3 `onHyperPlane` myLineAsNV
True
>>> Point2 1 5 `onHyperPlane` myLineAsNV
False

default onHyperPlane :: (Point_ point d r, Eq r, Num r) => point -> hyperPlane -> Bool

onSideTest :: (Point_ point d r, Ord r, Num r) => point -> hyperPlane -> Ordering

Test if a point lies on a hyperplane. Returns the sign when evaluating the hyperplane equation.

>>> Point2 0 2 `onSideTest` myLineAgain
EQ
>>> Point2 1 3 `onSideTest` myLineAgain
EQ
>>> Point2 1 5 `onSideTest` myLineAgain
GT
>>> Point2 4 5 `onSideTest` myLineAgain
LT
>>> Point2 0 0 `onSideTest` HyperPlane2 1 (-1) 0
LT

>>> Point2 1 1 `onSideTest` myVerticalLine
LT
>>> Point2 10 1 `onSideTest` myVerticalLine
GT
>>> Point2 5 20 `onSideTest` myVerticalLine
EQ

>>> Point2 0 1 `onSideTest` myOtherLine
LT
>>> Point2 0 (-2) `onSideTest` myOtherLine
EQ
>>> Point2 1 (-3.5) `onSideTest` myOtherLine
EQ
>>> Point2 1 (-4) `onSideTest` myOtherLine
GT

class HyperPlane_ hyperPlane d r => ConstructableHyperPlane_ hyperPlane (d :: Natural) r where

Class representing hyperplanes with methods to construct the hyperplane.

Minimal complete definition

hyperPlaneFromEquation

Associated Types

type HyperPlaneFromEquationConstraint hyperPlane (d :: Natural) r

Additional constraints for constructing a hyperplane from its equation.

type HyperPlaneFromEquationConstraint hyperPlane (d :: Natural) r = ()

Methods

hyperPlaneFromEquation :: Vector (d + 1) r -> hyperPlane

Given the coefficients \(a_0,..,a_d\) of the equation, i.e. so that

\( a_0 + \sum_i=1^d a_i*p_i = 0 \)

construct the hyperplane form it.

>>> hyperPlaneFromEquation (Vector3 10 2 1) :: HyperPlane 2 Int -- the line 2*x + 1*y + 10 = 0
HyperPlane [10,2,1]
>>> hyperPlaneFromEquation $ Vector4 100 5 3 (-1)  :: HyperPlane 3 Int -- the plane 5*x + 3*y + (-1)*z + 100= 0
HyperPlane [100,5,3,-1]

>>> myOtherLine == hyperPlaneFromEquation (Vector3 4 3 2)
True

fromPointAndNormal :: (Point_ point d r, Num r) => point -> Vector d r -> hyperPlane

Construct a Hyperplane from a point and a normal. The normal points into the halfplane for which the side-test is positive.

>>> myVerticalLine == fromPointAndNormal (Point2 5 30) (Vector2 1 0)
True

default fromPointAndNormal :: (Point_ point d r, Num r, ConstructableHyperPlane_ hyperPlane d r, HyperPlaneFromEquationConstraint hyperPlane d r, Has_ Metric_ d r) => point -> Vector d r -> hyperPlane

class HyperPlane_ hyperPlane d r => NonVerticalHyperPlane_ hyperPlane (d :: Natural) r where

Non-vertical hyperplanes.

Minimal complete definition

hyperPlaneCoefficients

Methods

evalAt :: (Num r, 1 <= d, Point_ point (d - 1) r) => point -> hyperPlane -> r

Get the coordinate in dimension \(d\) of the hyperplane at the given position.

>>> evalAt (Point1 1) myLineAsNV
3.0
>>> evalAt (Point1 10) myLineAsNV
12.0
>>> evalAt (Point1 5) <$> asNonVerticalHyperPlane myOtherLine
Just (-9.5)

default evalAt :: (Num r, 1 <= d, Point_ point (d - 1) r, (1 + (d - 1)) ~ d, Has_ Metric_ d r) => point -> hyperPlane -> r

hyperPlaneCoefficients :: Lens' hyperPlane (Vector d r)

The coefficients \( \langle a_1,..,a_d \rangle \) such that a point \(p = (p_1,..,p_d) \) lies on the hyperplane the given coefficients iff

\( a_d + \sum_i=1^{d-1} a_i*p_i = p_d \)

>>> view hyperPlaneCoefficients myLineAsNV
Vector2 1.0 2.0
>>> view hyperPlaneCoefficients myLine
Vector2 1.0 2.0
>>> view hyperPlaneCoefficients $ Plane 1 2 3
Vector3 1 2 3
>>> asNonVerticalHyperPlane myOtherLine ^?_Just.hyperPlaneCoefficients
Just (Vector2 (-1.5) (-2.0))

verticalSideTest :: (1 <= d, Point_ point d r, Ord r, Num r) => point -> hyperPlane -> Ordering

Test if a point q lies above a non-vertical hyperplane h (i.e. verticalSideTest q h == GT), on the hyperplane (== EQ), or below (LT).

>>> Point2 0 2 `verticalSideTest` myLineAsNV
EQ
>>> Point2 1 3 `verticalSideTest` myLineAsNV
EQ
>>> Point2 1 5 `verticalSideTest` myLineAsNV
GT
>>> Point2 4 5 `verticalSideTest` myLineAsNV
LT

>>> Point2 0 1 `verticalSideTest` myOtherNVLine
GT
>>> Point2 0 (-2) `verticalSideTest` myOtherNVLine
EQ
>>> Point2 1 (-3.5) `verticalSideTest` myOtherNVLine
EQ
>>> Point2 1 (-4) `verticalSideTest` myOtherNVLine
LT

default verticalSideTest :: (Point_ point d r, Num r, Ord r, 1 <= d, (1 + (d - 1)) ~ d, Has_ Metric_ d r, Has_ Additive_ (d - 1) r) => point -> hyperPlane -> Ordering

isParallelTo :: forall hyperPlane (d :: Natural) r hyperPlane'. (HyperPlane_ hyperPlane d r, HyperPlane_ hyperPlane' d r, Has_ Metric_ d r, Num r, Eq r, 1 <= d) => hyperPlane -> hyperPlane' -> Bool

Test if two hyperplanes are parallel.

pointOn :: forall hyperPlane (d :: Natural) r. (HyperPlane_ hyperPlane d r, Eq r, Fractional r, Has_ Additive_ d r, FoldableWithIndex Int (Vector d)) => hyperPlane -> Point d r

Produce a point that lies on the hyperplane. No gurantees are given about which point

>>> pointOn myLine
Point2 (-2.0) 0.0

class HyperPlaneFromPoints hyperPlane where

Class that tells us we can construct a hyperplane from a vector of points.

Methods

hyperPlaneThrough :: forall point (d :: Nat) r. (Point_ point d r, HyperPlane_ hyperPlane d r, Num r) => Vector d point -> hyperPlane

Construct a hyperplane through the given d points.