HGeometry.Intersection

Description

Defines a data type for representing intersections. Mostly useful for the more geometric types.

Documentation

type family Intersection g h :: Type

The result of interesecting two geometries,

data family IntersectionOf g h

The data family specifying to help implement the Intersection type family.

class HasIntersectionWith g h where

Class for types for which we can test if they intersect.

Minimal complete definition

Nothing

Methods

intersects :: g -> h -> Bool

g intersects h = The intersection of g and h is non-empty.

default intersects :: (Intersection g h ~ Maybe (IntersectionOf g h), IsIntersectableWith g h) => g -> h -> Bool

class HasIntersectionWith g h => IsIntersectableWith g h where

Class relationship between intersectable geometric objects.

Methods

intersect :: g -> h -> Intersection g h

Computes te intersection of two geometric objects.