HGeometry.Interval

Description

Intervals

Documentation

data Interval (endPoint :: k -> Type) (r :: k) where

Data type representing intervals

Bundled Patterns

pattern Interval :: endPoint r -> endPoint r -> Interval endPoint r
Construct an interval

pattern ClosedInterval :: r -> r -> ClosedInterval r
Construct a closed interval

pattern OpenInterval :: r -> r -> OpenInterval r
Construct an open ended interval

type ClosedInterval r = Interval (EndPoint 'Closed) r

Cosed intervals (using a boxed representation)

type OpenInterval r = Interval (EndPoint 'Open) r

Open intervals (using a boxed representation)

data HalfOpenInterval r

An interval that is open on the left and closed on the right

Constructors

HalfOpenInterval !(EndPoint 'Open r) !(EndPoint 'Closed r)

module HGeometry.Interval.Class

asClosedInterval :: Interval_ interval r => interval -> Maybe (ClosedInterval r)

Try to interpret the interval as a ClosedInterval, i.e. with both endpoints Closed

asOpenInterval :: Interval_ interval r => interval -> Maybe (OpenInterval r)

Try to interpret the interval as an OpenInterval, i.e. with both endpoints open

asAnInterval :: Interval_ interval r => interval -> Interval AnEndPoint r

convert into an interval whose endpoints are explicitly tagged.

data Interval_x_IntervalIntersection r interval

Data type representing intersections of intervals of the same type

Constructors

Interval_x_Interval_Point !r

Interval_x_Interval_Contained !interval

Interval_x_Interval_Partial !interval