HGeometry.Interval.Class

Description

Intervals

Documentation

class (IntervalLike_ interval r, NumType interval ~ r) => Interval_ interval r | interval -> r

A class for types representing Intervals

pattern Interval_ :: Interval_ interval r => StartPointOf interval -> EndPointOf interval -> interval

Pattern to match on intervals

class (HasStart interval point, HasStartPoint interval (StartPointOf interval), HasEnd interval point, HasEndPoint interval (EndPointOf interval), EndPoint_ (EndPointOf interval), IxValue (EndPointOf interval) ~ point, EndPoint_ (StartPointOf interval), IxValue (StartPointOf interval) ~ point) => IntervalLike_ interval point | interval -> point

A class for types representing interval like objects

class Interval_ interval point => ConstructableInterval_ interval point where

A class for constructable intervals

Minimal complete definition

mkInterval

Methods

mkInterval :: StartPointOf interval -> EndPointOf interval -> interval

Construct an interval given its start and end point.

pre: start < end

buildInterval :: StartPointOf interval -> EndPointOf interval -> interval

Construct an interval given two points. This makes sure the start-point comes before the endpoint.

pre: it is possible to construct a valid, non-empty interval this way. so if either endpoint is open the endpoints should not coincide.

class (Interval_ interval r, StartPointOf interval ~ EndPoint 'Closed r, EndPointOf interval ~ EndPoint 'Closed r) => ClosedInterval_ interval r

A class representing closed intervals, i.e. intervals that include their endpoints

pattern ClosedInterval_ :: ClosedInterval_ interval r => r -> r -> interval

Pattern matching on an arbitrary closed interval

class (ClosedInterval_ interval r, ConstructableInterval_ interval r) => ConstructableClosedInterval_ interval r where

A class representing constructable closed intervals

Minimal complete definition

Nothing

Methods

mkClosedInterval :: r -> r -> interval

Construct an interval given its start and end point.

buildClosedInterval :: r -> r -> interval

Construct an interval given two points. This makes sure the start-point comes before the endpoint.

clampTo :: (ClosedInterval_ interval r, Ord r) => interval -> r -> r

Clamps a value to an interval. I.e. if the value lies outside the range we report the closest value "in the range".

>>> clampTo (ClosedInterval 0 10) 20
10
>>> clampTo (ClosedInterval 0 10) (-20)
0
>>> clampTo (ClosedInterval 0 10) 5
5

class (Interval_ interval r, StartPointOf interval ~ EndPoint 'Open r, EndPointOf interval ~ EndPoint 'Open r) => OpenInterval_ interval r | interval -> r

A class representing open intervals, i.e. intervals that exclude their endpoints

pattern OpenInterval_ :: OpenInterval_ interval r => r -> r -> interval

Pattern matching on an arbitrary open interval

class (OpenInterval_ interval r, ConstructableInterval_ interval r) => ConstructableOpenInterval_ interval r | interval -> r where

Constructable Open intervals

Minimal complete definition

Nothing

Methods

mkOpenInterval :: r -> r -> interval

Construct an interval given its start s and end point t.

pre: s < t

class HasStart seg p | seg -> p where

Things that have a start point

Methods

start :: Lens' seg p

Lens to access the start point

class HasEnd seg p | seg -> p where

Things that have an end point

Methods

end :: Lens' seg p

Lens to access the ending point

startAndEnd :: (HasStart seg p, HasEnd seg p) => seg -> (p, p)

Get both the start and end of something that has a start and end.

type family StartPointOf interval

type family to declare the type of startpoint for an interval, the idea is to define this as one of the endpoinst form the Endpoints module

type family EndPointOf interval

type family to declare the type of endpoint for an interval, the idea is to define this as one of the endpoinst form the Endpoints module

class HasStartPoint seg p | seg -> p where

Things that have a start point

Methods

startPoint :: Lens' seg p

Lens to access the start point

class HasEndPoint seg p | seg -> p where

Things that have an end point

Methods

endPoint :: Lens' seg p

Lens to access the ending point

startAndEndPoint :: (HasStartPoint seg s, HasEndPoint seg e) => seg -> (s, e)

Get both the start and end of something that has a start and end.

inInterval :: forall interval r. (Ord r, Interval_ interval r) => r -> interval -> PointLocationResult

Compute where the given query value is with respect to the interval.

Note that even if the boundary of the interval is open we may return OnBoundary.

stabsInterval :: forall interval r. (Ord r, Interval_ interval r) => r -> interval -> Bool

Test if the point lies inside in,or on the boundary of, the interval.

Note that even if the boundary of the interval is open we may return OnBoundary.

compareInterval :: (Ord r, Interval_ interval r) => r -> interval -> Ordering

test if te point appears before (=LT), in (=EQ), or after (=GT) te interval.

>>> 1 `compareInterval` (OpenInterval 0 2)
EQ
>>> 1 `compareInterval` (OpenInterval 0 1)
GT
>>> 1 `compareInterval` (ClosedInterval 0 1)
EQ
>>> 10 `compareInterval` (OpenInterval 1 10)
GT
>>> 10 `compareInterval` (ClosedInterval 0 1)
GT

compareIntervalExact :: (Ord r, Interval_ interval r) => r -> interval -> CompareInterval

test if te point appears before, in, or after te interval, or on one of its endpoints.

>>> 1 `compareIntervalExact` (OpenInterval 0 2)
Interior
>>> 1 `compareIntervalExact` (OpenInterval 0 1)
OnEnd
>>> 1 `compareIntervalExact` (ClosedInterval 0 1)
OnEnd
>>> 10 `compareIntervalExact` (OpenInterval 1 10)
OnEnd
>>> 10 `compareIntervalExact` (ClosedInterval 0 1)
After

data CompareInterval

Data type to represent the result of where a point is with respect to an interval.

Constructors

Before

OnStart

Interior

OnEnd

After

shiftLeft :: (Num r, Interval_ interval r) => r -> interval -> interval

Shifts the interval to the left by delta

shiftRight :: (Num r, Interval_ interval r) => r -> interval -> interval

Shifts the interval to the right by delta

duration :: (Interval_ interval r, Num r) => interval -> r

Get the duration, or length of an interval.

module HGeometry.Interval.EndPoint