HGeometry.Interval.EndPoint

Description

Endpoints of intervals

Documentation

class IsEndPoint endPoint endPoint => EndPoint_ endPoint where

An endpoint storing values of some type r

Methods

endPointType :: endPoint -> EndPointType

Report the type of the endpoint

mkEndPoint :: IxValue endPoint -> endPoint

constructs a "default" enpoint

class IsEndPoint endPoint endPoint' where

Types that have an _endPoint field lens.

Methods

_endPoint :: Lens endPoint endPoint' (IxValue endPoint) (IxValue endPoint')

Lens to access the actual data value of the end point

data EndPointType

Possible endpoint types; open or closed

Constructors

Open

Closed

newtype EndPoint (et :: EndPointType) r

EndPoint with a type safe tag

Constructors

EndPoint r

Bundled Patterns

pattern OpenE :: r -> EndPoint Open r
Constructs an Open endpoint

pattern ClosedE :: r -> EndPoint Closed r
Constructs a closed endpoint

data AnEndPoint r

Data type modelling an endpoint that can both be open and closed.

Constructors

AnEndPoint !EndPointType !r

Bundled Patterns

pattern AnOpenE :: r -> AnEndPoint r
Constructs an Open endpoint

pattern AnClosedE :: r -> AnEndPoint r
Constructs a closed endpoint

asAnEndPoint :: EndPoint_ endPoint => endPoint -> AnEndPoint (IxValue endPoint)

Convert the endpoint into a, AnEndPoint