HGeometry.IntervalTree

Description

Documentation

data IntervalTree interval r

IntervalTree type, storing intervals of type interval

createTree :: (Foldable f, Ord r) => f r -> IntervalTree interval r

Given an ordered list of points, create an interval tree

\(O(n)\)

fromIntervals :: (Foldable f, Ord r, ClosedInterval_ interval r) => f interval -> IntervalTree interval r

Build an interval tree

\(O(n \log n)\)

insert :: (Ord r, ClosedInterval_ interval r) => interval -> IntervalTree interval r -> IntervalTree interval r

Insert the interval into the tree.

pre: the interval intersects some midpoint in the tree

\(O(\log n)\), where n is the number of endpoints the tree is built on

delete :: (Ord r, ClosedInterval_ interval r, Eq interval) => interval -> IntervalTree interval r -> IntervalTree interval r

Delete an interval from the Tree

\(O(\log n)\) (under some general position assumption)

stab :: Ord r => r -> IntervalTree interval r -> [interval]

Find all intervals that stab x

\(O(\log n + k)\), where k is the output size

search :: Ord r => r -> IntervalTree interval r -> [interval]

Find all intervals that stab x

\(O(\log n + k)\), where k is the output size

listIntervals :: IntervalTree interval r -> [interval]

Lists the intervals. We don't guarantee anything about the order

running time: \(O(n)\).