HGeometry.Ipe.Instances

Description

Some instances for HasDefaultIpeOut for geometry types that require basic data types (in kernel) as well as polygons or polylines.

Orphan instances

(HasDefaultIpeOut halfPlane, HasDefaultIpeOut (BoundingHyperPlane halfPlane 2 r), HalfPlane_ halfPlane r, Fractional r, Ord r, Show r, NumType halfPlane ~ r, Dimension halfPlane ~ 2, Dimension (BoundingHyperPlane halfPlane 2 r) ~ 2, NumType (BoundingHyperPlane halfPlane 2 r) ~ r) => HasDefaultIpeOut (HalfPlaneIntersection r halfPlane)

Instance details

Associated Types

type DefaultIpeOut (HalfPlaneIntersection r halfPlane)

Instance details

Defined in HGeometry.Ipe.Instances

type DefaultIpeOut (HalfPlaneIntersection r halfPlane) = Group

Methods

defIO :: IpeOut (HalfPlaneIntersection r halfPlane) (DefaultIpeOut (HalfPlaneIntersection r halfPlane)) (NumType (HalfPlaneIntersection r halfPlane))

(HasDefaultIpeOut line, Fractional r, Ord r, Show r, DefaultIpeOut line ~ Path, NumType line ~ r) => HasDefaultIpeOut (LineHalfPlaneIntersection r line)

Instance details

Associated Types

type DefaultIpeOut (LineHalfPlaneIntersection r line)

Instance details

Defined in HGeometry.Ipe.Instances

type DefaultIpeOut (LineHalfPlaneIntersection r line) = Path

Methods

defIO :: IpeOut (LineHalfPlaneIntersection r line) (DefaultIpeOut (LineHalfPlaneIntersection r line)) (NumType (LineHalfPlaneIntersection r line))

HasDefaultIpeOut (Slab r side)

Instance details

Associated Types

type DefaultIpeOut (Slab r side)

Instance details

Defined in HGeometry.Ipe.Instances

type DefaultIpeOut (Slab r side) = Group

Methods

defIO :: IpeOut (Slab r side) (DefaultIpeOut (Slab r side)) (NumType (Slab r side))

(Point_ point 2 r, Fractional r, Ord r, Show r) => HasDefaultIpeOut (Cone r point edge)

Instance details

Associated Types

type DefaultIpeOut (Cone r point edge)

Instance details

Defined in HGeometry.Ipe.Instances

type DefaultIpeOut (Cone r point edge) = Group

Methods

defIO :: IpeOut (Cone r point edge) (DefaultIpeOut (Cone r point edge)) (NumType (Cone r point edge))