HGeometry.Kernel

Description

Constant Complexity Geometric Primitives

Documentation

module HGeometry.Point

module HGeometry.Vector

module HGeometry.Line

module HGeometry.HyperPlane

module HGeometry.HalfSpace

module HGeometry.LineSegment

class IsBoxable g where

Types for which we can compute an axis parallel boundingbox

Minimal complete definition

Nothing

Methods

boundingBox :: forall (d :: Nat) r. (d ~ Dimension g, r ~ NumType g, Ord r) => g -> Box (Point d r)

Compute the axis-parallel boundingbox of the given geometry.

default boundingBox :: forall (d :: Nat) r point. (d ~ Dimension g, r ~ NumType g, Ord r, HasPoints' g point, Point_ point d r, Ord (Vector d r)) => g -> Box (Point d r)

class (HasMinPoint box point, HasMaxPoint box point, Point_ point (Dimension box) (NumType box)) => Box_ box point | box -> point where

d-dimensional Boxes

Methods

extent :: forall r (d :: Nat). (r ~ NumType box, d ~ Dimension box, Num r) => box -> Vector d (ClosedInterval r)

Get a vector with the extent of the box in each dimension. Note that the resulting vector is 0 indexed whereas one would normally count dimensions starting at zero.

data Box point where

D-dimensional boxes.

A box is represented by two points; a point with lexicographically minimal coordinates, and a point with lexicographically maximal coordinates.

Bundled Patterns

pattern Box :: point -> point -> Box point
Construct a box

pattern Rect :: Num r => r -> r -> r -> r -> Rectangle (Point 2 r)
Given x y w h construct the rectangle with bottom left corner (x,y), width w, and hegith h.

pattern Rectangle :: Dimension point ~ 2 => point -> point -> Box point
Construct a Rectangle

type Rectangle = Box

Defines a rectangle

module HGeometry.Ball

data Triangle point where

Triangles in d-dimensional space

Bundled Patterns

pattern Triangle :: point -> point -> point -> Triangle point
Construct a triangle from its three points

module HGeometry.Transformation

module HGeometry.Intersection

module HGeometry.Properties