HGeometry.LPType

Description

Implements Clarkson's expected linear time algorithm for solving LPType problems. (E.g. linear programming, smallest enclosing ball, etc.)

Documentation

data LPType t (basis :: Type -> Type) (set :: Type -> Type) a

A Description of an LPType problem

Constructors

LPType

Fields

costFunction :: basis a -> t

the function we are trying to minimize

combinatorialDimension :: Int

the combinatorial dimension of the problem

extendBasis :: a -> basis a -> Maybe (basis a)

function to extend the current basis into a new basis. it returns the new basis (if it has changed) and Nothing otherwise.

initialBasis :: set a -> basis a

function to construct some initial basis the input may be large.

clarkson

Arguments

:: (Foldable set, Monoid (set a), Vector set a, Foldable basis, Ord t, Eq a, SplitGen gen)

=> gen
random generator

-> LPType t basis set a
an LP-type problem

-> set a

-> (t, basis a)

Solves the LP Type problem. In particular, a minimization problem.

returns the cost of an optimal solution (if it exists), and the subset of elements realizing it

expected: O(n) time.

clarkson2

Arguments

:: (Foldable set, Semigroup (set a), Vector set a, Foldable basis, Ord t, Eq a, SplitGen gen)

=> gen
random generator

-> LPType t basis set a
an LP-type problem

-> set a

-> basis a

The clarkson2 algorithm.

subExp

Arguments

:: (Foldable set, Foldable basis, Ord t, RandomGen gen, Eq a)

=> gen
random generator

-> LPType t basis set a
an LP-type problem

-> set a

-> basis a

SubExponential time algorithm to solve the LP-type problem

input: \(n\) constraints, combinatorial dimension \(d\)

running time: \(O(d^2 + nd) * e^{O(\sqrt{d\ln n})}\)