HGeometry.LPType.LinearProgramming

Description

Algorithm for solving Low dimensional linear programs.

Documentation

lp1D :: (Ord r, Foldable1 set, HalfSpace_ halfSpace 1 r, Point_ (BoundingHyperPlane halfSpace 1 r) 1 r) => set halfSpace -> Basis1DLP halfSpace

Linear programming in 1D, minimizes the coordinate values.

\(O(n)\).

data Basis1DLP halfSpace

A solution for 1D Linear pgoramming

Constructors

InFeasible1 halfSpace halfSpace

Feasible1 halfSpace

Unbounded

linearProgrammingMinY :: forall r (set :: Type -> Type). (Fractional r, Ord r, Foldable set) => LPType (UnBounded r) (Basis2D r) set (HalfPlane r)

minimize the y-coordinate. (Lexicographically) pre: LP is feasible

data Basis2D r halfSpace

A Basis for 2D Linear programming problems.

Constructors

EmptyBasis
The solution must be unbounded

Basis1 r halfSpace
apparently we have a halfspace perpendicular to the optimization direction. So it just has a fixed value r.

Basis2 (Point 2 r) halfSpace halfSpace
The normal case in which a feasible solution (optimum)

Infeasible (Vector 3 halfSpace)
Infeasible

type HalfPlane r = HalfSpaceF (LineEQ r)