HGeometry.Line.BatchPointLocation

Description

Implementation of Batched point location among a set of lines.

Documentation

data PointLocationDS v e f

The Point Location Data structure

subdivision :: forall v e f1 f2. Functor f2 => (Subdiv v e f1 -> f2 (Subdiv v e f1)) -> PointLocationDS v e f1 -> f2 (PointLocationDS v e f1)

vrStructure :: forall v e f1 f2. Functor f2 => (VerticalRayShootingStructure (ClosedLineSegment v :+ DartIx (Subdiv v e f1)) -> f2 (VerticalRayShootingStructure (ClosedLineSegment v :+ DartIx (Subdiv v e f1)))) -> PointLocationDS v e f1 -> f2 (PointLocationDS v e f1)

outerFaceIx :: forall v e f1 f2. Functor f2 => (FaceIx (Subdiv v e f1) -> f2 (FaceIx (Subdiv v e f1))) -> PointLocationDS v e f1 -> f2 (PointLocationDS v e f1)

type PointLocationDS' r line = PointLocationDS (Point 2 r) (ClosedLineSegment (Point 2 r) :+ Maybe line) ()

buildPointLocationStructure :: (IsBoxable geom, NumType geom ~ r, Dimension geom ~ 2, Ord r, Fractional r, Line_ line 2 r, Foldable set, Eq line, IsIntersectableWith line (Rectangle (Point 2 r)), Intersection line (Rectangle (Point 2 r)) ~ Maybe (LineBoxIntersection 2 r)) => geom -> set line -> PointLocationDS' r line

Construct the point location structure (restricted to the) bounding box of the given object

pointLocationStructureIn

Arguments

:: (Foldable set, Line_ line 2 r, Fractional r, Ord r, Eq line, IsIntersectableWith line (Rectangle (Point 2 r)), Intersection line (Rectangle (Point 2 r)) ~ Maybe (LineBoxIntersection 2 r))

=> Rectangle (Point 2 r)
bounding rectangle

-> set line

-> PointLocationDS' r line

Construct a point location data structure on the given set of n lines in a given bounding rectangle.

Note that we can answer point location queries for points only within the given bounding rectangle.

O(n^2 log n)

groupQueries :: (Point_ queryPoint 2 r, Line_ line 2 r, Foldable set, Ord r, Fractional r, Eq line, IsBoxable queryPoint, IsIntersectableWith line (Rectangle (Point 2 r)), Intersection line (Rectangle (Point 2 r)) ~ Maybe (LineBoxIntersection 2 r)) => NonEmpty queryPoint -> set line -> MonoidalNEMap (FaceIx (PointLocationDS' r line)) (NonEmpty queryPoint)

Given a set of \(n\) query points, and a set of \(r\) lines H computes for each query q the subset of lines above q, in increasing order.

running time: \(O(n\log n + r^5\log r)\)