HGeometry.Line.Class

Description

A class of types that can act as lines in d-dimensional space.

Documentation

class Line_ line d r | line -> d, line -> r where

Types that can act as lines in d-dimensional space.

Methods

fromPointAndVec :: (Point_ point d r, Line_ line d r, Num r) => point -> Vector d r -> line

Constructs a line through a point and a vector

type Line2_ line r = (Line_ line 2 r, HyperPlane_ line 2 r)

Synonym for a line-like class.

lineThrough :: forall line point d r. (Line_ line d r, Point_ point d r, Num r) => point -> point -> line

Construct a line through two points.

>>> lineThrough origin (Point2 4 5) :: LinePV 2 Int
LinePV (Point2 0 0) (Vector2 4 5)

verticalLine :: forall r line. (Line_ line 2 r, Num r) => r -> line

Vertical line with a given X-coordinate.

horizontalLine :: forall r line. (Line_ line 2 r, Num r) => r -> line

Horizontal line with a given Y-coordinate.

class HasOnLine line d | line -> d where

Methods

onLine :: (Point_ point d r, Num r, Eq r, r ~ NumType line, d ~ Dimension line) => point -> line -> Bool

Test if the point lies on the line

>>> Point2 0 0 `onLine` lineThrough @(LinePV 2 Double) origin (Point2 1 0)
True
>>> Point2 10 10 `onLine` lineThrough @(LinePV 2 Double) origin (Point2 2 2)
True
>>> Point2 10 5 `onLine` lineThrough @(LinePV 2 Double) origin (Point2 2 2)
False