HGeometry.Line.LowerEnvelope

Description

Compute the lower envelope of a set of lines. The particular algorithm used here dualizes the lines to points, computes the upper hull, and dualizes everything back.

Documentation

newtype LowerEnvelopeF f vertex line

The lower envelope of a set of lines

Constructors

LowerEnvelope (Alternating f vertex line)

type LowerEnvelope = LowerEnvelopeF Vector

A lower envelope, where the data structure is a vector.

_Alternating :: Iso (LowerEnvelopeF f vertex line) (LowerEnvelopeF f' vertex' line') (Alternating f vertex line) (Alternating f' vertex' line')

projection function that turns a lower envelope into an alternating "list" of lines and vertices.

lowerEnvelope :: forall g f line r. (NonVerticalHyperPlane_ line 2 r, Fractional r, Ord r, Foldable1 f, Functor f, IsIntersectableWith line line, Intersection line line ~ Maybe (LineLineIntersection line), HasFromFoldable g) => f line -> LowerEnvelopeF g (Point 2 r) line

Computes the lower envelope of a set of \(n\) lines.

pre: the input is a set (so no duplicates)

running time: \(O(n\log n)\)

lineAt :: (Ord r, Point_ vertex d r, 1 <= d, BinarySearch (f (vertex, line)), Elem (f (vertex, line)) ~ (vertex, line)) => r -> LowerEnvelopeF f vertex line -> line

Commputes the line directly above the query point x

lookupLEVertex :: (Ord r, Point_ vertex d r, 1 <= d, BinarySearch (f (vertex, line)), Elem (f (vertex, line)) ~ (vertex, line)) => r -> LowerEnvelopeF f vertex line -> Maybe (vertex, line)

returns the rightmost vertex v for which v_x < x.