HGeometry.Line.PointAndVector

Description

\(d\)-dimensional lines.

Documentation

data LinePV (d :: Nat) r

A line is given by an anchor point and a vector indicating the direction.

Constructors

LinePV

Fields

_anchorPoint :: !(Point d r)

_direction :: !(Vector d r)

anchorPoint :: forall (d :: Nat) r f. Functor f => (Point d r -> f (Point d r)) -> LinePV d r -> f (LinePV d r)

Lens to access the anchor point of the line

class HasDirection t where

Types that have a Direction field

Methods

direction :: forall (d :: Nat) r. (Dimension t ~ d, NumType t ~ r) => Lens' t (Vector d r)

Lens to access the direction field

isIdenticalTo :: forall (d :: Nat) r. Eq (Vector d r) => LinePV d r -> LinePV d r -> Bool

Test if two lines are identical, meaning; if they have exactly the same anchor point and directional vector.

class HasSupportingLine t where

Types for which we can compute a supporting line, i.e. a line that contains the thing of type t.

Methods

supportingLine :: t -> LinePV (Dimension t) (NumType t)

fromLinearFunction :: Num r => r -> r -> LinePV 2 r

Create a line from the linear function ax + b

toLinearFunction :: (Fractional r, Ord r) => LinePV 2 r -> VerticalOrLineEQ r

get values a,b s.t. the input line is described by y = ax + b. returns Nothing if the line is vertical

data SideTestUpDown

Result of a side test

Constructors

Below

On

Above

class OnSideUpDownTest t where

Class for types that support testing on which side (above, on, below) a particular point is.

Methods

onSideUpDown :: forall (d :: Nat) r point. (d ~ Dimension t, r ~ NumType t, Ord r, Num r, Point_ point d r) => point -> t -> SideTestUpDown

Given a point q and a hyperplane h, compute to which side of h q lies. For vertical hyperplanes the left side of the hyperplane is interpeted as below.

liesAbove :: (Ord r, Num r, Point_ point 2 r) => point -> LinePV 2 r -> Bool

Test if the query point q lies (strictly) above line l

liesBelow :: (Ord r, Num r, Point_ point 2 r) => point -> LinePV 2 r -> Bool

Test if the query point q lies (strictly) above line l

data SideTest

Result of a side test

Constructors

LeftSide

OnLine

RightSide

onSide :: (Ord r, Num r, Point_ point 2 r) => point -> LinePV 2 r -> SideTest

Given a point q and a line l, compute to which side of l q lies. For vertical lines the left side of the line is interpeted as below.

>>> Point2 10 10 `onSide` (lineThrough origin $ Point2 10 5)
LeftSide
>>> Point2 10 10 `onSide` (lineThrough origin $ Point2 (-10) 5)
RightSide
>>> Point2 5 5 `onSide` (verticalLine 10)
LeftSide
>>> Point2 5 5 `onSide` (lineThrough origin $ Point2 (-3) (-3))
OnLine

leftHalfPlane :: (Num r, Ord r) => LinePV 2 r -> HalfSpaceF (LinePV 2 r)

Given the oriented line, computes the halfspace left of the line.

rightHalfPlane :: (Num r, Ord r) => LinePV 2 r -> HalfSpaceF (LinePV 2 r)

Given the oriented line, computes the halfspace right of the line.

bisector :: (Fractional r, Point_ point 2 r) => point -> point -> LinePV 2 r

Get the bisector between two points

perpendicularTo :: Num r => LinePV 2 r -> LinePV 2 r

Given a line l with anchor point p and vector v, get the line m perpendicular to l that also goes through p. The line is oriented *into* the right halfplane of l.

>>> perpendicularTo $ LinePV (Point2 3 4) (Vector2 (-1) 2)
LinePV (Point2 3 4) (Vector2 2 1)
>>> perpendicularTo $ LinePV (Point2 (-1000) 268) (Vector2 (-50) 8)
LinePV (Point2 (-1000) 268) (Vector2 8 50)

isPerpendicularTo :: (Num r, Eq r) => Vector 2 r -> LinePV 2 r -> Bool

Test if a vector is perpendicular to the line.

cmpSlope :: (Num r, Ord r) => LinePV 2 r -> LinePV 2 r -> Ordering

Compares the lines on slope. Vertical lines are considered larger than anything else.

>>> (LinePV origin (Vector2 5 1)) `cmpSlope` (LinePV origin (Vector2 3 3))
LT
>>> (LinePV origin (Vector2 5 1)) `cmpSlope` (LinePV origin (Vector2 (-3) 3))
GT
>>> (LinePV origin (Vector2 5 1)) `cmpSlope` (LinePV origin (Vector2 0 1))
LT