HGeometry.LineSegment.Class

Description

A class representing line segments

Documentation

class (IntervalLike_ lineSegment point, Point_ point (Dimension lineSegment) (NumType lineSegment)) => LineSegment_ lineSegment point | lineSegment -> point

A class representing line segments

pattern LineSegment_ :: forall lineSegment point. LineSegment_ lineSegment point => point -> point -> lineSegment

Deconstructs a line segment from the start and end point

class LineSegment_ lineSegment point => ConstructableLineSegment_ lineSegment point where

A class representing line segments

Minimal complete definition

uncheckedLineSegment

Methods

uncheckedLineSegment :: point -> point -> lineSegment

Create a segment

pre: the points are disjoint

mkLineSegment :: Eq (Vector (Dimension point) (NumType point)) => point -> point -> Maybe lineSegment

smart constructor that creates a valid segment, i.e. it validates that the endpoints are disjoint.

class (LineSegment_ lineSegment point, StartPointOf lineSegment ~ EndPoint Closed point, EndPointOf lineSegment ~ EndPoint Closed point) => ClosedLineSegment_ lineSegment point

A class representing Closed Linesegments

class (LineSegment_ lineSegment point, StartPointOf lineSegment ~ EndPoint Open point, EndPointOf lineSegment ~ EndPoint Open point) => OpenLineSegment_ lineSegment point

A Class representing Open ended linesegments

interpolate :: forall lineSegment d point r. (Fractional r, LineSegment_ lineSegment point, ConstructablePoint_ point d r) => r -> lineSegment -> point

Linearly interpolate the two endpoints with a value in the range [0,1]

>>> interpolate 0.5 $ ClosedLineSegment origin (Point2 10.0 10.0)
Point2 5.0 5.0
>>> interpolate 0.1 $ ClosedLineSegment origin (Point2 10.0 10.0)
Point2 1.0 1.0
>>> interpolate 0 $ ClosedLineSegment origin (Point2 10.0 10.0)
Point2 0.0 0.0
>>> interpolate 1 $ ClosedLineSegment origin (Point2 10.0 10.0)
Point2 10.0 10.0

class HasOnSegment lineSegment d | lineSegment -> d where

Class that expresses that we can test if the segment contains a point

Methods

onSegment :: (Ord r, Point_ point d r, r ~ NumType lineSegment, d ~ Dimension lineSegment) => point -> lineSegment -> Bool

Test if a point lies on a line segment.

As a user, you should typically just use intersects instead.

class HasStart seg p | seg -> p where

Things that have a start point

Methods

start :: Lens' seg p

Lens to access the start point

class HasEnd seg p | seg -> p where

Things that have an end point

Methods

end :: Lens' seg p

Lens to access the ending point

class HasStartPoint seg p | seg -> p where

Things that have a start point

Methods

startPoint :: Lens' seg p

Lens to access the start point

class HasEndPoint seg p | seg -> p where

Things that have an end point

Methods

endPoint :: Lens' seg p

Lens to access the ending point

type family StartPointOf interval

type family to declare the type of startpoint for an interval, the idea is to define this as one of the endpoinst form the Endpoints module

type family EndPointOf interval

type family to declare the type of endpoint for an interval, the idea is to define this as one of the endpoinst form the Endpoints module

ordAtY :: (Num r, Ord r, LineSegment_ lineSegment point, Point_ point 2 r) => r -> lineSegment -> lineSegment -> Ordering

Given a y-coordinate, compare the segments based on the x-coordinate of the intersection with the horizontal line through y

ordAtX :: (Num r, Ord r, LineSegment_ lineSegment point, Point_ point 2 r) => r -> lineSegment -> lineSegment -> Ordering

Given an x-coordinate, compare the segments based on the y-coordinate of the intersection with the horizontal line through y

xCoordAt :: (Fractional r, Ord r, LineSegment_ lineSegment point, Point_ point 2 r) => r -> lineSegment -> r

Given a y coord and a line segment that intersects the horizontal line through y, compute the x-coordinate of this intersection point.

note that we will pretend that the line segment is closed, even if it is not

yCoordAt :: (Fractional r, Ord r, LineSegment_ lineSegment point, Point_ point 2 r) => r -> lineSegment -> r

Given an x-coordinate and a line segment that intersects the vertical line through x, compute the y-coordinate of this intersection point.

note that we will pretend that the line segment is closed, even if it is not

orientLR :: (LineSegment_ lineSegment point, Point_ point d r, 1 <= d, Ord r) => lineSegment -> lineSegment

Orient the segment from left to right

orientBT :: (LineSegment_ lineSegment point, Point_ point d r, 2 <= d, Ord r) => lineSegment -> lineSegment

Orient the segment from bottom to top