HGeometry.LineSegment.Intersection

Documentation

data LineLineSegmentIntersection lineSegment

Line x LineSegment intersection

Constructors

Line_x_LineSegment_Point (Point 2 (NumType lineSegment))

Line_x_LineSegment_LineSegment lineSegment

data LineSegmentLineSegmentIntersection lineSegment

LineSegment x LineSegment intersection

Constructors

LineSegment_x_LineSegment_Point (Point 2 (NumType lineSegment))

LineSegment_x_LineSegment_LineSegment lineSegment

data HalfLineLineSegmentIntersection point segment

Data type representing the intersection of a HalfLine and a LineSegment

Constructors

HalfLine_x_LineSegment_Point point

HalfLine_x_LineSegment_LineSegment segment

compareColinearInterval :: (Ord r, Num r, Point_ point 2 r, IxValue (endPoint point) ~ point, EndPoint_ (endPoint point)) => LinePV 2 r -> LineSegment endPoint point -> CompareInterval

Given a line l, and a line segment seg that lies on l. Returns where the anchorPoint of the line is with respect to the line segment (which we can interpret as some interval along l).

Orphan instances

(Point_ point 2 r, Num r, Ord r) => HasIntersectionWith (ClosedLineSegment point) (OpenLineSegment point)

Instance details

Methods

intersects :: ClosedLineSegment point -> OpenLineSegment point -> Bool

(Point_ point 2 r, Num r, Ord r) => HasIntersectionWith (OpenLineSegment point) (ClosedLineSegment point)

Instance details

Methods

intersects :: OpenLineSegment point -> ClosedLineSegment point -> Bool

(Point_ point 2 r, Fractional r, Ord r) => IsIntersectableWith (LineEQ r) (ClosedLineSegment point)

Instance details

Methods

intersect :: LineEQ r -> ClosedLineSegment point -> Intersection (LineEQ r) (ClosedLineSegment point)

(Point_ point 2 r, Fractional r, Ord r) => IsIntersectableWith (LineEQ r) (OpenLineSegment point)

Instance details

Methods

intersect :: LineEQ r -> OpenLineSegment point -> Intersection (LineEQ r) (OpenLineSegment point)

(Ord r, Num r, Point_ point 2 r, IxValue (endPoint point) ~ point, EndPoint_ (endPoint point)) => HasIntersectionWith (HalfLine point) (LineSegment endPoint point)

Instance details

Methods

intersects :: HalfLine point -> LineSegment endPoint point -> Bool

(Point_ point 2 r, Num r, Ord r, IxValue (endPoint point) ~ point, EndPoint_ (endPoint point)) => HasIntersectionWith (LineEQ r) (LineSegment endPoint point)

Instance details

Methods

intersects :: LineEQ r -> LineSegment endPoint point -> Bool

(Point_ point 2 r, Num r, Ord r) => HasIntersectionWith (ClosedLineSegment point) (LineSegment AnEndPoint point)

Instance details

Methods

intersects :: ClosedLineSegment point -> LineSegment AnEndPoint point -> Bool

(Point_ point 2 r, Num r, Ord r) => HasIntersectionWith (OpenLineSegment point) (LineSegment AnEndPoint point)

Instance details

Methods

intersects :: OpenLineSegment point -> LineSegment AnEndPoint point -> Bool

(Ord r, Fractional r, Point_ point 2 r, IxValue (endPoint point) ~ point, EndPoint_ (endPoint point), IsIntersectableWith (LinePV 2 r) (LineSegment endPoint point), Intersection (LinePV 2 r) (LineSegment endPoint point) ~ Maybe (LineLineSegmentIntersection lineSegment'), NumType lineSegment' ~ r) => IsIntersectableWith (HalfLine point) (LineSegment endPoint point)

Instance details

Methods

intersect :: HalfLine point -> LineSegment endPoint point -> Intersection (HalfLine point) (LineSegment endPoint point)

(Point_ point 2 r, Fractional r, Ord r) => IsIntersectableWith (LineEQ r) (LineSegment AnEndPoint point)

Instance details

Methods

intersect :: LineEQ r -> LineSegment AnEndPoint point -> Intersection (LineEQ r) (LineSegment AnEndPoint point)

(Point_ point 2 r, Num r, Ord r) => HasIntersectionWith (LineSegment AnEndPoint point) (ClosedLineSegment point)

Instance details

Methods

intersects :: LineSegment AnEndPoint point -> ClosedLineSegment point -> Bool

(Point_ point 2 r, Num r, Ord r) => HasIntersectionWith (LineSegment AnEndPoint point) (OpenLineSegment point)

Instance details

Methods

intersects :: LineSegment AnEndPoint point -> OpenLineSegment point -> Bool

(Point_ point 2 r, Fractional r, Ord r) => IsIntersectableWith (LinePV 2 r) (ClosedLineSegment point)

Instance details

Methods

intersect :: LinePV 2 r -> ClosedLineSegment point -> Intersection (LinePV 2 r) (ClosedLineSegment point)

(Point_ point 2 r, Fractional r, Ord r) => IsIntersectableWith (LinePV 2 r) (OpenLineSegment point)

Instance details

Methods

intersect :: LinePV 2 r -> OpenLineSegment point -> Intersection (LinePV 2 r) (OpenLineSegment point)

(Point_ point 2 r, Num r, Ord r, IxValue (endPoint point) ~ point, EndPoint_ (endPoint point)) => HasIntersectionWith (LinePV 2 r) (LineSegment endPoint point)

Instance details

Methods

intersects :: LinePV 2 r -> LineSegment endPoint point -> Bool

(Point_ point 2 r, Num r, Ord r, Functor endPoint, IxValue (endPoint point) ~ point, EndPoint_ (endPoint point), IxValue (endPoint r) ~ r, EndPoint_ (endPoint r), HasIntersectionWith (Interval endPoint r) (Interval endPoint r)) => HasIntersectionWith (LineSegment endPoint point) (LineSegment endPoint point)

Instance details

Methods

intersects :: LineSegment endPoint point -> LineSegment endPoint point -> Bool

(Point_ point 2 r, Fractional r, Ord r) => IsIntersectableWith (LinePV 2 r) (LineSegment AnEndPoint point)

Instance details

Methods

intersect :: LinePV 2 r -> LineSegment AnEndPoint point -> Intersection (LinePV 2 r) (LineSegment AnEndPoint point)

(Point_ point 2 r, Num r, Ord r, Functor endPoint, IxValue (endPoint point) ~ point, EndPoint_ (endPoint point), IxValue (endPoint r) ~ r, EndPoint_ (endPoint r), HasIntersectionWith (Interval endPoint r) (Interval endPoint r), IxValue (endPoint point) ~ point, IxValue (endPoint (r :+ endPoint point)) ~ (r :+ endPoint point), EndPoint_ (endPoint point), IsEndPoint (endPoint point) (endPoint (r :+ endPoint point)), IsIntersectableWith (LinePV 2 r) (LineSegment endPoint point), Intersection (LinePV 2 r) (LineSegment endPoint point) ~ Maybe (LineLineSegmentIntersection (LineSegment endPoint point)), HasOnSegment (LineSegment endPoint point) 2, IsIntersectableWith (Interval endPoint (r :+ endPoint point)) (Interval endPoint (r :+ endPoint point)), Intersection (LineSegment endPoint point) (LineSegment endPoint point) ~ Maybe (LineSegmentLineSegmentIntersection (LineSegment endPoint point)), EndPoint_ (endPoint (r :+ endPoint point))) => IsIntersectableWith (LineSegment endPoint point) (LineSegment endPoint point)

Instance details

Methods

intersect :: LineSegment endPoint point -> LineSegment endPoint point -> Intersection (LineSegment endPoint point) (LineSegment endPoint point)

(IsIntersectableWith (LineSegment endPoint point) (LineSegment endPoint point), Intersection (LineSegment endPoint point) (LineSegment endPoint point) ~ Maybe (LineSegmentLineSegmentIntersection (LineSegment endPoint point))) => IsIntersectableWith (LineSegment endPoint point :+ extra) (LineSegment endPoint point :+ extra)

Instance details

Methods

intersect :: (LineSegment endPoint point :+ extra) -> (LineSegment endPoint point :+ extra) -> Intersection (LineSegment endPoint point :+ extra) (LineSegment endPoint point :+ extra)