HGeometry.LowerEnvelope

Description

Computing the lower envelope of a set of planes

Documentation

module HGeometry.LowerEnvelope.Naive

type VertexID = Int

VertexIds are just ints

data BoundedVertexF f plane

Data type representing the bounded vertices of the lower envelope.

In the vertexform, the f will be Const (), i.e. we don't have adjacency information. But in case of Adjacencylist form we can implement it using a Sequence.

Constructors

Vertex !(Point 3 (NumType plane)) (Set plane) (f (LEEdge plane))

location :: NumType plane ~ r => Lens' (BoundedVertexF f plane) (Point 3 r)

The location of the vertex

definers :: Lens' (BoundedVertexF f plane) (Set plane)

The three planes defining the vertex

location2 :: NumType plane ~ r => Getter (BoundedVertexF f plane) (Point 2 r)

Projected 2d location of the point

incidentEdgesB :: Lens (BoundedVertexF f plane) (BoundedVertexF g plane) (f (LEEdge plane)) (g (LEEdge plane))

Lens to access the incident edges of a vertex

traverseBoundedV :: (Traversable f, Applicative g, NumType plane ~ NumType plane', Ord plane') => (plane -> g plane') -> BoundedVertexF f plane -> g (BoundedVertexF f plane')

Traverse the bounded vertex.

\(O(k\log k)\), where \(k\) is the number of definers of v.

data LEEdge plane

An (half)Edge in the Lower envelope

Constructors

Edge !VertexID plane plane

destination :: Getter (LEEdge plane) VertexID

Getter to access the destination field of an edge.

leftPlane :: Lens' (LEEdge plane) plane

Lens to access the plane left of/above the edge

rightPlane :: Lens' (LEEdge plane) plane

Lens to access the plane right of/below the edge

flipEdge :: VertexID -> LEEdge plane -> LEEdge plane

Given some vertex u and an edge e from u towards some other vertex v, flip the edge e, s othat it is the edge from v to u.