HGeometry.LowerEnvelope.VertexForm

Description

A Representation of the Lower envelope of planes in vertex form; i.e. storing only the vertices.

Documentation

newtype VertexForm plane

The lower envelope in vertex form

Constructors

VertexForm (Map (Point 3 (NumType plane)) (Definers plane))

hasVertices :: VertexForm plane -> Bool

Test if there are vertices

vertices' :: IndexedTraversal' (VertexIx (VertexForm plane)) (VertexForm plane) (Vertex (VertexForm plane))

Traversal (rather than a traversal1) of the vertices.

singleton :: LEVertex plane -> VertexForm plane

Computes a lower envelope consisting of a single vertex.

type LEVertex = BoundedVertexF (Const ())

Vertices of a lower envelope in vertex form.

pattern LEVertex :: Point 3 (NumType plane) -> Definers plane -> LEVertex plane

Convenience Constructor for computing vertices of the lower envelope that are in Vertex form.

type Definers plane = Set plane

The definers of a vertex.

data BoundedVertexF f plane

Data type representing the bounded vertices of the lower envelope.

In the vertexform, the f will be Const (), i.e. we don't have adjacency information. But in case of Adjacencylist form we can implement it using a Sequence.

Constructors

Vertex !(Point 3 (NumType plane)) (Set plane) (f (LEEdge plane))

location :: NumType plane ~ r => Lens' (BoundedVertexF f plane) (Point 3 r)

The location of the vertex

definers :: Lens' (BoundedVertexF f plane) (Set plane)

The three planes defining the vertex

location2 :: NumType plane ~ r => Getter (BoundedVertexF f plane) (Point 2 r)

Projected 2d location of the point

data IntersectionLine r

the intersection line may be vertical or non-vertical

Constructors

Vertical r

NonVertical (LineEQ r)

intersectionLine :: (Plane_ plane r, Fractional r, Eq r) => plane -> plane -> Maybe (IntersectionLine r)

Given two planes, computes the line in which they intersect.

intersectionPoint :: (Plane_ plane r, Ord r, Fractional r) => Three plane -> Maybe (Point 3 r)

Computes there the three planes intersect