HGeometry.Map.NonEmpty.Monoidal

Description

A NonEmpty Monoidal Map, i.e. combining the monoidal maps combines the elements with their semigroup operation.

Documentation

newtype MonoidalNEMap k v

A NonEmpty Map in which we combine conflicting elements by using their semigroup operation rather than picking the left value (as is done in the default Data.Map)

Constructors

MonoidalNEMap

Fields

getNEMap :: NEMap k v

singleton :: k -> v -> MonoidalNEMap k v

Create a singleton MonoidalNE Map

fromSet :: (k -> v) -> NESet k -> MonoidalNEMap k v

Creates a non-empty Map from a nonempty Set \(O(n)\)

unions1WithKey :: (Foldable1 f, Ord k) => (k -> a -> a -> a) -> f (NEMap k a) -> NEMap k a

Merge a bunch of non-empty maps with the given mergin function

mapWithKeyMerge1 :: (Ord k', Semigroup v') => (k -> v -> NEMap k' v') -> NEMap k v -> NEMap k' v'

Merge the maps. When they share a key, combine their values using a semigroup.

mapAccumWithKey :: forall k a b c. (a -> k -> b -> (a, c)) -> a -> MonoidalNEMap k b -> (a, MonoidalNEMap k c)

Map with some accumulator function

insert :: (Ord k, Semigroup v) => k -> v -> MonoidalNEMap k v -> MonoidalNEMap k v

Insert an new key,value pair into the Map

\(O(\log n)\)

insertReplace :: Ord k => k -> v -> MonoidalNEMap k v -> MonoidalNEMap k v

Insert an new key,value pair into the Map \(O(\log n)\)

assocs :: MonoidalNEMap k v -> NonEmpty (k, v)

Return all key/value pairs in the map in ascending key order.

\(O(n)\)

toAscList :: MonoidalNEMap k v -> NonEmpty (k, v)

Return all key/value pairs in the map in ascending key order.

\(O(n)\)

keysSet :: MonoidalNEMap k v -> NESet k

Return all keys in ascending order

\(O(n)\)

(!?) :: Ord k => MonoidalNEMap k v -> k -> Maybe v infixl 9

Lookup in the Map \(O(\log n)\)

(!) :: Ord k => MonoidalNEMap k v -> k -> v infixl 9

unsafe lookup in the Map \(O(\log n)\)