HGeometry.Matrix.Class

Description

A class of types representing matrices

Documentation

class (r ~ NumType matrix, Ixed matrix, IxValue matrix ~ r, Index matrix ~ (Int, Int), HasElements matrix matrix) => Matrix_ matrix (n :: Nat) (m :: Nat) r | matrix -> n, matrix -> m, matrix -> r where

A matrix of n rows, each of m columns, storing values of type r.

Minimal complete definition

generateMatrix, matrixFromRows, rows

Methods

identityMatrix :: matrix

Produces the Identity Matrix.

>>> printMatrix $ identityMatrix @(Matrix 3 3 Int)
Vector3 1 0 0
Vector3 0 1 0
Vector3 0 0 1

generateMatrix :: ((Int, Int) -> r) -> matrix

Given a function that specifies the values, generate the matrix

matrixFromList :: [r] -> Maybe matrix

Given a list of the elements in the matrix, in row by row order, constructs the matrix. requires that there are exactly n*m elements.

>>> matrixFromList @(Matrix 2 3 Int) [1,2,3,4,5,6]
Just (Matrix (Vector2 (Vector3 1 2 3) (Vector3 4 5 6)))
>>> matrixFromList @(Matrix 2 3 Int) [1,2,3,4,5,6,7]
Nothing

matrixFromRows :: Vector_ rowVector n (Vector m r) => rowVector -> matrix

Given a list of the elements in the matrix, in row by row order, constructs the matrix. requires that there are exactly n*m elements.

>>> matrixFromRows @(Matrix 2 3 Int) (Vector2 (Vector3 1 2 3) (Vector3 4 5 6))
Matrix (Vector2 (Vector3 1 2 3) (Vector3 4 5 6))

(!*!) :: forall matrix' (m' :: Nat) matrix''. (Matrix_ matrix' m m' r, Matrix_ matrix'' n m' r, Num r, Has_ Additive_ m r) => matrix -> matrix' -> matrix'' infixl 7

Matrix multiplication

>>> let m1  = matrixFromRows @(Matrix 2 3 Int) (Vector2 (Vector3 1 2 3) (Vector3 4 5 6))
>>> let r i = Vector4 i (i*10) (i*100) (i*1000)
>>> let m2  = matrixFromRows @(Matrix 3 4 Int) (Vector3 (r 1) (r 2) (r 3))
>>> printMatrix $ (m1 !*! m2 :: Matrix 2 4 Int)
Vector4 14 140 1400 14000
Vector4 32 320 3200 32000

(!*) :: matrix -> Vector m r -> Vector n r infixl 7

Multiply a matrix and a vector.

>>> let m = matrixFromRows @(Matrix 2 3 Int) (Vector2 (Vector3 1 2 3) (Vector3 4 5 6))
>>> m !* Vector3 2 3 1
Vector2 11 29

(*!!) :: r -> matrix -> matrix infixl 7

Multiply a scalar and a matrix

(!!*) :: matrix -> r -> matrix infixl 7

Multiply a matrix and a scalar

rows :: Lens' matrix (Vector n (Vector m r))

Lens to access all rows

row :: Int -> matrix -> Maybe (Vector m r)

Access the i^th row in the matrix.

column :: Int -> matrix -> Maybe (Vector n r)

Access the j^th column in the matrix.

class HasElements matrix matrix' where

Types that have an elements field lens.

Methods

elements :: IndexedTraversal1 (Int, Int) matrix matrix' (NumType matrix) (NumType matrix')

IndexedTraversal over the elements of the matrix, each index is a (row,column) index pair.

class HasDeterminant (d :: Nat) where

Dimensions for which we can compute the determinant of a matrix

Methods

det :: (Num r, Matrix_ matrix d d r) => matrix -> r

class Invertible (n :: Nat) where

Class of matrices that are invertible.

Methods

inverseMatrix :: (Fractional r, Matrix_ matrix n n r, Has_ Vector_ n r) => matrix -> matrix

given an invertable square \(n \times n\) matrix A, computes the \(n \times n\) matrix B such that A !*! B = identityMatrix