HGeometry.Matrix.Internal

Helper functions to easily create matrices

mkRow :: forall d r. (Num r, Has_ Vector_ d r) => Int -> r -> Vector d r

Creates a row with zeroes everywhere, except at position i, where the value is the supplied value.