HGeometry.Measured.Class

Description

Documentation

class Measured (f :: Type -> Type) a where

Things that can be measured.

Methods

measure :: a -> f a

Given a single a, measure it into someting of type 'f a'.

class Measured f a => CanInsert (f :: Type -> Type) a where

Things that can be inserted.

Minimal complete definition

Nothing

Methods

insertMeasure :: a -> f a -> f a

Delete an element form the associated structure

default insertMeasure :: Semigroup (f a) => a -> f a -> f a

class Measured f a => CanDelete (f :: Type -> Type) a where

Things that can be deleted.

Methods

deleteMeasure :: a -> f a -> Maybe (f a)

Delete the element from the structure. Returns a Maybe since the structure may be empty after deletion (and that may not be representable by f).