HGeometry.Miso.Subscription.MouseExtra

Description

Extra mouse event events/decoders

Documentation

onMouseEnterAt :: (Point 2 Int -> action) -> Attribute action

onMouseEnter event, the position is relative to the target of the event

onMouseMoveAt :: (Point 2 Int -> action) -> Attribute action

onMouseMove event, the position is relative to the target of the event

onMouseClickAt :: (Point 2 Int -> action) -> Attribute action

onMouseEnter event, the position is relative to the target of the event

onTouchStartAt :: (Point 2 Int -> action) -> Attribute action

On start of a touch event,

onTouchMoveAt :: (Point 2 Int -> action) -> Attribute action

On touchMove event

onTouchEnd :: action -> Attribute action

onTouchEnd event