HGeometry.Miso.Svg

Description

Render geometric objects to Svg files through miso

Documentation

renderSvgToFile :: OsPath -> View action -> IO ()

Given an file path, and a view whose root is an svg element, render the output to the given file.

renderAsSvgText :: View action -> Text

Given an View whose root is an svg element, renders the view to a lazy Text

renderAsSvgByteString :: View action -> ByteString

Given an View whose root is an svg element, renders the view to a lazy ByteString.

withAts :: ([Attribute action] -> [View action] -> View action) -> [Attribute action] -> [Attribute action] -> View action

Helper function to construct drawing functions. I..e it allows you do pre-specify a bunch of attributes that should be drawn (ats1) yet allow more attributes to be added by the user later.

withAts' :: ([Attribute action] -> [View action] -> View action) -> [Attribute action] -> [Attribute action] -> [View action] -> View action

Helper function to construct a View. See withAts for its usage.

class Drawable t where

Default implementation for drawing geometric objects

Minimal complete definition

draw | drawWith

Methods

draw :: t -> [Attribute action] -> View action

Draws the given object with the given attributes

drawWith :: t -> [Attribute action] -> [View action] -> View action

draw the given object, as well as the given "children"

dPoint :: (Point_ point 2 r, ToMisoString r) => point -> [Attribute action] -> View action

Draw a point

dLineSegment :: (LineSegment_ lineSegment point, Point_ point 2 r, ToMisoString r) => lineSegment -> [Attribute action] -> View action

Draw a line segment

dRectangle :: (Rectangle_ rectangle point, Point_ point 2 r, ToMisoString r, Num r) => rectangle -> [Attribute action] -> View action

Draw a rectangle

dCircle :: (Point_ point 2 r, ToMisoString r) => Circle point -> [Attribute action] -> View action

Draw a circle

dDisk :: (Disk_ disk point, Point_ point 2 r, ToMisoString r, Floating r) => disk -> [Attribute action] -> View action

Draw a disk

dPolyLine :: (PolyLine_ polyLine point, Point_ point 2 r, ToMisoString r) => polyLine -> [Attribute action] -> View action

Draw a polyline

dSimplePolygon :: (SimplePolygon_ simplePolygon point r, ToMisoString r) => simplePolygon -> [Attribute action] -> View action

Draw a simple polygon