HGeometry.Miso.Svg.Canvas

Documentation

data Canvas r

Svg Canvas that has a "proper" Coordinate system whose origin is in the bottom left.

theViewport :: forall r s f. Functor f => (Viewport r -> f (Viewport s)) -> Canvas r -> f (Canvas s)

Lens to access the viewport

blankCanvas :: Num r => Int -> Int -> Canvas r

Crate a blank canvas, that has the origin in the bottom-left.

class HasDimensions s a | s -> a where

Class for types that have a dimensions field

Methods

dimensions :: Lens' s a

Lens to access the Dimensions field

class HasMousePosition s a | s -> a where

Class for types that have a mouse position

Methods

mousePosition :: Lens' s a

Lens to access the raw mouse position

mouseCoordinates :: Fractional r => Getter (Canvas r) (Maybe (Point 2 r))

Getter to access the mouse coordinates (in terms of the coordinate system as used by the canvas). Returns a Nothing if the mouse is not currently on/over the canvas.

data InternalCanvasAction

Actions that CanvasAction will handle itself.

handleInternalCanvasAction :: Canvas r -> InternalCanvasAction -> Effect (Canvas r) action

Handles InternalCanvas Actions

withCanvasEvents :: Map MisoString Bool -> Map MisoString Bool

Events a canvas wants to listen to

svgCanvas_ :: (RealFrac r, ToSvgCoordinate r) => Canvas r -> [Attribute action] -> [View action] -> View (Either InternalCanvasAction action)

Draws the actual canvas using an svg tag