HGeometry.Miso.Svg.StaticCanvas

Description

A Static canvas to be usedi nMiso applications

Documentation

data StaticCanvas r

Static Canvas that has a "proper" Coordinate system whose origin is in the bottom left.

staticCanvas :: Num r => Int -> Int -> StaticCanvas r

Create a canvas

class HasDimensions s a | s -> a where

Class for types that have a dimensions field

Methods

dimensions :: Lens' s a

Lens to access the Dimensions field

staticCanvas_ :: (RealFrac r, ToSvgCoordinate r) => StaticCanvas r -> [Attribute action] -> [View action] -> View action

Draws the actual canvas

type ToSvgCoordinate = ToMisoString

Types for which we can convert into a Svg coordinate.

matrixToMisoString :: ToSvgCoordinate r => Matrix 3 3 r -> MisoString

Renders a matrix as a MisoString