HGeometry.Number.Ratio.Generalized

Description

Generalized Ratio type that accepts arbitrary 'Num a' types rather than just Integral ones as in Data.Ratio

Documentation

data GRatio a

Generalized Ratio type that accepts more general "base" types than just Integral ones. That does mean we cannot normalize the intermediate expressions, so expect the numbers to become big quite quickly!

invariant: the denominator is not zero

(%) :: (Eq a, Num a) => a -> a -> GRatio a

smart constructor to construct a GRatio. Throws an exception if the denominator is zero.

numerator :: GRatio a -> a

Get the numerator

denominator :: GRatio a -> a

Get the denominator