HGeometry.Ord.Dynamic

Description

Documentation

newtype O s a

Values of type 'a' in our dynamically constructed Ord instance

Constructors

O

Fields

runO :: a

newtype OrdDict a

An Ord Dictionary

Constructors

OrdDict

Fields

compare_ :: a -> a -> Ordering

withOrd :: (a -> a -> Ordering) -> (forall s. Reifies s (OrdDict a) => O s b) -> b

Run a computation with a given ordering

Introducing and removing the dynamic order type

extractOrd1 :: f (O s a) -> O s (f a)

Lifts a container f whose values (of type a) depend on 's' into a more general computation in that produces a 'f a' (depending on s).

running time: \(O(1)\)

introOrd1 :: f a -> f (O s a)

Introduce dynamic order in a container 'f'.

running time: \(O(1)\)

liftOrd1 :: (f (O s a) -> g (O s a)) -> f a -> O s (g a)

Lifts a function that works on a container 'f' of orderable-things into one that works on dynamically ordered ones.

extractOrd2 :: forall {k1} f s k2 (v :: k1). f (O s k2) v -> O s (f k2 v)

Lifts a container f whose keys (of type k) depend on 's' into a more general computation in that produces a `f k v` (depending on s).

running time: \(O(1)\)

introOrd2 :: forall {k1} f k2 (v :: k1) s. f k2 v -> f (O s k2) v

Introduce dynamic order in a container 'f' that has keys of type k.

running time: \(O(1)\)