HGeometry.Permutation

Description

Data type for representing a non-empty Permutation

Documentation

data Permutation a

Cyclic representation of a non-empty permutation.

Constructors

Permutation (NonEmptyVector (Orbit a)) (Vector (Int, Int))

orbits :: forall a b f. Functor f => (NonEmptyVector (Orbit a) -> f (NonEmptyVector (Orbit b))) -> Permutation a -> f (Permutation b)

Lens to access the orbits of the permutation

indices :: (Indexable i p, Applicative f) => (i -> Bool) -> Optical' p (Indexed i) f a a

This allows you to filter an IndexedFold, IndexedGetter, IndexedTraversal or IndexedLens based on a predicate on the indices.

>>> ["hello","the","world","!!!"]^..traversed.indices even
["hello","world"]

>>> over (traversed.indices (>0)) Prelude.reverse $ ["He","was","stressed","o_O"]
["He","saw","desserts","O_o"]

type Orbit a = NonEmptyVector a

Orbits (Cycles) are represented by vectors

elems :: Permutation a -> NonEmptyVector a

Get the elements of the permutation

size :: Permutation a -> Int

Get the size of a permutation

running time: \(O(1)\)

cycleOf :: Enum a => Permutation a -> a -> Orbit a

The cycle containing a given item

next :: NonEmptyVector a -> Int -> a

Next item in a cyclic permutation

previous :: NonEmptyVector a -> Int -> a

Previous item in a cyclic permutation

lookupIdx :: Enum a => Permutation a -> a -> (Int, Int)

Lookup the indices of an element, i.e. in which orbit the item is, and the index within the orbit.

runnign time: \(O(1)\)

apply :: Enum a => Permutation a -> a -> a

Apply the permutation, i.e. consider the permutation as a function.

orbitFrom :: Eq a => a -> (a -> a) -> NonEmpty a

Find the cycle in the permutation starting at element s

cycleRep :: (Enum a, Eq a) => NonEmptyVector a -> (a -> a) -> Permutation a

Given a vector with items in the permutation, and a permutation (by its functional representation) construct the cyclic representation of the permutation.

toCycleRep :: Enum a => Int -> NonEmpty (NonEmpty a) -> Permutation a

Given the size n, and a list of Cycles, turns the cycles into a cyclic representation of the Permutation.