HGeometry.Plane.BatchPointLocation

Description

Implementation of Batched point location among a set of planes.

Documentation

batchedPointLocation :: (Point_ queryPoint 3 r, Plane_ plane r, Foldable set, Foldable1 set', Ord r, Fractional r, Ord queryPoint) => set' queryPoint -> set plane -> NEMap queryPoint (Vector plane)

Given a set of \(n\) queries, and a set of \(r\) planes H computes for each query q the subset of planes above q, in increasing order.

running time: \(O(n\log n + r^5\log r)\)

groupQueries :: (Point_ queryPoint 3 r, Plane_ plane r, Foldable set, Foldable1 nonEmpty, Ord r, Fractional r) => nonEmpty queryPoint -> set plane -> NonEmpty (NonEmpty queryPoint)

Given a set of \(n\) query points, and the set of \(r\) planes, partitions the queries into at most \(O(r^4)\) non-empty subsets Q_F so that

for every pair of query points q,q' in a set Q_F the vertical (w.r.t z) line through q intersects the planes in the input set in the same order as the vertical line through q'.

running time: \(O(n\log n + r^4 \log r)\)

answerBatch :: forall set set' plane queryPoint r. (Foldable set, Foldable1 set', Point_ queryPoint 3 r, Plane_ plane r, Ord r, Num r, Ord queryPoint) => set plane -> set' queryPoint -> NEMap queryPoint (Vector plane)

Ansers a batch of queries

pre: the vertical (w.r.t z) order of the planes is the same for all queries in the batch

running time: \(O(n\log n + r\log r + K)\), where \(K\) is the output size.