HGeometry.Plane.LowerEnvelope

Description

Computing the lower envelope of a set of planes

Documentation

module HGeometry.Plane.LowerEnvelope.Naive

pointLocateParallel :: (Plane_ plane r, Point_ point 2 r, Ord r, Num r) => point -> Alternating Vector (VerticalOrLineEQ r) plane -> plane

Point locates the given point among the parralel strips.

\(O(\log h)\), where \(h\) is the number of planes on the lower envelope

data LowerEnvelope plane vtxData

The lower enevelope of planes in R^3. (Or rather, its minimization diagram)

Constructors

ParallelStrips !(Alternating Vector (VerticalOrLineEQ (NumType plane)) plane)

ConnectedEnvelope !(MinimizationDiagram (NumType plane) (MDVertex (NumType plane) plane vtxData) plane)

module HGeometry.Plane.LowerEnvelope.Connected