HGeometry.Plane.LowerEnvelope.Connected

Description

A Representation of the Lower envelope of planes as a bunch of convex regions

Documentation

newtype MinimizationDiagram r vertex plane

A minimization daigram just maps every plane on the lower envelope to the region above which it is minimal. Every plane has at most one such a region.

Constructors

MinimizationDiagram (NEMap plane (Region r vertex))

asMap :: MinimizationDiagram r vertex plane -> NEMap plane (Region r vertex)

Get the underlying Map that relates every plane in the envelope to its projected region

mapVertices :: (vertex -> vertex') -> MinimizationDiagram r vertex plane -> MinimizationDiagram r vertex' plane

Apply some mapping function to the vertex data of each vertex

data Region r vertex

A region in the minimization diagram. The boundary is given in CCW order; i.e. the region is to the left of the boundary.

Constructors

BoundedRegion (ConvexPolygonF (Cyclic NonEmpty) vertex)

UnboundedRegion (UnboundedConvexRegionF r NonEmpty vertex)

data MDVertex r plane a

A vertex of the minimzation Diagram that is defined by the intersection of a number of planes.

Note that we interpet this vertex as a 2-dimensional thing.

Constructors

MDVertex (Point 3 r) (Definers plane) a

location :: forall r1 plane a r2 f. Functor f => (Point 3 r1 -> f (Point 3 r2)) -> MDVertex r1 plane a -> f (MDVertex r2 plane a)

vertexData :: forall r plane a1 a2 f. Functor f => (a1 -> f a2) -> MDVertex r plane a1 -> f (MDVertex r plane a2)

toConvexPolygonIn :: (Rectangle_ rectangle corner, Point_ corner 2 r, Point_ point 2 r, Ord r, Fractional r) => rectangle -> Region r point -> Either (ConvexPolygonF (Cyclic NonEmpty) point) (ConvexPolygonF (Cyclic NonEmpty) (OriginalOrExtra point (Point 2 r)))

Computes a convex polygon corresponding to the region.

pre: the bounding box (strictly) contains all vertices in its interior

intersectionLine :: (Plane_ plane r, Fractional r, Eq r) => plane -> plane -> Maybe (VerticalOrLineEQ r)

Given two planes, computes the line in which they intersect.

intersectionPoint :: (Plane_ plane r, Ord r, Fractional r) => Three plane -> Maybe (Point 3 r)

Computes there the three planes intersect

intersectionVector :: (Plane_ plane r, Ord r, Fractional r) => plane -> plane -> Maybe (Vector 2 r)

Computes the direction vector v of the directed line l in which the two planes h and h' intersect, and so that h will be to the left of the directed line

fromMinimizationDiagramIn :: (Plane_ plane r, Ord plane, Ord r, Fractional r, Point_ corner 2 r, Foldable1 set, Ord vtxData, Show r, Show corner, Show plane, Show vtxData) => Triangle corner -> set plane -> Maybe (MinimizationDiagram r (MDVertex r plane vtxData) plane) -> NEMap plane (ClippedMDCell r plane vtxData)

Compute a clipped minimization diagram within the given triangle

The result is essentially a ClippedMinimizationDiagram.

fromVertexForm :: (Plane_ plane r, Ord plane, Ord r, Fractional r, Show r, Show plane, Ord vtxData, HasDefiners vtxData plane) => NEMap (Point 3 r) vtxData -> MinimizationDiagram r (MDVertex r plane vtxData) plane

Given the vertices of the lower envelope; compute the minimization diagram.

\(O(h\log h)\) assuming that the input is non-degenerate.

fromVertexFormIn :: (Plane_ plane r, Ord plane, Ord r, Fractional r, Show r, Show plane, Point_ corner 2 r, Foldable1 set, Show r, Show corner, Show vertexData, Ord vertexData, HasDefiners vertexData plane) => Triangle corner -> set plane -> NEMap (Point 3 r) vertexData -> NEMap plane (ClippedMDCell r plane vertexData)

Computes the part of the minimization diagram inside the given triangle.

This function takes both the set of planes, as well as the vertex form of the lower envelope of these planes.

mergeDefiners :: (Plane_ plane r, Ord plane, Ord r, Fractional r, Show plane, Show r) => Point 3 r -> Definers plane -> Definers plane -> Definers plane

Merge two lists of definers.

\(O(n\log n)\), where \(n\) is the total number of planes involved.

definers :: (Plane_ plane r, Ord r, Fractional r) => Three plane -> Maybe (Point 3 r, Definers plane)

Smart constructor for creating the definers of three planes

fromCCWList :: NonEmpty plane -> Definers plane

Given the planes in order, starting with the one that is closest in the up direction, construct the Definers.

type ClippedBoundedRegion (r :: k) vertex corner = ConvexPolygonF (Cyclic NonEmpty) (OriginalOrExtra vertex corner)

A Convex bounded region, which may be clipped using vertices of type corner.

data Definers plane

in CCW order, starting with the plane that is minimal at the vertical up direction from their common vertex.

type VertexForm (map :: Type -> Type -> k) r plane = map (Point 3 r) (Definers plane)

The vertices of a lower envelope is just a Map with every vertex its definers, i.e. the planes that define the vertex in CCW order around it.

module HGeometry.Plane.LowerEnvelope.Connected.BruteForce

definers :: (Plane_ plane r, Ord r, Fractional r) => Three plane -> Maybe (Point 3 r, Definers plane)

Smart constructor for creating the definers of three planes

fromCCWList :: NonEmpty plane -> Definers plane

Given the planes in order, starting with the one that is closest in the up direction, construct the Definers.

newtype Definers plane

in CCW order, starting with the plane that is minimal at the vertical up direction from their common vertex.

Constructors

Definers (NonEmpty plane)

class HasDefiners vertex plane | vertex -> plane where

Methods

definersOf :: vertex -> Definers plane

Access the definers of some object

type VertexForm (map :: Type -> Type -> k) r plane = map (Point 3 r) (Definers plane)

The vertices of a lower envelope is just a Map with every vertex its definers, i.e. the planes that define the vertex in CCW order around it.