HGeometry.Plane.LowerEnvelope.Connected.Graph

Description

A Representation of the Lower envelope of planes as a bunch of convex regions

Documentation

type PlaneGraph k v e = Map k (Map e k, v)

A Plane graph storing vertices of type v that are identified by keys of type k, and some ordered sequence of edges (which are ordered using e).

newtype E r

Constructors

E (Vector 2 r)

toPlaneGraph :: (Plane_ plane r, Num r, Ord r) => MinimizationDiagram r plane -> PlaneGraph (Point 2 r) (First r) (E r)

Produce a triangulated plane graph on the bounded vertices. every vertex is represented by its point, it stores a list of its outgoing edges, and some data.