HGeometry.Plane.LowerEnvelope.Connected.Randomized

Description

expected O(n polylog n) divide and conquer implementation

Documentation

computeVertexForm :: (Plane_ plane r, Ord plane, Ord r, Fractional r, Foldable1 set, RandomGen gen, Show plane, Show r) => gen -> set plane -> VertexForm Map r plane

withConflictLists :: (Plane_ plane r, Ord r, Num r, Ord plane, FunctorWithIndex (Point 3 r) (map (Point 3 r)), Show r, Show plane) => Set plane -> VertexForm map r plane -> map (Point 3 r) (Definers plane, Set plane)

Computes conflict list

takeSample :: (Foldable1 set, Ord plane) => Int -> set plane -> (NonEmpty plane, Set plane)

Given a size r > 0; take a sample of the planes from the given size (essentially by just taking the first r planes.)

pre: r > 0

withExtraConflictLists :: (Plane_ plane r, Ord r, Num r, Show plane, Show r) => Set plane -> NEMap plane (ClippedMDCell r plane (a, Set plane)) -> NEMap plane (ClippedMDCell'' r (MDVertex r plane (a, Set plane)) (Point 2 r :+ (ExtraDefiners plane, Set plane)))

Compute the conflict lists for the extra vertices we added. In addition, compute the definers for the given position.