HGeometry.Plane.LowerEnvelope.Naive

Documentation

lowerEnvelope :: (Plane_ plane r, Ord r, Fractional r, Foldable1 f, Functor f, Ord plane, Show plane, Show r) => f plane -> LowerEnvelope plane

Brute force implementation that computes the lower envelope, by explicitly considering every triple of planes.

pre: the input forms a *set* of planes, i.e. no duplicates

running time: \(O(n^4 )\)

lowerEnvelopeVertexForm :: (Plane_ plane r, Ord r, Fractional r, Foldable f, Ord plane) => f plane -> VertexForm plane

Brute force implementation that computes the vertex form of the lower envelope, by explicitly considering every triple of planes.

pre: the input forms a *set* of planes, i.e. no duplicates

general position assumptions: None. Though keep in mind that e.g. if all planes are parallel there are no vertices.

running time: \(O(n^4 )\)

asVertex :: (Plane_ plane r, Foldable f, Ord plane, Ord r, Fractional r) => f plane -> Three plane -> Maybe (LEVertex plane)

Given all planes, and a triple, computes if the triple defines a vertex of the lower envelope, and if so returns it.

belowAll :: (Plane_ plane r, Ord r, Num r, Foldable f) => Point 3 r -> f plane -> Bool

test if v lies below (or on) all the planes in hs