HGeometry.PlaneGraph

Description

Plane Graphs, i.e. embedded planar graphs.

Documentation

module Hiraffe.Graph.Class

module HGeometry.PlaneGraph.Class

module Hiraffe.PlanarGraph.Class

data PlaneGraph (s :: k) v e f

An Embedded, *connected*, planar graph

fromCPlaneGraph :: forall {k} vertex r (s :: k) e f. (Point_ vertex 2 r, Ord r, Num r) => CPlaneGraph s vertex e f -> PlaneGraph s vertex e f

Given a connected Plane graph, constructs a PlaneGraph.

\(O(n)\)

fromCPlaneGraphWith :: forall {k} (s :: k) v e f. Dart s -> CPlaneGraph s v e f -> PlaneGraph s v e f

Given a dart that thas the outer face on its left, and a Connected Plane Graph, construct a PlaneGraph.

_PlanarGraph :: forall {k} (s :: k) v e f1 v' e' f' p f2. (Profunctor p, Functor f2) => p (PlanarGraph 'Primal s v e f1) (f2 (PlanarGraph 'Primal s v' e' f')) -> p (PlaneGraph s v e f1) (f2 (PlaneGraph s v' e' f'))

Iso to access the underlying PlanarGraph. Use at your own risk.

module HGeometry.PlaneGraph.Connected