HGeometry.PlaneGraph.Class

Description

Type type for planar graphs embedded in \(\mathbb{R}^2\). For functions that export faces and edges etc, we assume the graph has a (planar) straight line embedding.

Documentation

class (PlanarGraph_ planeGraph, vertex ~ Vertex planeGraph, Point_ vertex 2 (NumType vertex), NumType vertex ~ NumType planeGraph, HasEdges planeGraph planeGraph, HasOuterFace planeGraph) => PlaneGraph_ planeGraph vertex | planeGraph -> vertex

A class representing Plane graphs, i.e. planar graphs that have a straight line embedding in the plane.

class PlaneGraph_ planeGraph vertex => ConstructablePlaneGraph_ planeGraph vertex where

A class representing constructable Plane graphs, i.e. planar graphs that have a straight line embedding in the plane.

Methods

fromEmbedding :: (Foldable1 f, Functor f, Foldable h, Functor h, vi ~ VertexIx planeGraph, v ~ Vertex planeGraph, e ~ Edge planeGraph, GraphFromAdjListExtraConstraints planeGraph h) => f (vi, v, h (vi, e)) -> planeGraph

Build a graph from its embedding; i.e. for each vertex we expect its adjacencies in CCW order.

If the, in the list of neighbours of vertex u we see a vertex v that itself does not appear in the adjacencylist, we may drop it. In other words if u has a neighbour v, then v better have a specification of its neighbours somewhere.

defaultOuterFaceDart :: (PlanarGraph_ planeGraph, vertex ~ Vertex planeGraph, r ~ NumType planeGraph, Point_ vertex 2 r, Ord r, Num r) => planeGraph -> DartIx planeGraph

Computes a dart that is incident to the outer face (i.e. in | particular, that has the outer face to its left.)

running time: \(O(n)\)

dartSegmentAt :: (PlaneGraph_ planeGraph vertex, Point_ vertex 2 (NumType vertex)) => DartIx planeGraph -> IndexedFold (DartIx planeGraph) planeGraph (ClosedLineSegment vertex)

Get the line segment representing a particular dart.

Note that this is a fold rather than a getter for the same reason dartAt is a traversal rather than a lens: i.e. if you pass some nonsensical DartIx the dart may not exist.

edgeSegmentAt :: (PlaneGraph_ planeGraph vertex, Point_ vertex 2 (NumType vertex)) => EdgeIx planeGraph -> IndexedFold (EdgeIx planeGraph) planeGraph (ClosedLineSegment vertex)

Get the line segment representing a particular edge.

Note that this is a fold rather than a getter for the same reason edgeAt is a traversal rather than a lens: i.e. if you pass some nonsensical EdgeIx the edge may not exist.

dartSegments :: (PlaneGraph_ planeGraph vertex, Point_ vertex 2 (NumType vertex)) => IndexedFold (DartIx planeGraph) planeGraph (ClosedLineSegment vertex)

Renders all darts as line segments. Thesegments are all oriented in the direction of the dart.

edgeSegments :: (HasEdges planeGraph planeGraph, BidirGraph_ planeGraph, vertex ~ Vertex planeGraph, Point_ vertex 2 (NumType vertex)) => IndexedFold (EdgeIx planeGraph) planeGraph (ClosedLineSegment vertex)

Renders all edges as line segments.

interiorFacePolygonAt :: (PlaneGraph_ planeGraph vertex, HasOuterBoundaryOf planeGraph, HasInnerComponents planeGraph, Point_ vertex 2 (NumType vertex)) => FaceIx planeGraph -> IndexedFold (FaceIx planeGraph) planeGraph (PolygonalDomain (vertex :+ VertexIx planeGraph))

Render a given interior face as a polygon (which may cotnain holes)

interiorFacePolygons :: (PlaneGraph_ planeGraph vertex, HasOuterBoundaryOf planeGraph, HasInnerComponents planeGraph, Point_ vertex 2 r, Ord r, Num r, Eq (FaceIx planeGraph)) => IndexedFold (FaceIx planeGraph) planeGraph (PolygonalDomain (vertex :+ VertexIx planeGraph))

Renders all interior faces as polygons (which may possibly contain holes)

outerBoundaryPolygonAt :: (PlaneGraph_ planeGraph vertex, HasOuterBoundaryOf planeGraph, Point_ vertex 2 (NumType vertex)) => FaceIx planeGraph -> IndexedFold (FaceIx planeGraph) planeGraph (SimplePolygon (vertex :+ VertexIx planeGraph))

Renders a single interior face as a simple polygon.

Note that this is a fold rather than a getter for the same reason faceAt is a traversal rather than a lens: i.e. if you pass some nonsensical FaceIx the face may not exist.

outerBoundaryPolygons :: (PlaneGraph_ planeGraph vertex, HasOuterBoundaryOf planeGraph, Point_ vertex 2 r, Ord r, Num r, Eq (FaceIx planeGraph)) => IndexedFold (FaceIx planeGraph) planeGraph (SimplePolygon (vertex :+ VertexIx planeGraph))

Renders all interior faces as simple polygons.