HGeometry.PlaneGraph.Connected

Description

Plane Graphs, i.e. embedded planar graphs.

Documentation

module Hiraffe.Graph.Class

module HGeometry.PlaneGraph.Class

module Hiraffe.PlanarGraph.Class

data CPlaneGraph (s :: k) v e f

An Embedded, *connected*, planar graph

fromAdjacencyRep :: forall {k} vertex r i (f :: Type -> Type) proxy (s :: k) e. (Point_ vertex 2 r, Ord i, Foldable1 f) => proxy s -> GGraph f i vertex e -> CPlaneGraph s vertex e ()

Given a connected plane graph in adjacency list format; convert it into an actual CPlaneGraph.

\(O(n\log n)\)

fromConnectedSegments :: forall {k} (s :: k) nonEmpty lineSegment point r. (Foldable1 nonEmpty, Ord r, Num r, LineSegment_ lineSegment point, Point_ point 2 r) => nonEmpty lineSegment -> CPlaneGraph s (NonEmpty point) lineSegment ()

Constructs a connected plane graph

pre: The segments form a single connected component No two segments partially overlap.

running time: \(O(n\log n)\)

fromIntersectingSegments :: forall {k} (s :: k) nonEmpty ix lineSegment segment r point. (Foldable1 nonEmpty, FunctorWithIndex ix nonEmpty, LineSegment_ lineSegment point, Point_ point 2 r, Ord r, Fractional r, Intersection lineSegment lineSegment ~ Maybe (LineSegmentLineSegmentIntersection segment), LineSegment_ segment point, IsIntersectableWith lineSegment lineSegment, Ord ix, HasOnSegment lineSegment 2, StartPointOf lineSegment ~ EndPointOf lineSegment, Eq lineSegment) => nonEmpty lineSegment -> CPlaneGraph s (Point 2 r :+ Seq point) (ViewL1 lineSegment) ()

Construct A connected PlaneGraph from a set of intersecting segments.

\( O((n+k)\log n) \), where \(n\) is the number of segments, and \(k\) is the number of intersections.

pre: the segments actually form a connected graph.

fromSimplePolygon :: forall {k} simplePolygon vertex r (s :: k). (SimplePolygon_ simplePolygon vertex r, EdgeIx simplePolygon ~ Int) => simplePolygon -> CPlaneGraph s (vertex :+ VertexIx simplePolygon) (EdgeIx simplePolygon) PointLocationResult

Construct a connected PlaneGraph from a single simple polygon.

the interior of the polygon will have faceId 0

running time: \(O(n)\).

fromSimplePolygonWith

Arguments

:: forall {k} simplePolygon vertex r face (s :: k). (SimplePolygon_ simplePolygon vertex r, EdgeIx simplePolygon ~ Int)

=> simplePolygon

-> face
Data for the face representing the inside

-> face
Data for the face representing the outside

-> CPlaneGraph s (vertex :+ VertexIx simplePolygon) (EdgeIx simplePolygon) face

Construct a plane graph from a simple polygon.

the interior of the polygon will have faceId 0

running time: \(O(n)\).

Unchecked ways of accessing the underlying PlanarGraph

_CPlanarGraph :: forall {k} (s :: k) v e f1 v' e' f' p f2. (Profunctor p, Functor f2) => p (CPlanarGraph 'Primal s v e f1) (f2 (CPlanarGraph 'Primal s v' e' f')) -> p (CPlaneGraph s v e f1) (f2 (CPlaneGraph s v' e' f'))

Iso to access the underlying planar graph.