HGeometry.PlaneGraph.Connected.PolygonOverlay

Description

Computes the Overlay of a bunch of simple polygon

Documentation

polygonOverlay :: forall {k} nonEmpty (s :: k) simplePolygon vertex r. (Foldable1 nonEmpty, Point_ vertex 2 r, SimplePolygon_ simplePolygon vertex r, Fractional r, Ord r, Ord vertex, HasIntersectionWith (Point 2 r) simplePolygon, Eq simplePolygon) => nonEmpty simplePolygon -> CPlaneGraph s (V simplePolygon) (E simplePolygon) (F simplePolygon)

Given a set of polygons, constructs a plane graph whose vertices, edges, and faces are tagged with the polygons that cover that face.

The current running time is \(O( (N+k) + \log (N+k) + (N+k)*m )\), where \(N\) is the total complexity of all polygons, \(k\) is the number of intersections between the polygons, and \(m\) is the maximum complexity of a single polygon.

pre: everything forms a connected graph

definingEdges :: forall polygon f. Functor f => (ViewL1 (ClosedLineSegment (Vertex polygon) :+ polygon) -> f (ViewL1 (ClosedLineSegment (Vertex polygon) :+ polygon))) -> E polygon -> f (E polygon)

definingVertices :: forall polygon f. Functor f => (Seq (Vertex polygon) -> f (Seq (Vertex polygon))) -> V polygon -> f (V polygon)

location :: forall polygon f. Functor f => (Point 2 (NumType polygon) -> f (Point 2 (NumType polygon))) -> V polygon -> f (V polygon)

pointInEdge :: forall polygon f. Functor f => (Point 2 (NumType polygon) -> f (Point 2 (NumType polygon))) -> E polygon -> f (E polygon)

pointInFace :: forall polygon f. Functor f => (Maybe (Point 2 (NumType polygon)) -> f (Maybe (Point 2 (NumType polygon)))) -> F polygon -> f (F polygon)

data E polygon

The data associated with edges in the overlay.

An edge in the overlay corresponds to some line segment uv.

Constructors

E (ViewL1 (ClosedLineSegment (Vertex polygon) :+ polygon)) (Seq polygon) (Point 2 (NumType polygon))

data F polygon

The data associated with a face in the overlay.

Constructors

F (Seq polygon) (Maybe (Point 2 (NumType polygon)))

class HasCoveringPolygons geom polygon | geom -> polygon where

Class for types that have some polygons covering the object.

Methods

coveringPolygons :: Lens' geom (Seq polygon)

Get the polygons covering this geometry

data V polygon

Vertices in the overlay

Constructors

V (Point 2 (NumType polygon)) (Seq (Vertex polygon)) (Seq polygon)