HGeometry.Point.Quadrants

Description

Documentation

data Quadrant

Quadrants of two dimensional points. in CCW order

Constructors

TopRight

TopLeft

BottomLeft

BottomRight

quadrantWith :: (Ord r, Point_ point 2 r) => point -> point -> Quadrant

Quadrants around point c; quadrants are closed on their "previous" boundary (i..e the boundary with the previous quadrant in the CCW order), open on next boundary. The origin itself is assigned the topRight quadrant

quadrant :: (Ord r, Num r, Point_ point 2 r) => point -> Quadrant

Quadrants with respect to the origin

partitionIntoQuadrants :: (Ord r, Point_ point 2 r) => point -> [point] -> ([point], [point], [point], [point])

Given a center point c, and a set of points, partition the points into quadrants around c (based on their x and y coordinates). The quadrants are reported in the order topLeft, topRight, bottomLeft, bottomRight. The points are in the same order as they were in the original input lists. Points with the same x-or y coordinate as p, are "rounded" to above.