HGeometry.PolyLine

Description

A Polyline and some basic functions to interact with them.

Documentation

newtype PolyLineF (f :: k -> Type) (point :: k)

Simple polygons just store their vertices in CCCW order

Constructors

PolyLine (f point)

type PolyLine = PolyLineF NonEmptyVector

By default we store simple poylline as non-empty vectors.

_PolyLineF :: forall {k1} {k2} f1 (point :: k1) f' (point' :: k2) p f2. (Profunctor p, Functor f2) => p (f1 point) (f2 (f' point')) -> p (PolyLineF f1 point) (f2 (PolyLineF f' point'))

Access the container

module HGeometry.PolyLine.Class

class HasVertices' graph => HasVertices graph graph' where

Class that expresses that we have a non-empty type changing traversal of all vertices.

Methods

vertices :: IndexedTraversal1 (VertexIx graph) graph graph' (Vertex graph) (Vertex graph')

Traversal of all vertices in the graph

class HasEdges' graph => HasEdges graph graph' where

Class for types that have a type changing traversal of the edges

Methods

edges :: IndexedTraversal (EdgeIx graph) graph graph' (Edge graph) (Edge graph')

Traversal of all edges in the graph