HGeometry.PolyLine.Class

Description

A Polyline class and some basic functions to interact with them.

Documentation

class (HasVertices polyLine polyLine, HasPoints' polyLine point, HasStart polyLine point, HasEnd polyLine point, Vertex polyLine ~ point, Point_ point (Dimension point) (NumType point), NumType polyLine ~ NumType point, Dimension polyLine ~ Dimension point) => PolyLine_ polyLine point | polyLine -> point

A class representing PolyLines

class PolyLine_ polyLine point => ConstructablePolyLine_ polyLine point where

Class for constructable polylglines

Methods

polyLineFromPoints :: Foldable1 f => f point -> polyLine

Constructs a polyline from a given sequence of points.

pre: there should be at least two distinct points

_PolyLineLineSegment :: (ConstructableLineSegment_ lineSegment point, ConstructablePolyLine_ polyLine point) => Prism' polyLine lineSegment

Prism between a polyline and a line segment