HGeometry.PolyLine.Frechet.Discrete

Description

Computes the discrete frechet distance

Documentation

frechetDistanceWith :: (HasVertices' polyLine, HasVertices' polyLine', Ord r, VertexIx polyLine ~ Int, VertexIx polyLine' ~ Int) => (Vertex polyLine -> Vertex polyLine' -> r) -> polyLine -> polyLine' -> r

Computes the discrete frechet distance with respect to the given distance function.

O(nm*(I+T)), where I is the time to index one of the vertices and T is the time to evaluate the distance function.