HGeometry.PolyLine.Simplification.DouglasPeucker

Description

Documentation

douglasPeucker :: forall polyLine point d r. (PolyLine_ polyLine point, Point_ point d r, Ord r, Fractional r, HasSquaredEuclideanDistance point) => r -> polyLine -> PolyLineF ViewR1 point

Line simplification with the well-known Douglas Peucker alogrithm. Given a distance value eps and a polyline pl, constructs a simplification of pl (i.e. with vertices from pl) s.t. all other vertices are within dist eps to the original polyline.

Running time: \( O(n^2) \) worst case, \( O(n log n) \) on average.