HGeometry.PolyLine.Simplification.ImaiIri

Description

Documentation

simplify :: forall polyLine point (d :: Nat) r. (PolyLine_ polyLine point, Point_ point d r, Ord r, Fractional r, HasSquaredEuclideanDistance point) => r -> polyLine -> PolyLine point

Line simplification with the Imai-Iri alogrithm. Given a distance value eps and a polyline pl, constructs a simplification of pl (i.e. with vertices from pl) s.t. all other vertices are within dist eps to the original polyline.

Running time: \( O(n^2) \) time.

simplifyWith :: forall polyLine point (d :: Nat) r. (PolyLine_ polyLine point, Point_ point d r, Ord r, Fractional r) => (ClosedLineSegment point -> PolyLine point -> Bool) -> polyLine -> PolyLine point

Given a function that tests if the shortcut is valid, compute a simplification using the Imai-Iri algorithm.

Running time: \( O(Tn^2 \) time, where \(T\) is the time to evaluate the predicate.