HGeometry.Polygon

Description

A polygon and some basic functions to interact with them.

Documentation

module HGeometry.Polygon.Class

asTriangle :: Polygon_ polygon point r => polygon -> Maybe (Triangle point)

Try to convert the polygon into a triangle (whose vertices are given in CCW order).