HGeometry.Polygon.Class

Description

A polygon and some basic functions to interact with them.

Documentation

class HasVertices polygon polygon => HasOuterBoundary polygon where

A class for items that have an outer boundary.

Minimal complete definition

outerBoundaryVertexAt, ccwOuterBoundaryFrom, cwOuterBoundaryFrom

Methods

outerBoundary :: IndexedTraversal1' (VertexIx polygon) polygon (Vertex polygon)

A fold over all vertices of the outer boundary, the vertices are traversed in CCW order.

running time :: \(O(n)\)

default outerBoundary :: Num (VertexIx polygon) => IndexedFold1 (VertexIx polygon) polygon (Vertex polygon)

ccwOuterBoundaryFrom :: VertexIx polygon -> IndexedTraversal1' (VertexIx polygon) polygon (Vertex polygon)

A CCW-traversal over all vertices of the outer boundary, starting from the vertex with the given index.

running time :: \(O(n)\)

cwOuterBoundaryFrom :: VertexIx polygon -> IndexedTraversal1' (VertexIx polygon) polygon (Vertex polygon)

A CW-fold over all vertices of the outer boundary, starting from the vertex with the given index.

running time :: \(O(n)\)

outerBoundaryVertexAt :: VertexIx polygon -> IndexedGetter (VertexIx polygon) polygon (Vertex polygon)

A particular vertex of the outer polygon

running time: \(O(1)\)

outerBoundaryEdges :: IndexedFold1 (VertexIx polygon, VertexIx polygon) polygon (Vertex polygon, Vertex polygon)

A fold over all edges in the polygon. The edges are folded over in CCW order, and each edge is associated with the index of its start vertex outerBoundaryEdges :: IndexedFold (VertexIx polygon) polygon (Vertex polygon, Vertex polygon)

running time :: \(O(n)\)

default outerBoundaryEdges :: Enum (VertexIx polygon) => IndexedFold1 (VertexIx polygon, VertexIx polygon) polygon (Vertex polygon, Vertex polygon)

outerBoundaryEdgeAt :: VertexIx polygon -> IndexedGetter (VertexIx polygon, VertexIx polygon) polygon (Vertex polygon, Vertex polygon)

Get the edge that has the given vertex as its starting edge

running time: \(O(1)\)

default outerBoundaryEdgeAt :: Enum (VertexIx polygon) => VertexIx polygon -> IndexedGetter (VertexIx polygon, VertexIx polygon) polygon (Vertex polygon, Vertex polygon)

signedArea2X :: (Num r, HasOuterBoundary simplePolygon, Point_ point 2 r, Vertex simplePolygon ~ point) => simplePolygon -> r

Compute the signed area times 2 of a simple polygon. The the vertices are in clockwise order, the signed area will be negative, if the verices are given in counter clockwise order, the area will be positive.

running time: \(O(n)\)

minimumVertexBy :: HasOuterBoundary polygon => (Vertex polygon -> Vertex polygon -> Ordering) -> IndexedFold1 (VertexIx polygon) polygon (Vertex polygon)

Yield the minimum vertex of a polygon according to the given comparison function.

running time \( O(n) \)

Yield the minimum vertex of a polygon according to the given comparison function.

running time \( O(n) \)

maximumVertexBy :: HasOuterBoundary polygon => (Vertex polygon -> Vertex polygon -> Ordering) -> IndexedFold1 (VertexIx polygon) polygon (Vertex polygon)

Yield the maximum vertex of a polygon according to the given comparison function.

running time \( O(n) \)

outerBoundaryEdgeSegmentAt :: (HasOuterBoundary polygon, Vertex polygon ~ point, Point_ point 2 r) => VertexIx polygon -> IndexedGetter (VertexIx polygon, VertexIx polygon) polygon (ClosedLineSegment point)

Get the line segment representing the edge that has the given vertex as its starting edge

running time: \(O(1)\)

outerBoundaryEdgeSegments :: (HasOuterBoundary polygon, Vertex polygon ~ point, Point_ point 2 r) => IndexedFold1 (VertexIx polygon, VertexIx polygon) polygon (ClosedLineSegment point)

Get the line segments representing the outer boundary of the polygon.

outerBoundaryWithNeighbours :: (HasOuterBoundary polygon, VertexIx polygon ~ Int) => IndexedFold1 (VertexIx polygon, (VertexIx polygon, VertexIx polygon)) polygon (Vertex polygon, (Vertex polygon, Vertex polygon))

A fold that associates each vertex on the boundary with its predecessor and successor (in that order) along the boundary.

type Hole polygon = SimplePolygonF (HoleF polygon) (Vertex polygon)

A hole is a simple polygon

class VertexContainer (HoleF polygon) (Vertex polygon) => HasHoles polygon where

Accessing the holes in a polygon (if there are any.)

the default implementation assumes there are no holes

Minimal complete definition

Nothing

Associated Types

type HoleIx polygon

Type we use to index holes.

type HoleIx polygon = Void

type HoleF polygon :: Type -> Type

The functor used in the holes

Traversal over the holes in the polygon. Each hole is a simple polygon.

type HoleF polygon = Cyclic NonEmptyVector

Methods

holes :: IndexedTraversal' (HoleIx polygon) polygon (Hole polygon)

holeAt :: HoleIx polygon -> IndexedTraversal' (HoleIx polygon) polygon (Hole polygon)

class (HasOuterBoundary polygon, Vertex polygon ~ point, Point_ point 2 r, NumType polygon ~ r, Dimension polygon ~ 2, HasHoles polygon) => Polygon_ polygon point r where

A class representing (planar) polygons. The edges of the polygon may not intersect.

Minimal complete definition

ccwPredecessorOf, ccwSuccessorOf

Methods

extremes :: Vector 2 r -> polygon -> (point, point)

Finds the extreme points, minimum and maximum, in a given direction

running time: \(O(n)\)

ccwPredecessorOf :: VertexIx polygon -> IndexedLens' (VertexIx polygon) polygon (Vertex polygon)

Given a vertexIdx v; get an IndexedLens to access the CCW predecessor of v

ccwSuccessorOf :: VertexIx polygon -> IndexedLens' (VertexIx polygon) polygon (Vertex polygon)

Given a vertexIdx v; get an IndexedLens to access the CCW predecessor of v

class HasVertices' graph => HasVertices graph graph' where

Class that expresses that we have a non-empty type changing traversal of all vertices.

Methods

vertices :: IndexedTraversal1 (VertexIx graph) graph graph' (Vertex graph) (Vertex graph')

Traversal of all vertices in the graph

class HasVertices' graph where

A class representing types that have vertices.

Minimal complete definition

vertexAt

Associated Types

type Vertex graph

Vertices of the graph are of this type

type VertexIx graph

Vertices are indexed by elements of type VertexIx

Methods

vertexAt :: VertexIx graph -> IndexedTraversal' (VertexIx graph) graph (Vertex graph)

Accessor to a given vertex.

numVertices :: graph -> Int

Number of vertices in the graph.

running time: O(1)

default numVertices :: HasVertices graph graph => graph -> Int

class HasEdges' graph => HasEdges graph graph' where

Class for types that have a type changing traversal of the edges

Methods

edges :: IndexedTraversal (EdgeIx graph) graph graph' (Edge graph) (Edge graph')

Traversal of all edges in the graph

class HasEdges' graph where

A class for things that have edges

Minimal complete definition

edgeAt

Associated Types

type Edge graph

The edges of the graph are of this type

type EdgeIx graph

The Edges are indexed by something of type EdgeIx

Methods

edgeAt :: EdgeIx graph -> IndexedTraversal' (EdgeIx graph) graph (Edge graph)

Indexed traversal of a given edge.

numEdges :: graph -> Int

Number of edges in the graph.

running time: O(1)

default numEdges :: HasEdges graph graph => graph -> Int