HGeometry.Polygon.Convex

Description

A type for representing Convex polygons, and some basic functions on Convex polygons.

Documentation

class SimplePolygon_ convexPolygon point r => ConvexPolygon_ convexPolygon point r

Class modelling convex polygons.

type ConvexPolygon = ConvexPolygonF (Cyclic NonEmptyVector)

By default we use a cyclic non-empty vector to represent convex polygons.

data ConvexPolygonF (f :: k -> Type) (point :: k)

Convex polygons

fromSimplePolygon :: forall r point (f :: Type -> Type). (Ord r, Num r, Point_ point 2 r, VertexContainer f point) => SimplePolygonF f point -> Maybe (ConvexPolygonF f point)

Smart constructor to construct a strictly convex polygon from a simple polygon.

toSimplePolygon :: ConvexPolygonF f point -> SimplePolygonF f point

Convert to a simple polygon, i.e. forget the polygon is convex.

_ConvexPolygon :: forall (f :: Type -> Type) point r. (Num r, Ord r, Point_ point 2 r, VertexContainer f point) => Prism' (SimplePolygonF f point) (ConvexPolygonF f point)

Prism that can forget that the polygon is convex

isStrictlyConvex :: forall r point (f :: Type -> Type). (Ord r, Num r, Point_ point 2 r, VertexContainer f point) => SimplePolygonF f point -> Bool

\( O(n) \) Check if a polygon is strictly convex.

isConvex :: forall r point (f :: Type -> Type). (Ord r, Num r, Point_ point 2 r, VertexContainer f point) => SimplePolygonF f point -> Bool

\( O(n) \) Check if a polygon is convex.

verifyConvex :: forall r point (f :: Type -> Type). (Ord r, Num r, Point_ point 2 r, VertexContainer f point) => ConvexPolygonF f point -> Bool

Verify that a convex polygon is strictly convex.

running time \( O(n) \)

maxInDirection :: (Num r, Ord r, ConvexPolygon_ convexPolygon point r) => Vector 2 r -> convexPolygon -> point

Finds the extreme maximum point in the given direction. Based on http://geomalgorithms.com/a14-_extreme_pts.html

pre: The input polygon is strictly convex.

running time: \(O(\log n)\)

_UncheckedConvexPolygon :: forall {k1} {k2} (f1 :: k1 -> Type) (point :: k1) (f' :: k2 -> Type) (point' :: k2) p f2. (Profunctor p, Functor f2) => p (SimplePolygonF f1 point) (f2 (SimplePolygonF f' point')) -> p (ConvexPolygonF f1 point) (f2 (ConvexPolygonF f' point'))

ConvexPolygons are isomorphic to SimplePolygons with the added constraint that all vertices are strictly convex.

Note that this is unchecked; i.e. one can turn an arbitrary simple polygon into a suposedly convex one.

findMaxWith :: ConvexPolygon_ convexPolygon point r => (point -> point -> Ordering) -> convexPolygon -> point

Find the maximum vertex in a convex polygon using a binary search. \( O(\log n) \)