HGeometry.Polygon.Convex.Internal

Description

A Simple polygon and some basic functions to interact with them.

Documentation

type ConvexPolygon = ConvexPolygonF (Cyclic NonEmptyVector)

By default we use a cyclic non-empty vector to represent convex polygons.

newtype ConvexPolygonF (f :: k -> Type) (point :: k)

Convex polygons

Constructors

ConvexPolygon

Fields

toSimplePolygon :: SimplePolygonF f point

Convert to a simple polygon, i.e. forget the polygon is convex.

fromSimplePolygon :: forall r point (f :: Type -> Type). (Ord r, Num r, Point_ point 2 r, VertexContainer f point) => SimplePolygonF f point -> Maybe (ConvexPolygonF f point)

Smart constructor to construct a strictly convex polygon from a simple polygon.

_ConvexPolygon :: forall (f :: Type -> Type) point r. (Num r, Ord r, Point_ point 2 r, VertexContainer f point) => Prism' (SimplePolygonF f point) (ConvexPolygonF f point)

Prism that can forget that the polygon is convex

_UncheckedConvexPolygon :: forall {k1} {k2} (f1 :: k1 -> Type) (point :: k1) (f' :: k2 -> Type) (point' :: k2) p f2. (Profunctor p, Functor f2) => p (SimplePolygonF f1 point) (f2 (SimplePolygonF f' point')) -> p (ConvexPolygonF f1 point) (f2 (ConvexPolygonF f' point'))

ConvexPolygons are isomorphic to SimplePolygons with the added constraint that all vertices are strictly convex.

Note that this is unchecked; i.e. one can turn an arbitrary simple polygon into a suposedly convex one.

isStrictlyConvex :: forall r point (f :: Type -> Type). (Ord r, Num r, Point_ point 2 r, VertexContainer f point) => SimplePolygonF f point -> Bool

\( O(n) \) Check if a polygon is strictly convex.

isConvex :: forall r point (f :: Type -> Type). (Ord r, Num r, Point_ point 2 r, VertexContainer f point) => SimplePolygonF f point -> Bool

\( O(n) \) Check if a polygon is convex.

verifyConvex :: forall r point (f :: Type -> Type). (Ord r, Num r, Point_ point 2 r, VertexContainer f point) => ConvexPolygonF f point -> Bool

Verify that a convex polygon is strictly convex.

running time \( O(n) \)

maxInDirection :: (Num r, Ord r, ConvexPolygon_ convexPolygon point r) => Vector 2 r -> convexPolygon -> point

Finds the extreme maximum point in the given direction. Based on http://geomalgorithms.com/a14-_extreme_pts.html

pre: The input polygon is strictly convex.

running time: \(O(\log n)\)

findMaxWith :: ConvexPolygon_ convexPolygon point r => (point -> point -> Ordering) -> convexPolygon -> point

Find the maximum vertex in a convex polygon using a binary search. \( O(\log n) \)

inConvex :: (ConvexPolygon_ convexPolygon point r, Point_ queryPoint 2 r, Num r, Ord r) => queryPoint -> convexPolygon -> PointLocationResultWith (VertexIx convexPolygon)

Check if a point lies inside a convex polygon, on the boundary, or outside of the convex polygon.

\( O(\log n) \)

type HalfPlaneConvexPolygonIntersection (f :: Type -> Type) r vertex = PossiblyDegenerateSimplePolygon vertex (ConvexPolygonF f (OriginalOrExtra vertex (Point 2 r)))

A single Component of a HalfPlane x ConvexPolygon intersection.

Orphan instances

(Point_ corner 2 r, Num r, Ord r) => HasIntersectionWith (HalfSpaceF (LinePV 2 r)) (Rectangle corner)

Instance details

Methods

intersects :: HalfSpaceF (LinePV 2 r) -> Rectangle corner -> Bool

(Point_ corner 2 r, Num r, Ord r) => HasIntersectionWith (HalfSpaceF (LinePV 2 r)) (Triangle corner)

Instance details

Methods

intersects :: HalfSpaceF (LinePV 2 r) -> Triangle corner -> Bool

(Point_ corner 2 r, Fractional r, Ord r) => IsIntersectableWith (HalfSpaceF (LinePV 2 r)) (Rectangle corner)

Instance details

Methods

intersect :: HalfSpaceF (LinePV 2 r) -> Rectangle corner -> Intersection (HalfSpaceF (LinePV 2 r)) (Rectangle corner)

(Point_ corner 2 r, Fractional r, Ord r) => IsIntersectableWith (HalfSpaceF (LinePV 2 r)) (Triangle corner)

Instance details

Methods

intersect :: HalfSpaceF (LinePV 2 r) -> Triangle corner -> Intersection (HalfSpaceF (LinePV 2 r)) (Triangle corner)