HGeometry.Polygon.Convex.Merge

Description

Algorithm for merging two disjoint convex polygons; i.e. computing their tangents and combining them.

Documentation

merge :: (ConvexPolygon_ convexPolygon point r, Num r, Ord r) => convexPolygon -> convexPolygon -> (convexPolygon, ClosedLineSegment (point :+ VertexIx convexPolygon), ClosedLineSegment (point :+ VertexIx convexPolygon))

Merging two convex hulls, based on the paper:

Two Algorithms for Constructing a Delaunay Triangulation Lee and Schachter International Journal of Computer and Information Sciences, Vol 9, No. 3, 1980

: (combined hull, lower tangent that was added, upper tangent that was added)

pre: - lp and rp are disjoint, and there is a vertical line separating the two polygons.

Running time: O(n+m), where n and m are the sizes of the two polygons respectively

FIXME: This currently returns the VertexIx's in the *original* polygons, rather than in the new polygon.