HGeometry.Polygon.Convex.MinkowskiSum

Description

Computes the Minkowski Sum of two convex polygons. Based on the implementation in "Computational Geometry" by de Berg, van Kreveld, Overmars, and Schwartskopf.

Documentation

minkowskiSum :: (Ord r, Num r, ConvexPolygon_ convexPolygon point r, ConvexPolygon_ convexPolygon' point' r) => convexPolygon -> convexPolygon' -> ConvexPolygon (point :+ point')

Computes the Minkowski sum of the two input polygons with \(n\) and \(m\) vertices respectively.

Note: it takes te position of te second polyon as is; you may want to translate it to te oriin first.

running time: \(O(n+m)\).