HGeometry.Polygon.Convex.Random

Description

Constructs a random convex polygon, based on the initial implementation of David Himmelstrup.

Documentation

randomConvex :: (MonadState g m, RandomGen g) => Int -> Int -> m (ConvexPolygon (Point 2 Rational))

Generate a random ConvexPolygon with N vertices and a granularity of vMax.

\( O(n \log n) \)