HGeometry.Polygon.Convex.Tangents

Description

Computing tangents of convex polygons.

Documentation

leftTangent :: (Ord r, Num r, ConvexPolygon_ convexPolygon point r) => convexPolygon -> point -> point

Given a convex polygon poly, and a point outside the polygon, find the left tangent of q and the polygon, i.e. the vertex v of the convex polygon s.t. the polygon lies completely to the right of the line from q to v.

running time: \(O(\log n)\).

rightTangent :: (Ord r, Num r, ConvexPolygon_ convexPolygon point r) => convexPolygon -> point -> point

Given a convex polygon poly, and a point outside the polygon, find the right tangent of q and the polygon, i.e. the vertex v of the convex polygon s.t. the polygon lies completely to the left of the line from q to v.

running time: \(O(\log n)\).

lowerTangent :: (Num r, Ord r, ConvexPolygon_ convexPolygon point r, Point_ point 2 r) => convexPolygon -> convexPolygon -> ClosedLineSegment point

Compute the lower tangent of the two polgyons

pre: - polygons lp and rp have at least 1 vertex - lp and rp are disjoint, and there is a vertical line separating the two polygons. - The vertices of the polygons are given in clockwise order

Running time: O(n+m), where n and m are the sizes of the two polygons respectively

lowerTangent' :: forall point r f. (Ord r, Num r, Foldable1 f, Point_ point 2 r) => f point -> f point -> Vector 2 (point :+ [point])

Compute the lower tangent of the two convex chains lp and rp

pre: - the chains lp and rp have at least 1 vertex - lp and rp are disjoint, and there is a vertical line having lp on the left and rp on the right. - The vertices in the left-chain are given in clockwise order, (right to left) - The vertices in the right chain are given in counterclockwise order (left-to-right)

The result returned is the two endpoints l and r of the tangents, and the remainders lc and rc of the chains (i.e.) such that the lower hull of both chains is: (reverse lc) ++ [l,h] ++ rc

Running time: \(O(n+m)\), where n and m are the sizes of the two chains respectively

upperTangent :: forall convexPolygon point r. (Num r, Ord r, ConvexPolygon_ convexPolygon point r) => convexPolygon -> convexPolygon -> ClosedLineSegment point

Compute the upper tangent of the two polgyons

pre: - polygons lp and rp have at least 1 vertex - lp and rp are disjoint, and there is a vertical line separating the two polygons. - The vertices of the polygons are given in clockwise order

Running time: \( O(n+m) \), where n and m are the sizes of the two polygons respectively

upperTangent' :: (Ord r, Num r, Foldable1 f, Point_ point 2 r) => f point -> f point -> Vector 2 (point :+ [point])

Compute the upper tangent of the two convex chains lp and rp

pre: - the chains lp and rp have at least 1 vertex - lp and rp are disjoint, and there is a vertical line having lp on the left and rp on the right. - The vertices in the left-chain are given in clockwise order, (right to left) - The vertices in the right chain are given in counterclockwise order (left-to-right)

The result returned is the two endpoints l and r of the tangents, and the remainders lc and rc of the chains (i.e.) such that the upper hull of both chains is: (reverse lc) ++ [l,h] ++ rc

Running time: \(O(n+m)\), where n and m are the sizes of the two chains respectively

findMaxWith :: (ConvexPolygon_ convexPolygon point r, Point_ point 2 r) => (point -> point -> Ordering) -> convexPolygon -> point

Find the maximum vertex in a convex polygon using a binary search. \( O(\log n) \)