HGeometry.Polygon.Convex.Unbounded

Description

A type for representing unbounded, Convex, polygonal regions.

Documentation

type UnboundedConvexRegion vertex = UnboundedConvexRegionF (NumType vertex) NonEmpty vertex

An unbounded polygonal Convex Region

data UnboundedConvexRegionF r (nonEmpty :: Type -> Type) vertex

An unbounded polygonal ConvexRegion whose vertices are stored in an nonEmpty

Constructors

Unbounded

Fields

(Vector 2 r)

vector indicating the direction of the unbounded edge incident to the first vertex. Note that this vector thus points INTO vertex v.

(nonEmpty vertex)

the vertices in CCW order,

(Vector 2 r)

the vector indicating the direction of the unbounded edge incident to the last vertex. The vector points away from the vertex (i.e. towards +infty).

chain :: forall r nonEmpty vertex nonEmpty' vertex' f. Functor f => (nonEmpty vertex -> f (nonEmpty' vertex')) -> UnboundedConvexRegionF r nonEmpty vertex -> f (UnboundedConvexRegionF r nonEmpty' vertex')

Lens to access the chain of vertices in CCW order

extremalVertices :: UnboundedConvexRegionF r NonEmpty vertex -> Either vertex (Vector 2 vertex)

Compute the first and last vertex of the chain. Returns a Left if the first and last are the same.

mapChain :: (nonEmpty vertex -> nonEmpty' vertex') -> UnboundedConvexRegionF r nonEmpty vertex -> UnboundedConvexRegionF r nonEmpty' vertex'

map a function over the sequence of points

boundedCore :: forall (nonEmpty :: Type -> Type) vertex r. (VertexContainer nonEmpty vertex, Point_ vertex 2 r, HasFromFoldable1 nonEmpty) => UnboundedConvexRegionF r nonEmpty vertex -> PossiblyDegenerateSimplePolygon vertex (ConvexPolygonF (Cyclic nonEmpty) vertex)

Computes the core of the unbounded region; i.e. the convex hull of the vertices of the region.

boundingRays :: (Point_ vertex 2 r, Num r) => UnboundedConvexRegionF r NonEmpty vertex -> Vector 2 (HalfLine vertex)

Computes the two bounding rays of the unbounded region. Note that the rays now are both pointing toward infinity. The second ray has the unbounded region to its left, whereas the first one has it to its right.

unboundedBoundingHalfplanes :: (Point_ vertex 2 r, Num r, Ord r) => UnboundedConvexRegionF r NonEmpty vertex -> TwoOrThree (HalfPlaneF (LinePV 2 r))

the 2 or three halfplanes bounding the unbounded part of the region in particular: the region minus its bounded core

toBoundedFrom :: (Foldable nonEmpty, Point_ point 2 r, Point_ vertex 2 r, Ord r, Fractional r) => nonEmpty point -> UnboundedConvexRegionF r NonEmpty vertex -> ConvexPolygonF (Cyclic NonEmpty) vertex

Given some convex shape S, construct some "big enough" convex polygon B out of the unbounded convex polygon U so that the intersection \(U \cap S\) is the same as (B cap S)

note: this creates some new vertices; which are "copies" from the extremal vertices. this is to avoid having to introduce yet another level of OriginalOrExtra's

data LineUnboundedConvexRegionIntersection r

Intersection between a halfline and a unbounded convex region

Constructors

Line_x_UnboundedConvexRegion_HalfLine (HalfLine (Point 2 r))

Line_x_UnboundedConvexRegion_LineSegment (ClosedLineSegment (Point 2 r))

Line_x_UnboundedConvexRegion_Point (Point 2 r)