HGeometry.Polygon.Instances

Description

Arbitrary instances for the polygon types in hgeometry

Documentation

allSimplePolygons :: [SimplePolygon (Point 2 Double)]

shrinkPolygon :: (Ord r, Fractional r, Real r) => SimplePolygon (Point 2 r) -> [SimplePolygon (Point 2 r)]

Shrink a simple polygon.

The main idea is to try and remove every vertex (making sure that we still get a valid polygon). At the end, we try to simplify the coordinates of the points involved as well.

Orphan instances

(Uniform r, Ord r, Num r) => Arbitrary (MonotonePolygon (Point 2 r))

Instance details

Methods

arbitrary :: Gen (MonotonePolygon (Point 2 r))

shrink :: MonotonePolygon (Point 2 r) -> [MonotonePolygon (Point 2 r)]

Arbitrary (SimplePolygon (Point 2 Rational))

Instance details

Methods

arbitrary :: Gen (SimplePolygon (Point 2 Rational))

shrink :: SimplePolygon (Point 2 Rational) -> [SimplePolygon (Point 2 Rational)]

Arbitrary (SimplePolygon (Point 2 (RealNumber p)))

Instance details

Methods

arbitrary :: Gen (SimplePolygon (Point 2 (RealNumber p)))

shrink :: SimplePolygon (Point 2 (RealNumber p)) -> [SimplePolygon (Point 2 (RealNumber p))]

Arbitrary (SimplePolygon (Point 2 Double))

Instance details

Methods

arbitrary :: Gen (SimplePolygon (Point 2 Double))

shrink :: SimplePolygon (Point 2 Double) -> [SimplePolygon (Point 2 Double)]