HGeometry.Polygon.Monotone

Description

A polygon is monotone in a certain direction if rays orthogonal to that direction intersects the polygon at most twice. See https://en.wikipedia.org/wiki/Monotone_polygon

Documentation

type MonotonePolygon point = MonotonePolygonF (SimplePolygon point)

A type representing montone polygons

data MonotonePolygonF simplePolygon

Monotone polygons. A polygon is monotone in direction d if any line perpendicular to direction d intersects the polygon in a single interval.

monotoneDirection :: MonotonePolygonF simplePolygon -> Vector 2 (NumType simplePolygon)

the direction in which we are monotone

toSimplePolygon :: MonotonePolygonF simplePolygon -> simplePolygon

Convert to a simple polygon, i.e. forget the polygon is monotone

asMonotonePolygon :: (SimplePolygon_ simplePolygon point r, Num r, Ord r) => Vector 2 r -> simplePolygon -> Maybe (MonotonePolygonF simplePolygon)

Try to establish that the given simple polygon is monotone with respect to the given direction.

\( O(n \log n) \)

uncheckedMontonePolygon :: SimplePolygon_ simplePolygon point r => Vector 2 r -> simplePolygon -> MonotonePolygonF simplePolygon

Given a direction v, and a simple polygon that is monotone in direction v, creates a Monotone polygon out of it. All of this is unchecked.

running time: O(1)

randomMonotone :: (MonadState g m, RandomGen g, Uniform r, Ord r, Num r) => Int -> m (MonotonePolygon (Point 2 r))

Generate a random N-sided polygon that is monotone in a random direction.

- pre: n >= 3

\( O(n \log n) \)

randomMonotoneDirected :: (MonadState g m, RandomGen g, Uniform r, Ord r, Num r) => Int -> Vector 2 r -> m (MonotonePolygon (Point 2 r))

Generate a random n-sided polygon that is monotone in the given direction.

pre: n >= 3

\( O(n \log n) \)

isMonotone :: (SimplePolygon_ simplePolygon point r, Num r, Ord r) => Vector 2 r -> simplePolygon -> Bool

A polygon is monotone if a straight line in a given direction cannot have more than two intersections.

\( O(n \log n) \)

randomNonZeroVector :: (MonadState g m, RandomGen g, Uniform r, Eq r, Num r) => m (Vector 2 r)

\( O(1) \) Create a random 2D vector which has a non-zero magnitude.