HGeometry.Polygon.Simple

Description

A Simple polygon and some basic functions to interact with them.

Documentation

class (Polygon_ simplePolygon point r, VertexIx simplePolygon ~ Int, Point_ point 2 r) => SimplePolygon_ simplePolygon point r where

A class representing simple polygons; i.e. polygons without holes (and without self intersections in the boundary.)

Minimal complete definition

uncheckedFromCCWPoints, fromPoints

Associated Types

type ConstructableSimplePolygon simplePolygon point r

Constraints that allow us to construct a simple polygon

type ConstructableSimplePolygon simplePolygon point r = (Eq r, Num r)

Methods

uncheckedFromCCWPoints :: Foldable1 f => f point -> simplePolygon

Given the vertices of the polygon, in CCW order, constructs the polygon. The vertices are numbered in the order they are given.

pre: - the vertices are given in CCW order - at least 3 vertices, not all colinear - no repeated vertices - no self-inttersections

fromPoints :: (Foldable f, ConstructableSimplePolygon simplePolygon point r) => f point -> Maybe simplePolygon

Given the vertices of the polygon, constructs the polygon. This function will make sure the polygon is a valid simple polygon, i.e. it has at least three vertices, is given in CCW order, no repeated vertices etc.

In particular, it will drop repeated vertices.

centroid :: (Fractional r, ConstructablePoint_ point' 2 r) => simplePolygon -> point'

Compute the centroid of a simple polygon.

running time: \(O(n)\)

type SimplePolygon = SimplePolygonF (Cyclic NonEmptyVector)

By default we store simple polygons as non-empty circular vectors.

newtype SimplePolygonF (f :: k -> Type) (point :: k)

Simple polygons just store their vertices in CCW order

Constructors

MkSimplePolygon (f point)

toCyclic :: forall {k} (v :: k -> Type) (point :: k). SimplePolygonF (Cyclic v) point -> Cyclic v point

Get the underlying cyclic vector.

type VertexContainer (f :: Type -> Type) point = (IxValue (f point) ~ point, Index (f point) ~ Int, TraversableWithIndex Int f, Traversable1 f, Ixed (f point), HasDirectedTraversals f)

shortcut for all default properties of f we need to store the vertices.

class HasInPolygon polygon point r | polygon -> point, point -> r where

Types that implement a point-in-polygon test.

Minimal complete definition

Nothing

Methods

inPolygon :: (Num r, Ord r, Point_ queryPoint 2 r) => queryPoint -> polygon -> PointLocationResultWith (VertexIx polygon)

Check if a point lies inside a polygon, on the boundary, or outside of the polygon. If the point lies on the boundary we also return an edge e containing the point (identified by the vertex v so that e follows v in the CCW order along the boundary). No guarantees are given about which edge is returned if the query point lies on multiple edges (i.e. when it coincides with a vertex.)

default inPolygon :: (Num r, Ord r, Point_ point 2 r, Point_ queryPoint 2 r, SimplePolygon_ polygon point r) => queryPoint -> polygon -> PointLocationResultWith (VertexIx polygon)

inSimplePolygon :: forall queryPoint simplePolygon point r. (Num r, Ord r, Point_ point 2 r, Point_ queryPoint 2 r, SimplePolygon_ simplePolygon point r) => queryPoint -> simplePolygon -> PointLocationResultWith (VertexIx simplePolygon)

Check if a point lies inside a polygon, on the boundary, or outside of the polygon. If the point lies on the boundary we also return an edge e containing the point (identified by the vertex v so that e follows v in the CCW order along the boundary). No guarantees are given about which edge is returned if the query point lies on multiple edges (i.e. when it coincides with a vertex.)

Running time: O(n).

>>> Point2 1 1 `inPolygon` simplePoly
StrictlyInside
>>> Point2 0 0 `inPolygon` simplePoly
OnBoundaryEdge 0
>>> Point2 10 0 `inPolygon` simplePoly
OnBoundaryEdge 1
>>> Point2 5 13 `inPolygon` simplePoly
StrictlyInside
>>> Point2 5 10 `inPolygon` simplePoly
StrictlyInside
>>> Point2 10 5 `inPolygon` simplePoly
OnBoundaryEdge 1
>>> Point2 20 5 `inPolygon` simplePoly
StrictlyOutside

containedIn :: (ClosedLineSegment_ lineSegment point, SimplePolygon_ simplePolygon vertex r, Intersection (ClosedLineSegment vertex) lineSegment ~ Maybe (LineSegmentLineSegmentIntersection lineSegment'), IsIntersectableWith (ClosedLineSegment vertex) lineSegment, NumType lineSegment' ~ r, HasInPolygon simplePolygon vertex r, Point_ point 2 r, Point_ vertex 2 r, Ord r, Fractional r) => lineSegment -> simplePolygon -> Bool

test if the given line segment is contained in the polygon. It is also ok if the segment lies partially on the boundary

hasNoSelfIntersections :: (Foldable1 f, Functor f, Point_ point 2 r, Ord r, Real r) => f point -> Bool

verify that some sequence of points has no self intersecting edges.

module HGeometry.Polygon.Simple.Class

Orphan instances

HasIntersectionWith geom (SimplePolygonF f vertex) => HasIntersectionWith geom (SimplePolygonF f vertex :+ extra)

Instance details

Methods

intersects :: geom -> (SimplePolygonF f vertex :+ extra) -> Bool

(Point_ point 2 r, Num r, Ord r, VertexContainer f point, HyperPlane_ line 2 r) => HasIntersectionWith (HalfSpaceF line) (SimplePolygonF f point)

Instance details

Methods

intersects :: HalfSpaceF line -> SimplePolygonF f point -> Bool

(Num r, Ord r, SimplePolygon_ (SimplePolygonF nonEmpty vertex) vertex r) => HasIntersectionWith (LinePV 2 r) (SimplePolygonF nonEmpty vertex)

Instance details

Methods

intersects :: LinePV 2 r -> SimplePolygonF nonEmpty vertex -> Bool

(SimplePolygon_ (SimplePolygonF f point) point r, Num r, Ord r) => HasIntersectionWith (Point 2 r) (SimplePolygonF f point)

Instance details

Methods

intersects :: Point 2 r -> SimplePolygonF f point -> Bool

(SimplePolygon_ (SimplePolygonF f point) point r, Num r, Ord r) => IsIntersectableWith (Point 2 r) (SimplePolygonF f point)

Instance details

Methods

intersect :: Point 2 r -> SimplePolygonF f point -> Intersection (Point 2 r) (SimplePolygonF f point)

(Read point, SimplePolygon_ (SimplePolygonF f point) point r) => Read (SimplePolygonF f point)

Instance details

Methods

readsPrec :: Int -> ReadS (SimplePolygonF f point)

readList :: ReadS [SimplePolygonF f point]

readPrec :: ReadPrec (SimplePolygonF f point)

readListPrec :: ReadPrec [SimplePolygonF f point]

(Show point, SimplePolygon_ (SimplePolygonF f point) point r) => Show (SimplePolygonF f point)

Instance details

Methods

showsPrec :: Int -> SimplePolygonF f point -> ShowS

show :: SimplePolygonF f point -> String

showList :: [SimplePolygonF f point] -> ShowS

(VertexContainer f point, DefaultTransformByConstraints (SimplePolygonF f point) 2 r, Point_ point 2 r, IsTransformable point, HasFromFoldable1 f, Eq r) => IsTransformable (SimplePolygonF f point)

Instance details

Methods

transformBy :: Transformation (Dimension (SimplePolygonF f point)) (NumType (SimplePolygonF f point)) -> SimplePolygonF f point -> SimplePolygonF f point

HasHoles (SimplePolygonF f point)

Instance details

Associated Types

type HoleIx (SimplePolygonF f point)

Instance details

Defined in HGeometry.Polygon.Simple

type HoleIx (SimplePolygonF f point) = Void

type HoleF (SimplePolygonF f point)

Instance details

Defined in HGeometry.Polygon.Simple

type HoleF (SimplePolygonF f point) = Cyclic NonEmptyVector

Methods

holes :: IndexedTraversal' (HoleIx (SimplePolygonF f point)) (SimplePolygonF f point) (Hole (SimplePolygonF f point))

holeAt :: HoleIx (SimplePolygonF f point) -> IndexedTraversal' (HoleIx (SimplePolygonF f point)) (SimplePolygonF f point) (Hole (SimplePolygonF f point))

VertexContainer f point => HasOuterBoundary (SimplePolygonF f point)

Instance details

Methods

outerBoundary :: IndexedTraversal1' (VertexIx (SimplePolygonF f point)) (SimplePolygonF f point) (Vertex (SimplePolygonF f point))

ccwOuterBoundaryFrom :: VertexIx (SimplePolygonF f point) -> IndexedTraversal1' (VertexIx (SimplePolygonF f point)) (SimplePolygonF f point) (Vertex (SimplePolygonF f point))

cwOuterBoundaryFrom :: VertexIx (SimplePolygonF f point) -> IndexedTraversal1' (VertexIx (SimplePolygonF f point)) (SimplePolygonF f point) (Vertex (SimplePolygonF f point))

outerBoundaryVertexAt :: VertexIx (SimplePolygonF f point) -> IndexedGetter (VertexIx (SimplePolygonF f point)) (SimplePolygonF f point) (Vertex (SimplePolygonF f point))

outerBoundaryEdges :: IndexedFold1 (VertexIx (SimplePolygonF f point), VertexIx (SimplePolygonF f point)) (SimplePolygonF f point) (Vertex (SimplePolygonF f point), Vertex (SimplePolygonF f point))

outerBoundaryEdgeAt :: VertexIx (SimplePolygonF f point) -> IndexedGetter (VertexIx (SimplePolygonF f point), VertexIx (SimplePolygonF f point)) (SimplePolygonF f point) (Vertex (SimplePolygonF f point), Vertex (SimplePolygonF f point))

(VertexContainer nonEmpty vertex, HasFromFoldable1 nonEmpty, Point_ vertex 2 r, Fractional r) => HasPickPoint (SimplePolygonF nonEmpty vertex) r

Instance details

Methods

pointInteriorTo :: SimplePolygonF nonEmpty vertex -> Point 2 r

(Point_ point 2 r, HasFromFoldable1 f, VertexContainer f point) => Polygon_ (SimplePolygonF f point) point r

Instance details

Methods

extremes :: Vector 2 r -> SimplePolygonF f point -> (point, point)

ccwPredecessorOf :: VertexIx (SimplePolygonF f point) -> IndexedLens' (VertexIx (SimplePolygonF f point)) (SimplePolygonF f point) (Vertex (SimplePolygonF f point))

ccwSuccessorOf :: VertexIx (SimplePolygonF f point) -> IndexedLens' (VertexIx (SimplePolygonF f point)) (SimplePolygonF f point) (Vertex (SimplePolygonF f point))

(Point_ point 2 r, VertexContainer f point, HasFromFoldable1 f) => SimplePolygon_ (SimplePolygonF f point) point r

Instance details

Associated Types

type ConstructableSimplePolygon (SimplePolygonF f point) point r

Instance details

Defined in HGeometry.Polygon.Simple

type ConstructableSimplePolygon (SimplePolygonF f point) point r = (Eq r, Num r)

Methods

uncheckedFromCCWPoints :: Foldable1 f0 => f0 point -> SimplePolygonF f point

fromPoints :: (Foldable f0, ConstructableSimplePolygon (SimplePolygonF f point) point r) => f0 point -> Maybe (SimplePolygonF f point)

centroid :: (Fractional r, ConstructablePoint_ point' 2 r) => SimplePolygonF f point -> point'

SimplePolygon_ (SimplePolygonF f point) point r => HasInPolygon (SimplePolygonF f point) point r

Instance details

Methods

inPolygon :: (Num r, Ord r, Point_ queryPoint 2 r) => queryPoint -> SimplePolygonF f point -> PointLocationResultWith (VertexIx (SimplePolygonF f point))