HGeometry.Polygon.Simple.DualTree

Description

The Dual tree of a simple polygon

Documentation

data DualTree a e b

A dualTree of a polygon; essentially a binary tree (with the exception) that the root may actually have three children.

Constructors

RootZero a

RootOne a (e, BinaryTrie e b)

RootTwo a (e, BinaryTrie e b) (e, BinaryTrie e b)

RootThree a (e, BinaryTrie e b) (e, BinaryTrie e b) (e, BinaryTrie e b)

rootVertex :: forall a e b a' f. Functor f => (a -> f a') -> DualTree a e b -> f (DualTree a' e b)

Access the root of the dual tree

trimap :: (a -> a') -> (e -> e') -> (b -> b') -> DualTree a e b -> DualTree a' e' b'

Map over the dual tree

dualTreeFrom :: forall {k} source r point (s :: k) f. (Point_ source 2 r, Point_ point 2 r, Ord r, Num r) => source -> CPlaneGraph s point PolygonEdgeType f -> Maybe (DualTree (FaceIx (CPlaneGraph s point PolygonEdgeType f)) (DartIx (CPlaneGraph s point PolygonEdgeType f)) (FaceIx (CPlaneGraph s point PolygonEdgeType f), (VertexIx (CPlaneGraph s point PolygonEdgeType f), point)))

Computes the dual tree of the polygon, starting with the triangle containing the source point.

returns a Maybe in case the polygon actually lies outside the polygon

dualTreeFromTriangle

Arguments

:: (PlaneGraph_ triangulatedPolygon vertex, HasOuterBoundaryOf triangulatedPolygon, Dart triangulatedPolygon ~ PolygonEdgeType)

=> FaceIx triangulatedPolygon
the (index of the) triangle from which to produce the dual tree.

-> triangulatedPolygon

-> Maybe (DualTree (FaceIx triangulatedPolygon) (DartIx triangulatedPolygon) (FaceIx triangulatedPolygon, (VertexIx triangulatedPolygon, vertex)))

Construct the dual tree from a given triangle