HGeometry.Polygon.Simple.Sample

Description

Functionality to sample points uniformly at random from within a simple polygon.

Documentation

samplePolygon :: (SimplePolygon_ polygon point r, Ord r, Real r, UniformRange r, StatefulGen g m) => polygon -> g -> m (Point 2 Double)

Uniformly samples a point in a polygon.

\(O(n\log n)\)

samplePolygons :: (SimplePolygon_ polygon point r, StatefulGen g m, Foldable1 nonEmpty, Real r, Ord r, UniformRange r) => nonEmpty polygon -> g -> m (Point 2 Double)

Unfiormly samples a point from a set of polygons. You may want to build a pointSampler if the goal is to sample multiple points.

\(O(N\log N)\), where \(N\) is the total size of all polygons.

data Sampler w v

A data structure that can be used to efficiently sample values of type v.

samplePoint :: (Point_ point 2 r, StatefulGen g m, Real r, Ord r, UniformRange r) => Sampler r (Triangle point) -> g -> m (Point 2 Double)

Sample a point

triangleSampler :: (SimplePolygon_ polygon point r, Num r, Ord r, Foldable1 nonEmpty) => nonEmpty polygon -> Sampler r (Triangle point)

Build a triangle sampler; i.e. samples a triagnle from the polygons with probability proportional to its area.

\(O(N\log N)\), where \(N\) is the total size of all polygons.