HGeometry.Polygon.Simple.ShortestPath.Tree

Description

Computing a shortest path tree in a simple polygon

Documentation

labelWithShortestPaths

Arguments

:: forall {k} source r vertex (s :: k) f. (Point_ source 2 r, Point_ vertex 2 r, Num r, Ord r)

=> source
the source point.

-> CPlaneGraph s vertex PolygonEdgeType f

-> CPlaneGraph s (vertex :+ Either source (VertexId s)) PolygonEdgeType f

Labels each vertex ith its parent (if it has one). computeShortestPaths :: ( SimplePolygon_ simplePolygon point r , SimplePolygon_ poly (point :+ Either source (VertexIx poly)) r , Point_ source 2 r , Ord r, Num r ) => source -> simplePolygon -> poly computeShortestPaths s poly = computeShortestPaths' s poly (triangulate poly)

Labels each vertex ith its parent.

computeShortestPaths' :: forall {k} source r vertex (s :: k) f. (Point_ source 2 r, Point_ vertex 2 r, Num r, Ord r) => source -> CPlaneGraph s vertex PolygonEdgeType f -> [(vertex :+ VertexId s) :+ Either source (vertex :+ VertexId s)]

Computes the shortest paths form the source point to all vertices of the polygon.