HGeometry.Polygon.Triangulation

Description

Documentation

triangulate :: forall {k} (s :: k) polygon point r. (SimplePolygon_ polygon point r, Ord r, Num r) => polygon -> CPlaneGraph s point PolygonEdgeType PolygonFaceData

Triangulates a polygon of \(n\) vertices

running time: \(O(n \log n)\)

computeDiagonals :: (SimplePolygon_ polygon point r, Ord r, Num r) => polygon -> [Diagonal polygon]

Computes a set of diagaonals that together triangulate the input polygon of \(n\) vertices.

running time: \(O(n \log n)\)

data PolygonEdgeType

After triangulation, edges are either from the original polygon or a new diagonal.

Constructors

Original

Diagonal

data PolygonFaceData

Data type that expresses whether or not we are inside or outside the polygon.

Constructors

Inside

Outside

type Diagonal polygon = Vector 2 (VertexIx polygon)