HGeometry.Polygon.Triangulation.TriangulateMonotone

Description

Documentation

type YMonotonePolygon_ = SimplePolygon_

Y-monotone polygon. All straight horizontal lines intersects the polygon no more than twice.

triangulate :: forall {k} (s :: k) yMonotonePolygon point r. (YMonotonePolygon_ yMonotonePolygon point r, Ord r, Num r) => yMonotonePolygon -> CPlaneGraph s point PolygonEdgeType PolygonFaceData

Triangulates a polygon of \(n\) vertices

running time: \(O(n \log n)\)

computeDiagonals :: (YMonotonePolygon_ yMonotonePolygon point r, Ord r, Num r) => yMonotonePolygon -> [Diagonal yMonotonePolygon]

Given a y-monotone polygon in counter clockwise order computes the diagonals to add to triangulate the polygon

pre: the input polygon is y-monotone and has \(n \geq 3\) vertices

running time: \(O(n)\)