HGeometry.Polygon.Visibility.Naive

Description

Naive algorithms to compute the Visibility graph and the Visibility polygon of a particular point in a simple polgyon

Documentation

visibilityGraph :: (SimplePolygon_ simplePolygon point r, HasIntersectionWith point simplePolygon, Ord r, Num r) => simplePolygon -> [Vector 2 (VertexIx simplePolygon)]

Naive algorithm to compute the visibilityGraph of a simple polygon.

\(O(n^3)\)

visibilityGraphWith

Arguments

:: (SimplePolygon_ simplePolygon point r, HasIntersectionWith point simplePolygon, Ord r, Num r)

=> simplePolygon

-> [Vector 2 (point :+ VertexIx simplePolygon)]
candidate edges

-> [Vector 2 (VertexIx simplePolygon)]

Given a polygon, and candidate visibility edges (edges guaranteed not to intersect) the boundary; test if the edges are actually inside the polygon (as well as adding the actual polygon edges as visible).

\(O(n + m)\), where m is the number of candidate edges.

visibilityPolygon :: forall point vertex polygon r. (Point_ point 2 r, Point_ vertex 2 r, HasSquaredEuclideanDistance vertex, SimplePolygon_ polygon vertex r, Ord r, Fractional r) => point -> polygon -> SimplePolygon (Definer polygon)

Naive O(n^2) time algorithm to compute the visibility polygon of a point inside a simple polygon.

type Definer polygon = DefinerF (NumType polygon) (VertexIx polygon) (Vertex polygon :+ VertexIx polygon)

The second parameter of DefinerF; the edge, actually stores the index of its starting vertex.

data DefinerF r edge orig

A vertex of a visibility polygon; which is either an original vertex of type orig or a vertex that is the intersection point of a ray through an original vertex and an edge.

Constructors

OriginalVertex orig

NewVertex (Point 2 r) edge orig
the intersection point on the edge, defined by the edge and the original vertex