HGeometry.Properties

Description

Defines some generic geometric properties e.g. Dimensions, NumType, and Intersection types.

Documentation

type family Dimension t :: Nat

A type family for types that are associated with a dimension. The dimension is the dimension of the geometry they are embedded in.

type family NumType t

A type family for types that have an associated numeric type.