HGeometry.RangeTree.Base

Description

Range Tree implementation

Documentation

data RangeTree f point

Type Representing a generic (1D) range-tree.

buildRangeTree :: (Foldable1 g, Point_ point d r, Ord r, 1 <= d, Semigroup (f point), Measured f point) => g point -> RangeTree f point

Construct a rangetree on n points.

\(O(n\log n)\)

fromAscList :: (Foldable1 g, Functor g, Point_ point d r, 1 <= d, Semigroup (f point), Measured f point) => g point -> RangeTree f point

Given the points in sorted order, builds the range tree.

The running time and space depends on the monoid used. In particlar, we use (<>) at every internal node.

fromGroupedAscList :: (Foldable1 g, Foldable1 h, Point_ point d r, 1 <= d, Semigroup (f point), Measured f point) => g (h point) -> RangeTree f point

Given the points in groups per x-coordinate, in increasing order of x-coordinate

rangeQuery :: (Interval_ interval r, Point_ point d r, 1 <= d, Ord r) => interval -> RangeTree f point -> [f point]

Report the canonical subsets of the nodes that together represent the query interval.

\(O(\log n)\)

query :: (Interval_ interval r, Point_ point d r, 1 <= d, Ord r, Monoid (f point)) => interval -> RangeTree f point -> f point

Query the range tree

\(O(\log n + k)\), where \(k\) is somehow depends on the output size (and the monoid used).