HGeometry.SegmentTree

Description

Segment Tree implementation

Documentation

data SegmentTree (f :: Type -> Type) interval

A Segment tree

buildSegmentTree :: forall (f :: Type -> Type) interval r g. (ClosedInterval_ interval r, Ord r, Monoid (f interval), CanInsert f interval, Foldable1 g) => g interval -> SegmentTree f interval

Given a set of intervals, build a segment tree

\(O(n\log n)\)

buildSkeleton :: forall (f :: Type -> Type) interval r g. (ClosedInterval_ interval r, Ord r, Monoid (f interval), Foldable1 g) => g r -> SegmentTree f interval

Given te endpoints in ascending order, build an empty segment-tree

\(O(n)\)

stab :: (Ord r, ClosedInterval_ interval r) => r -> SegmentTree f interval -> [f interval]

Query the segment tree, returns a list of canonical subsets; one for each relevant node. Each element of the resulting list represents all intervals stored at that node that are stabbed by q.

\(O(\log n)\)

query :: (Ord r, ClosedInterval_ interval r, Monoid (f interval)) => r -> SegmentTree f interval -> f interval

Query the segment tree

\(O(\log n + k)\), where \(k\) is somehow depends on the output size (and the monoid used).

insert :: forall interval r (f :: Type -> Type). (ClosedInterval_ interval r, Ord r, CanInsert f interval) => interval -> SegmentTree f interval -> SegmentTree f interval

Insert a segment into the segment tree, whose endpoints are already in the tree

\(O(\log n)\)

ascEndPoints :: forall (f :: Type -> Type) interval. SegmentTree f interval -> [NumType interval]

Report te list of endpoints that the segment tree is built on in left to right order.