HGeometry.Sequence.NonEmpty

Description

Documentation

data ViewL1 a

NonEmpty ViewL

Constructors

a :<< (Seq a)

data ViewR1 a

NonEmpty ViewR

Constructors

(Seq a) :>> a

viewl1 :: ViewR1 a -> ViewL1 a

View the leftmost element

viewr1 :: ViewL1 a -> ViewR1 a

View the rightmost element

(|>>) :: ViewR1 a -> a -> ViewR1 a infixl 5

snoc an element to the right

(<>>) :: ViewR1 a -> Seq a -> ViewR1 a infixl 5

Append a Sequence to a ViewR1

splitL1At :: Int -> ViewL1 a -> Maybe (Seq a, a, Seq a)

Given an index i, and a viewL1 s, attempts to split s at index i. Returns nothing if the index is out of range.

splitR1At :: Int -> ViewR1 a -> Maybe (Seq a, a, Seq a)

Given an index i, and a viewL1 s, attempts to split s at index i. Returns nothing if the index is out of range.