HGeometry.Small.TwoOrThree

Description

A type expressing that some value may have two or three values

Documentation

newtype TwoOrThree a

Either two or three elements

Constructors

TwoOrThree (Either (Two a) (Three a))

type Two = V2

Strict pair whose elements are of the same type.

Strict pair with both items the same

type Three = V3

Strict Triple with all items the same

pattern Two :: a -> a -> Two a

Pattern synonym for strict pairs.

pattern Three :: a -> a -> a -> Three a

Pattern synonym for strict triples.