HGeometry.Transformation

Description

Documentation

newtype Transformation d r

A type representing a Transformation for d dimensional objects

Constructors

Transformation (Matrix (d + 1) (d + 1) r)

transformationMatrix :: Iso (Transformation d r) (Transformation d s) (Matrix (d + 1) (d + 1) r) (Matrix (d + 1) (d + 1) s)

Transformations and Matrices are isomorphic.

(|.|) :: (Num r, OptMatrix_ (d + 1) r) => Transformation d r -> Transformation d r -> Transformation d r

Compose transformations (right to left)

identity :: (Num r, OptMatrix_ (d + 1) r) => Transformation d r

Identity transformation; i.e. the transformation which does not change anything.

inverseOf :: (Fractional r, OptMatrix_ (d + 1) r, Invertible (d + 1)) => Transformation d r -> Transformation d r

Compute the inverse transformation

>>> inverseOf $ translation (Vector2 (10.0) (5.0))
Transformation {_transformationMatrix = Matrix (Vector3 (Vector3 1.0 0.0 (-10.0)) (Vector3 0.0 1.0 (-5.0)) (Vector3 0.0 0.0 1.0))}

class IsTransformable g where

A class representing types that can be transformed using a transformation

Minimal complete definition

Nothing

Methods

transformBy :: Transformation (Dimension g) (NumType g) -> g -> g

default transformBy :: forall d r. DefaultTransformByConstraints g d r => Transformation (Dimension g) (NumType g) -> g -> g

type TransformationConstraints d r = (KnownNat d, Has_ Vector_ d r, OptMatrix_ (d + 1) r, HasComponents (Vector (d + 1) r) (Vector (d + 1) (Vector (d + 1) r)), Has_ Additive_ (d + 1) r)

Some constraints that we will pretty much need to transform a d dimensional object whose numtype is r

type DefaultTransformByConstraints g d r = (d ~ Dimension g, r ~ NumType g, HasPoints g g (Point d r) (Point d r), OptMatrix_ (d + 1) r, Fractional r, Has_ Additive_ d r, HasComponents (Vector (1 + d) (Vector (1 + d) r)) (Vector (1 + d) r))

Bunch of constraints we need for the default implementation of transformBy

type OptMatrix_ d r = (Has_ Additive_ d r, Has_ Vector_ d (Vector d r), Ixed (Vector d r), Ixed (Vector d (Vector d r)))

shorthand for square matrixecs

translation :: forall d r vector. (Num r, Vector_ vector d r, TransformationConstraints d r) => vector -> Transformation d r

Create translation transformation from a vector.

>>> transformBy (translation $ Vector2 1 2) $ Point2 2 3
Point2 3.0 5.0

scaling :: forall d r vector. (Num r, Vector_ vector d r, TransformationConstraints d r) => vector -> Transformation d r

Create scaling transformation from a vector.

>>> transformBy (scaling $ Vector2 2 (-1)) $ Point2 2 3
Point2 4.0 (-3.0)

uniformScaling :: forall d r. (Num r, TransformationConstraints d r) => r -> Transformation d r

Create scaling transformation from a scalar that is applied to all dimensions.

>>> transformBy (uniformScaling 5) $ Point2 2 3
Point2 10.0 15.0
>>> uniformScaling 5 == scaling (Vector2 5 5)
True
>>> uniformScaling 5 == scaling (Vector3 5 5 5)
True

translateBy :: (IsTransformable g, Num (NumType g), Vector_ vector (Dimension g) (NumType g), TransformationConstraints (Dimension g) (NumType g)) => vector -> g -> g

Translate a given point.

>>> translateBy (Vector2 1 2) $ Point2 2 3
Point2 3.0 5.0

scaleBy :: (IsTransformable g, Num (NumType g), Vector_ vector (Dimension g) (NumType g), TransformationConstraints (Dimension g) (NumType g)) => vector -> g -> g

Scale a given point.

>>> scaleBy (Vector2 2 (-1)) $ Point2 2 3
Point2 4.0 (-3.0)

scaleUniformlyBy :: (IsTransformable g, Num (NumType g), TransformationConstraints (Dimension g) (NumType g)) => NumType g -> g -> g

Scale a given point uniformly in all dimensions.

>>> scaleUniformlyBy 5 $ Point2 2 3
Point2 10.0 15.0

rotateTo :: Num r => Vector 3 (Vector 3 r) -> Transformation 3 r

Given three new unit-length basis vectors (u,v,w) that map to (x,y,z), construct the appropriate rotation that does this.

skewX :: Num r => r -> Transformation 2 r

Skew transformation that keeps the y-coordinates fixed and shifts the x coordinates.

rotation :: Floating r => r -> Transformation 2 r

Create a matrix that corresponds to a rotation by a radians counter-clockwise around the origin.

reflection :: Floating r => r -> Transformation 2 r

Create a matrix that corresponds to a reflection in a line through the origin which makes an angle of a radians with the positive x-asis, in counter-clockwise orientation.

reflectionV :: Num r => Transformation 2 r

Vertical reflection

reflectionH :: Num r => Transformation 2 r

Horizontal reflection

fitToBox :: forall rectangle g point r. (NumType g ~ r, Dimension g ~ 2, IsTransformable g, IsBoxable g, Rectangle_ rectangle point, Point_ point 2 r, Ord r, Fractional r) => rectangle -> g -> g

Given a rectangle r and a geometry g with its boundingbox, transform the g to fit r.

fitToBoxTransform :: forall rectangle g point r. (NumType g ~ r, Dimension g ~ 2, IsTransformable g, IsBoxable g, Rectangle_ rectangle point, Point_ point 2 r, Ord r, Fractional r) => rectangle -> g -> Transformation 2 r

Given a rectangle r and a geometry g with its boundingbox, compute a transformation can fit g to r.