HGeometry.Triangle.Class

Description

A class of types representing Triangles

Documentation

class Point_ point (Dimension point) (NumType point) => Triangle_ triangle point | triangle -> point where

Class representing triangles

Methods

mkTriangle :: point -> point -> point -> triangle

Construct a triangle from its three vertices.

corners :: Lens' triangle (Vector 3 point)

Lens to access the corners of the triangle.

pattern Triangle_ :: Triangle_ triangle point => point -> point -> point -> triangle

Constructs a triangle

toCounterClockwiseTriangle :: (Num r, Eq r, Point_ point 2 r, Triangle_ triangle point) => triangle -> triangle

Make sure that the triangles vertices are given in counter clockwise order

>>> let t = Triangle origin (Point2 0 (-1)) (Point2 (-1) 0) :: Triangle (Point 2 Int)
>>> toCounterClockwiseTriangle t
Triangle (Point2 0 0) (Point2 (-1) 0) (Point2 0 (-1))

triangleSignedArea2X :: (Num r, Point_ point 2 r, Triangle_ triangle point) => triangle -> r

Computes the double-signed area of a triangle

intersectingHalfPlanes :: (Triangle_ triangle point, Point_ point 2 r, Num r, Ord r) => triangle -> Vector 3 (HalfSpace 2 r)

Get the three halfplanes such that the triangle is the intersection of those halfspaces.

>>> let t = Triangle origin (Point2 0 (-1)) (Point2 (-1) 0) :: Triangle (Point 2 Int)
>>> mapM_ print $ intersectingHalfPlanes t
HalfSpace Negative (HyperPlane [0,0,1])
HalfSpace Positive (HyperPlane [-1,-1,-1])
HalfSpace Negative (HyperPlane [0,-1,0])

toBarricentric :: (Fractional r, Point_ point 2 r, Triangle_ triangle point) => point -> triangle -> Vector 3 r

Given a point q and a triangle, q inside the triangle, get the baricentric cordinates of q

fromBarricentric :: (Triangle_ triangle point, Point_ point d r, Num r) => Vector 3 r -> triangle -> Point d r

Given a vector of barricentric coordinates and a triangle, get the corresponding point in the same coordinate sytsem as the vertices of the triangle.