HGeometry.Trie.Binary

Description

A Trie type

Documentation

data TrieF (f :: k -> Type -> Type) (e :: k) v

The Trie data type, parameterized by the data structure storing the children.

Constructors

Node v (f e (TrieF f e v))

type BinaryTrie e v = TrieF (KV AtMostTwoOriented) e v

A binary Trie type

pattern Leaf :: v -> BinaryTrie e v

Pattern match on a leaf node

pattern LeftNode :: v -> (e, BinaryTrie e v) -> BinaryTrie e v

Pattern match on a node with one child, namely on the left

pattern RightNode :: v -> (e, BinaryTrie e v) -> BinaryTrie e v

Pattern match on a node with one child, namely on the left

pattern TwoNode :: v -> (e, BinaryTrie e v) -> (e, BinaryTrie e v) -> BinaryTrie e v

Pattern match on a node with two children

asBinaryTrie :: forall (f :: Type -> Type) e v. Traversable f => TrieF (KV f) e v -> Maybe (BinaryTrie e v)

Trie to convert the trie into a binary trie. If we only have one element, we alwasy produce a left node.

data AtMostTwoOriented a

At most two elements. A one element can either be left or right

Constructors

Zero

OneLeft !a

OneRight !a

Two !a !a