HGeometry.VerticalRayShooting.PersistentSweep

Description

Documentation

type VerticalRayShootingStructure lineSegment = VerticalRayShootingStructure' (NumType lineSegment) lineSegment

The vertical ray shooting data structure

type StatusStructure lineSegment = Set lineSegment

The status structure

Building the Data Structure

verticalRayShootingStructure :: (LineSegment_ lineSegment point, Point_ point 2 r, Ord r, Fractional r, Foldable1 nonEmpty) => nonEmpty lineSegment -> VerticalRayShootingStructure lineSegment

Given a set of \(n\) interiorly pairwise disjoint *closed* segments, compute a vertical ray shooting data structure. (i.e. the endpoints of the segments may coincide).

pre: no vertical segments

running time: \(O(n\log n)\). space: \(O(n\log n)\).

verticalRayShootingStructure' :: (LineSegment_ lineSegment point, Point_ point 2 r, Ord r, Fractional r, Foldable1 nonEmpty) => nonEmpty lineSegment -> Maybe (VerticalRayShootingStructure lineSegment)

Given a set of \(n\) interiorly pairwise disjoint *closed* segments, try to construct a vertical ray shooting data structure. This function will remove all vertical line segments. If there are no non-vertical segments, we return a Nothing.

The endpoints of the segments are allowed to coincide.

running time: \(O(n\log n)\). space: \(O(n\log n)\).

Querying the Data Structure

segmentAbove :: (LineSegment_ lineSegment point, Point_ point 2 r, Point_ queryPoint 2 r, Ord r, Num r, HasSupportingLine lineSegment) => queryPoint -> VerticalRayShootingStructure lineSegment -> Maybe lineSegment

Find the segment vertically strictly above query point q, if it exists.

\(O(\log n)\)

segmentAboveOrOn :: (LineSegment_ lineSegment point, Point_ point 2 r, Point_ queryPoint 2 r, Ord r, Num r, HasSupportingLine lineSegment) => queryPoint -> VerticalRayShootingStructure lineSegment -> Maybe lineSegment

Find the segment vertically query point q, if it exists.

\(O(\log n)\)

findSlab :: (LineSegment_ lineSegment point, Point_ point 2 r, Point_ queryPoint 2 r, Ord r, Num r, HasSupportingLine lineSegment) => queryPoint -> VerticalRayShootingStructure lineSegment -> Maybe (StatusStructure lineSegment)

Given a query point, find the (data structure of the) slab containing the query point

\(O(\log n)\)

lookupAbove :: (LineSegment_ lineSegment point, Point_ point 2 r, Point_ queryPoint 2 r, Ord r, Num r, HasSupportingLine lineSegment) => queryPoint -> StatusStructure lineSegment -> Maybe lineSegment

Finds the first segment strictly above q

\(O(\log n)\)

lookupAboveOrOn :: (LineSegment_ lineSegment point, Point_ point 2 r, Point_ queryPoint 2 r, Ord r, Num r, HasSupportingLine lineSegment) => queryPoint -> StatusStructure lineSegment -> Maybe lineSegment

Finds the segment containing or above the query point q

\(O(\log n)\)

searchInSlab :: (LineSegment_ lineSegment point, Point_ point 2 r, HasSupportingLine lineSegment, Num r) => (LinePV 2 r -> Bool) -> StatusStructure lineSegment -> Maybe lineSegment

generic searching function