HGeometry.VoronoiDiagram

Description

Voronoi Diagrams of points in the plane

Documentation

data VoronoiDiagram point

A Voronoi diagram

Constructors

AllColinear ()

ConnectedVD (VoronoiDiagram' point)

newtype VoronoiDiagram' point

A connected VoronoiDiagram

Constructors

VoronoiDiagram (LowerEnvelope' (Plane (NumType point) :+ point))

voronoiDiagram :: (Point_ point 2 r, Functor f, Default point, Ord point, Ord r, Fractional r, Foldable f, Show point, Show r) => f point -> VoronoiDiagram point

Computes the Voronoi Diagram, by lifting the points to planes, and computing the lower envelope of these planes.

\(O(n\log n)\)

voronoiVertices :: (Point_ point 2 r, Functor f, Default point, Ord point, Ord r, Fractional r, Foldable f) => f point -> [Point 2 r]

Computes all voronoi vertices

edgeGeometries :: (Point_ point 2 r, Ord r, Fractional r, Default point, Show point, Show r) => Fold (VoronoiDiagram' point) (EdgeGeometry (Point 2 r))

Get the halflines and line segments representing the VoronoiDiagram