HGeometry.VoronoiDiagram.Clipped

Description

Representing a Voronoi Diagram of a set of points in R^2 that is clipped to some triangle, this means all cells are bounded convex polygons. (The points may be outside the triangle).

Documentation

data ClippedVoronoiDiagram point

A Voronoi diagram clipped to a triangle

_ClippedVoronoiDiagram :: NumType point ~ r => Iso' (ClippedVoronoiDiagram point) (NEMap point (ClippedMDCell r point))

Get access to the underling nonEmpty Map

voronoiDiagramIn :: (Point_ point 2 r, Point_ corner 2 r, Functor nonEmpty, Ord point, Ord r, Fractional r, Foldable1 nonEmpty, Show r, Show corner, Show point) => Triangle corner -> nonEmpty point -> ClippedVoronoiDiagram point

Computes the Voronoi inside the given triangle.

O(n^4) (for now).

voronoiDiagramInWith :: (Point_ point 2 r, Functor nonEmpty, Ord point, Ord r, Fractional r, Foldable1 nonEmpty) => (Triangle corner -> nonEmpty (Plane r :+ point) -> ClippedMinimizationDiagram (Plane r :+ point)) -> Triangle corner -> nonEmpty point -> ClippedVoronoiDiagram point

Given a function to compute a lower envelope; construct use it to construct the Voronoi diagram inside the given triangle.