Ipe

Description

Reexports the functionality for reading and writing Ipe files.

Ipe Files

data IpeFile r

A complete ipe file

Constructors

IpeFile (Maybe IpePreamble) [IpeStyle] (NonEmpty (IpePage r))

preamble :: forall r f. Functor f => (Maybe IpePreamble -> f (Maybe IpePreamble)) -> IpeFile r -> f (IpeFile r)

Lens to access the preamble of an ipe file

styles :: forall r f. Functor f => ([IpeStyle] -> f [IpeStyle]) -> IpeFile r -> f (IpeFile r)

Lens to access the styles of an ipe file

pages :: forall r r' f. Functor f => (NonEmpty (IpePage r) -> f (NonEmpty (IpePage r'))) -> IpeFile r -> f (IpeFile r')

Lens to access the pages of an ipe file

ipeFile :: NonEmpty (IpePage r) -> IpeFile r

Convenience constructor for creating an ipe file without preamble and with the default stylesheet.

singlePageFile :: IpePage r -> IpeFile r

Convenience function to construct an ipe file consisting of a single page.

singlePageFromContent :: [IpeObject r] -> IpeFile r

Create a single page ipe file from a list of IpeObjects

Reading Ipe files

readIpeFile :: (Coordinate r, Eq r) => OsPath -> IO (Either ConversionError (IpeFile r))

Given a file path, tries to read an ipe file.

This function applies all matrices to objects.

readSinglePageFile :: (Coordinate r, Eq r) => OsPath -> IO (Either ConversionError (IpePage r))

Since most Ipe file contain only one page, we provide a shortcut for that as well.

This function applies all matrices, and it makes sure there is at least one layer and view in the page.

readSinglePageFileThrow :: (Coordinate r, Eq r) => OsPath -> IO (IpePage r)

Tries to read a single page file, throws an error when this fails. See readSinglePageFile for further details.

readRawIpeFile :: (Coordinate r, Eq r) => OsPath -> IO (Either ConversionError (IpeFile r))

Given a file path, tries to read an ipe file

type ConversionError = Text

Reading all Geometries from a single page ipe file

readAll :: (HasDefaultFromIpe g, r ~ NumType g) => IpePage r -> [g :+ IpeAttributes (DefaultFromIpe g) r]

Read all g's from some ipe page(s).

readAllDeep :: (HasDefaultFromIpe g, r ~ NumType g) => IpePage r -> [g :+ IpeAttributes (DefaultFromIpe g) r]

Read all as looking into groups

readAllFrom :: (HasDefaultFromIpe g, r ~ NumType g, Coordinate r, Eq r) => OsPath -> IO [g :+ IpeAttributes (DefaultFromIpe g) r]

Convenience function from reading all g's from an ipe file. If there is an error reading or parsing the file the error is "thrown away".

Writing ipe files

writeIpeFile :: IpeWriteText r => OsPath -> IpeFile r -> IO ()

Write an IpeFiele to file.

writeIpeFile' :: IpeWrite t => t -> OsPath -> IO ()

Convert to ipe XML and write the output to a file.

writeIpePage :: IpeWriteText r => OsPath -> IpePage r -> IO ()

Creates a single page ipe file with the given page

toIpeXML :: IpeWrite t => t -> Maybe ByteString

Convert to Ipe xml

printAsIpeSelection :: IpeWrite t => t -> IO ()

Convert the input to ipeXml, and prints it to standard out in such a way that the copied text can be pasted into ipe as a geometry object.

toIpeSelectionXML :: IpeWrite t => t -> Maybe ByteString

Convert input into an ipe selection.

Ipe Pages

data IpePage r

An IpePage is essentially a Group, together with a list of layers and a list of views.

Constructors

IpePage [LayerName] [View] [IpeObject r]

layers :: forall r f. Functor f => ([LayerName] -> f [LayerName]) -> IpePage r -> f (IpePage r)

Lens to access the layers of an ipe page

views :: forall r f. Functor f => ([View] -> f [View]) -> IpePage r -> f (IpePage r)

Lens to access the views of an ipe page

content :: forall r r' f. Functor f => ([IpeObject r] -> f [IpeObject r']) -> IpePage r -> f (IpePage r')

Lens to access the content of an ipe page

emptyPage :: IpePage r

Creates an empty page with one layer and view.

fromContent :: [IpeObject r] -> IpePage r

Creates a simple page with a single view.

onLayer :: LayerName -> Getting (Endo [IpeObject r]) [IpeObject r] (IpeObject r)

This allows you to filter the objects on some layer.

>>> let page = IpePage [] [] []
>>> page^..content.onLayer "myLayer"
[]

contentInView :: Word -> Getter (IpePage r) [IpeObject r]

Gets all objects that are visible in the given view.

Note that views are indexed starting from 0. If the page does not have any explicit view definitions, this function returns an empty list.

>>> let page = IpePage [] [] []
>>> page^.contentInView 0
[]

withDefaults :: IpePage r -> IpePage r

Makes sure that the page has at least one layer and at least one view, essentially matching the behaviour of ipe. In particular,

if the page does not have any layers, it creates a layer named "alpha", and

if the page does not have any views, it creates a view in which all layers are visible.

Content: Ipe Objects

data IpeObject r

Constructors

IpeGroup (IpeObject' Group r)

IpeImage (IpeObject' Image r)

IpeTextLabel (IpeObject' TextLabel r)

IpeMiniPage (IpeObject' MiniPage r)

IpeUse (IpeObject' IpeSymbol r)

IpePath (IpeObject' Path r)

_IpePath :: forall r p f. (Choice p, Applicative f) => p (IpeObject' Path r) (f (IpeObject' Path r)) -> p (IpeObject r) (f (IpeObject r))

_IpeUse :: forall r p f. (Choice p, Applicative f) => p (IpeObject' IpeSymbol r) (f (IpeObject' IpeSymbol r)) -> p (IpeObject r) (f (IpeObject r))

_IpeGroup :: forall r p f. (Choice p, Applicative f) => p (IpeObject' Group r) (f (IpeObject' Group r)) -> p (IpeObject r) (f (IpeObject r))

_IpeTextLabel :: forall r p f. (Choice p, Applicative f) => p (IpeObject' TextLabel r) (f (IpeObject' TextLabel r)) -> p (IpeObject r) (f (IpeObject r))

_IpeMiniPage :: forall r p f. (Choice p, Applicative f) => p (IpeObject' MiniPage r) (f (IpeObject' MiniPage r)) -> p (IpeObject r) (f (IpeObject r))

_IpeImage :: forall r p f. (Choice p, Applicative f) => p (IpeObject' Image r) (f (IpeObject' Image r)) -> p (IpeObject r) (f (IpeObject r))

type IpeObject' (g :: Type -> Type) r = g r :+ IpeAttributes g r

An IpeObject' is essentially the oject ogether with its attributes

ipeObject' :: ToObject i => i r -> IpeAttributes i r -> IpeObject r

Shorthand for constructing ipeObjects

class ToObject (i :: Type -> Type) where

Minimal complete definition

_IpeObject

Methods

_IpeObject :: Prism' (IpeObject r) (i r :+ IpeAttributes i r)

mkIpeObject :: IpeObject' i r -> IpeObject r

Specific Ipe-Objects

newtype Path r

A path is a non-empty sequence of PathSegments.

Constructors

Path (Seq (PathSegment r))

pathSegments :: forall r r' p f. (Profunctor p, Functor f) => p (Seq (PathSegment r)) (f (Seq (PathSegment r'))) -> p (Path r) (f (Path r'))

Lens/Iso to access the sequcne of segments of the path

data PathSegment r

Paths consist of Path Segments. PathSegments come in the following forms:

Constructors

PolyLineSegment (PolyLine (Point 2 r))

PolygonPath !Orientation (SimplePolygon (Point 2 r))

CubicBezierSegment (CubicBezier (Point 2 r))

QuadraticBezierSegment (QuadraticBezier (Point 2 r))

EllipseSegment (Ellipse r)

ArcSegment

SplineSegment

ClosedSplineSegment

data IpeSymbol r

Ipe Symbols, i.e. Points

A symbol (point) in ipe

Constructors

Symbol (Point 2 r) Text

symbolPoint :: forall r r' f. Functor f => (Point 2 r -> f (Point 2 r')) -> IpeSymbol r -> f (IpeSymbol r')

Lens to access the position of the symbol

symbolName :: forall r f. Functor f => (Text -> f Text) -> IpeSymbol r -> f (IpeSymbol r)

Lens to access the name of the symbol

newtype Group r

Groups and Objects

Group Attributes

A group is essentially a list of IpeObjects.

Constructors

Group [IpeObject r]

groupItems :: forall r s f. Functor f => ([IpeObject r] -> f [IpeObject s]) -> Group r -> f (Group s)

data TextLabel r

Text Objects

A text label

Constructors

Label Text (Point 2 r)

data MiniPage r

A Minipage

Constructors

MiniPage Text (Point 2 r) r

width :: MiniPage t -> t

data Image r

Image Objects

bitmap image objects in Ipe

Constructors

Image () (Rectangle (Point 2 r))

imageData :: forall r f. Functor f => (() -> f ()) -> Image r -> f (Image r)

Lens to access the image data

imageRect :: forall r r' f. Functor f => (Rectangle (Point 2 r) -> f (Rectangle (Point 2 r'))) -> Image r -> f (Image r')

Lens to access the rectangle of the image

type IpeBitmap = Text

Ipe Bitmap data

Attributes

type family IpeAttributes (g :: Type -> Type) r where ...

Equations

IpeAttributes Group r = GroupAttributes r

IpeAttributes Image r = ImageAttributes r

IpeAttributes TextLabel r = TextAttributes r

IpeAttributes MiniPage r = TextAttributes r

IpeAttributes Path r = PathAttributes r

IpeAttributes IpeSymbol r = SymbolAttributes r

attributes :: forall i r f. Functor f => (IpeAttributes i r -> f (IpeAttributes i r)) -> IpeObject' i r -> f (IpeObject' i r)

Access the attributes

Layers and Views

newtype LayerName

Defines an Layer in Ipe.

Constructors

LayerName Text

layerName :: Iso' LayerName Text

data View

The definition of a view make active layer into an index ?

Constructors

View [LayerName] LayerName

layerNames :: Lens' View [LayerName]

Lens to access the layers in this view

activeLayer :: Lens' View LayerName

Lens to access the active layer

Ipe Syles and Preamble

data IpeStyle

for now we pretty much ignore these

Constructors

IpeStyle (Maybe Text) (Node Text Text)

styleName :: Lens' IpeStyle (Maybe Text)

Lens to access the style name

styleData :: Lens' IpeStyle (Node Text Text)

Lens to access the style data

basicIpeStyle :: IpeStyle

The "basic" ipe stylesheet

opacitiesStyle :: IpeStyle

A stylesheet with some convenient predefined opacities. In particular the opacities "10%","20%",..,"90%".

readIpeStylesheet :: OsPath -> IO (Either ConversionError IpeStyle)

Reads an Ipe stylesheet from Disk.

addStyleSheet :: IpeStyle -> IpeFile r -> IpeFile r

Adds a stylesheet to the ipe file. This will be the first stylesheet, i.e. it has priority over all previously imported stylesheets.

addStyleSheetFrom :: OsPath -> IpeFile r -> IO (IpeFile r)

Given a path to a stylesheet, add it to the ipe file with the highest priority. Throws an error when this fails.

data IpePreamble

The maybe string is the encoding

Constructors

IpePreamble (Maybe Text) Text

encoding :: Lens' IpePreamble (Maybe Text)

Lens to access the encoding

preambleData :: Lens' IpePreamble Text

Lens to access the preambleData

Reading Geometries *From* Ipe

class IpeRead t where

Reading an ipe lement from Xml

Methods

ipeRead :: Node Text Text -> Either ConversionError t

Try to convert an xml tree into a value of type t

Converting *from* IpeObjects

_asPoint :: forall r p f. (Choice p, Applicative f) => p (Point 2 r) (f (Point 2 r)) -> p (IpeSymbol r) (f (IpeSymbol r))

Extracts the point from a Symbol. When creating a symbol this creates a disk that supports a stroke color.

_asLineSegment :: forall r p f. (Choice p, Applicative f) => p (LineSegment AnEndPoint (Point 2 r)) (f (LineSegment AnEndPoint (Point 2 r))) -> p (Path r) (f (Path r))

Try to convert a path into a line segment, fails if the path is not a line segment or a polyline with more than two points.

_asClosedLineSegment :: forall r p f. (Choice p, Applicative f) => p (ClosedLineSegment (Point 2 r)) (f (ClosedLineSegment (Point 2 r))) -> p (Path r) (f (Path r))

Try to convert a path into a line segment, fails if the path is not a line segment or a polyline with more than two points.

_asHalfLine :: (Fractional r, Ord r, Show r) => Prism' (IpeObject r) (HalfLine (Point 2 r) :+ IpeAttributes Path r)

Try to parse an Line segment with an arrow head as a HalfLine

_asRectangle :: (Num r, Ord r) => Prism' (Path r) (Rectangle (Point 2 r))

Tries to convert a path into a rectangle.

_asTriangle :: forall r p f. (Choice p, Applicative f) => p (Triangle (Point 2 r)) (f (Triangle (Point 2 r))) -> p (Path r) (f (Path r))

Convert to a triangle

_asPolyLine :: forall r p f. (Choice p, Applicative f) => p (PolyLine (Point 2 r)) (f (PolyLine (Point 2 r))) -> p (Path r) (f (Path r))

Convert to a polyline. Ignores all non-polyline parts

>>> testPath ^? _asPolyLine
Just (PolyLine [Point2 0 0,Point2 10 10,Point2 200 100])

_asSimplePolygon :: forall (f :: Type -> Type) r. Foldable1 f => Prism (Path r) (Path r) (SimplePolygon (Point 2 r)) (SimplePolygonF f (Point 2 r))

Convert to a simple polygon

_asConvexPolygon :: (Num r, Ord r) => Prism' (Path r) (ConvexPolygon (Point 2 r))

Convert to a convex polygon

_asPolygonalDomain :: forall r p f. (Choice p, Applicative f) => p (PolygonalDomain (Point 2 r)) (f (PolygonalDomain (Point 2 r))) -> p (Path r) (f (Path r))

Convert to a polygonal domain

Dealing with Attributes

_withAttrs :: forall r (i :: Type -> Type) g. Prism' (IpeObject r) (i r :+ IpeAttributes i r) -> Prism' (i r) g -> Prism' (IpeObject r) (g :+ IpeAttributes i r)

Use the first prism to select the ipe object to depicle with, and the second how to select the geometry object from there on. Then we can select the geometry object, directly with its attributes here.

>>> testObject ^? _withAttrs _IpePath _asPolyLine
Just (PolyLine [Point2 0 0,Point2 10 10,Point2 200 100] :+ PathAttributes {_commonAttrs = CommonAttributes {_layer = Nothing, _matrix = Nothing, _pin = Nothing, _transformations = Nothing}, _stroke = Just (IpeColor (Named "red")), _fill = Nothing, _pen = Nothing, _dash = Nothing, _lineCap = Nothing, _lineJoin = Nothing, _fillRule = Nothing, _arrow = Nothing, _rArrow = Nothing, _strokeOpacity = Nothing, _opacity = Nothing, _tiling = Nothing, _gradient = Nothing})

Default readers

class HasDefaultFromIpe g where

Associated Types

type DefaultFromIpe g :: Type -> Type

Methods

defaultFromIpe :: r ~ NumType g => Prism' (IpeObject r) (g :+ IpeAttributes (DefaultFromIpe g) r)

Converting *to* IpeObjects

IpeWrite

class IpeWrite t where

Types that correspond to an XML Element. All instances should produce an Element. If the type should produce a Node with the Text constructor, use the IpeWriteText typeclass instead.

Methods

ipeWrite :: t -> Maybe (Node Text Text)

class IpeWriteText t where

For types that can produce a text value

Methods

ipeWriteText :: t -> Maybe Text

IpeOut

module Ipe.IpeOut

Batch reexports

module Ipe.Attributes

arrowName :: forall r f. Functor f => (Text -> f Text) -> IpeArrow r -> f (IpeArrow r)

arrowSize :: forall r1 r2 f. Functor f => (IpeSize r1 -> f (IpeSize r2)) -> IpeArrow r1 -> f (IpeArrow r2)

normalArrow :: IpeArrow r

A normal arrow

data FillType

Allowed Fill types

Constructors

Wind

EOFill

data HorizontalAlignment

Constructors

AlignLeft

AlignHCenter

AlignRight

data IpeArrow r

Possible values for an ipe arrow

Constructors

IpeArrow

Fields

_arrowName :: Text

_arrowSize :: IpeSize r

data IpeDash r

Possible values for Dash

Constructors

DashNamed Text

DashPattern [r] r

type IpeGradient = Text

type IpeOpacity = Text

IpeOpacity, IpeTyling, and IpeGradient are all symbolic values

newtype IpePen r

Pen/Thickness

Constructors

IpePen (IpeValue r)

newtype IpeSize r

The optional Attributes for a symbol data SymbolAttributeUniverse = SymbolStroke | SymbolFill | SymbolPen | Size deriving (Show,Eq)

Size

Constructors

IpeSize (IpeValue r)

type IpeTiling = Text

data PinType

Possible values for Pin

Constructors

No

Yes

Horizontal

Vertical

type TeXStyle = Text

Should be a symbolic name.

newtype TextSizeUnit r

size of text in points

Constructors

TextSizeUnit r

data TransformationTypes

Possible values for Transformation

Constructors

Affine

Rigid

Translations

data VerticalAlignment

Constructors

AlignTop

AlignVCenter

AlignBottom

AlignBaseline

module Ipe.Value

newtype IpeColor r

Defines a color in Ipe. Colors are either RGB Values or Named values.

Constructors

IpeColor (IpeValue (RGB r))

named :: Text -> IpeColor r

Creates a named color

Orphan instances

(Fractional r, Ord r, Show r) => HasDefaultFromIpe (HalfLine (Point 2 r))

Instance details

Associated Types

type DefaultFromIpe (HalfLine (Point 2 r))

Instance details

Defined in Ipe

type DefaultFromIpe (HalfLine (Point 2 r)) = Path

Methods

defaultFromIpe :: r0 ~ NumType (HalfLine (Point 2 r)) => Prism' (IpeObject r0) (HalfLine (Point 2 r) :+ IpeAttributes (DefaultFromIpe (HalfLine (Point 2 r))) r0)