Ipe.AllColors

Documentation

createIpeStyle

Arguments

:: IpeWriteText r

=> Text
the name of the stylesheet

-> Map Text (IpeValue (RGB r))

-> IpeStyle

Given a name and a map of colors, construct an ipe style

namedSet :: Map Text v -> Set (IpeColor r)

Helper for just producing a set of names

allColors :: Fractional r => Map Text (IpeValue (RGB r))

Defines all colors with their RGB value as defined in the colors.isy style file that comes with ipe

nonRepeated :: Map Text v -> Map Text v

only the non-repeated names; i.e. the ones without numerical suffix.

allColorNames :: Map Text ()

A map with all colors

allColorsWith :: (Rational -> Rational -> Rational -> r) -> Map Text r

The actual colors; needs some parsing function to produce an appropriate value

data RGB a

An RGB triple for an unspecified colour space.

Constructors

RGB

Fields

channelRed :: !a

channelGreen :: !a

channelBlue :: !a

cssStandardColors :: Fractional r => Map Text (IpeValue (RGB r))

The default CSS colors and their values

cssExtraColors :: Fractional r => Map Text (IpeValue (RGB r))

Extra CSS colors and their values