Ipe.Content

Documentation

data Image r

Image Objects

bitmap image objects in Ipe

Constructors

Image () (Rectangle (Point 2 r))

imageData :: forall r f. Functor f => (() -> f ()) -> Image r -> f (Image r)

Lens to access the image data

rect :: forall r r' f. Functor f => (Rectangle (Point 2 r) -> f (Rectangle (Point 2 r'))) -> Image r -> f (Image r')

Lens to access the rectangle of the image

data TextLabel r

Text Objects

A text label

Constructors

Label Text (Point 2 r)

data MiniPage r

A Minipage

Constructors

MiniPage Text (Point 2 r) r

width :: MiniPage t -> t

data IpeSymbol r

Ipe Symbols, i.e. Points

A symbol (point) in ipe

Constructors

Symbol (Point 2 r) Text

symbolPoint :: forall r r' f. Functor f => (Point 2 r -> f (Point 2 r')) -> IpeSymbol r -> f (IpeSymbol r')

Lens to access the position of the symbol

symbolName :: forall r f. Functor f => (Text -> f Text) -> IpeSymbol r -> f (IpeSymbol r)

Lens to access the name of the symbol

newtype Path r

A path is a non-empty sequence of PathSegments.

Constructors

Path (Seq (PathSegment r))

pathSegments :: forall r r' p f. (Profunctor p, Functor f) => p (Seq (PathSegment r)) (f (Seq (PathSegment r'))) -> p (Path r) (f (Path r'))

Lens/Iso to access the sequcne of segments of the path

data PathSegment r

Paths

Paths consist of Path Segments. PathSegments come in the following forms:

Constructors

PolyLineSegment (PolyLine (Point 2 r))

PolygonPath (SimplePolygon (Point 2 r))

CubicBezierSegment (CubicBezier (Point 2 r))

QuadraticBezierSegment (QuadraticBezier (Point 2 r))

EllipseSegment (Ellipse r)

ArcSegment

SplineSegment

ClosedSplineSegment

newtype Group r

Groups and Objects

Group Attributes

A group is essentially a list of IpeObjects.

Constructors

Group [IpeObject r]

groupItems :: forall r s f. Functor f => ([IpeObject r] -> f [IpeObject s]) -> Group r -> f (Group s)

data IpeObject r

Constructors

IpeGroup (IpeObject' Group r)

IpeImage (IpeObject' Image r)

IpeTextLabel (IpeObject' TextLabel r)

IpeMiniPage (IpeObject' MiniPage r)

IpeUse (IpeObject' IpeSymbol r)

IpePath (IpeObject' Path r)

_IpeGroup :: forall r p f. (Choice p, Applicative f) => p (IpeObject' Group r) (f (IpeObject' Group r)) -> p (IpeObject r) (f (IpeObject r))

_IpeImage :: forall r p f. (Choice p, Applicative f) => p (IpeObject' Image r) (f (IpeObject' Image r)) -> p (IpeObject r) (f (IpeObject r))

_IpeTextLabel :: forall r p f. (Choice p, Applicative f) => p (IpeObject' TextLabel r) (f (IpeObject' TextLabel r)) -> p (IpeObject r) (f (IpeObject r))

_IpeMiniPage :: forall r p f. (Choice p, Applicative f) => p (IpeObject' MiniPage r) (f (IpeObject' MiniPage r)) -> p (IpeObject r) (f (IpeObject r))

_IpeUse :: forall r p f. (Choice p, Applicative f) => p (IpeObject' IpeSymbol r) (f (IpeObject' IpeSymbol r)) -> p (IpeObject r) (f (IpeObject r))

_IpePath :: forall r p f. (Choice p, Applicative f) => p (IpeObject' Path r) (f (IpeObject' Path r)) -> p (IpeObject r) (f (IpeObject r))

type IpeObject' (g :: Type -> Type) r = g r :+ IpeAttributes g r

An IpeObject' is essentially the oject ogether with its attributes

ipeObject' :: ToObject i => i r -> IpeAttributes i r -> IpeObject r

Shorthand for constructing ipeObjects

class ToObject (i :: Type -> Type) where

Methods

mkIpeObject :: IpeObject' i r -> IpeObject r

type IpeAttributes (g :: Type -> Type) r = Attributes' r (AttributesOf g)

type Attributes' r = Attributes (AttrMapSym1 r)

Attributes' :: Type -> [AttributeUniverse] -> Type

type family AttributesOf (t :: Type -> Type) :: [AttributeUniverse] where ...

Equations

AttributesOf Group = GroupAttributes

AttributesOf Image = CommonAttributes

AttributesOf TextLabel = CommonAttributes

AttributesOf MiniPage = CommonAttributes

AttributesOf IpeSymbol = SymbolAttributes

AttributesOf Path = PathAttributes

type family AttrMap r (l :: AttributeUniverse) where ...

The mapping between the labels of the the attributes and the types of the attributes with these labels. For example, the Matrix label/attribute should have a value of type 'Matrix 3 3 r'.

Equations

AttrMap r 'Layer = LayerName

AttrMap r 'Matrix = Matrix 3 3 r

AttrMap r 'Pin = PinType

AttrMap r 'Transformations = TransformationTypes

AttrMap r 'Stroke = IpeColor r

AttrMap r 'Pen = IpePen r

AttrMap r 'Fill = IpeColor r

AttrMap r 'Size = IpeSize r

AttrMap r 'Dash = IpeDash r

AttrMap r 'LineCap = Int

AttrMap r 'LineJoin = Int

AttrMap r 'FillRule = FillType

AttrMap r 'Arrow = IpeArrow r

AttrMap r 'RArrow = IpeArrow r

AttrMap r 'StrokeOpacity = IpeOpacity

AttrMap r 'Opacity = IpeOpacity

AttrMap r 'Tiling = IpeTiling

AttrMap r 'Gradient = IpeGradient

AttrMap r 'Width = TextSizeUnit r

AttrMap r 'Height = TextSizeUnit r

AttrMap r 'Depth = TextSizeUnit r

AttrMap r 'VAlign = VerticalAlignment

AttrMap r 'HAlign = HorizontalAlignment

AttrMap r 'Style = TeXStyle

AttrMap r 'Clip = Path r

data AttrMapSym1 a6989586621679259897 (b :: TyFun AttributeUniverse Type)

attributes :: forall g r f. Functor f => (IpeAttributes g r -> f (IpeAttributes g r)) -> IpeObject' g r -> f (IpeObject' g r)

traverseIpeAttrs :: forall f (g :: Type -> Type) proxy r s. (Applicative f, AllConstrained TraverseIpeAttr (AttributesOf g)) => proxy g -> (r -> f s) -> IpeAttributes g r -> f (IpeAttributes g s)

traverse for ipe attributes

commonAttributes :: forall r f. Functor f => (Attributes (AttrMapSym1 r) CommonAttributes -> f (Attributes (AttrMapSym1 r) CommonAttributes)) -> IpeObject r -> f (IpeObject r)

flattenGroups :: [IpeObject r] -> [IpeObject r]

collect all non-group objects