Ipe.IpeOut

Description

Functions that help drawing geometric values in ipe. An IpeOut is essenitally a function that converts a geometric type g into an IpeObject.

We also proivde a HasDefaultIpeOut typeclass that defines a default conversion function from a geometry type g to an ipe type.

Documentation

>>> :set -XOverloadedStrings
>>> :{
let Just (myPolygon :: SimplePolygon (Point 2 Int)) = fromPoints $ [origin, Point2 10 10, Point2 100 200]
:}

The IpeOut type and the default combinator to use it

type IpeOut g (i :: Type -> Type) r = g -> IpeObject' i r

type IpeOut' (f :: Type -> Type) g (i :: Type -> Type) r = g -> f (IpeObject' i r)

Give the option to draw zero, one or more things, i.e. by choosing f ~ Maybe or f ~ []

(!) :: IpeObject' i r -> IpeAttributes i r -> IpeObject' i r

Add attributes to an IpeObject'

iO :: ToObject i => IpeObject' i r -> IpeObject r

Render an ipe object

>>> :{
  iO $ defIO myPolygon ! attr SFill (IpeColor "blue")
                       ! attr SLayer "alpha"
                       ! attr SLayer "beta"
:}
IpePath (Path {_pathSegments = fromList [PolygonPath (SimplePolygon [Point2 0 0,Point2 10 10,Point2 100 200])]} :+ Attrs {Attr LayerName {_layerName = "beta"}, NoAttr, NoAttr, NoAttr, NoAttr, Attr IpeColor (Named "blue"), NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr})

>>> :{
  iO $ ipeGroup [ iO $ ipePolygon myPolygon ! attr SFill (IpeColor "red")
                ] ! attr SLayer "alpha"
:}
IpeGroup (Group [IpePath (Path {_pathSegments = fromList [PolygonPath (SimplePolygon [Point2 0 0,Point2 10 10,Point2 100 200])]} :+ Attrs {NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, Attr IpeColor (Named "red"), NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr})] :+ Attrs {Attr LayerName {_layerName = "alpha"}, NoAttr, NoAttr, NoAttr, NoAttr})

iO'' :: forall g r (i :: Type -> Type). (HasDefaultIpeOut g, NumType g ~ r, DefaultIpeOut g ~ i, ToObject i) => g -> IpeAttributes i r -> IpeObject r

Render to an ipe object using the defIO IpeOut

>>> :{
  iO'' myPolygon $  attr SFill (IpeColor "red")
                 <> attr SLayer "alpha"
                 <> attr SLayer "beta"
:}
IpePath (Path {_pathSegments = fromList [PolygonPath (SimplePolygon [Point2 0 0,Point2 10 10,Point2 100 200])]} :+ Attrs {Attr LayerName {_layerName = "beta"}, NoAttr, NoAttr, NoAttr, NoAttr, Attr IpeColor (Named "red"), NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr})

>>> iO'' [ myPolygon , myPolygon ] $ attr SLayer "alpha"
IpeGroup (Group [IpePath (Path {_pathSegments = fromList [PolygonPath (SimplePolygon [Point2 0 0,Point2 10 10,Point2 100 200])]} :+ Attrs {NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr}),IpePath (Path {_pathSegments = fromList [PolygonPath (SimplePolygon [Point2 0 0,Point2 10 10,Point2 100 200])]} :+ Attrs {NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr})] :+ Attrs {Attr LayerName {_layerName = "alpha"}, NoAttr, NoAttr, NoAttr, NoAttr})

iO' :: HasDefaultIpeOut g => g -> IpeObject (NumType g)

generate an ipe object without any specific attributes

Default Conversions

class ToObject (DefaultIpeOut g) => HasDefaultIpeOut g where

Class that specifies a default conversion from a geometry type g into an ipe object.

Associated Types

type DefaultIpeOut g :: Type -> Type

Methods

defIO :: IpeOut g (DefaultIpeOut g) (NumType g)

Point Converters

ipeMark :: Text -> IpeOut (Point 2 r) IpeSymbol r

ipeDiskMark :: IpeOut (Point 2 r) IpeSymbol r

Path Converters

defaultBox :: Num r => Rectangle (Point 2 r)

Size of the default bounding box used to clip lines and half-lines in the default IpeOuts.

ipeLine :: (Ord r, Fractional r, Show r) => IpeOut (LinePV 2 r) Path r

Renders a line as a Path. The line is clipped to the defaultBox

ipeLineIn :: (Ord r, Fractional r, Show r) => Rectangle (Point 2 r) -> IpeOut (LinePV 2 r) Path r

Renders the line in the given box.

pre: the intersection of the box with the line is non-empty

ipeHalfLine :: (Ord r, Fractional r, Point_ point 2 r, Show r, Show point) => IpeOut (HalfLine point) Path r

Renders an Halfine.

pre: the intersection of the box with the line is non-empty

ipeRay :: (Ord r, Fractional r, Point_ point 2 r, Show r, Show point) => IpeOut (HalfLine point) Path r

Renders a ray, i.e. a half line drawing an arrow in the direction of the ray.

pre: the intersection of the box with the line is non-empty

ipeHalfLineIn :: (Ord r, Fractional r, Point_ point 2 r, Show r, Show point) => Rectangle point -> IpeOut (HalfLine point) Path r

Renders the HalfLine in the given box.

pre: the intersection of the box with the halfline is a line segment

ipeLineSegment :: (LineSegment_ lineSegment point, Point_ point 2 r) => IpeOut lineSegment Path r

Renders an line segment to a Path

ipePolyLine :: IpeOut (PolyLine (Point 2 r)) Path r

Renders a polyline to a Path

ipeEllipse :: IpeOut (Ellipse r) Path r

Renders an Ellipse to a Path

ipeCircle :: Radical r => IpeOut (Circle (Point 2 r)) Path r

Renders a circle to a Path

ipeDisk :: Radical r => IpeOut (Disk (Point 2 r)) Path r

Renders a Disk to a Path

ipeBezier :: IpeOut (CubicBezier (Point 2 r)) Path r

Renders a Bezier curve to a Path

path :: PathSegment r -> Path r

Helper to construct a path from a singleton item

pathSegment :: (LineSegment_ lineSegment point, Point_ point 2 r) => lineSegment -> PathSegment r

Construct a PolyLine path segment

ipePolygon :: Polygon_ polygon point r => IpeOut polygon Path r

Render as a polygon

toPolygonPathSegment :: (HasOuterBoundary polygon, Point_ point 2 r, Vertex polygon ~ point) => polygon -> PathSegment r

Helper type to build a path segment

ipeSimplePolygon :: IpeOut (SimplePolygon (Point 2 r)) Path r

Draw a polygon

ipeRectangle :: Num r => IpeOut (Rectangle (Point 2 r)) Path r

Draw a Rectangle

Group Converters

ipeGroup :: Foldable f => IpeOut (f (IpeObject r)) Group r

Text Converters

ipeLabel :: IpeOut (Text :+ Point 2 r) TextLabel r

Creates an text label

labelled

Arguments

:: (Show lbl, NumType g ~ r, ToObject i)

=> (g -> Point 2 r)
where to place the label

-> IpeOut g i r
how to draw the geometric object

-> IpeOut (g :+ lbl) Group r

Annotate an IpeOut with a label

labelledWith

Arguments

:: (Show lbl, NumType g ~ r, ToObject i)

=> IpeAttributes TextLabel r
attributes for the label

-> (g -> Point 2 r)
where to place the label

-> IpeOut g i r
how to draw the geometric object

-> IpeOut (g :+ lbl) Group r

Annotate an IpeOut with a label