Ipe.IpeToIpe

Description

Use ipetoipe to generate pdf files.

Note that all functions in this module require that ipetoipe is installed (it is bundled with ipe) and available on the path.

Documentation

ipeToIpeWith

Arguments

:: Options
the options to use

-> FileType
output file type

-> OsPath
input file path

-> OsPath
output file path

-> IO ()

Call ipetoipe to produce an image of the specified type.

ipeToIpe :: FileType -> OsPath -> OsPath -> IO ()

Call ipetoipe with the default options.

writeIpeFileAsPdf :: IpeWriteText r => OsPath -> IpeFile r -> IO ()

Write an ipe file to disk as an ipe readable pdf file.

Note that like all other functions in this module, this requires ipetoipe to be installed (it is bundled with ipe) and available on the path.

Note this will write soem intermediate file to your system temp dir.

data FileType

Constructors

PDF

XML

data Export

Constructors

RetainIpeInfo

Export

type PageNumber = Int

type ViewNumber = Int

data MarkedView

Constructors

All

OnlyMarkedViews

data NoZip

Constructors

NoZip

Zip

data PageRange

Constructors

EntireFile

PageRange (ClosedInterval PageNumber)

data Options

Constructors

Options

Fields

export :: Export

pages :: PageRange

singleView :: Maybe (PageNumber, ViewNumber)

markedView :: MarkedView

runLatex :: Bool

nozip :: NoZip

defaultOptions :: Options