Ipe.ParserPrimitives

Documentation

runP :: Parser a -> Text -> Either ParseError (a, Text)

runP' :: Parser a -> Text -> (a, Text)

pMany :: Parser a -> Parser [a]

reexporting some standard combinators

pMany1 :: Parser a -> Parser [a]

pChoice :: [Parser a] -> Parser a

pChar :: Char -> Parser Char

pSpace :: Parser Char

pWhiteSpace :: Parser [Char]

pInteger :: Parser Integer

pNatural :: Parser Integer

pPaddedNatural :: Parser (Int, Integer)

parses an integer with a prefix of zeros. Returns the total length of the string parced (i.e. number of digits) and the resulting antural number.

(<*><>) :: (Reversable s, Stream s m t) => ParsecT s u m (a -> b) -> ParsecT s u m a -> ParsecT s u m b infix 1

Running parsers in reverse

Runs parser q ``in reverse'' on the end of the input stream

(<*><) :: (Stream s m t, Reversable s) => ParsecT s u m b -> ParsecT s u m a -> ParsecT s u m b infix 1

(<***>) :: Monad m => m (t -> b) -> m t -> m b infixr 2

run the parsers in reverse order, first q, then p

(<***) :: Monad m => m b -> m t -> m b infixr 2

(***>) :: Monad m => m a -> m b -> m b infixr 2

the variants with missing brackets

pMaybe :: Parser a -> Parser (Maybe a)

pCount :: Int -> Parser a -> Parser [a]

pSepBy :: Parser a -> Parser b -> Parser [a]

type Parser = Parsec Text ()

data ParseError

The abstract data type ParseError represents parse errors. It provides the source position (SourcePos) of the error and a list of error messages (Message). A ParseError can be returned by the function parse. ParseError is an instance of the Show and Eq classes.

pNotFollowedBy :: Parser a -> Parser b -> Parser a

infix variant of notfollowed by