Ipe.Types

Description

Data type modeling the various elements in Ipe files.

Ipe Files

data IpeFile r

A complete ipe file

Constructors

IpeFile (Maybe IpePreamble) [IpeStyle] (NonEmpty (IpePage r))

preamble :: forall r f. Functor f => (Maybe IpePreamble -> f (Maybe IpePreamble)) -> IpeFile r -> f (IpeFile r)

Lens to access the preamble of an ipe file

styles :: forall r f. Functor f => ([IpeStyle] -> f [IpeStyle]) -> IpeFile r -> f (IpeFile r)

Lens to access the styles of an ipe file

pages :: forall r r' f. Functor f => (NonEmpty (IpePage r) -> f (NonEmpty (IpePage r'))) -> IpeFile r -> f (IpeFile r')

Lens to access the pages of an ipe file

ipeFile :: NonEmpty (IpePage r) -> IpeFile r

Convenience constructor for creating an ipe file without preamble and with the default stylesheet.

singlePageFile :: IpePage r -> IpeFile r

Convenience function to construct an ipe file consisting of a single page.

singlePageFromContent :: [IpeObject r] -> IpeFile r

Create a single page ipe file from a list of IpeObjects

Ipe Pages

data IpePage r

An IpePage is essentially a Group, together with a list of layers and a list of views.

Constructors

IpePage [LayerName] [View] [IpeObject r]

layers :: forall r f. Functor f => ([LayerName] -> f [LayerName]) -> IpePage r -> f (IpePage r)

Lens to access the layers of an ipe page

views :: forall r f. Functor f => ([View] -> f [View]) -> IpePage r -> f (IpePage r)

Lens to access the views of an ipe page

content :: forall r r' f. Functor f => ([IpeObject r] -> f [IpeObject r']) -> IpePage r -> f (IpePage r')

Lens to access the content of an ipe page

emptyPage :: IpePage r

Creates an empty page with one layer and view.

fromContent :: [IpeObject r] -> IpePage r

Creates a simple page with a single view.

onLayer :: LayerName -> Getting (Endo [IpeObject r]) [IpeObject r] (IpeObject r)

This allows you to filter the objects on some layer.

>>> let page = IpePage [] [] []
>>> page^..content.onLayer "myLayer"
[]

contentInView :: Word -> Getter (IpePage r) [IpeObject r]

Gets all objects that are visible in the given view.

Note that views are indexed starting from 0. If the page does not have any explicit view definitions, this function returns an empty list.

>>> let page = IpePage [] [] []
>>> page^.contentInView 0
[]

withDefaults :: IpePage r -> IpePage r

Makes sure that the page has at least one layer and at least one view, essentially matching the behaviour of ipe. In particular,

if the page does not have any layers, it creates a layer named "alpha", and

if the page does not have any views, it creates a view in which all layers are visible.

Content: Ipe Objects

data IpeObject r

Constructors

IpeGroup (IpeObject' Group r)

IpeImage (IpeObject' Image r)

IpeTextLabel (IpeObject' TextLabel r)

IpeMiniPage (IpeObject' MiniPage r)

IpeUse (IpeObject' IpeSymbol r)

IpePath (IpeObject' Path r)

_IpeGroup :: forall r p f. (Choice p, Applicative f) => p (IpeObject' Group r) (f (IpeObject' Group r)) -> p (IpeObject r) (f (IpeObject r))

_IpeImage :: forall r p f. (Choice p, Applicative f) => p (IpeObject' Image r) (f (IpeObject' Image r)) -> p (IpeObject r) (f (IpeObject r))

_IpeTextLabel :: forall r p f. (Choice p, Applicative f) => p (IpeObject' TextLabel r) (f (IpeObject' TextLabel r)) -> p (IpeObject r) (f (IpeObject r))

_IpeMiniPage :: forall r p f. (Choice p, Applicative f) => p (IpeObject' MiniPage r) (f (IpeObject' MiniPage r)) -> p (IpeObject r) (f (IpeObject r))

_IpeUse :: forall r p f. (Choice p, Applicative f) => p (IpeObject' IpeSymbol r) (f (IpeObject' IpeSymbol r)) -> p (IpeObject r) (f (IpeObject r))

_IpePath :: forall r p f. (Choice p, Applicative f) => p (IpeObject' Path r) (f (IpeObject' Path r)) -> p (IpeObject r) (f (IpeObject r))

type IpeObject' (g :: Type -> Type) r = g r :+ IpeAttributes g r

An IpeObject' is essentially the oject ogether with its attributes

ipeObject' :: ToObject i => i r -> IpeAttributes i r -> IpeObject r

Shorthand for constructing ipeObjects

class ToObject (i :: Type -> Type) where

Methods

mkIpeObject :: IpeObject' i r -> IpeObject r

Specific Ipe-Objects

newtype Path r

A path is a non-empty sequence of PathSegments.

Constructors

Path (Seq (PathSegment r))

pathSegments :: forall r r' p f. (Profunctor p, Functor f) => p (Seq (PathSegment r)) (f (Seq (PathSegment r'))) -> p (Path r) (f (Path r'))

Lens/Iso to access the sequcne of segments of the path

data PathSegment r

Paths

Paths consist of Path Segments. PathSegments come in the following forms:

Constructors

PolyLineSegment (PolyLine (Point 2 r))

PolygonPath (SimplePolygon (Point 2 r))

CubicBezierSegment (CubicBezier (Point 2 r))

QuadraticBezierSegment (QuadraticBezier (Point 2 r))

EllipseSegment (Ellipse r)

ArcSegment

SplineSegment

ClosedSplineSegment

data IpeSymbol r

Ipe Symbols, i.e. Points

A symbol (point) in ipe

Constructors

Symbol (Point 2 r) Text

symbolPoint :: forall r r' f. Functor f => (Point 2 r -> f (Point 2 r')) -> IpeSymbol r -> f (IpeSymbol r')

Lens to access the position of the symbol

symbolName :: forall r f. Functor f => (Text -> f Text) -> IpeSymbol r -> f (IpeSymbol r)

Lens to access the name of the symbol

newtype Group r

Groups and Objects

Group Attributes

A group is essentially a list of IpeObjects.

Constructors

Group [IpeObject r]

groupItems :: forall r s f. Functor f => ([IpeObject r] -> f [IpeObject s]) -> Group r -> f (Group s)

data TextLabel r

Text Objects

A text label

Constructors

Label Text (Point 2 r)

data MiniPage r

A Minipage

Constructors

MiniPage Text (Point 2 r) r

width :: MiniPage t -> t

data Image r

Image Objects

bitmap image objects in Ipe

Constructors

Image () (Rectangle (Point 2 r))

imageData :: forall r f. Functor f => (() -> f ()) -> Image r -> f (Image r)

Lens to access the image data

rect :: forall r r' f. Functor f => (Rectangle (Point 2 r) -> f (Rectangle (Point 2 r'))) -> Image r -> f (Image r')

Lens to access the rectangle of the image

type IpeBitmap = Text

Ipe Bitmap data

Attributes

type IpeAttributes (g :: Type -> Type) r = Attributes' r (AttributesOf g)

type Attributes' r = Attributes (AttrMapSym1 r)

Attributes' :: Type -> [AttributeUniverse] -> Type

type family AttributesOf (t :: Type -> Type) :: [AttributeUniverse] where ...

Equations

AttributesOf Group = GroupAttributes

AttributesOf Image = CommonAttributes

AttributesOf TextLabel = CommonAttributes

AttributesOf MiniPage = CommonAttributes

AttributesOf IpeSymbol = SymbolAttributes

AttributesOf Path = PathAttributes

type family AttrMap r (l :: AttributeUniverse) where ...

The mapping between the labels of the the attributes and the types of the attributes with these labels. For example, the Matrix label/attribute should have a value of type 'Matrix 3 3 r'.

Equations

AttrMap r 'Layer = LayerName

AttrMap r 'Matrix = Matrix 3 3 r

AttrMap r 'Pin = PinType

AttrMap r 'Transformations = TransformationTypes

AttrMap r 'Stroke = IpeColor r

AttrMap r 'Pen = IpePen r

AttrMap r 'Fill = IpeColor r

AttrMap r 'Size = IpeSize r

AttrMap r 'Dash = IpeDash r

AttrMap r 'LineCap = Int

AttrMap r 'LineJoin = Int

AttrMap r 'FillRule = FillType

AttrMap r 'Arrow = IpeArrow r

AttrMap r 'RArrow = IpeArrow r

AttrMap r 'StrokeOpacity = IpeOpacity

AttrMap r 'Opacity = IpeOpacity

AttrMap r 'Tiling = IpeTiling

AttrMap r 'Gradient = IpeGradient

AttrMap r 'Width = TextSizeUnit r

AttrMap r 'Height = TextSizeUnit r

AttrMap r 'Depth = TextSizeUnit r

AttrMap r 'VAlign = VerticalAlignment

AttrMap r 'HAlign = HorizontalAlignment

AttrMap r 'Style = TeXStyle

AttrMap r 'Clip = Path r

data AttrMapSym1 a6989586621679259243 (b :: TyFun AttributeUniverse Type)

attributes :: forall g r f. Functor f => (IpeAttributes g r -> f (IpeAttributes g r)) -> IpeObject' g r -> f (IpeObject' g r)

traverseIpeAttrs :: forall f (g :: Type -> Type) proxy r s. (Applicative f, AllConstrained TraverseIpeAttr (AttributesOf g)) => proxy g -> (r -> f s) -> IpeAttributes g r -> f (IpeAttributes g s)

traverse for ipe attributes

commonAttributes :: forall r f. Functor f => (Attributes (AttrMapSym1 r) CommonAttributes -> f (Attributes (AttrMapSym1 r) CommonAttributes)) -> IpeObject r -> f (IpeObject r)

Layers and Views

newtype LayerName

Defines an Layer in Ipe.

Constructors

LayerName Text

layerName :: Iso' LayerName Text

data View

The definition of a view make active layer into an index ?

Constructors

View [LayerName] LayerName

layerNames :: Lens' View [LayerName]

Lens to access the layers in this view

activeLayer :: Lens' View LayerName

Lens to access the active layer

Styles and Preamble

addStyleSheet :: IpeStyle -> IpeFile r -> IpeFile r

Adds a stylesheet to the ipe file. This will be the first stylesheet, i.e. it has priority over all previously imported stylesheets.

data IpeStyle

for now we pretty much ignore these

Constructors

IpeStyle (Maybe Text) (Node Text Text)

styleName :: Lens' IpeStyle (Maybe Text)

Lens to access the style name

styleData :: Lens' IpeStyle (Node Text Text)

Lens to access the style data

basicIpeStyle :: IpeStyle

The "basic" ipe stylesheet

opacitiesStyle :: IpeStyle

A stylesheet with some convenient predefined opacities. In particular the opacities "10%","20%",..,"90%".

data IpePreamble

The maybe string is the encoding

Constructors

IpePreamble (Maybe Text) Text

encoding :: Lens' IpePreamble (Maybe Text)

Lens to access the encoding

preambleData :: Lens' IpePreamble Text

Lens to access the preambleData