Algorithms.Geometry.LineSegmentIntersection

Description

Documentation

hasIntersections :: (Ord r, Num r) => [LineSegment 2 p r :+ e] -> Bool

Tests if there are any intersections.

\(O(n\log n)\)

intersections :: forall p r e. (Ord r, Fractional r) => [LineSegment 2 p r :+ e] -> Intersections p r e

Compute all intersections

\(O((n+k)\log n)\), where \(k\) is the number of intersections.

interiorIntersections :: (Ord r, Fractional r) => [LineSegment 2 p r :+ e] -> Intersections p r e

Computes all intersection points p s.t. p lies in the interior of at least one of the segments.

\(O((n+k)\log n)\), where \(k\) is the number of intersections.

type Intersections p r e = Map (Point 2 r) (Associated p r e)

For each intersection point the segments intersecting there.

data Associated p r e

The line segments that contain a given point p may either have p as the endpoint or have p in their interior.

if somehow the segment is degenerate, and p is both the start and end it is reported only as the start point.

Constructors

Associated

Fields

_startPointOf :: Set (AroundEnd (LineSegment 2 p r :+ e))

segments for which the intersection point is the start point (i.e. s^.start.core == p)

_endPointOf :: Set (AroundStart (LineSegment 2 p r :+ e))

segments for which the intersection point is the end point (i.e. s^.end.core == p)

_interiorTo :: Set (AroundIntersection (LineSegment 2 p r :+ e))

startPointOf :: forall p r e. Lens' (Associated p r e) (Set (AroundEnd ((:+) (LineSegment 2 p r) e)))

endPointOf :: forall p r e. Lens' (Associated p r e) (Set (AroundStart ((:+) (LineSegment 2 p r) e)))

interiorTo :: forall p r e. Lens' (Associated p r e) (Set (AroundIntersection ((:+) (LineSegment 2 p r) e)))

data IntersectionPoint p r e

An intersection point together with all segments intersecting at this point.

Constructors

IntersectionPoint

Fields

_intersectionPoint :: !(Point 2 r)

_associatedSegs :: !(Associated p r e)

mkIntersectionPoint

Arguments

:: (Ord r, Fractional r)

=> Point 2 r

-> [LineSegment 2 p r :+ e]
uncategorized

-> [LineSegment 2 p r :+ e]
segments we know contain p,

-> IntersectionPoint p r e

Given a point p, and a bunch of segments that suposedly intersect at p, correctly categorize them.

intersectionPoint :: forall p r e. Lens' (IntersectionPoint p r e) (Point 2 r)

associatedSegs :: forall p r e p e. Lens (IntersectionPoint p r e) (IntersectionPoint p r e) (Associated p r e) (Associated p r e)

hasSelfIntersections :: (Ord r, Fractional r) => Polygon t p r -> Bool

Test if the polygon has self intersections.

\(O(n \log n)\)

newtype AroundStart a

Assumes that two segments have the same start point

Constructors

AroundStart a

newtype AroundEnd a

Assumes that two segments have the same end point

Constructors

AroundEnd a

newtype AroundIntersection a

Assumes that two segments intersect in a single point.

Constructors

AroundIntersection a