Algorithms.Geometry.LineSegmentIntersection.BentleyOttmann

Description

The \(O((n+k)\log n)\) time line segment intersection algorithm by Bentley and Ottmann.

Documentation

intersections :: (Ord r, Fractional r) => [LineSegment 2 p r] -> Intersections p r

Compute all intersections

\(O((n+k)\log n)\), where \(k\) is the number of intersections.

interiorIntersections :: (Ord r, Fractional r) => [LineSegment 2 p r] -> Intersections p r

Computes all intersection points p s.t. p lies in the interior of at least one of the segments.

\(O((n+k)\log n)\), where \(k\) is the number of intersections.

asEventPts :: Ord r => LineSegment 2 p r -> [Event p r]

Computes the event points for a given line segment

merge :: Ord r => [IntersectionPoint p r] -> Intersections p r

Group the segments with the intersection points

groupStarts :: Eq r => [Event p r] -> [Event p r]

Group the startpoints such that segments with the same start point correspond to one event.

Data type for Events

data EventType s

Type of segment

Constructors

Start !(NonEmpty s)

Intersection

End !s

data Event p r

The actual event consists of a point and its type

Constructors

Event

Fields

eventPoint :: !(Point 2 r)

eventType :: !(EventType (LineSegment 2 p r))

ordPoints :: Ord r => Point 2 r -> Point 2 r -> Ordering

An ordering that is decreasing on y, increasing on x

startSegs :: Event p r -> [LineSegment 2 p r]

Get the segments that start at the given event point

ordAt :: (Fractional r, Ord r) => r -> Compare (LineSegment 2 p r)

Compare based on the x-coordinate of the intersection with the horizontal line through y

xCoordAt :: (Fractional r, Ord r) => r -> LineSegment 2 p r -> r

Given a y coord and a line segment that intersects the horizontal line through y, compute the x-coordinate of this intersection point.

note that we will pretend that the line segment is closed, even if it is not

The Main Sweep

type EventQueue p r = Set (Event p r)

type StatusStructure p r = Set (LineSegment 2 p r)

sweep :: (Ord r, Fractional r) => EventQueue p r -> StatusStructure p r -> [IntersectionPoint p r]

Run the sweep handling all events

isClosedStart :: Eq r => Point 2 r -> LineSegment 2 p r -> Bool

handle :: forall r p. (Ord r, Fractional r) => Event p r -> EventQueue p r -> StatusStructure p r -> [IntersectionPoint p r]

Handle an event point

extractContains :: (Fractional r, Ord r) => Point 2 r -> StatusStructure p r -> (StatusStructure p r, [LineSegment 2 p r], StatusStructure p r)

split the status structure, extracting the segments that contain p. the result is (before,contains,after)

toStatusStruct :: (Fractional r, Ord r) => Point 2 r -> [LineSegment 2 p r] -> StatusStructure p r

Given a point and the linesegements that contain it. Create a piece of status structure for it.

rightEndpoint :: Ord r => LineSegment 2 p r -> r

Get the right endpoint of a segment

endsAt :: Ord r => Point 2 r -> LineSegment 2 p r -> Bool

Test if a segment ends at p

Finding New events

findNewEvent :: (Ord r, Fractional r) => Point 2 r -> LineSegment 2 p r -> LineSegment 2 p r -> Maybe (Event p r)

Find all events