Data.Geometry.Vector.VectorFamilyPeano

Description

Natural number stuff

type One = S Z

type Two = S One

type Three = S Two

type Four = S Three

type Many d = S (S (S (S (S d))))

type family FromPeano (d :: PeanoNum) :: Nat where ...

Equations

FromPeano Z = 0

FromPeano (S d) = 1 + FromPeano d

data SingPeano (d :: PeanoNum) where

Constructors

SZ :: SingPeano Z

SS :: !(SingPeano d) -> SingPeano (S d)

class ImplicitPeano (d :: PeanoNum) where

Methods

implicitPeano :: SingPeano d

d dimensional Vectors

newtype VectorFamily (d :: PeanoNum) (r :: *)

Datatype representing d dimensional vectors. The default implementation is based n VectorFixed. However, for small vectors we automatically select a more efficient representation.

Constructors

VectorFamily

Fields

_unVF :: VectorFamilyF d r

type family VectorFamilyF (d :: PeanoNum) :: * -> * where ...

Mapping between the implementation type, and the actual implementation.

Equations

VectorFamilyF Z = Const ()

VectorFamilyF One = Identity

VectorFamilyF Two = V2

VectorFamilyF Three = V3

VectorFamilyF Four = V4

VectorFamilyF (Many d) = Vector (FromPeano (Many d))

unVF :: Lens (VectorFamily d r) (VectorFamily d t) (VectorFamilyF d r) (VectorFamilyF d t)

class (ImplicitPeano d, Arity (FromPeano d)) => ImplicitArity d

element' :: forall d r. ImplicitArity d => Int -> Traversal' (VectorFamily d r) r

elem0 :: Int -> Traversal' (VectorFamily Z r) r

elem1 :: Int -> Traversal' (VectorFamily One r) r

elem2 :: Int -> Traversal' (VectorFamily Two r) r

elem3 :: Int -> Traversal' (VectorFamily Three r) r

elem4 :: Int -> Traversal' (VectorFamily Four r) r

elemD :: Arity (FromPeano (Many d)) => Int -> Traversal' (VectorFamily (Many d) r) r

vectorFromList :: ImplicitArity d => [r] -> Maybe (VectorFamily d r)

vectorFromListUnsafe :: ImplicitArity d => [r] -> VectorFamily d r

destruct :: (ImplicitArity d, ImplicitArity (S d)) => VectorFamily (S d) r -> (r, VectorFamily d r)

Get the head and tail of a vector