HGeometry.Ball.CenterAndRadius

Description

Balls in d-dimensional space, represented by their center and squared radius.

Documentation

data Ball point

d-dimensional ball, specified by center and squared radius.

Constructors

Ball !point !(NumType point)

Bundled Patterns

pattern Disk :: point -> NumType point -> Disk point
Construct a disk

squaredRadius' :: forall point f. Functor f => (NumType point -> f (NumType point)) -> Ball point -> f (Ball point)

Lens to access the squared radius of a ball

type Disk = Ball

Balls in 2D are also known as Disks

newtype Sphere point

A sphere, i.e. the boudary of a Ball.

Constructors

MkSphere (Ball point)

Bundled Patterns

pattern Sphere :: point -> NumType point -> Sphere point
Construct a Sphere; the boundary of a ball

pattern Circle :: point -> NumType point -> Circle point
Construct a Circle

type Circle = Sphere

A circle, i.e. the boundary of a Disk

_BallSphere :: forall point point' p f. (Profunctor p, Functor f) => p (Sphere point) (f (Sphere point')) -> p (Ball point) (f (Ball point'))

Iso to convert between a ball and a sphere.

_DiskCircle :: forall point point' p f. (Profunctor p, Functor f) => p (Circle point) (f (Circle point')) -> p (Disk point) (f (Disk point'))

Iso to convert between a Disk and a Circle

fromDiametralPair :: forall r point (d :: Nat). (Fractional r, Point_ point d r, Has_ Metric_ d r) => point -> point -> Ball (Point d r)

Given two points on the diameter of the ball, construct a ball.

fromCenterAndPoint :: forall r point (d :: Nat) center. (Num r, Point_ point d r, Point_ center d r, Has_ Metric_ d r) => center -> point -> Ball center

Construct a ball given the center point and a point p on the boundary.

fromBoundaryPoints :: (Point_ point 2 r, Fractional r, Ord r) => Vector 3 point -> Maybe (Disk (Point 2 r))

Tries to create a disk from three points on the boundary.

data family IntersectionOf g h

The data family specifying to help implement the Intersection type family.