HGeometry.Ball.Class

Description

A class of types that can act as balls in \(d\)-dimensional space.

Documentation

class (HasCenter ball point, NumType ball ~ NumType point, Point_ point (Dimension ball) (NumType ball)) => Ball_ ball point | ball -> point where

Balls in d-dimensional space

Minimal complete definition

squaredRadius, fromCenterAndSquaredRadius

Methods

squaredRadius :: Lens' ball (NumType ball)

Lens to access the squared radius of a ball

fromCenterAndSquaredRadius :: point -> NumType ball -> ball

Construct a ball from the center and *squared* radius

unitBall :: (r ~ NumType ball, d ~ Dimension ball, Num r, ConstructablePoint_ point d r) => ball

A d dimensional unit ball centered at the origin.

radius :: Radical (NumType ball) => Lens' ball (NumType ball)

A lens to get/set the radius of a Ball

pattern Ball_ :: Ball_ ball point => point -> NumType ball -> ball

Constructs a Ball from its center and squared radius.

class HasCenter geom point | geom -> point where

Types that have a center field lens

Methods

center :: Lens' geom point

Lens to access the center point of a geometry

fromDiametralPair :: (Fractional r, Point_ point d r, Ball_ ball point, Has_ Metric_ d r) => point -> point -> ball

Given two points on the diameter of the ball, construct a ball.

fromCenterAndPoint :: (Num r, Point_ point d r, Ball_ ball point, Has_ Metric_ d r) => point -> point -> ball

Construct a ball given the center point and a point p on the boundary.

class (Ball_ disk point, Dimension disk ~ 2) => Disk_ disk point

Disks are two dimensional balls

fromBoundaryPoints :: (Disk_ disk (Point 2 r), Point_ point 2 r, Fractional r) => Vector 3 point -> disk

Creates a circle from three points on the boundary

pattern Disk_ :: Disk_ ball point => point -> NumType ball -> ball

Constructs a Disk by its center and squared radius.