HGeometry.Box

Description

d-dimensional Boxes

Documentation

module HGeometry.Box.Class

data Box point where

D-dimensional boxes.

A box is represented by two points; a point with lexicographically minimal coordinates, and a point with lexicographically maximal coordinates.

Bundled Patterns

pattern Box :: point -> point -> Box point
Construct a box

pattern Rectangle :: Dimension point ~ 2 => point -> point -> Box point
Construct a Rectangle

type Rectangle = Box

Defines a rectangle

fromExtent :: forall vector (d :: Nat) interval r. (Vector_ vector d interval, ClosedInterval_ interval r, Has_ Additive_ d r, Has_ Vector_ d interval, HasComponents (Vector d interval) (Vector d r)) => vector -> Box (Point d r)

Given a vector of intervals, construct a box.

>>> fromExtent (Vector2 (ClosedInterval 1 2) (ClosedInterval 10 20))
Box (Point2 1 10) (Point2 2 20)

data Corners a

A data type rperesenting the corners of a box. the order of the Corners is 'northWest, northEast, southEast, southWest', i.e. in clockwise order starting from the topleft.

Constructors

Corners !a !a !a !a

northWest :: forall a f. Functor f => (a -> f a) -> Corners a -> f (Corners a)

Access the northwest corner

northEast :: forall a f. Functor f => (a -> f a) -> Corners a -> f (Corners a)

Access the northeast corner

southEast :: forall a f. Functor f => (a -> f a) -> Corners a -> f (Corners a)

Access the southeast corner

southWest :: forall a f. Functor f => (a -> f a) -> Corners a -> f (Corners a)

Access the southwest corner

corners :: (Num r, Rectangle_ rectangle point, Point_ point 2 r) => rectangle -> Corners point

Get the corners of a rectangle, the order is:

Any additional information from in the top points is taken from the Top point, Any additional information from in the bottom points is taken from the Bottom point,

cornersInDirection :: CardinalDirection -> Corners p -> Vector 2 p

Gets the corners in a particular direction

data Sides a

The four sides of a rectangle

Constructors

Sides !a !a !a !a

north :: forall a f. Functor f => (a -> f a) -> Sides a -> f (Sides a)

Access the north side

east :: forall a f. Functor f => (a -> f a) -> Sides a -> f (Sides a)

Access the east side

south :: forall a f. Functor f => (a -> f a) -> Sides a -> f (Sides a)

Access the south side

west :: forall a f. Functor f => (a -> f a) -> Sides a -> f (Sides a)

Access the west side

topSide :: (Num r, Rectangle_ rectangle point, Point_ point 2 r) => rectangle -> ClosedLineSegment point

The top side of the box, from left to right.

bottomSide :: (Num r, Rectangle_ rectangle point, Point_ point 2 r) => rectangle -> ClosedLineSegment point

Oriented from *left to right*

leftSide :: (Num r, Rectangle_ rectangle point, Point_ point 2 r) => rectangle -> ClosedLineSegment point

Left side of the box, from bottom to top

rightSide :: (Num r, Rectangle_ rectangle point, Point_ point 2 r) => rectangle -> ClosedLineSegment point

The right side, oriented from *bottom* to top

sides :: (Num r, Rectangle_ rectangle point, Point_ point 2 r) => rectangle -> Sides (ClosedLineSegment point)

The sides of the rectangle, in order (Top, Right, Bottom, Left). The sides themselves are also oriented in clockwise order. If, you want them in the same order as the functions topSide, bottomSide, leftSide, and rightSide, use sides' instead.

sides' :: (Num r, Rectangle_ rectangle point, Point_ point 2 r) => rectangle -> Sides (ClosedLineSegment point)

The sides of the rectangle. The order of the segments is (Top, Right, Bottom, Left). Note that the segments themselves, are oriented as described by the functions topSide, bottomSide, leftSide, rightSide (basically: from left to right, and from bottom to top). If you want the segments oriented along the boundary of the rectangle, use the sides function instead.

sideDirections :: Sides CardinalDirection

Constructs a Sides value that indicates the appropriate direction.

sideValues :: (Rectangle_ rectangle point, Point_ point 2 r) => rectangle -> Sides r

The side values of the rectangle, i.e. the coordinates of the top, right, bottom, and left sides.

class IsBoxable g where

Types for which we can compute an axis parallel boundingbox

Minimal complete definition

Nothing

Methods

boundingBox :: forall (d :: Nat) r. (d ~ Dimension g, r ~ NumType g, Ord r) => g -> Box (Point d r)

Compute the axis-parallel boundingbox of the given geometry.

default boundingBox :: forall (d :: Nat) r point. (d ~ Dimension g, r ~ NumType g, Ord r, HasPoints' g point, Point_ point d r, Ord (Vector d r)) => g -> Box (Point d r)

data LineBoxIntersection (d :: Nat) r

Data type representing the intersection of a Box and a line

Constructors

Line_x_Box_Point (Point d r)

Line_x_Box_LineSegment (ClosedLineSegment (Point d r))

data HalfLineBoxIntersection point

Constructors

HalfLine_x_Box_Point point

HalfLine_x_Box_LineSegment (ClosedLineSegment point)

Orphan instances

(Point_ point d r, IsTransformable point) => IsTransformable (Box point)

Instance details

Methods

transformBy :: Transformation (Dimension (Box point)) (NumType (Box point)) -> Box point -> Box point

(Point_ point 2 r, Ord r, Num r) => HasIntersectionWith (HalfLine point) (Rectangle point)

Instance details

Methods

intersects :: HalfLine point -> Rectangle point -> Bool

(Num r, Ord r, Point_ point 2 r) => HasIntersectionWith (VerticalOrLineEQ r) (Rectangle point)

Instance details

Methods

intersects :: VerticalOrLineEQ r -> Rectangle point -> Bool

(Num r, Ord r, Point_ point 2 r) => HasIntersectionWith (LineEQ r) (Rectangle point)

Instance details

Methods

intersects :: LineEQ r -> Rectangle point -> Bool

(Point_ point 2 r, Ord r, Fractional r) => IsIntersectableWith (HalfLine point) (Rectangle point)

Instance details

Methods

intersect :: HalfLine point -> Rectangle point -> Intersection (HalfLine point) (Rectangle point)

(Fractional r, Ord r, Point_ point 2 r) => IsIntersectableWith (VerticalOrLineEQ r) (Rectangle point)

Instance details

Methods

intersect :: VerticalOrLineEQ r -> Rectangle point -> Intersection (VerticalOrLineEQ r) (Rectangle point)

(Fractional r, Ord r, Point_ point 2 r) => IsIntersectableWith (LineEQ r) (Rectangle point)

Instance details

Methods

intersect :: LineEQ r -> Rectangle point -> Intersection (LineEQ r) (Rectangle point)

(Num r, Ord r, Point_ point 2 r) => HasIntersectionWith (LinePV 2 r) (Rectangle point)

Instance details

Methods

intersects :: LinePV 2 r -> Rectangle point -> Bool

(Ord r, Num r, Point_ point 2 r, Point_ point' 2 r, HasIntersectionWith (LineSegment endPoint point) (ClosedLineSegment point')) => HasIntersectionWith (LineSegment endPoint point) (Boundary (Rectangle point'))

Instance details

Methods

intersects :: LineSegment endPoint point -> Boundary (Rectangle point') -> Bool

(Ord r, Num r, Point_ point 2 r, Point_ point' 2 r, IxValue (endPoint point) ~ point, EndPoint_ (endPoint point), HasIntersectionWith (LineSegment endPoint point) (ClosedLineSegment point')) => HasIntersectionWith (LineSegment endPoint point) (Rectangle point')

Instance details

Methods

intersects :: LineSegment endPoint point -> Rectangle point' -> Bool

(Point_ point d r, Num r, Ord r, Foldable1 (Vector d), Zip (Vector d)) => HasIntersectionWith (Point d r) (Box point)

Instance details

Methods

intersects :: Point d r -> Box point -> Bool

(Fractional r, Ord r, Point_ point 2 r) => IsIntersectableWith (LinePV 2 r) (Rectangle point)

Instance details

Methods

intersect :: LinePV 2 r -> Rectangle point -> Intersection (LinePV 2 r) (Rectangle point)

(Point_ point d r, Num r, Ord r, Foldable1 (Vector d), Zip (Vector d)) => IsIntersectableWith (Point d r) (Box point)

Instance details

Methods

intersect :: Point d r -> Box point -> Intersection (Point d r) (Box point)