HGeometry.Graphics.Camera

Description

Documentation

data Camera r

A basic camera data type. The fields stored are:

the camera position,

the raw camera normal, i.e. a unit vector into the center of the screen,

the raw view up vector indicating which side points "upwards" in the scene,

the focal depth (i.e. the distance from the camera position to the plane on which we project),

the near distance (everything closer than this is clipped),

the far distance (everything further away than this is clipped), and

the viewport dimensions; i.e. the viewport (on the view plane) on which we draw

Constructors

Camera !(Point 3 r) !(Vector 3 r) !(Vector 3 r) !r !r !r !(Vector 2 r)

cameraPosition :: forall r f. Functor f => (Point 3 r -> f (Point 3 r)) -> Camera r -> f (Camera r)

Camera position.

rawCameraNormal :: forall r f. Functor f => (Vector 3 r -> f (Vector 3 r)) -> Camera r -> f (Camera r)

Raw camera normal, i.e. a vector into the center of the screen.

rawViewUp :: forall r f. Functor f => (Vector 3 r -> f (Vector 3 r)) -> Camera r -> f (Camera r)

Raw view up vector indicating which side points "upwards" in the scene.

focalDepth :: forall r f. Functor f => (r -> f r) -> Camera r -> f (Camera r)

The focal length (i.e. distance from the camera position to the plane on which we project.) Also known as viewplane depth.

nearDist :: forall r f. Functor f => (r -> f r) -> Camera r -> f (Camera r)

Near distance (everything closer than this is clipped).

farDist :: forall r f. Functor f => (r -> f r) -> Camera r -> f (Camera r)

Far distance (everything further away than this is clipped).

viewportDimensions :: forall r f. Functor f => (Vector 2 r -> f (Vector 2 r)) -> Camera r -> f (Camera r)

The viewport dimensions. (i.e. the size of the screen on which we draw, in terms of world coordinates)

cameraNormal :: (Radical r, Fractional r) => Lens' (Camera r) (Vector 3 r)

Lens to get and set the Camera normal, makes sure that the vector remains normalized.

viewUp :: (Radical r, Fractional r) => Lens' (Camera r) (Vector 3 r)

Lens to get and set the viewUp vector. Makes sure the vector remains normalized.

blenderCamera :: (Floating r, Radical r) => Camera r

This is the default camera position used in Blender

blenderCameraAt

Arguments

:: (Floating r, Radical r)

=> Point 3 r
Initial position

-> Vector 3 r
Rotation angles in degrees

-> Camera r

Construct a "default" blender camera at the given position and rotation angles

The rotation angles are given in degrees with respect to:

x=0 means looking towards z=-infty, so along Vector3 0 0 (-1) z=0 means looking towards y=infy, so along Vecto3 0 1 0 (-- the default view dir)

cameraTransform :: Fractional r => Camera r -> Transformation 3 r

Full transformation that renders the figure

worldToView :: Fractional r => Camera r -> Transformation 3 r

Translates world coordinates into view coordinates

toViewportTransform :: Fractional r => Camera r -> Transformation 3 r

Transformation into viewport coordinates

perspectiveProjection :: Fractional r => Camera r -> Transformation 3 r

constructs a perspective projection

rotateCoordSystem :: Num r => Camera r -> Transformation 3 r

Rotates coordinate system around the camera, such that we look in the negative z direction

flipAxes :: Num r => Transformation 3 r

Flips the y and z axis.

viewportInWorld :: (Radical r, Floating r) => Camera r -> Corners (Point 3 r)

Computes the corners of the viewport; in world coordinates.