HGeometry.Graphics.Camera

Description

Documentation

data Camera r

A basic camera data type. The fields stored are:

the camera position,

the raw camera normal, i.e. a unit vector into the center of the screen,

the raw view up vector indicating which side points "upwards" in the scene,

the viewplane depth (i.e. the distance from the camera position to the plane on which we project),

the near distance (everything closer than this is clipped),

the far distance (everything further away than this is clipped), and

the screen dimensions.

Constructors

Camera !(Point 3 r) !(Vector 3 r) !(Vector 3 r) !r !r !r !(Vector 2 r)

cameraPosition :: Lens' (Camera r) (Point 3 r)

Camera position.

rawCameraNormal :: Lens' (Camera r) (Vector 3 r)

Raw camera normal, i.e. a unit vector into the center of the screen.

rawViewUp :: Lens' (Camera r) (Vector 3 r)

Raw view up vector indicating which side points "upwards" in the scene.

viewPlaneDepth :: Lens' (Camera r) r

Viewplane depth (i.e. the distance from the camera position to the plane on which we project).

nearDist :: Lens' (Camera r) r

Near distance (everything closer than this is clipped).

farDist :: Lens' (Camera r) r

Far distance (everything further away than this is clipped).

screenDimensions :: Lens' (Camera r) (Vector 2 r)

Screen dimensions.

cameraNormal :: (Radical r, Fractional r) => Lens' (Camera r) (Vector 3 r)

Lens to get and set the Camera normal, makes sure that the vector remains normalized.

viewUp :: (Radical r, Fractional r) => Lens' (Camera r) (Vector 3 r)

Lens to get and set the viewUp vector. Makes sure the vector remains normalized.

cameraTransform :: Fractional r => Camera r -> Transformation 3 r

Full transformation that renders the figure

worldToView :: Fractional r => Camera r -> Transformation 3 r

Translates world coordinates into view coordinates

toViewPort :: Fractional r => Camera r -> Transformation 3 r

Transformation into viewport coordinates

perspectiveProjection :: Fractional r => Camera r -> Transformation 3 r

constructs a perspective projection

rotateCoordSystem :: Num r => Camera r -> Transformation 3 r

Rotates coordinate system around the camera, such that we look in the negative z direction

flipAxes :: Num r => Transformation 3 r

Flips the y and z axis.