HGeometry.HalfLine

Description

Half-lines in \(d\)-dimensional space.

Documentation

data HalfLine point

A Halfline in R^d

Constructors

HalfLine !point !(Vector (Dimension point) (NumType point))

asOrientedLine :: forall point (d :: Nat) r. Point_ point d r => HalfLine point -> LinePV d r

Convert the Halfline into an oriented line.

halfLineThrough :: forall point (d :: Nat) r. (Point_ point d r, Num r) => point -> point -> HalfLine point

Given two points p and q, create a halfline from p through q.

>>> halfLineThrough (Point2 5 10) (Point2 10 30 :: Point 2 Int)
HalfLine (Point2 5 10) (Vector2 5 20)

data LineHalfLineIntersection point halfLine

Data type representing the intersection of a Line and a HalfLine

Constructors

Line_x_HalfLine_Point point

Line_x_HalfLine_HalfLine halfLine

class HasDirection t where

Types that have a Direction field

Methods

direction :: forall (d :: Nat) r. (Dimension t ~ d, NumType t ~ r) => Lens' t (Vector d r)

Lens to access the direction field