HGeometry.HalfPlane.CommonIntersection

Documentation

data CommonIntersection halfPlane r

Common intersection of a bunch of halfplanes

Constructors

EmptyIntersection

SingletonPoint (Point 2 r) (Vector 3 halfPlane)
Common intersection is a singleton point, defined by the three given halfplanes

InSubLine (VerticalOrLineEQ r) (Vector 2 halfPlane) (SubLine halfPlane r)
The two halfPlanes that define the line, and the other halfplanes furthe restricitng the line.

Slab halfPlane halfPlane
two parallel halfPlanes l and u that form a slab;

BoundedRegion (ConvexPolygon (Point 2 r :+ halfPlane))

UnboundedRegion (Chain Seq halfPlane r)
each vertex stores the interior halfplane of the CCW-edge it is incident to.

data SubLine halfPlane r

Part of the line

Constructors

EntireLine

InHalfLine (HalfLine (Point 2 r)) halfPlane
the third halfPlane

InSegment (ClosedLineSegment (Point 2 r)) halfPlane halfPlane
the remaining two halfplanes

newtype Chain (f :: Type -> Type) halfPlane r

A polygonal chain bounding an unbounded convex region, in CCW order.

Constructors

Chain (Alternating f (Point 2 r) halfPlane)

commonIntersection :: (Foldable1 f, Functor f, HalfPlane_ halfPlane r, Fractional r, Ord r, HyperPlane_ (BoundingHyperPlane halfPlane 2 r) 2 r, Show halfPlane, Show r) => f halfPlane -> CommonIntersection halfPlane r

Computes the common intersection of a \(n\) halfplanes.

running time: \(O(n\log n)\)