HGeometry.HalfSpace

Description

Representing halfspaces

Documentation

data HalfSpaceF boundingHyperPlane

A HalfSpace bounded by a hyperplane. The sign indicates which side of the bounding hyperplane is indicated. Iff the sign is positive, we mean the points for which onSideTest returns a positive value.

Half spaces include their bounding hyperplane.

Constructors

HalfSpace !Sign boundingHyperPlane

type HalfSpace (d :: Natural) r = HalfSpaceF (HyperPlane d r)

Arbitrary halfspaces in r^d.

data Sign

The sign of an expression

Constructors

Negative

Positive

module HGeometry.HalfSpace.Class