HGeometry.HalfSpace

Description

Representing halfspaces

Documentation

module HGeometry.HalfSpace.Class

boundingHyperPlaneLens :: forall boundingHyperPlane boundingHyperPlane' f. Functor f => (boundingHyperPlane -> f boundingHyperPlane') -> HalfSpaceF boundingHyperPlane -> f (HalfSpaceF boundingHyperPlane')

Lens to access the hyperplane bounding the halfspace

Generally, one can just use boundingHyperPlane from the HalfSpace_ class. But this one is useful if you need a lens to change the type of the bounding hyperplane.

type family BoundingHyperPlane halfSpace (d :: Nat) r

type HalfPlane_ halfPlane r = HalfSpace_ halfPlane 2 r

Type synonym for halfplanes in R^2

class (Dimension halfSpace ~ d, Dimension (BoundingHyperPlane halfSpace d r) ~ d, NumType halfSpace ~ r, NumType (BoundingHyperPlane halfSpace d r) ~ r) => HalfSpace_ halfSpace (d :: Nat) r | halfSpace -> d, halfSpace -> r where

Types modelling halfspaces.

Associated Types

type BoundingHyperPlane halfSpace (d :: Nat) r

Methods

boundingHyperPlane :: Lens' halfSpace (BoundingHyperPlane halfSpace d r)

Access the bounding hyperplane

halfSpaceSign :: Lens' halfSpace Sign

Lens to access the sign of the halfspace.

type HalfPlaneF line = HalfSpaceF line

A Halfplane (which is just a Halfspace in R^2)

type HalfSpace (d :: Natural) r = HalfSpaceF (HyperPlane d r)

Arbitrary halfspaces in r^d.

data HalfSpaceF boundingHyperPlane

A HalfSpace bounded by a hyperplane. The sign indicates which side of the bounding hyperplane is indicated. Iff the sign is positive, we mean the points for which onSideTest returns a positive value.

Half spaces include their bounding hyperplane.

Constructors

HalfSpace !Sign boundingHyperPlane

data Point_x_HalfSpace_Intersection boundaryPoint interiorPoint

Point x HalfSpace intersection

Constructors

Point_x_HalfSpace_OnBoundary boundaryPoint

Point_x_HalfSpace_Interior interiorPoint

data Sign

The sign of an expression

Constructors

Negative

Positive