HGeometry.Kernel

Description

Constant Complexity Geometric Primitives

Documentation

asPoint :: forall point (d :: Nat) r. Point_ point d r => Lens' point (Point d r)

projectPoint :: forall (i :: Nat) point (d :: Nat) r. (Point_ point d r, i <= d, Has_ Vector_ i r) => point -> Point i r

pattern Point1_ :: Point_ point 1 r => r -> point

pattern Point2_ :: Point_ point 2 r => r -> r -> point

pattern Point3_ :: Point_ point 3 r => r -> r -> r -> point

pattern Point4_ :: Point_ point 4 r => r -> r -> r -> r -> point

coord :: forall (i :: Natural) point (d :: Natural) r. (1 <= i, i <= d, KnownNat i, Point_ point d r) => IndexedLens' Int point r

dCoord :: forall point (d :: Natural) r. (1 <= d, Point_ point d r) => IndexedLens' Int point r

origin :: forall point (d :: Nat) r. (Num r, ConstructablePoint_ point d r) => point

pointFromList :: forall point (d :: Nat) r. (ConstructablePoint_ point d r, Vector_ (Vector d r) d r) => [r] -> Maybe point

wCoord :: forall (d :: Natural) point r. (4 <= d, Point_ point d r) => IndexedLens' Int point r

xCoord :: forall (d :: Natural) point r. (1 <= d, Point_ point d r) => IndexedLens' Int point r

yCoord :: forall (d :: Natural) point r. (2 <= d, Point_ point d r) => IndexedLens' Int point r

zCoord :: forall (d :: Natural) point r. (3 <= d, Point_ point d r) => IndexedLens' Int point r

cmpByDistanceTo :: forall r point (d :: Nat) center. (Ord r, Num r, Point_ point d r, Point_ center d r, Metric_ (Vector d r) d r) => center -> point -> point -> Ordering

euclideanDist :: forall r point (d :: Nat). (Radical r, Point_ point d r, Metric_ (Vector d r) d r) => point -> point -> r

squaredEuclideanDist :: forall r point (d :: Nat) point'. (Num r, Point_ point d r, Point_ point' d r, Metric_ (Vector d r) d r) => point -> point' -> r

cmpInDirection2 :: (Num r, Ord r, Point_ point 2 r) => Vector 2 r -> point -> point -> Ordering

ccw :: (Point_ point 2 r, Point_ point' 2 r, Point_ point'' 2 r, Num r, Ord r) => point -> point' -> point'' -> CCW

ccwCmpAround :: (Point_ center 2 r, Point_ point 2 r, Ord r, Num r) => center -> point -> point -> Ordering

ccwCmpAroundWith :: (Point_ center 2 r, Point_ point 2 r, Ord r, Num r) => Vector 2 r -> center -> point -> point -> Ordering

cwCmpAround :: (Point_ center 2 r, Point_ point 2 r, Ord r, Num r) => center -> point -> point -> Ordering

cwCmpAroundWith :: (Point_ center 2 r, Point_ point 2 r, Ord r, Num r) => Vector 2 r -> center -> point -> point -> Ordering

insertIntoCyclicOrder :: (Point_ center 2 r, Point_ point 2 r, Ord r, Num r) => center -> point -> CList point -> CList point

isCoLinear :: (Point_ point 2 r, Num r, Eq r) => point -> point -> point -> Bool

sortAround :: (Point_ center 2 r, Point_ point 2 r, Num r, Ord r) => center -> [point] -> [point]

partitionIntoQuadrants :: (Ord r, Point_ point 2 r) => point -> [point] -> ([point], [point], [point], [point])

quadrant :: (Ord r, Num r, ConstructablePoint_ point 2 r) => point -> Quadrant

quadrantWith :: (Ord r, Point_ point 2 r) => point -> point -> Quadrant

type CanonicalPoint geom = Point (Dimension geom) (NumType geom)

type Point (d :: Nat) r = PointF (Vector d r)

class (Additive_ (Vector d r) d r, HasCoordinates point point, d ~ Dimension point, r ~ NumType point) => Affine_ point (d :: Nat) r | point -> d, point -> r where

Minimal complete definition

Nothing

Methods

(.-.) :: point -> point -> Vector d r

(.+^) :: point -> Vector d r -> point

(.-^) :: point -> Vector d r -> point

class Point_ point d r => ConstructablePoint_ point (d :: Nat) r where

Methods

fromVector :: Vector d r -> point

class (Has_ Vector_ (Dimension point) (NumType point), Has_ Vector_ (Dimension point') (NumType point'), HasComponents (Vector (Dimension point') (NumType point)) (Vector (Dimension point') (NumType point')), Dimension point ~ Dimension point', HasVector point point') => HasCoordinates point point' where

Minimal complete definition

Nothing

Methods

coordinates :: IndexedTraversal1 Int point point' (NumType point) (NumType point')

class HasPoints s t point point' | s -> point, t -> point' where

Methods

allPoints :: forall (d :: Nat) r r'. (Point_ point d r, Point_ point' d r', NumType s ~ r, NumType t ~ r', Dimension s ~ d, Dimension t ~ d) => Traversal1 s t point point'

type HasPoints' s point = HasPoints s s point point

class HasVector point point' where

Methods

vector :: forall (d :: Nat) r s. (Dimension point ~ d, NumType point ~ r, Dimension point' ~ d, NumType point' ~ s) => Lens point point' (Vector d r) (Vector d s)

class (Affine_ point d r, HasVector point point) => Point_ point (d :: Nat) r where

Minimal complete definition

Nothing

Methods

coord' :: Int -> IndexedTraversal' Int point r

class Metric_ (Vector (Dimension g) (NumType g)) (Dimension g) (NumType g) => HasSquaredEuclideanDistance g where

Minimal complete definition

pointClosestToWithDistance | pointClosestTo

Methods

squaredEuclideanDistTo :: forall r (d :: Nat) point. (r ~ NumType g, d ~ Dimension g, Num r, Point_ point d r) => point -> g -> r

pointClosestTo :: forall r (d :: Nat) point. (r ~ NumType g, d ~ Dimension g, Num r, Point_ point d r) => point -> g -> Point d r

pointClosestToWithDistance :: forall r (d :: Nat) point. (r ~ NumType g, d ~ Dimension g, Num r, Point_ point d r) => point -> g -> (Point d r, r)

data CCW where

Bundled Patterns

pattern CCW :: CCW

pattern CW :: CCW

pattern CoLinear :: CCW

newtype PointF v

Constructors

Point

Fields

toVec :: v

Bundled Patterns

pattern Point1 :: r -> Point 1 r

pattern Point2 :: r -> r -> Point 2 r

pattern Point3 :: r -> r -> r -> Point 3 r

pattern Point4 :: r -> r -> r -> r -> Point 4 r

data Quadrant

Constructors

TopRight

TopLeft

BottomLeft

BottomRight

isScalarMultipleOf :: forall r vector (d :: Nat). (Eq r, Num r, Metric_ vector d r) => vector -> vector -> Bool

sameDirection :: forall vector (d :: Nat) r. (Additive_ vector d r, Num r, Eq r) => vector -> vector -> Bool

scalarMultiple :: forall r vector (d :: Nat). (Eq r, Fractional r, Additive_ vector d r) => vector -> vector -> Maybe r

(*^) :: forall r vector (d :: Nat). (Num r, Vector_ vector d r) => r -> vector -> vector

(^*) :: forall r vector (d :: Nat). (Num r, Vector_ vector d r) => vector -> r -> vector

(^+^) :: forall r vector (d :: Nat). (Num r, Additive_ vector d r) => vector -> vector -> vector

(^-^) :: forall r vector (d :: Nat). (Num r, Additive_ vector d r) => vector -> vector -> vector

(^/) :: forall vector (d :: Nat) r. (Vector_ vector d r, Fractional r) => vector -> r -> vector

basis :: forall vector (d :: Nat) r. (Additive_ vector d r, Num r) => [vector]

component :: forall (i :: Natural) vector r (d :: Natural). (i <= (d - 1), KnownNat i, Vector_ vector d r) => IndexedLens' Int vector r

cons :: forall vector' vector (d :: Nat) r. (Vector_ vector d r, Vector_ vector' (d + 1) r) => r -> vector -> vector'

foldMapZip :: forall vector (d :: Nat) r m. (Additive_ vector d r, Semigroup m) => (r -> r -> m) -> vector -> vector -> m

generate :: forall vector (d :: Nat) r. Vector_ vector d r => (Int -> r) -> vector

lerp :: forall r vector (d :: Nat). (Num r, Additive_ vector d r) => r -> vector -> vector -> vector

liftI2 :: forall vector (d :: Nat) r. Additive_ vector d r => (r -> r -> r) -> vector -> vector -> vector

negated :: forall r vector (d :: Nat). (Num r, Vector_ vector d r) => vector -> vector

prefix :: forall (i :: Nat) (d :: Nat) vector vector' r. (i <= d, Vector_ vector d r, Vector_ vector' i r) => vector -> vector'

snoc :: forall vector' vector (d :: Nat) r. (Vector_ vector d r, Vector_ vector' (d + 1) r) => vector -> r -> vector'

suffix :: forall (i :: Nat) (d :: Nat) vector vector' r. (i <= d, Vector_ vector d r, Vector_ vector' i r) => vector -> vector'

sumV :: forall f vector (d :: Nat) r. (Foldable f, Additive_ vector d r, Num r) => f vector -> vector

uncons :: forall vector' vector (d :: Natural) r. (Vector_ vector (d + 1) r, Vector_ vector' d r, 0 <= ((d + 1) - 1), d <= Dimension vector) => vector -> (r, vector')

unit :: forall vector (d :: Nat) r. (Additive_ vector d r, Num r) => vector

unsnoc :: forall vector' vector (d :: Natural) r. (Vector_ vector (d + 1) r, Vector_ vector' d r, d <= ((d + 1) - 1), d <= Dimension vector) => vector -> (vector', r)

vectorFromList :: forall vector (d :: Nat) r. Vector_ vector d r => [r] -> Maybe vector

wComponent :: forall vector (d :: Nat) r. (Vector_ vector d r, 4 <= d) => IndexedLens' Int vector r

xComponent :: forall vector (d :: Nat) r. (Vector_ vector d r, 1 <= d) => IndexedLens' Int vector r

yComponent :: forall vector (d :: Nat) r. (Vector_ vector d r, 2 <= d) => IndexedLens' Int vector r

zComponent :: forall vector (d :: Nat) r. (Vector_ vector d r, 3 <= d) => IndexedLens' Int vector r

zero :: forall r vector (d :: Nat). (Num r, Additive_ vector d r) => vector

cross :: Num r => Vector 3 r -> Vector 3 r -> Vector 3 r

class Vector_ vector d r => Additive_ vector (d :: Nat) r where

Methods

liftU2 :: (r -> r -> r) -> vector -> vector -> vector

liftI2A :: Apply f => (r -> r -> f r) -> vector -> vector -> f vector

class (r ~ IxValue vector, s ~ IxValue vector', d ~ Dimension vector, d ~ Dimension vector') => AsVector_ vector vector' (d :: Nat) r s | vector -> d, vector -> r, vector' -> s where

Methods

_Vector :: Iso vector vector' (Vector d r) (Vector d s)

type Has_ (c :: Type -> Nat -> Type -> k) (d :: Nat) r = c (Vector d r) d r

class Additive_ vector d r => Metric_ vector (d :: Nat) r where

Minimal complete definition

Nothing

Methods

dot :: vector -> vector -> r

quadrance :: vector -> r

qd :: vector -> vector -> r

norm :: vector -> r

signorm :: vector -> vector

class (HasComponents vector vector, AsVector_ vector vector d r r, KnownNat d) => Vector_ vector (d :: Nat) r where

Minimal complete definition

generateA

Methods

generateA :: Applicative f => (Int -> f r) -> f vector

component' :: Int -> IndexedTraversal' Int vector r

class HasComponents vector vector' where

Methods

components :: IndexedTraversal1 Int vector vector' (IxValue vector) (IxValue vector')

data family Vector (d :: Nat) r

module HGeometry.Line

module HGeometry.HyperPlane

module HGeometry.HalfSpace

module HGeometry.LineSegment

class IsBoxable g where

Types for which we can compute an axis parallel boundingbox

Minimal complete definition

Nothing

Methods

boundingBox :: forall (d :: Nat) r. (d ~ Dimension g, r ~ NumType g, Ord r) => g -> Box (Point d r)

Compute the axis-parallel boundingbox of the given geometry.

default boundingBox :: forall (d :: Nat) r point. (d ~ Dimension g, r ~ NumType g, Ord r, HasPoints' g point, Point_ point d r, Ord (Vector d r)) => g -> Box (Point d r)

class (HasMinPoint box point, HasMaxPoint box point, Point_ point (Dimension box) (NumType box)) => Box_ box point | box -> point where

d-dimensional Boxes

Methods

extent :: forall r (d :: Nat). (r ~ NumType box, d ~ Dimension box, Num r) => box -> Vector d (ClosedInterval r)

Get a vector with the extent of the box in each dimension. Note that the resulting vector is 0 indexed whereas one would normally count dimensions starting at zero.

data Box point where

D-dimensional boxes.

A box is represented by two points; a point with lexicographically minimal coordinates, and a point with lexicographically maximal coordinates.

Bundled Patterns

pattern Box :: point -> point -> Box point
Construct a box

pattern Rect :: Num r => r -> r -> r -> r -> Rectangle (Point 2 r)
Given x y w h construct the rectangle with bottom left corner (x,y), width w, and hegith h.

pattern Rectangle :: Dimension point ~ 2 => point -> point -> Box point
Construct a Rectangle

type Rectangle = Box

Defines a rectangle

module HGeometry.Ball

data Triangle point where

Triangles in d-dimensional space

Bundled Patterns

pattern Triangle :: point -> point -> point -> Triangle point
Construct a triangle from its three points

module HGeometry.Transformation

module HGeometry.Intersection

type family Dimension t :: Nat

type family NumType t