HGeometry.LineSegment

Description

Line segment data type and some basic functions on line segments

Documentation

data LineSegment endPoint point where

Data type representing intervals

Bundled Patterns

pattern LineSegment :: endPoint point -> endPoint point -> LineSegment endPoint point
Construct a line Segment

pattern ClosedLineSegment :: point -> point -> ClosedLineSegment point
Construct a closed interval

pattern OpenLineSegment :: point -> point -> OpenLineSegment point
Construct an open ended interval

type ClosedLineSegment point = LineSegment (EndPoint Closed) point

Default implementation of Closed LineSegments

type OpenLineSegment point = LineSegment (EndPoint Open) point

Default implementation of Open LineSegments

newtype EndPoint (et :: EndPointType) r

EndPoint with a type safe tag

Constructors

EndPoint r

Bundled Patterns

pattern OpenE :: r -> EndPoint Open r
Constructs an Open endpoint

pattern ClosedE :: r -> EndPoint Closed r
Constructs a closed endpoint

data AnEndPoint r

Data type modelling an endpoint that can both be open and closed.

Constructors

AnEndPoint !EndPointType !r

Bundled Patterns

pattern AnOpenE :: r -> AnEndPoint r
Constructs an Open endpoint

pattern AnClosedE :: r -> AnEndPoint r
Constructs a closed endpoint

module HGeometry.LineSegment.Class

spanIn :: (Point_ point d r, Ord r, IxValue (endPoint point) ~ point, EndPoint_ (endPoint point)) => Getter point r -> LineSegment endPoint point -> Interval AnEndPoint r

Computes the span of the interval in the given direction. Note that the returned interval is a proper interval, i.e. with the start smaller than the end.

>>> spanIn xCoord (ClosedLineSegment (Point2 5 (10 :: Int)) (Point2 20 0))
Interval (AnEndPoint Closed 5) (AnEndPoint Closed 20)
>>> spanIn yCoord (ClosedLineSegment (Point2 5 (10 :: Int)) (Point2 20 0))
Interval (AnEndPoint Closed 0) (AnEndPoint Closed 10)

class IsEndPoint endPoint endPoint => EndPoint_ endPoint where

An endpoint storing values of some type r

Methods

endPointType :: endPoint -> EndPointType

Report the type of the endpoint

mkEndPoint :: IxValue endPoint -> endPoint

constructs a "default" enpoint

data LineLineSegmentIntersection lineSegment

Line x LineSegment intersection

Constructors

Line_x_LineSegment_Point (Point 2 (NumType lineSegment))

Line_x_LineSegment_LineSegment lineSegment

data LineSegmentLineSegmentIntersection lineSegment

LineSegment x LineSegment intersection

Constructors

LineSegment_x_LineSegment_Point (Point 2 (NumType lineSegment))

LineSegment_x_LineSegment_LineSegment lineSegment

data HalfLineLineSegmentIntersection point segment

Data type representing the intersection of a HalfLine and a LineSegment

Constructors

HalfLine_x_LineSegment_Point point

HalfLine_x_LineSegment_LineSegment segment