HGeometry.Plane.LowerEnvelope.Clipped

Description

A Representation of the Lower envelope of planes inside a given triangle. This means now all regions are actually bounded.

Documentation

type ClippedMinimizationDiagram plane = ClippedMinimizationDiagram' (NumType plane) plane

The representing (the minimization diagram) of the lower envelope of a set of planes, clipped to some bounding triangle.

A ClippedMinimizationDiagram is essentially just a non-empty Map mapping planes to cells.

data ClippedMinimizationDiagram' r plane

Implementatino of the ClippedMinimizationDiagram type; r is the numeric type of the planes

_ClippedMinimizationDiagramMap :: NumType plane ~ r => Iso' (ClippedMinimizationDiagram plane) (NEMap plane (ClippedMDCell r plane ()))

Get access to the underlying NonEmpty Map

type ClippedMDCell r plane a = ClippedMDCell' r (MDVertex r plane a)

Cells in the Minimization diagram (i.e. the projected lower envelope of planes) parameterized by the numeric type and the planes

type ClippedMDCell' r vertex = ClippedMDCell'' r vertex (Point 2 r)

We mostly use ClippedMDCell'''s where the Extra is just a Point

newtype ClippedMDCell'' (r :: k) vertex extra

Helper type for representing cells in a minimzation diagram. These cells are possibly degenerate convex polygons, whose vertices are either of type vertex or of type 'Point 2 r'.

Constructors

ClippedMDCell (PossiblyDegenerateSimplePolygon (OriginalOrExtra vertex extra) (ClippedBoundedRegion r vertex extra))

lowerEnvelopeIn :: (Plane_ plane r, Ord plane, Ord r, Fractional r, Point_ corner 2 r, Foldable1 set, Show r, Show corner, Show plane) => Triangle corner -> set plane -> ClippedMinimizationDiagram plane

Compute the lower envelope of planes inside a given triangle.

\(O(n^4)\) (as it currently uses the brute force algorithm).

foldMapVertices :: forall {k} m vertex extra (r :: k). Semigroup m => (OriginalOrExtra vertex extra -> m) -> ClippedMDCell'' r vertex extra -> m

Fold a given function over all vertices