HGeometry.Plane.LowerEnvelope.Connected

Description

A Representation of the Lower envelope of planes as a bunch of convex regions

Documentation

data MinimizationDiagram r plane

A minimization daigram just maps every plane on the lower envelope to the region above which it is minimal. Every plane has at most one such a region.

asMap :: MinimizationDiagram r plane -> Map plane (Region r (Point 2 r))

Get the underlying Map that relates every plane in the envelope to its projected region

data Region r point

A region in the minimization diagram. The boundary is given in CCW order; i.e. the region is to the left of the boundary.

Constructors

Bounded (CircularList point)

Unbounded

Fields

(Vector 2 r)

vector indicating the direction of the unbounded edge incident to the first vertex. Note that this vector thus points INTO vertex v.

(NonEmpty point)

the vertices in CCW order,

(Vector 2 r)

the vector indicating the direction of the unbounded edge incident to the last vertex. The vector points away from the vertex (i.e. towards +infty).

type CircularList a = [a]

bounded regions are really circular lists, but we just represent them as lists for now.

intersectionPoint :: (Plane_ plane r, Ord r, Fractional r) => Three plane -> Maybe (Point 3 r)

Computes there the three planes intersect

intersectionLine :: (Plane_ plane r, Fractional r, Eq r) => plane -> plane -> Maybe (VerticalOrLineEQ r)

Given two planes, computes the line in which they intersect.

intersectionVector :: (Plane_ plane r, Ord r, Fractional r) => plane -> plane -> Maybe (Vector 2 r)

Computes the direction vector v of the directed line l in which the two planes h and h' intersect, and so that h will be to the left of the directed line

type VertexForm r plane = Map (Point 3 r) (Definers plane)

The vertices of a lower envelope is just a Map with every vertex its definers, i.e. the planes that define the vertex in CCW order around it.

data Definers plane

in CCW order, starting with the plane that is minimal at the vertical up direction from their common vertex.

fromCCWList :: [plane] -> Definers plane

Given the planes in order, starting with the one that is closest in the up direction, construct the Definers.

definers :: (Plane_ plane r, Ord r, Fractional r) => Three plane -> Maybe (Point 3 r, Definers plane)

Smart constructor for creating the definers of three planes

fromVertexForm :: (Plane_ plane r, Ord plane, Ord r, Fractional r) => VertexForm r plane -> MinimizationDiagram r plane

Given the vertices of the lower envelope; compute the minimization diagram.

\(O(h\log h)\) assuming that the input is degenerate.

mergeDefiners :: (Plane_ plane r, Eq plane, Ord r, Fractional r) => Point 3 r -> Definers plane -> Definers plane -> Definers plane

Merge two lists of definers.

O(n^2), since we are using incremental insertion.

bruteForceLowerEnvelope :: (Plane_ plane r, Ord plane, Ord r, Fractional r, Foldable set) => set plane -> MinimizationDiagram r plane

Computes the lower envelope in O(n^4) time.

computeVertexForm :: (Plane_ plane r, Ord plane, Ord r, Fractional r, Foldable set) => set plane -> VertexForm r plane

Computes the vertices of the lower envelope

O(n^4) time.