HGeometry.Plane.LowerEnvelope.Connected

Description

A Representation of the Lower envelope of planes as a bunch of convex regions

Documentation

newtype MinimizationDiagram r vertex plane

A minimization daigram just maps every plane on the lower envelope to the region above which it is minimal. Every plane has at most one such a region.

Constructors

MinimizationDiagram (NEMap plane (Region r vertex))

asMap :: MinimizationDiagram r vertex plane -> NEMap plane (Region r vertex)

Get the underlying Map that relates every plane in the envelope to its projected region

mapVertices :: (vertex -> vertex') -> MinimizationDiagram r vertex plane -> MinimizationDiagram r vertex' plane

Apply some mapping function to the vertex data of each vertex

data Region r vertex

A region in the minimization diagram. The boundary is given in CCW order; i.e. the region is to the left of the boundary.

Constructors

BoundedRegion (ConvexPolygonF (Cyclic NonEmpty) vertex)

UnboundedRegion (UnboundedConvexRegionF r NonEmpty vertex)

data MDVertex r plane

A vertex of the minimzation Diagram that is defined by the intersection of a number of planes.

Note that we interpet this vertex as a 2-dimensional thing.

Constructors

MDVertex (Point 3 r) (Definers plane)

location :: forall r1 plane r2 f. Functor f => (Point 3 r1 -> f (Point 3 r2)) -> MDVertex r1 plane -> f (MDVertex r2 plane)

toConvexPolygonIn :: (Rectangle_ rectangle corner, Point_ corner 2 r, Point_ point 2 r, Ord r, Fractional r) => rectangle -> Region r point -> Either (ConvexPolygonF (Cyclic NonEmpty) point) (ConvexPolygonF (Cyclic NonEmpty) (OriginalOrExtra point (Point 2 r)))

Computes a convex polygon corresponding to the region.

pre: the bounding box (strictly) contains all vertices in its interior

intersectionPoint :: (Plane_ plane r, Ord r, Fractional r) => Three plane -> Maybe (Point 3 r)

Computes there the three planes intersect

intersectionLine :: (Plane_ plane r, Fractional r, Eq r) => plane -> plane -> Maybe (VerticalOrLineEQ r)

Given two planes, computes the line in which they intersect.

intersectionVector :: (Plane_ plane r, Ord r, Fractional r) => plane -> plane -> Maybe (Vector 2 r)

Computes the direction vector v of the directed line l in which the two planes h and h' intersect, and so that h will be to the left of the directed line

type VertexForm (map :: Type -> Type -> k) r plane = map (Point 3 r) (Definers plane)

The vertices of a lower envelope is just a Map with every vertex its definers, i.e. the planes that define the vertex in CCW order around it.

data Definers plane

in CCW order, starting with the plane that is minimal at the vertical up direction from their common vertex.

fromCCWList :: NonEmpty plane -> Definers plane

Given the planes in order, starting with the one that is closest in the up direction, construct the Definers.

definers :: (Plane_ plane r, Ord r, Fractional r) => Three plane -> Maybe (Point 3 r, Definers plane)

Smart constructor for creating the definers of three planes

type ClippedBoundedRegion (r :: k) vertex corner = ConvexPolygonF (Cyclic NonEmpty) (OriginalOrExtra vertex corner)

A Convex bounded region, which may be clipped using vertices of type corner.

fromVertexForm :: (Plane_ plane r, Ord plane, Ord r, Fractional r, Show r, Show plane) => NEMap (Point 3 r) (Definers plane) -> MinimizationDiagram r (MDVertex r plane) plane

Given the vertices of the lower envelope; compute the minimization diagram.

\(O(h\log h)\) assuming that the input is non-degenerate.

mergeDefiners :: (Plane_ plane r, Eq plane, Ord r, Fractional r, Show plane, Show r) => Point 3 r -> Definers plane -> Definers plane -> Definers plane

Merge two lists of definers.

\(O(n\log n)\), where \(n\) is the total number of planes involved.

fromVertexFormIn :: (Plane_ plane r, Ord plane, Ord r, Fractional r, Show r, Show plane, Point_ corner 2 r, Show r, Show corner) => Triangle corner -> NEMap (Point 3 r) (Definers plane) -> NEMap plane (ClippedBoundedRegion r (MDVertex r plane) (Point 2 r))

Pre: the triangle is big enough to contain all vertices of the lower envelope

extraPoints :: (Triangle_ triangle corner, Point_ corner 2 r, Point_ point 2 r, Fractional r, Ord r, IsIntersectableWith (HalfLine point) (ClosedLineSegment corner), Intersection (HalfLine point) (ClosedLineSegment corner) ~ Maybe (HalfLineLineSegmentIntersection (Point 2 r) (ClosedLineSegment corner)), HasOuterBoundary triangle, Eq (VertexIx triangle), Show point, Show r, Show corner, Show triangle) => HalfLine point -> HalfLine point -> triangle -> NonEmpty (Point 2 r)

computes the extra vertices that we have to insert to make an unbounded region bounded

module HGeometry.Plane.LowerEnvelope.Connected.BruteForce