HGeometry.Plane.LowerEnvelope.Connected

Description

A Representation of the Lower envelope of planes in Adjacency-list form.

Documentation

data LowerEnvelope' plane

A connected lower envelope in adjacencylist form.

invariant: there is always at least one bounded vertex.

Constructors

LowerEnvelope !(UnboundedVertex plane) (Seq (BoundedVertex plane))

theUnboundedVertex :: Lens' (LowerEnvelope' plane) (UnboundedVertex plane)

lens to access the unbounded vertex

boundedVertices :: Lens' (LowerEnvelope' plane) (Seq (BoundedVertex plane))

Lens to access the sequence of bounded vertices.

singleton :: (Plane_ plane r, Ord r, Fractional r, Ord plane) => LEVertex plane -> LowerEnvelope' plane

Given a single Vertex, construct a LowerEnvelope' out of it.

fromVertexForm' :: forall plane r. (Plane_ plane r, Ord plane, Ord r, Fractional r, Show plane, Show r) => VertexForm plane -> LowerEnvelope' plane

Given a Lower envelope in vertex form, construct the AdjacencyList representation out of it.

pre: The set of vertices is non-empty

\(O(n\log n)\)

data BoundedVertexF f plane

Data type representing the bounded vertices of the lower envelope.

In the vertexform, the f will be Const (), i.e. we don't have adjacency information. But in case of Adjacencylist form we can implement it using a Sequence.

Constructors

Vertex !(Point 3 (NumType plane)) (Set plane) (f (LEEdge plane))

location :: NumType plane ~ r => Lens' (BoundedVertexF f plane) (Point 3 r)

The location of the vertex

definers :: Lens' (BoundedVertexF f plane) (Set plane)

The three planes defining the vertex

location2 :: NumType plane ~ r => Getter (BoundedVertexF f plane) (Point 2 r)

Projected 2d location of the point

newtype UnboundedVertex plane

The unbounded vertex.

Constructors

UnboundedVertex (Seq (LEEdge plane))

unboundedVertexId :: VertexID

The vertexID of the unbounded vertex

class HasUnboundedEdges t e | t -> e where

Methods

unboundedEdges :: Lens' t (Seq e)

Lens to access the unbounded edges.

type EdgeGeometry point = Either (HalfLine point) (ClosedLineSegment point)

Edges are either halflines or Linesegments

projectedEdgeGeometries :: (Plane_ plane r, Ord r, Fractional r, Show plane, Show r) => IndexedFold (EdgeIx (LowerEnvelope' plane)) (LowerEnvelope' plane) (EdgeGeometry (Point 2 r))

Get the projected geometries (halflines and edges line segments ) representing the edges.

projectedEdgeGeometry :: (Plane_ plane r, Ord r, Fractional r, Show plane, Show r) => LowerEnvelope' plane -> EdgeIx (LowerEnvelope' plane) -> Edge (LowerEnvelope' plane) -> EdgeGeometry (Point 2 r)

Computes the (projected) LineSegment or HalfLine representing a given edge.