HGeometry.Plane.LowerEnvelope.Naive

Documentation

lowerEnvelope :: (Plane_ plane r, Ord r, Fractional r, Foldable1 f, Functor f, Ord plane, Show plane, Show r) => f plane -> LowerEnvelope plane

Brute force implementation that computes the lower envelope, by explicitly considering every triple of planes.

pre: the input forms a *set* of planes, i.e. no duplicates

running time: \(O(n^4 )\)

lowerEnvelopeWith :: forall nonEmpty plane r. (Plane_ plane r, Ord r, Fractional r, Foldable1 nonEmpty) => (nonEmpty plane -> Maybe (MinimizationDiagram r (Point 2 r :+ Definers plane) plane)) -> nonEmpty plane -> LowerEnvelope plane

Computes the lower envelope of the planes; possibly using the given function to construct the minimization diagram.

In particular, we distinguish whether the planes define a connected lower envelope (in that case, we actually call the algorithm on it). Or that the planes are degenerate and are all parallel and thus define parallel strips. In that case, we use a two dimensional lower envelope algorithm to compute the lower envelope of the (parallel) planes.

\(O(T(n) + n \log n)\).