HGeometry.PlaneGraph

Description

Plane Graphs, i.e. embedded planar graphs.

Documentation

module Hiraffe.Graph.Class

module HGeometry.PlaneGraph.Class

module Hiraffe.PlanarGraph.Class

newtype CPlaneGraph (s :: k) v e f

An Embedded, *connected*, planar graph

Constructors

CPlaneGraph (CPlanarGraph 'Primal s v e f)

fromAdjacencyRep :: forall {k} vertex r i (f :: Type -> Type) proxy (s :: k) e. (Point_ vertex 2 r, Ord i, Foldable1 f) => proxy s -> GGraph f i vertex e -> CPlaneGraph s vertex e ()

Given a connected plane graph in adjacency list format; convert it into an actual CPlaneGraph.

\(O(n\log n)\)

fromConnectedSegments :: forall {k} f r lineSegment point (s :: k). (Foldable1 f, Ord r, Num r, LineSegment_ lineSegment point, Point_ point 2 r) => f lineSegment -> CPlaneGraph s (NonEmpty point) lineSegment ()

Constructs a connected plane graph

pre: The segments form a single connected component No two segments partially overlap.

running time: \(O(n\log n)\)