HGeometry.PlaneGraph.Class

Description

Type type for planar graphs embedded in \(\mathbb{R}^2\). For functions that export faces and edges etc, we assume the graph has a (planar) straight line embedding.

Documentation

class (PlanarGraph_ planeGraph, vertex ~ Vertex planeGraph, Point_ vertex 2 (NumType vertex), NumType vertex ~ NumType planeGraph, HasEdges planeGraph planeGraph) => PlaneGraph_ planeGraph vertex | planeGraph -> vertex where

A class representing Plane graphs, i.e. planar graphs that have a straight line embedding in the plane.

Minimal complete definition

fromEmbedding

Methods

fromEmbedding :: (Foldable1 f, Functor f, Foldable h, Functor h, vi ~ VertexIx planeGraph, v ~ Vertex planeGraph, e ~ Edge planeGraph, GraphFromAdjListExtraConstraints planeGraph h) => f (vi, v, h (vi, e)) -> planeGraph

Build a graph from its embedding; i.e. for each vertex we expect its adjacencies in CCW order.

If the, in the list of neighbours of vertex u we see a vertex v that itself does not appear in the adjacencylist, we may drop it. In other words if u has a neighbour v, then v better have a specification of its neighbours somewhere.

outerFace :: IndexedLens' (FaceIx planeGraph) planeGraph (Face planeGraph)

Getter to access the outer face

interiorFaces :: IndexedTraversal' (FaceIx planeGraph) planeGraph (Face planeGraph)

Traversal of all interior faces in the graph

outerFaceId :: planeGraph -> FaceIx planeGraph

gets the id of the outer face

outerFaceDart :: planeGraph -> DartIx planeGraph

gets a dart incident to the outer face (in particular, that has the outerface on its left)

running time: \(O(n)\)

default outerFaceDart :: (r ~ NumType planeGraph, Ord r, Num r) => planeGraph -> DartIx planeGraph

dartSegmentAt :: (PlaneGraph_ planeGraph vertex, Point_ vertex 2 (NumType vertex)) => DartIx planeGraph -> IndexedFold (DartIx planeGraph) planeGraph (ClosedLineSegment vertex)

Get the line segment representing a particular dart.

Note that this is a fold rather than a getter for the same reason dartAt is a traversal rather than a lens: i.e. if you pass some nonsensical DartIx the dart may not exist.

edgeSegmentAt :: (PlaneGraph_ planeGraph vertex, Point_ vertex 2 (NumType vertex)) => EdgeIx planeGraph -> IndexedFold (EdgeIx planeGraph) planeGraph (ClosedLineSegment vertex)

Get the line segment representing a particular edge.

Note that this is a fold rather than a getter for the same reason edgeAt is a traversal rather than a lens: i.e. if you pass some nonsensical EdgeIx the edge may not exist.

dartSegments :: (PlaneGraph_ planeGraph vertex, Point_ vertex 2 (NumType vertex)) => IndexedFold (DartIx planeGraph) planeGraph (ClosedLineSegment vertex)

Renders all darts as line segments. Thesegments are all oriented in the direction of the dart.

edgeSegments :: (HasEdges planeGraph planeGraph, BidirGraph_ planeGraph, vertex ~ Vertex planeGraph, Point_ vertex 2 (NumType vertex)) => IndexedFold (EdgeIx planeGraph) planeGraph (ClosedLineSegment vertex)

Renders all edges as line segments.

interiorFacePolygonAt :: (PlaneGraph_ planeGraph vertex, Point_ vertex 2 (NumType vertex)) => FaceIx planeGraph -> IndexedFold (FaceIx planeGraph) planeGraph (SimplePolygon (vertex :+ VertexIx planeGraph))

Renders a single interior face as a simple polygon.

Note that this is a fold rather than a getter for the same reason faceAt is a traversal rather than a lens: i.e. if you pass some nonsensical FaceIx the face may not exist.

interiorFacePolygons :: (PlaneGraph_ planeGraph vertex, Point_ vertex 2 r, Ord r, Num r, Eq (FaceIx planeGraph)) => IndexedFold (FaceIx planeGraph) planeGraph (SimplePolygon (vertex :+ VertexIx planeGraph))

Renders all interior faces as simple polygons.