HGeometry.Point.Class

Description

A class of types that can act as \(d\)-dimensional points.

Documentation

class HasVector point point' where

Type class for types, usually points, that have a Lens to interpret the point as a vector.

Methods

vector :: forall (d :: Nat) r s. (Dimension point ~ d, NumType point ~ r, Dimension point' ~ d, NumType point' ~ s) => Lens point point' (Vector d r) (Vector d s)

Lens to access the vector corresponding to this point.

>>> myPoint ^. vector
Vector3 1 2 3
>>> ( myPoint & vector .~ Vector3 3 2 1  ) :: Point 3 Int
Point3 3 2 1
>>> (myPoint & coordinates %~ show ) :: Point 3 String
Point3 "1" "2" "3"

class (Has_ Vector_ (Dimension point) (NumType point), Has_ Vector_ (Dimension point') (NumType point'), HasComponents (Vector (Dimension point') (NumType point)) (Vector (Dimension point') (NumType point')), Dimension point ~ Dimension point', HasVector point point') => HasCoordinates point point' where

Class for point types that have a type changing traversal over all coordinates.

Minimal complete definition

Nothing

Methods

coordinates :: IndexedTraversal1 Int point point' (NumType point) (NumType point')

Traversal over *all* coordinates of the points. Coordinates are 1-indexed.

>>> imapMOf_ coordinates (\i x -> print (i,x)) (Point2 10 20 :: Point 2 Int)
(1,10)
(2,20)
>>> itraverseOf coordinates (\i x -> print (i,x)) (Point2 10 20) :: IO (Point 2 ())
(1,10)
(2,20)
Point2 () ()
>>> over coordinates (+1) $ Point2 10 20 :: Point 2 Int
Point2 11 21

class (Additive_ (Vector d r) d r, HasCoordinates point point, d ~ Dimension point, r ~ NumType point) => Affine_ point (d :: Nat) r | point -> d, point -> r where

Affine space; essentially the same as Linear.Affine, but for points of kind Type rather than (Type -> Type).

Minimal complete definition

Nothing

Methods

(.-.) :: point -> point -> Vector d r

p .-. q represents the vector from q to p

default (.-.) :: (HasVector point point, Num r) => point -> point -> Vector d r

(.+^) :: point -> Vector d r -> point

add a vector to a point

>>> myPoint .+^ Vector3 100 200 300
Point3 101 202 303

default (.+^) :: (HasVector point point, Num r) => point -> Vector d r -> point

(.-^) :: point -> Vector d r -> point

subtract a vector from a point

>>> myPoint .-^ Vector3 100 200 300
Point3 (-99) (-198) (-297)

class (Affine_ point d r, HasVector point point) => Point_ point (d :: Nat) r where

A class representing points in d-dimensional space.

Minimal complete definition

Nothing

Methods

coord' :: Int -> IndexedTraversal' Int point r

Get the coordinate in a given dimension. This operation is unsafe in the sense that no bounds are checked. Consider using coord instead.

>>> myPoint ^.. coord' 2
[2]

pattern Point1_ :: Point_ point 1 r => r -> point

A pattern synonym for 1 dimensional points.

pattern Point2_ :: Point_ point 2 r => r -> r -> point

A pattern synonym for 2 dimensional points.

pattern Point3_ :: Point_ point 3 r => r -> r -> r -> point

A pattern synonym for 3 dimensional points.

pattern Point4_ :: Point_ point 4 r => r -> r -> r -> r -> point

A bidirectional pattern synonym for 4 dimensional points.

class Point_ point d r => ConstructablePoint_ point (d :: Nat) r where

Type class for constructable points

Methods

fromVector :: Vector d r -> point

Construct a point from a vector

>>> fromVector (Vector4 1 2 3 4) :: Point 4 Int
Point4 1 2 3 4

origin :: forall point (d :: Nat) r. (Num r, ConstructablePoint_ point d r) => point

Point representing the origin in d dimensions

>>> origin :: Point 4 Int
Point4 0 0 0 0

pointFromList :: forall point (d :: Nat) r. (ConstructablePoint_ point d r, Vector_ (Vector d r) d r) => [r] -> Maybe point

Constructs a point from a list of coordinates. The length of the list has to match the dimension exactly.

>>> pointFromList [1,2,3] :: Maybe (Point 3 Int)
Just (Point3 1 2 3)
>>> pointFromList [1] :: Maybe (Point 3 Int)
Nothing
>>> pointFromList [1,2,3,4] :: Maybe (Point 3 Int)
Nothing

coord :: forall (i :: Natural) point (d :: Natural) r. (1 <= i, i <= d, KnownNat i, Point_ point d r) => IndexedLens' Int point r

Get the coordinate in a given dimension

>>> myPoint ^. coord @2
2
>>> myPoint & coord @1 .~ 10
Point3 10 2 3
>>> myPoint & coord @3 %~ (+1)
Point3 1 2 4

xCoord :: forall (d :: Natural) point r. (1 <= d, Point_ point d r) => IndexedLens' Int point r

Shorthand to access the first coordinate

>>> myPoint ^. xCoord
1
>>> Point2 1 (2 :: Int) & xCoord .~ 10
Point2 10 2

yCoord :: forall (d :: Natural) point r. (2 <= d, Point_ point d r) => IndexedLens' Int point r

Shorthand to access the second coordinate

>>> Point2 1 (2 :: Int) ^. yCoord
2
>>> myPoint & yCoord %~ (+1)
Point3 1 3 3

zCoord :: forall (d :: Natural) point r. (3 <= d, Point_ point d r) => IndexedLens' Int point r

Shorthand to access the third coordinate

>>> myPoint ^. zCoord
3
>>> myPoint & zCoord %~ (+1)
Point3 1 2 4

wCoord :: forall (d :: Natural) point r. (4 <= d, Point_ point d r) => IndexedLens' Int point r

Shorthand to access the fourth coordinate

>>> (Point4 1 2 3 4 :: Point 4 Int) ^. wCoord
4
>>> (Point4 1 2 3 4 :: Point 4 Int) & wCoord %~ (+1)
Point4 1 2 3 5

dCoord :: forall point (d :: Natural) r. (1 <= d, Point_ point d r) => IndexedLens' Int point r

Shorthand to access the last coordinate

>>> (Point2 1 2 :: Point 2 Int) ^. dCoord
2
>>> (Point4 1 2 3 4 :: Point 4 Int) ^. dCoord
4
>>> (Point4 1 2 3 4 :: Point 4 Int) & dCoord %~ (+1)
Point4 1 2 3 5

class HasPoints s t point point' | s -> point, t -> point' where

Data types that store points

Methods

allPoints :: forall (d :: Nat) r r'. (Point_ point d r, Point_ point' d r', NumType s ~ r, NumType t ~ r', Dimension s ~ d, Dimension t ~ d) => Traversal1 s t point point'

Traversal over all points in the structure

>>> let xs = NonEmpty.fromList [Point2 10 10, Point2 20 (30 :: Int)]
>>> xs^..allPoints
[Point2 10 10,Point2 20 30]
>>> over allPoints (.+^ Vector2 10 10) xs :: NonEmpty.NonEmpty (Point 2 Int)
Point2 20 20 :| [Point2 30 40]

type HasPoints' s point = HasPoints s s point point

Shorthand for 'HasPoints s s point point'