HGeometry.Point.EuclideanDistance

Description

Points that have to do with Euclidean distances between \(d\)-dimensional points.

Documentation

cmpByDistanceTo :: (Ord r, Num r, Point_ point d r, Metric_ (Vector d r) d r) => point -> point -> point -> Ordering

Compare two points by their distance to the first argument

squaredEuclideanDist :: (Num r, Point_ point d r, Metric_ (Vector d r) d r) => point -> point -> r

Squared Euclidean distance between two points

euclideanDist :: (Radical r, Point_ point d r, Metric_ (Vector d r) d r) => point -> point -> r

Euclidean distance between two points

class Metric_ (Vector (Dimension g) (NumType g)) (Dimension g) (NumType g) => HasSquaredEuclideanDistance g where

Types for which we can compute the squared Euclidean distance.

Minimal complete definition

pointClosestToWithDistance | pointClosestTo

Methods

squaredEuclideanDistTo :: (r ~ NumType g, d ~ Dimension g, Num r, Point_ point d r) => point -> g -> r

Given a point q and a geometry g, the squared Euclidean distance between q and g.

pointClosestTo :: (r ~ NumType g, d ~ Dimension g, Num r, Point_ point d r) => point -> g -> Point d r

Given q and g, computes the point p in g closest to q according to the Squared Euclidean distance.

pointClosestToWithDistance :: (r ~ NumType g, d ~ Dimension g, Num r, Point_ point d r) => point -> g -> (Point d r, r)

Given q and g, computes the point p in g closest to q according to the Squared Euclidean distance. Returns both the point and the distance realized by this point.