HGeometry.Polygon

Description

A polygon and some basic functions to interact with them.

Documentation

module HGeometry.Polygon.Class

asTriangle :: Polygon_ polygon point r => polygon -> Maybe (Triangle point)

Try to convert the polygon into a triangle (whose vertices are given in CCW order).

area2X :: (Polygon_ polygon point r, Num r, Point_ point 2 r, SimplePolygon_ (Hole polygon) point r) => polygon -> r

Computes the double area of a polygon

area :: (Polygon_ polygon point r, SimplePolygon_ (Hole polygon) point r, Fractional r) => polygon -> r

The area of a polygon

running time: \(O(n)\)