HGeometry.Polygon.Convex

Description

A type for representing Convex polygons, and some basic functions on Convex polygons.

Documentation

class SimplePolygon_ convexPolygon point r => ConvexPolygon_ convexPolygon point r

Class modelling convex polygons.

type ConvexPolygon = ConvexPolygonF (Cyclic NonEmptyVector)

By default we use a cyclic non-empty vector to represent convex polygons.

data ConvexPolygonF (f :: k -> Type) (point :: k)

Convex polygons

fromSimplePolygon :: forall r point (f :: Type -> Type). (Ord r, Num r, Point_ point 2 r, VertexContainer f point) => SimplePolygonF f point -> Maybe (ConvexPolygonF f point)

Smart constructor to construct a strictly convex polygon from a simple polygon.

toSimplePolygon :: ConvexPolygonF f point -> SimplePolygonF f point

Convert to a simple polygon, i.e. forget the polygon is convex.

_ConvexPolygon :: forall (f :: Type -> Type) point r. (Num r, Ord r, Point_ point 2 r, VertexContainer f point) => Prism' (SimplePolygonF f point) (ConvexPolygonF f point)

Prism that can forget that the polygon is convex

isStrictlyConvex :: forall r point (f :: Type -> Type). (Ord r, Num r, Point_ point 2 r, VertexContainer f point) => SimplePolygonF f point -> Bool

\( O(n) \) Check if a polygon is strictly convex.

isConvex :: forall r point (f :: Type -> Type). (Ord r, Num r, Point_ point 2 r, VertexContainer f point) => SimplePolygonF f point -> Bool

\( O(n) \) Check if a polygon is convex.

verifyConvex :: forall r point (f :: Type -> Type). (Ord r, Num r, Point_ point 2 r, VertexContainer f point) => ConvexPolygonF f point -> Bool

Verify that a convex polygon is strictly convex.

running time \( O(n) \)

minkowskiSum :: (Ord r, Num r, ConvexPolygon_ convexPolygon point r, ConvexPolygon_ convexPolygon' point' r) => convexPolygon -> convexPolygon' -> ConvexPolygon (point :+ point')

Computes the Minkowski sum of the two input polygons with \(n\) and \(m\) vertices respectively.

Note: it takes te position of te second polyon as is; you may want to translate it to te oriin first.

running time: \(O(n+m)\).

maxInDirection :: (Num r, Ord r, ConvexPolygon_ convexPolygon point r) => Vector 2 r -> convexPolygon -> point

Finds the extreme maximum point in the given direction. Based on http://geomalgorithms.com/a14-_extreme_pts.html

pre: The input polygon is strictly convex.

running time: \(O(\log n)\)