HGeometry.Polygon.Convex.Unbounded

Description

A type for representing unbounded, Convex, polygonal regions.

Documentation

type UnboundedConvexRegion vertex = UnboundedConvexRegionF (NumType vertex) NonEmpty vertex

An unbounded polygonal Convex Region

data UnboundedConvexRegionF r (nonEmpty :: k -> Type) (vertex :: k)

An unbounded polygonal ConvexRegion whose vertices are stored in an nonEmpty

Constructors

Unbounded

Fields

(Vector 2 r)

vector indicating the direction of the unbounded edge incident to the first vertex. Note that this vector thus points INTO vertex v.

(nonEmpty vertex)

the vertices in CCW order,

(Vector 2 r)

the vector indicating the direction of the unbounded edge incident to the last vertex. The vector points away from the vertex (i.e. towards +infty).

extremalVertices :: UnboundedConvexRegionF r NonEmpty vertex -> Either vertex (Vector 2 vertex)

Compute the first and last vertex of the chain. Returns a Left if the first and last are the same.

mapChain :: forall {k1} {k2} nonEmpty (vertex :: k1) nonEmpty' (vertex' :: k2) r. (nonEmpty vertex -> nonEmpty' vertex') -> UnboundedConvexRegionF r nonEmpty vertex -> UnboundedConvexRegionF r nonEmpty' vertex'

map a function over the sequence of points

toBoundedFrom :: (Foldable nonEmpty, Point_ point 2 r, Point_ vertex 2 r, Ord r, Fractional r) => nonEmpty point -> UnboundedConvexRegionF r NonEmpty vertex -> ConvexPolygonF (Cyclic NonEmpty) vertex

Given some convex shape S, construct some "big enough" convex polygon B out of the unbounded convex polygon U so that the intersection \(U \cap S\) is the same as (B cap S)

note: this creates some new vertices; which are "copies" from the extremal vertices. this is to avoid having to introduce yet another level of OriginalOrExtra's