HGeometry.Polygon.Simple

Description

A Simple polygon and some basic functions to interact with them.

Documentation

class (Polygon_ simplePolygon point r, VertexIx simplePolygon ~ Int, Point_ point 2 r) => SimplePolygon_ simplePolygon point r where

A class representing simple polygons; i.e. polygons without holes (and without self intersections in the boundary.)

Minimal complete definition

uncheckedFromCCWPoints, fromPoints

Associated Types

type ConstructableSimplePolygon simplePolygon point r :: Constraint

Constraints that allow us to construct a simple polygon

type ConstructableSimplePolygon simplePolygon point r = (Eq r, Num r)

Methods

uncheckedFromCCWPoints :: Foldable f => f point -> simplePolygon

Given the vertices of the polygon, in CCW order, constructs the polygon. The vertices are numbered in the order they are given.

pre: - the vertices are given in CCW order - at least 3 vertices, not all colinear - no repeated vertices - no self-inttersections

fromPoints :: (Foldable f, ConstructableSimplePolygon simplePolygon point r) => f point -> Maybe simplePolygon

Given the vertices of the polygon, constructs the polygon. This function will make sure the polygon is a valid simple polygon, i.e. it has at least three vertices, is given in CCW order, no repeated vertices etc.

centroid :: (Fractional r, ConstructablePoint_ point' 2 r) => simplePolygon -> point'

Compute the centroid of a simple polygon.

running time: \(O(n)\)

type SimplePolygon = SimplePolygonF (Cyclic NonEmptyVector)

By default we store simple polygons as non-empty circular vectors.

newtype SimplePolygonF f point

Simple polygons just store their vertices in CCCW order

Constructors

MkSimplePolygon (f point)

toCyclic :: SimplePolygonF (Cyclic v) point -> Cyclic v point

Get the underlying cyclic vector.

type VertexContainer f point = (IxValue (f point) ~ point, Index (f point) ~ Int, TraversableWithIndex Int f, Traversable1 f, Ixed (f point), HasDirectedTraversals f)

shortcut for all default properties of f we need to store the vertices.

class HasInPolygon polygon point r | polygon -> point, point -> r where

Types that implement a point-in-polygon test.

Minimal complete definition

Nothing

Methods

inPolygon :: (Num r, Ord r, Point_ queryPoint 2 r) => queryPoint -> polygon -> PointLocationResultWith (VertexIx polygon)

Check if a point lies inside a polygon, on the boundary, or outside of the polygon. If the point lies on the boundary we also return an edge e containing the point (identified by the vertex v so that e follows v in the CCW order along the boundary). No guarantees are given about which edge is returned if the query point lies on multiple edges (i.e. when it coincides with a vertex.)

default inPolygon :: (Num r, Ord r, Point_ point 2 r, Point_ queryPoint 2 r, SimplePolygon_ polygon point r) => queryPoint -> polygon -> PointLocationResultWith (VertexIx polygon)

inSimplePolygon :: forall queryPoint simplePolygon point r. (Num r, Ord r, Point_ point 2 r, Point_ queryPoint 2 r, SimplePolygon_ simplePolygon point r) => queryPoint -> simplePolygon -> PointLocationResultWith (VertexIx simplePolygon)

Check if a point lies inside a polygon, on the boundary, or outside of the polygon. If the point lies on the boundary we also return an edge e containing the point (identified by the vertex v so that e follows v in the CCW order along the boundary). No guarantees are given about which edge is returned if the query point lies on multiple edges (i.e. when it coincides with a vertex.)

Running time: O(n).

>>> Point2 1 1 `inPolygon` simplePoly
StrictlyInside
>>> Point2 0 0 `inPolygon` simplePoly
OnBoundaryEdge 0
>>> Point2 10 0 `inPolygon` simplePoly
OnBoundaryEdge 1
>>> Point2 5 13 `inPolygon` simplePoly
StrictlyInside
>>> Point2 5 10 `inPolygon` simplePoly
StrictlyInside
>>> Point2 10 5 `inPolygon` simplePoly
OnBoundaryEdge 1
>>> Point2 20 5 `inPolygon` simplePoly
StrictlyOutside

hasNoSelfIntersections :: forall f point r. (Foldable f, Functor f, Point_ point 2 r, Ord r, Real r) => f point -> Bool

verify that some sequence of points has no self intersecting edges.