HGeometry.Polygon.Triangulation.MakeMonotone

Description

Documentation

computeDiagonals :: (Polygon_ polygon point r, Point_ point 2 r, Ord r, Num r, Ord (VertexIx polygon), Default (VertexIx polygon)) => polygon -> [Diagonal polygon]

Given a polygon, find a set of non-intersecting diagonals that partition the polygon into y-monotone pieces.

running time: \(O(n\log n)\)

classifyVertices :: (Polygon_ polygon point r, Num r, Ord r, Point_ point 2 r) => polygon -> [(VertexIx polygon, VertexType)]

Classify all vertices of the polygon; i.e. compute their vertex types.

data VertexType

The various vertex types

Constructors

Start

Merge

Split

End

Regular

cmpX :: (Point_ point 2 r, LineSegment_ lineSegment point', Point_ point' 2 r, Ord r, Num r) => point -> lineSegment -> Ordering

compare the point and the segment horizontally. I.e. returs LT if the segment lies left of the point, EQ if the point is on the segment, and GT if the segment lies right of the point.