HGeometry.Polygon.Visibility.Naive

Description

Naive algorithm to compute the Visibility graph in a simple polgyon

Documentation

visibilityGraph :: (SimplePolygon_ simplePolygon point r, HasIntersectionWith point simplePolygon, Ord r, Num r) => simplePolygon -> [Vector 2 (VertexIx simplePolygon)]

Naive algorithm to compute the visibilityGraph of a simple polygon.

\(O(n^3)\)

visibilityGraphWith

Arguments

:: (SimplePolygon_ simplePolygon point r, HasIntersectionWith point simplePolygon, Ord r, Num r)

=> simplePolygon

-> [Vector 2 (point :+ VertexIx simplePolygon)]
candidate edges

-> [Vector 2 (VertexIx simplePolygon)]

Given a polygon, and candidate visibility edges (edges guaranteed not to intersect) the boundary; test if the edges are actually inside the polygon (as well as adding the actual polygon edges as visible).

\(O(n + m)\), where m is the number of candidate edges.