HGeometry.Triangle.Class

Description

A class of types representing Triangles

Documentation

class (Point_ point (Dimension point) (NumType point), HasVertices triangle triangle, Vertex triangle ~ point) => Triangle_ triangle point | triangle -> point where

Class representing triangles

Methods

corners :: Lens' triangle (Vector 3 point)

Lens to access the corners of the triangle.

pattern Triangle_ :: Triangle_ triangle point => point -> point -> point -> triangle

Pattern match on a triangle

class Triangle_ triangle point => ConstructableTriangle_ triangle point where

Class representing constructable triangles

Methods

mkTriangle :: point -> point -> point -> triangle

Construct a triangle from its three vertices.

toCounterClockwiseTriangle :: (Num r, Eq r, Point_ point 2 r, Triangle_ triangle point) => triangle -> triangle

Make sure that the triangles vertices are given in counter clockwise order

>>> let t = Triangle origin (Point2 0 (-1)) (Point2 (-1) 0) :: Triangle (Point 2 Int)
>>> toCounterClockwiseTriangle t
Triangle (Point2 0 0) (Point2 (-1) 0) (Point2 0 (-1))

triangleSignedArea2X :: (Num r, Point_ point 2 r, Triangle_ triangle point) => triangle -> r

Computes the double-signed area of a triangle

intersectingHalfPlanes :: (Triangle_ triangle point, Point_ point 2 r, Num r, Ord r) => triangle -> Vector 3 (HalfSpaceF (LinePV 2 r))

Get the three halfplanes such that the triangle is the intersection of those halfspaces.

>>> let t = Triangle origin (Point2 0 (-1)) (Point2 (-1) 0) :: Triangle (Point 2 Int)
>>> mapM_ print $ intersectingHalfPlanes t
HalfSpace Positive (LinePV (Point2 0 0) (Vector2 (-1) 0))
HalfSpace Positive (LinePV (Point2 (-1) 0) (Vector2 1 (-1)))
HalfSpace Positive (LinePV (Point2 0 (-1)) (Vector2 0 1))

toBarricentric :: (Fractional r, Point_ point 2 r, Triangle_ triangle point) => point -> triangle -> Vector 3 r

Given a point q and a triangle, q inside the triangle, get the baricentric cordinates of q

fromBarricentric :: forall triangle point (d :: Nat) r. (Triangle_ triangle point, Point_ point d r, Num r) => Vector 3 r -> triangle -> Point d r

Given a vector of barricentric coordinates and a triangle, get the corresponding point in the same coordinate sytsem as the vertices of the triangle.

Re-exports from Hiraffe

class HasVertices' graph => HasVertices graph graph' where

Class that expresses that we have a non-empty type changing traversal of all vertices.

Methods

vertices :: IndexedTraversal1 (VertexIx graph) graph graph' (Vertex graph) (Vertex graph')

Traversal of all vertices in the graph

class HasVertices' graph where

A class representing types that have vertices.

Minimal complete definition

vertexAt

Associated Types

type Vertex graph

Vertices of the graph are of this type

type VertexIx graph

Vertices are indexed by elements of type VertexIx

Methods

vertexAt :: VertexIx graph -> IndexedTraversal' (VertexIx graph) graph (Vertex graph)

Accessor to a given vertex.

numVertices :: graph -> Int

Number of vertices in the graph.

running time: O(1)

default numVertices :: HasVertices graph graph => graph -> Int