HGeometry.VoronoiDiagram.ViaLowerEnvelope

Description

Compute a Voronoi Diagram of a set of points in R^2 using the lower envelope of planes in R^3.

Documentation

data VoronoiDiagram point

A Voronoi diagram

Constructors

AllColinear !(Alternating Vector (VerticalOrLineEQ (NumType point)) point)

ConnectedVD !(VoronoiDiagram' (Point 2 (NumType point)) point)

newtype VoronoiDiagram' vertex point

A connected VoronoiDiagram

Constructors

VoronoiDiagram (MinimizationDiagram (NumType point) vertex point)

asMap :: (Point_ point 2 r, Ord point) => VoronoiDiagram' vertex point -> NEMap point (Region r vertex)

Get, for each point, its Voronoi region

voronoiDiagram :: (Point_ point 2 r, Functor f, Ord point, Ord r, Fractional r, Foldable1 f, Show point, Show r) => f point -> VoronoiDiagram point

Computes the Voronoi Diagram, by lifting the points to planes, and computing the lower envelope of these planes.

O(n^4) ( we currently use the brute force implentation...) TODO: switch to something faster

voronoiDiagramWith :: (Point_ point 2 r, Functor nonEmpty, Ord point, Ord r, Fractional r, Foldable1 nonEmpty) => (nonEmpty (Plane r :+ point) -> LowerEnvelope (Plane r :+ point)) -> nonEmpty point -> VoronoiDiagram point

Given a function to compute a lower envelope; construct use it to construct the Voronoi diagram.

voronoiVertices :: (Point_ point 2 r, Functor f, Ord point, Ord r, Fractional r, Foldable1 f, Show point, Show r, Ord point) => f point -> Set (Point 2 r)

Compute the vertices of the Voronoi diagram

pointToPlane :: (Point_ point 2 r, Num r) => point -> Plane r

lifts the point to a plane; so that the lower envelope corresponds to the VD