Ipe.Draw

Description

Class and setup for rendering/drawing

Documentation

type family Rendered backend

The type of objects a backend renders

type Attr backend geom = AttrOf backend geom -> AttrOf backend geom

An Attribute Assignment

class Monoid (Rendered backend) => IsDrawable backend geom where

A class that expresses that something is drawable using a particular backend

Associated Types

type AttrOf backend geom

A GADT that expresses possible attributes for a particular object

Methods

draw :: [Attr backend geom] -> geom -> Rendered backend

Draw some objects

The Ipe backend

data Ipe r

The Ipe backend