Ipe.Reader

Reading ipe Files

readRawIpeFile :: (Coordinate r, Eq r) => OsPath -> IO (Either ConversionError (IpeFile r))

Given a file path, tries to read an ipe file

readIpeFile :: (Coordinate r, Eq r) => OsPath -> IO (Either ConversionError (IpeFile r))

Given a file path, tries to read an ipe file.

This function applies all matrices to objects.

readSinglePageFile :: (Coordinate r, Eq r) => OsPath -> IO (Either ConversionError (IpePage r))

Since most Ipe file contain only one page, we provide a shortcut for that as well.

This function applies all matrices, and it makes sure there is at least one layer and view in the page.

readSinglePageFileThrow :: (Coordinate r, Eq r) => OsPath -> IO (IpePage r)

Tries to read a single page file, throws an error when this fails. See readSinglePageFile for further details.

type ConversionError = Text

Readiing ipe style files

readIpeStylesheet :: OsPath -> IO (Either ConversionError IpeStyle)

Reads an Ipe stylesheet from Disk.

addStyleSheetFrom :: OsPath -> IpeFile r -> IO (IpeFile r)

Given a path to a stylesheet, add it to the ipe file with the highest priority. Throws an error when this fails.

Reading XML directly

fromIpeXML :: forall {k} t (r :: k). IpeRead (t r) => ByteString -> Either ConversionError (t r)

Given a Bytestring, try to parse the bytestring into anything that is IpeReadable, i.e. any of the Ipe elements.

readXML :: ByteString -> Either ConversionError (Node Text Text)

Reads the data from a Bytestring into a proper Node

Read classes

class IpeReadText t where

Reading an ipe elemtn from a Text value

Methods

ipeReadText :: Text -> Either ConversionError t

class IpeRead t where

Reading an ipe lement from Xml

Methods

ipeRead :: Node Text Text -> Either ConversionError t

class IpeReadAttr t where

Basically IpeReadText for attributes. This class is not really meant to be implemented directly. Just define an IpeReadText instance for the type (Apply f at), then the generic instance below takes care of looking up the name of the attribute, and calling the right ipeReadText value. This class is just so that reifyConstraint in ipeReadRec can select the right typeclass when building the rec.

Methods

ipeReadAttr :: Text -> Node Text Text -> Either ConversionError t

Some low level implementation functions

ipeReadTextWith :: (Text -> Either t v) -> Text -> Either ConversionError (IpeValue v)

ipeReadObject :: forall i r (f :: TyFun AttributeUniverse Type -> Type) (ats :: [AttributeUniverse]) proxy. (IpeRead (i r), f ~ AttrMapSym1 r, ats ~ AttributesOf i, RecApplicative ats, ReifyConstraint IpeReadAttr (Attr f) ats, RecAll (Attr f) ats IpeReadAttr, AllConstrained IpeAttrName ats) => Proxy i -> proxy r -> Node Text Text -> Either ConversionError (i r :+ IpeAttributes i r)

If we can ipeRead an ipe element, and we can ipeReadAttrs its attributes we can properly read an ipe object using ipeReadObject

ipeReadAttrs :: forall proxy proxy' (i :: Type -> Type) r (f :: TyFun AttributeUniverse Type -> Type) (ats :: [AttributeUniverse]). (f ~ AttrMapSym1 r, ats ~ AttributesOf i, ReifyConstraint IpeReadAttr (Attr f) ats, RecApplicative ats, RecAll (Attr f) ats IpeReadAttr, AllConstrained IpeAttrName ats) => proxy i -> proxy' r -> Node Text Text -> Either ConversionError (IpeAttributes i r)

Reader for records. Given a proxy of some ipe type i, and a proxy of an coordinate type r, read the IpeAttributes for i from the xml node.

ipeReadRec :: forall (f :: TyFun AttributeUniverse Type -> Type) (ats :: [AttributeUniverse]). (RecApplicative ats, ReifyConstraint IpeReadAttr (Attr f) ats, RecAll (Attr f) ats IpeReadAttr, AllConstrained IpeAttrName ats) => Proxy f -> Proxy ats -> Node Text Text -> Either ConversionError (Rec (Attr f) ats)

Reading the Attributes into a Rec (Attr f), all based on the types of f (the type family mapping labels to types), and a list of labels (ats).

class Fractional r => Coordinate r where

Represent stuff that can be used as a coordinate in ipe. (similar to show/read)

Minimal complete definition

Nothing

Methods

fromSeq :: Integer -> Maybe (Int, Integer) -> r

default fromSeq :: (Ord r, Fractional r) => Integer -> Maybe (Int, Integer) -> r